% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{Abiotic.no_forcing}
\alias{Abiotic.no_forcing}
\title{Set up exogenous variables for abiotic forcing}
\usage{
\method{Abiotic}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Abiotic} for the no_forcing model of exogenous forcing (do nothing)
}
