% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_case_detection.R
\name{ActiveCaseDetection.none}
\alias{ActiveCaseDetection.none}
\title{Methods for distributing interventions during active_case_detectional visits}
\usage{
\method{ActiveCaseDetection}{none}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{ActiveCaseDetection} for the none model (do nothing)
}
