% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{add_lines_X_SEIS}
\alias{add_lines_X_SEIS}
\title{Add lines for the density of infected individuals for the SEIS model}
\usage{
add_lines_X_SEIS(time, XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{a list with the outputs of parse_Xorbits_SEIS}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SEIS model
}
