% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{add_lines_X_SEISd}
\alias{add_lines_X_SEISd}
\title{Add lines for the density of infected individuals for the SEISd model}
\usage{
add_lines_X_SEISd(time, XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{a list with the outputs of parse_Xorbits_SEISd}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SEISd model
}
