% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-dts.R
\name{add_lines_X_SIP_dts}
\alias{add_lines_X_SIP_dts}
\title{Add lines for the density of infected individuals for the SIP_dts model}
\usage{
add_lines_X_SIP_dts(
  XH,
  pars,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X.SIP_dts}

\item{pars}{a list that defines an \code{ramp.dts} model (\emph{e.g.},  generated by \code{dts_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIP_dts model
}
