% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{add_lines_X_SIPd}
\alias{add_lines_X_SIPd}
\title{Add lines for the density of infected individuals for the SIPd model}
\usage{
add_lines_X_SIPd(
  time,
  XH,
  nStrata,
  clrs = c("darkblue", "darkred", "darkgreen"),
  llty = 1
)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{a list with the outputs of parse_Xorbits.SIPd}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIPd model
}
