% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{add_lines_X_SIS}
\alias{add_lines_X_SIS}
\title{Add lines for the density of infected individuals for the SIS model}
\usage{
add_lines_X_SIS(time, XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{time}{time points for the observations}

\item{XH}{parsed outputs, from parse_Xorbits_SIS}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIS model
}
\keyword{internal}
