% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{add_lines_X_SIS_dts}
\alias{add_lines_X_SIS_dts}
\title{Add lines for the density of infected individuals for the SIS_dts model}
\usage{
add_lines_X_SIS_dts(XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of parse_outputs_X_SIS_dts}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIS_dts model
}
