% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{add_variable}
\alias{add_variable}
\title{Add a new variable}
\usage{
add_variable(pars, var_object)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{var_object}{an object to dispatch and compute a variable}
}
\value{
the \code{xds} object
}
\description{
The description for each method should include the equations.
}
