% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-concepts.R
\name{aquatic_habitats}
\alias{aquatic_habitats}
\title{Aquatic Habitats}
\description{
Aquatic habitats are distributed among patches on a landscape.
\describe{
\item{\code{membership}}{a vector: the patch index of each aquatic habitat}
\item{\code{nHabitats}}{the number of habitats (\code{length(membership)}) }
\item{\code{searchQ}}{a vector: a search weight for each aquatic habitat}
}
}
\section{The Habitat Matrix}{


The \emph{habitat matrix} is an \code{nPatches} \eqn{\times} \code{nHabitats} matrix called \eqn{N}.
It is constructed from the \code{membership} vector.
In the matrix, \eqn{N_{i,j}} is 1 if the \eqn{i^{th}} patch contains the \eqn{j^{th}} habitat.
For example, a membership vector \eqn{(1,1,2)} has a corresponding habitat matrix:
\deqn{
N = \left[ 
\begin{array}{ccc}
1&1&0\\
0&0&1\\
\end{array}
\right]
}
If \eqn{\alpha} denotes the emergence rate of adult mosquitoes
from each habitat, then the emergence rate at the patch, \eqn{\Lambda} is:
\deqn{\Lambda = N \cdot \alpha}
}

\section{The Egg Distribution Matrix}{


Let \eqn{\omega} denote the habitat search weights. Habitat availability (\eqn{Q}) is computed as:
\deqn{Q = N \cdot \omega.}
Mosquitoes could also lay some eggs in available bad habitats (\eqn{B}) or in available ovitraps (\eqn{O}).
The egg distribution matrix (\eqn{U}) is computed as
\deqn{U = \left( \mbox{diag} \left(\omega \right) \cdot N^T \right) \cdot \mbox{diag}\left( (Q+B+O)^{-1} \right).}
In models where there are patches with no available habitat, \eqn{(Q+B+O)^{-1}} is computed using the function \link{diag_inverse}, which
fixes the divide-by-zero problem.
}

