% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying-concepts.R
\name{aquatic_habitats_bad}
\alias{aquatic_habitats_bad}
\title{Unproductive Aquatic Habitats}
\description{
To model egg laying, a port
was developed to model availability
of unproductive aquatic habitats: water
where mosquitoes
lay eggs, but no mature adults ever emerge.
}
