% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{ar2eir.nb}
\alias{ar2eir.nb}
\title{A negative binomial model for the daily EIR. as a function of the daily attack rate}
\usage{
\method{ar2eir}{nb}(ar, b, env_het_obj)
}
\arguments{
\item{ar}{the attack rate}

\item{b}{the probability of infection, per bite}

\item{env_het_obj}{an environmental heterogeneity model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{ar2eir} for a negative binomial model
}
\seealso{
\link{F_ar.nb}
}
\keyword{internal}
