% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{ar2eir.pois}
\alias{ar2eir.pois}
\title{A Poisson model for the daily attack rate as a function of the daily EIR.}
\usage{
\method{ar2eir}{pois}(ar, b, pars)
}
\arguments{
\item{ar}{the attack rate}

\item{b}{the probability of infection, per bite}

\item{pars}{an \code{xds} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{ar2eir} for a Poisson model
}
