% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{ar2eir}
\alias{ar2eir}
\title{Convert AR to EIR}
\usage{
ar2eir(ar, b, pars)
}
\arguments{
\item{ar}{the attack rate}

\item{b}{the probability of infection, per bite}

\item{pars}{an \code{xds} object}
}
\value{
the daily EIR as a \link{numeric} vector
}
\description{
This function computes the inverse of \link{F_ar}
under a model for environmentally heterogeneous exposure.
}
