% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{AreaSpray.none}
\alias{AreaSpray.none}
\title{Set no AreaSpray}
\usage{
\method{AreaSpray}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for AreaSpray
}
