% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{AreaSpray}
\alias{AreaSpray}
\title{Set the AreaSpray}
\usage{
AreaSpray(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
AreaSpray
}
