% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-coverage-func.R
\name{AreaSprayCoverage.func}
\alias{AreaSprayCoverage.func}
\title{Set no area_spray_coverage}
\usage{
\method{AreaSprayCoverage}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for area_spray_coverage
}
