% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-coverage-none.R
\name{AreaSprayCoverage.none}
\alias{AreaSprayCoverage.none}
\title{Set no area spray coverage}
\usage{
\method{AreaSprayCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for area spray coverage
}
