% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-coverage-interface.R
\name{AreaSprayCoverage}
\alias{AreaSprayCoverage}
\title{Set the area spray coverage}
\usage{
AreaSprayCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
area spray coverage
}
