% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-effects-none.R
\name{AreaSprayEffects.none}
\alias{AreaSprayEffects.none}
\title{Set no area spray effects}
\usage{
\method{AreaSprayEffects}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for area spray effects
}
