% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-effectsizes-none.R
\name{AreaSprayEffectSizes.none}
\alias{AreaSprayEffectSizes.none}
\title{Set no area_spray_effectsizes}
\usage{
\method{AreaSprayEffectSizes}{none}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
\link{list}
}
\description{
The null model for area_spray_effectsizes
}
