% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{average_EIR}
\alias{average_EIR}
\title{Compute the average EIR}
\usage{
average_EIR(xds_obj, i = 1, members = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the human species index}

\item{members}{a membership matrix}
}
\value{
a PfPR
}
\description{
This is a generic way of computing the average EIR
for some subset(s) of the population strata. The object members is a
membership matrix, like the residence matrix (see \link{make_residency_matrix}).
If the residence matrix were passed to members, average_EIR returns
the average EIR for the residents of each patch, which is could be different
from what would be computed.

The default, \verb{members=NULL,} creates a \eqn{1 \times} \code{nStrata} matrix,
which takes the matrix over the whole population. If \code{members} is a
matrix with 1 column and \code{nStrata} rows, with only some elements set to one,
then it returns the average over those strata.
}
