% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_terms.R
\name{average_PR_true}
\alias{average_PR_true}
\title{Compute the average True PR}
\usage{
average_PR_true(xds_obj, i = 1, members = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the human species index}

\item{members}{a membership matrix}
}
\value{
a PfPR
}
\description{
This is a generic way of computing the average PR
for some subset(s) of the population. The object \code{members} is a
membership matrix, like the residence matrix (see \link{make_residency_matrix}).
If the residence matrix were passed to members, average_PR_true returns
the average PR for each patch by residency.

The default, \verb{members=NULL,} creates a \eqn{1 \times} \code{nStrata} matrix,
which takes the matrix over the whole population. If \code{members} is a
matrix with 1 column and \code{nStrata} rows, with only some elements set to one,
then it returns the average over those strata.
}
