% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{bad_habitats.static}
\alias{bad_habitats.static}
\title{Compute availability of other blood hosts}
\usage{
\method{bad_habitats}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up...
}
\keyword{internal}
