% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{basicL}
\alias{basicL}
\title{The \code{basicL} Module for the L Component}
\description{
A basic model of aquatic mosquito
ecology with density-dependent mortality and delayed maturation, based
on Smith DL, \emph{et al.} (2013).
}
\section{State Variables}{

\describe{
\item{\code{L}}{density of mosquito larvae in each habitat}
}
}

\section{Parameters}{

\describe{
\item{\code{psi}}{maturation rate (\eqn{\psi})}
\item{\code{xi}}{delayed maturation parameter in response to mean crowding (\eqn{\xi})}
\item{\code{phi}}{density-independent death rate (\eqn{\phi})}
\item{\code{theta}}{slope of density-dependent mortality (\eqn{\theta})}
}
}

\section{Egg Laying}{


\describe{
\item{\code{eta}}{the egg laying rate, (\eqn{\eta})}
}
}

\section{Dynamics}{


\deqn{dL/dt = \eta - (\psi e^{-\xi L} + \phi + \theta L) L}
}

\section{Emergence}{


The emergence rate of adult, female mosquitoes from each patch is:
\deqn{\alpha = \psi e^{-\xi L} L}
}

\section{Notes}{

Setting \eqn{\xi = 0} recovers the model of Smith DL, \emph{et al.} (2013).
}

\references{
Smith DL, Perkins TA, Tusting LS, Scott TW, Lindsay SW (2013) Mosquito Population Regulation and Larval Source Management in Heterogeneous Environments. PLOS ONE 8(8): e71247. \href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0071247}{https://doi.org/10.1371/journal.pone.0071247}.
}
