% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{basicM}
\alias{basicM}
\title{The \code{basicM} Module for the MY Component}
\description{
Implements the \strong{MY} component using a basic model of adult mosquito
ecology without explicit infection dynamics.
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{P}}{density of adult mosquitoes that have fed at least once}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma}}{emigration rate}
\item{\code{mu}}{emigration loss rate}
\item{\code{K}}{mosquito dispersal matrix}
\item{\code{Omega}}{adult mosquito demographic matrix (mortality + migration)}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dP/dt &= f(M-P) - \Omega \cdot P \\
\end{array}}
}

