% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{BedNet.none}
\alias{BedNet.none}
\title{Set no bednet}
\usage{
\method{BedNet}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for bednet
}
