% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{BedNet}
\alias{BedNet}
\title{Set the bednet}
\usage{
BedNet(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
bednet
}
