% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-none.R
\name{BedNetCoverage.none}
\alias{BedNetCoverage.none}
\title{Set no bednet_coverage}
\usage{
\method{BedNetCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for bednet_coverage
}
