% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-interface.R
\name{BedNetCoverage}
\alias{BedNetCoverage}
\title{Set the bednet_coverage}
\usage{
BedNetCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
bednet_coverage
}
