% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effects-interface.R
\name{BedNetEffects}
\alias{BedNetEffects}
\title{Set the bednet_effects}
\usage{
BedNetEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
bednet_effects
}
