% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effectsizes-func.R
\name{BedNetEffectSizes.func}
\alias{BedNetEffectSizes.func}
\title{Set no bednet}
\usage{
\method{BedNetEffectSizes}{func}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
The null model for bednet
}
