% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-no_net.R
\name{BedNetEffectSizes.no_nets}
\alias{BedNetEffectSizes.no_nets}
\title{Bed net ownership}
\usage{
\method{BedNetEffectSizes}{no_nets}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Implements \link{BedNetEffectSizes} for the no_nets model of bed nets (do nothing)
}
