% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{BedNetEffectSizes.none}
\alias{BedNetEffectSizes.none}
\title{Set no bednet_effectsizes}
\usage{
\method{BedNetEffectSizes}{none}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
The null model for bednet_effectsizes
}
