% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{BedNetEffectSizes}
\alias{BedNetEffectSizes}
\title{Set the bednet_effectsizes}
\usage{
BedNetEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
bednet_effectsizes
}
