% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{Behavior.no_behavior}
\alias{Behavior.no_behavior}
\title{Methods for dynamic human and mosquito behaviors}
\usage{
\method{Behavior}{no_behavior}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Behavior} for the no_behavior model (no changes)
}
