% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography-births.R
\name{Births.static}
\alias{Births.static}
\title{Derivatives of demographic changes in human populations}
\usage{
\method{Births}{static}(t, y, Hpar)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Hpar}{a \link{list}}
}
\value{
a \link{numeric} vector of zeros
}
\description{
Implements \link{Births} for static population models
}
