% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography.R
\name{Births.zero}
\alias{Births.zero}
\title{Human (or Host) population birth rate}
\usage{
\method{Births}{zero}(t, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
see help pages for specific methods
}
\description{
a function
}
\keyword{internal}
