% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{blood_feeding}
\alias{blood_feeding}
\title{Blood Feeding Overview}
\description{
Blood feeding is an interaction among humans and mosquitoes:
in this model, humans are spending time, and
mosquitoes are searching. To model exposure, each human (or host)
population stratum is assigned a \emph{search weight,} a
number used to weigh \emph{time spent} and get a measure of
\emph{availability.}
}
