% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{blood_feeding_dynamics}
\alias{blood_feeding_dynamics}
\title{Compute blood feeding objects dynamically}
\usage{
blood_feeding_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Compute host available, \eqn{W},
total blood host available, \eqn{B},
the time spent matrix \eqn{(\Theta)}, and the time-at-risk matrix \eqn{(\Psi)}
for static models.
}
