% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{blood_hosts_dynamics}
\alias{blood_hosts_dynamics}
\title{blood_hosts Dynamics}
\usage{
blood_hosts_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
on the \code{XY_interface.} Here is defined as time
spent in one of the patches in this model, so
travle is time spent elsewhere.

A function \code{F_blood_hosts} computes \code{TiSp_frac_blood_hostsing} and
this function sets \code{TiSp_frac_here.}
}
\note{
The quantity \code{TiSp_frac_here} is part of the \code{XY_interface}
It is used to compute the Time at Risk matrix, and it is also
used to weight local exposure \emph{vs.}  blood_hostsing exposure.
}
