% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{blood_search_weights}
\alias{blood_search_weights}
\title{blood_search_weights}
\usage{
blood_search_weights(t, y, xds_obj, i, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{S3} definition for the
function that computes blood_search_weights
}
\keyword{internal}
