% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{BloodFeeding.dynamic}
\alias{BloodFeeding.dynamic}
\title{Compute blood feeding objects dynamically}
\usage{
\method{BloodFeeding}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Compute host availability, \eqn{W},
total blood host availability, \eqn{B},
the time spent matrix \eqn{\Theta}, and the time-at-risk matrix \eqn{\Psi}
for static models.
}
