% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{BloodFeeding}
\alias{BloodFeeding}
\title{Blood feeding}
\usage{
BloodFeeding(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Compute and store host available, \eqn{W},
total blood host available, \eqn{B},
the time spent matrix \eqn{\Theta}, and the time-at-risk matrix \eqn{\Psi},
}
\keyword{internal}
