% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{BloodFeeding.static}
\alias{BloodFeeding.static}
\title{Compute blood feeding objects: static models}
\usage{
\method{BloodFeeding}{static}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
the unmodified \code{xds} object
}
\description{
Return the blood feeding objects unmodified
}
