% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{BloodHosts.dynamic}
\alias{BloodHosts.dynamic}
\title{Time Spent F_blood_hosts}
\usage{
\method{BloodHosts}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a parameter
describing time spent F_blood_hosts. It is computed in \link{compute_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
