% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{BloodHosts}
\alias{BloodHosts}
\title{Other Blood Hosts}
\usage{
BloodHosts(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a term
describing the availability of ovi traps
}
