% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{BloodHosts.static}
\alias{BloodHosts.static}
\title{Time Spent F_blood_hosts}
\usage{
\method{BloodHosts}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For a static model for time spent F_blood_hosts, the function
does not update anything.
}
