% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\title{Burn-in}
\usage{
burnin(model, neg_inf_yr = 10)
}
\arguments{
\item{model}{a \strong{\code{ramp.xds}} model object}

\item{neg_inf_yr}{run from -neg_inf_yr years to the present}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Use the defined hindcast to
run a model from the past up to the present.
Reset the initial conditions.
}
