% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{change_bad_habitat}
\alias{change_bad_habitat}
\title{Change Bad Habitat Availability}
\usage{
change_bad_habitat(Qbad, xds_obj, s = 1)
}
\arguments{
\item{Qbad}{availability of bad habitat}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the availability of bad habitat
}
