% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{change_bad_habitats}
\alias{change_bad_habitats}
\title{Change the bad_habitats}
\usage{
change_bad_habitats(bad_habitats, xds_obj, s = 1)
}
\arguments{
\item{bad_habitats}{bad habitat availability}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the bad habitats for the \eqn{s^{th}} vector species, and
trigger updates for the \code{ML} interface.
}
