% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{change_blood_hosts}
\alias{change_blood_hosts}
\title{Set static blood feeding search weights}
\usage{
change_blood_hosts(blood_hosts, xds_obj, s)
}
\arguments{
\item{blood_hosts}{availability of other blood hosts}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \code{xds} object
}
\description{
Set the blood feeding search weights, \eqn{\omega}, for a set of host strata
}
