% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{change_blood_search_weights}
\alias{change_blood_search_weights}
\title{Change the blood_search_weights}
\usage{
change_blood_search_weights(blood_search_weights, xds_obj, i = 1)
}
\arguments{
\item{blood_search_weights}{a blood_search_weights}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the blood_search_weights for the \eqn{i^{th}} host species, and
trigger updates for the \code{XY} interface.
}
