% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{change_blood_traps}
\alias{change_blood_traps}
\title{Change the blood_traps}
\usage{
change_blood_traps(blood_traps, xds_obj, s = 1)
}
\arguments{
\item{blood_traps}{a blood_traps}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the blood_traps for the \eqn{s^{th}} vector species, and
trigger updates for the \code{XY} interface.
}
