% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{change_blood_weights}
\alias{change_blood_weights}
\title{Set static blood feeding search weights}
\usage{
change_blood_weights(pars, search_weights = 1, s = 1, i = 1)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{search_weights}{the blood feeding search weights}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
an \code{xds} object
}
\description{
Set the blood feeding search weights, \eqn{\omega}, for a set of host strata
}
