% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-dispersal.R
\name{change_calK}
\alias{change_calK}
\title{Change the time spent matrix}
\usage{
change_calK(calK, pars, s = 1)
}
\arguments{
\item{calK}{a mosquito dispersal \link{matrix}}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Change the time spent matrix
}
