% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{change_eip}
\alias{change_eip}
\title{Change the extrinsic incubation period}
\usage{
change_eip(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Update the extrinsic incubation period for the \eqn{s^{th}}
vector species.
}
