% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{change_f}
\alias{change_f}
\title{Change the blood feeding rate}
\usage{
change_f(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
For models where the baseline value is constant,
update the blood feeding rate.
}
