% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_season.R
\name{change_F_season.MY}
\alias{change_F_season.MY}
\title{Change F_season}
\usage{
\method{change_F_season}{MY}(F, xds_obj, ix = 1)
}
\arguments{
\item{F}{a seasonal pattern function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the season function
when \code{forced_by = "MY"}
}
\keyword{internal}
