% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_shock.R
\name{change_F_shock.eir}
\alias{change_F_shock.eir}
\title{Change F_shock}
\usage{
\method{change_F_shock}{eir}(F, xds_obj, ix = 1)
}
\arguments{
\item{F}{new shock function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the shock function
when \code{forced_by = "eir"}
}
\keyword{internal}
