% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{change_g}
\alias{change_g}
\title{Change mosquito mortality rate}
\usage{
change_g(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the mosquito mortality rate for the \eqn{s^{th}} vector
species and refresh the demographic matrix.
}
