% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{change_habitat_search_weights}
\alias{change_habitat_search_weights}
\title{Change the habitat_search_weights}
\usage{
change_habitat_search_weights(search_weights, xds_obj, s = 1)
}
\arguments{
\item{search_weights}{habitat search weights}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the habitat_search_weights for the \eqn{i^{th}} host species, and
trigger updates for the \code{XY} interface.
}
