% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{change_habitat_weights}
\alias{change_habitat_weights}
\title{Change the habitat search weights}
\usage{
change_habitat_weights(pars, searchQ = 1, s = 1)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{searchQ}{the habitat search weights}

\item{s}{the vector species index}
}
\value{
the modified \code{xds} object
}
\description{
Set the search weights, \eqn{\omega}, for a set of aquatic habitats
}
