% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{change_K_matrix.K}
\alias{change_K_matrix.K}
\title{Change Mosquito Dispersal Matrix}
\usage{
\method{change_K_matrix}{K}(K_matrix, xds_obj, which_K = "K", s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{which_K}{which K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run \link{check_K_matrix} then

After passing checks, \code{xds_obj} is updated.

In models with multiple species, use \code{s} to
specify the species to update.
}
\seealso{
\link{xds_info_mosquito_dispersal}; \link{setup_K_matrix}
}
\keyword{internal}
