% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{change_L_inits.basicL}
\alias{change_L_inits.basicL}
\title{Set the Initial Values for \code{basicL} (\strong{L} Component)}
\usage{
\method{change_L_inits}{basicL}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Initial values of the variable \eqn{L} are reset if they are
passed as a named component of \code{options}
}
