% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{change_L_inits.trivial}
\alias{change_L_inits.trivial}
\title{Set the Initial Values for \code{trivial} (\strong{L} Component)}
\usage{
\method{change_L_inits}{trivial}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns the unmodified \strong{\code{xds}} object
}
