% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{change_L_pars.basicL}
\alias{change_L_pars.basicL}
\title{Set \strong{L} Component parameters for \code{basicL}}
\usage{
\method{change_L_pars}{basicL}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of \strong{L} component parameters
\itemize{
\item \code{psi} or \eqn{\psi}
\item \code{xi}  or \eqn{\xi}
\item \code{phi} or \eqn{\phi}
\item \code{theta} or \eqn{\theta}
}
}
\seealso{
\link{dLdt.basicL} or \link{make_L_obj_basicL}
}
