% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{change_L_pars}
\alias{change_L_pars}
\title{Set \strong{L} Component Parameters}
\usage{
change_L_pars(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
Set \strong{L} Component Parameters
}
