% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{change_L_pars.trivial}
\alias{change_L_pars.trivial}
\title{Change parameters for \code{trivial} (\strong{L})}
\usage{
\method{change_L_pars}{trivial}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change \code{Lambda}, \code{F_season}, \code{F_trend}, or \code{F_shock}
}
\note{
Utilities for trace functions are in \code{ramp.func}
}
\keyword{internal}
