% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_mean_forcing.R
\name{change_mean_forcing.Lambda}
\alias{change_mean_forcing.Lambda}
\title{Change mean forcing}
\usage{
\method{change_mean_forcing}{Lambda}(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the mean emergence rate
}
\keyword{internal}
