% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_mean_forcing.R
\name{change_mean_forcing}
\alias{change_mean_forcing}
\title{Change mean forcing}
\usage{
change_mean_forcing(X, xds_obj, ix = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the mean forcing parameter to \code{X}
}
