% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{change_mu}
\alias{change_mu}
\title{Change emigration-loss fraction}
\usage{
change_mu(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the emigration-loss fraction for the \eqn{s^{th}} vector
species and refresh the demographic matrix.
}
