% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{change_MY_inits.basicM}
\alias{change_MY_inits.basicM}
\title{Set new MY parameter values}
\usage{
\method{change_MY_inits}{basicM}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
