% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{change_MY_inits.GeM}
\alias{change_MY_inits.GeM}
\title{Change initial values for \code{GeM} (\strong{MY})}
\usage{
\method{change_MY_inits}{GeM}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
\keyword{internal}
