% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{change_MY_inits.RMdts}
\alias{change_MY_inits.RMdts}
\title{Change initial values for \code{RMdts} (\strong{MY})}
\usage{
\method{change_MY_inits}{RMdts}(xds_obj, s, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
a \link{list}
}
\description{
Change initial values for \code{RMdts} (\strong{MY})
}
\keyword{internal}
