% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{change_MY_inits.SI}
\alias{change_MY_inits.SI}
\title{change initial values for the macdonald model}
\usage{
\method{change_MY_inits}{SI}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
a \link{list}
}
\description{
Implements \link{change_MY_inits} for the macdonald model
}
