% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{change_MY_pars.GeRM}
\alias{change_MY_pars.GeRM}
\title{Return the parameters as a list}
\usage{
\method{change_MY_pars}{GeRM}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
