% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{change_MY_pars}
\alias{change_MY_pars}
\title{Set new MY parameter values}
\usage{
change_MY_pars(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
