% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{change_nu}
\alias{change_nu}
\title{Change F_nu}
\usage{
change_nu(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update F_nu for the \eqn{s^{th}} vector species.
}
