% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{change_Omega.dts}
\alias{change_Omega.dts}
\title{Change Omega}
\usage{
\method{change_Omega}{dts}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Update the demographic matrix for
discrete time systems
}
\keyword{internal}
