% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{change_other_blood_hosts}
\alias{change_other_blood_hosts}
\title{Change the other_blood_hosts}
\usage{
change_other_blood_hosts(other_blood_hosts, xds_obj, s = 1)
}
\arguments{
\item{other_blood_hosts}{a other_blood_hosts}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the other_blood_hosts for the \eqn{i^{th}} host species, and
trigger updates for the \code{XY} interface.
}
