% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{change_ovitraps}
\alias{change_ovitraps}
\title{Change ovitrap availability}
\usage{
change_ovitraps(ovitraps, xds_obj, s = 1)
}
\arguments{
\item{ovitraps}{ovitrap availability}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the ovitraps for the \eqn{s^{th}} vector species
}
