% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-change.R
\name{change_phi}
\alias{change_phi}
\title{Change phi}
\usage{
change_phi(phi, xds_obj, s)
}
\arguments{
\item{phi}{the density independent mortality rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the density independent
mortality rate for basicL for a static model
}
