% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-change.R
\name{change_psi}
\alias{change_psi}
\title{Change psi}
\usage{
change_psi(psi, xds_obj, s)
}
\arguments{
\item{psi}{the maturation rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the maturation
rate for basicL for a static model
}
