% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{change_q}
\alias{change_q}
\title{Change human feeding fraction}
\usage{
change_q(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the human feeding fraction for the \eqn{s^{th}} vector
species.
}
