% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season_bottom.eir}
\alias{change_season_bottom.eir}
\title{Set bottom}
\usage{
\method{change_season_bottom}{eir}(bottom, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{bottom}{the new bottom parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the bottom for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
