% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season_phase.Lambda}
\alias{change_season_phase.Lambda}
\title{Set phase}
\usage{
\method{change_season_phase}{Lambda}(phase, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{phase}{the new phase parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
the \strong{\code{ramp.xds}} model object with an updated phase parameter
}
\description{
Set the phase for the seasonal pattern for the
\eqn{s^{th}} species
and return the \strong{\code{ramp.xds}} model object
}
