% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season_pw.none}
\alias{change_season_pw.none}
\title{Set pw, a seasonality shape parameter}
\usage{
\method{change_season_pw}{none}(pw, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species inde}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{change_season_pw} for a model
with no forcing
}
