% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_shock.R
\name{change_shock.none}
\alias{change_shock.none}
\title{Set yy}
\usage{
\method{change_shock}{none}(shock_par, xds_obj, s = 1)
}
\arguments{
\item{shock_par}{parameters for \link{make_function}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{change_shock} for a model
with no forcing
}
