% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_shock.R
\name{change_shock}
\alias{change_shock}
\title{Change shock parameters}
\usage{
change_shock(shock_par, xds_obj, s = 1)
}
\arguments{
\item{shock_par}{parameters for \link{make_function}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change parameters for the shock function
}
