% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{change_sigma}
\alias{change_sigma}
\title{Change the patch emigration rate}
\usage{
change_sigma(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update and validate the patch emigration rate for the
\eqn{s^{th}} vector species, then update the demographic matrix.
}
