% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_trend.R
\name{change_spline}
\alias{change_spline}
\title{Set the interpolating points}
\usage{
change_spline(X, xds_obj, s = 1)
}
\arguments{
\item{X}{new interpolation points, as a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the interpolating points for F_trend
}
