% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_trend.R
\name{change_spline_y.none}
\alias{change_spline_y.none}
\title{Change y values for spline interpolation points}
\usage{
\method{change_spline_y}{none}(yy, xds_obj, s = 1)
}
\arguments{
\item{yy}{new y values for the interpolation points}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set new \eqn{y}-values for the spline interpolating points
and update \code{F_trend}
when \code{forced_by = "none"}
}
\keyword{internal}
