% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-change.R
\name{change_theta}
\alias{change_theta}
\title{Change theta, density dependent mortality}
\usage{
change_theta(theta, xds_obj, s)
}
\arguments{
\item{theta}{the density independent mortality rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the density dependent mortality
parameter for basicL for a static model
}
