% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{change_time_away}
\alias{change_time_away}
\title{Change the time spent}
\usage{
change_time_away(time_away, xds_obj, i = 1)
}
\arguments{
\item{time_away}{a time spent vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the time spent  for the \eqn{i^{th}} host species, and
trigger updates for the \code{XY} interface.
}
