% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{change_TimeSpent}
\alias{change_TimeSpent}
\title{Change the time spent matrix}
\usage{
change_TimeSpent(TiSp, pars, i = 1)
}
\arguments{
\item{TiSp}{a time spent \link{matrix}}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Change the time spent matrix
}
