% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{change_travel_eir}
\alias{change_travel_eir}
\title{Change the travel eir}
\usage{
change_travel_eir(travel_eir, xds_obj, i = 1)
}
\arguments{
\item{travel_eir}{a travel eir vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the travel eir  for the \eqn{i^{th}} host species
}
