% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{change_Upsilon.dde}
\alias{change_Upsilon.dde}
\title{Change Upsilon}
\usage{
\method{change_Upsilon}{dde}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Update \eqn{\Upsilon}, the matrix describing
mosquito survival and dispersal through an EIP in delay equations
}
\keyword{internal}
