% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{change_V_inits}
\alias{change_V_inits}
\title{Set new X parameter values}
\usage{
change_V_inits(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{xds_obj$V_obj[[s]]}.
}
