% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{change_vis_kappa}
\alias{change_vis_kappa}
\title{Change the vis_kappa}
\usage{
change_vis_kappa(vis_kappa, xds_obj, s = 1)
}
\arguments{
\item{vis_kappa}{a vis_kappa}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the vis_kappa for the \eqn{s^{th}} vector species, and
trigger updates for the \code{XY} interface.
}
