% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{change_visitor_kappa}
\alias{change_visitor_kappa}
\title{Change Visitor's NI}
\usage{
change_visitor_kappa(visitor_kappa, xds_obj, s)
}
\arguments{
\item{visitor_kappa}{net infectiousness of visitors}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of the parameter describing
available visitors
}
