% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{change_visitors}
\alias{change_visitors}
\title{Change the visitors}
\usage{
change_visitors(visitors, xds_obj, s = 1)
}
\arguments{
\item{visitors}{a visitors}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the visitors for the \eqn{s^{th}} vector species.
}
