% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{change_XH_inits.hMoI}
\alias{change_XH_inits.hMoI}
\title{Return the parameters as a list}
\usage{
\method{change_XH_inits}{hMoI}(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
