% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{change_XH_pars}
\alias{change_XH_pars}
\title{Set new X parameter values}
\usage{
change_XH_pars(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[s]]}.
}
