% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{change_XH_pars.SIS}
\alias{change_XH_pars.SIS}
\title{Set new values for the SIS model}
\usage{
\method{change_XH_pars}{SIS}(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} model object
}
\description{
A utility to change parameters
values for the SIS model (\strong{XH} component)

\code{options} is a named list, and:
\itemize{
\item \eqn{b} is set to \code{options$b}
\item \eqn{c} is set to \code{options$c}
\item \eqn{r} is set to \code{options$r}
\item \eqn{q_lm} is set to \code{options$q_lm}
\item \eqn{q_rdt} is set to \code{options$q_rdt}
\item \eqn{q_pcr} is set to \code{options$q_pcr}
}
}
\seealso{
\code{\link[=get_XH_pars.SIS]{get_XH_pars.SIS()}} and the github.io essay on \href{https://dd-harp.github.io/ramp.xds/articles/human_sis.html}{SIS Dynamics}
}
