% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-change.R
\name{change_xi}
\alias{change_xi}
\title{Change xi, the maturation delay parameter}
\usage{
change_xi(xi, xds_obj, s)
}
\arguments{
\item{xi}{the density independent mortality rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the maturation delay
parameter for basicL for a static model
}
