% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abiotic.R
\name{check_abiotic.no_forcing}
\alias{check_abiotic.no_forcing}
\title{Check abiotic}
\usage{
\method{check_abiotic}{no_forcing}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Check abiotic
}
