% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{check_bad_habitats}
\alias{check_bad_habitats}
\title{Check other blood hosts}
\usage{
check_bad_habitats(bad_habitats, nP)
}
\arguments{
\item{bad_habitats}{bad habitat availability}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(bad_habitats) == nPatches
}
}
\seealso{
\link{xds_info_blood_feeding}
}
