% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{check_blood_feeding}
\alias{check_blood_feeding}
\title{Check the XY Interface}
\usage{
check_blood_feeding(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run a set of consistency checks for the \code{blood_feeding}
}
\keyword{internal}
