% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{check_blood_traps}
\alias{check_blood_traps}
\title{Check other blood hosts}
\usage{
check_blood_traps(blood_traps, nP)
}
\arguments{
\item{blood_traps}{traps availability}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(blood_traps) == nPatc
}
}
\seealso{
\link{xds_info_blood_feeding}
}
