% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{check_civil.no_forcing}
\alias{check_civil.no_forcing}
\title{Check civil}
\usage{
\method{check_civil}{no_forcing}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Check civil
}
