% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{check_eip}
\alias{check_eip}
\title{Check the extrinsic incubation period}
\usage{
check_eip(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated extrinsic incubation period in days
}
\description{
Check that EIP values are numeric, non-negative, and have
length \code{nPatches}.
}
