% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{check_f}
\alias{check_f}
\title{Check blood feeding rate}
\usage{
check_f(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated blood feeding rate
}
\description{
Check that blood feeding rates are numeric, between zero and
one, and have length \code{nPatches}.
}
\seealso{
\link{xds_info_blood_feeding}
}
