% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{check_F_circadian}
\alias{check_F_circadian}
\title{Check circadian}
\usage{
check_F_circadian(F_circadian, tol = 1e-05)
}
\arguments{
\item{F_circadian}{a circadian pattern function}

\item{tol}{a tolerance function}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(circadian) == nPatc
}
}
\seealso{
\link{xds_info_blood_feeding}
}
