% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{check_g}
\alias{check_g}
\title{Check mosquito mortality rate}
\usage{
check_g(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated mosquito mortality rate
}
\description{
Check that the mosquito mortality rates are non-negative and have length
\code{nPatches}.
}
\seealso{
\link{xds_info_mosquito_demography}
}
