% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{check_habitat_search_weights}
\alias{check_habitat_search_weights}
\title{Check habitat search weights}
\usage{
check_habitat_search_weights(search_weights, nQ)
}
\arguments{
\item{search_weights}{habitat search weights}

\item{nQ}{the number of habitats}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(habitat_search_weights) == nHabitats
}
}
\seealso{
\link{xds_info_egg_laying}
}
