% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{check_L}
\alias{check_L}
\title{Run checks before solving (\strong{L})}
\usage{
check_L(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run checks before solving (\strong{L})
}
\keyword{internal}
