% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{check_ML_interface}
\alias{check_ML_interface}
\title{Check the ML Interface}
\usage{
check_ML_interface(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Run a set of consistency checks for the \code{ML_interface}
}
