% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_models}
\alias{check_models}
\title{Run Checks}
\usage{
check_models(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Run Checks
}
