% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{check_mu}
\alias{check_mu}
\title{Check emigration-loss fraction}
\usage{
check_mu(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated emigration-loss fraction
}
\description{
Check that the emigration-loss fractions are between zero and one and have
length \code{nPatches}.
}
\seealso{
\link{xds_info_mosquito_demography}
}
