% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{check_MY.GeM}
\alias{check_MY.GeM}
\title{Run a check before solving}
\usage{
\method{check_MY}{GeM}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run a check before solving
}
\keyword{internal}
