% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{check_MY.trivial}
\alias{check_MY.trivial}
\title{Check the \code{trivial} module (\strong{M})}
\usage{
\method{check_MY}{trivial}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run no consistency checks
}
\keyword{internal}
