% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{check_nu}
\alias{check_nu}
\title{Check egg laying rate}
\usage{
check_nu(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the egg laying rate
}
\description{
Check that the egg laying rates are in the expected range and length:
\itemize{
\item all elements are non-negative
\item length(F_nu) == nPatches
}
}
\seealso{
\link{xds_info_egg_laying}
}
