% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{check_other_blood_hosts}
\alias{check_other_blood_hosts}
\title{Check other blood hosts}
\usage{
check_other_blood_hosts(other_blood_hosts, nP)
}
\arguments{
\item{other_blood_hosts}{availbility of other blood hosts}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(other_blood_hosts) == nPatc
}
}
\seealso{
\link{xds_info_blood_feeding}
}
