% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{check_ovitraps}
\alias{check_ovitraps}
\title{Check ovitraps}
\usage{
check_ovitraps(ovitraps, nP)
}
\arguments{
\item{ovitraps}{ovitrap availability}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(ovitraps) == nPatches
}
}
\seealso{
\link{xds_info_blood_feeding}
}
