% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{check_q}
\alias{check_q}
\title{Check human feeding fraction}
\usage{
check_q(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated human feeding fraction
}
\description{
Check that the human feeding fractions are between zero and one and have
length \code{nPatches}.
}
\seealso{
\link{xds_info_blood_feeding}
}
