% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{check_sigma}
\alias{check_sigma}
\title{Check patch emigration rate}
\usage{
check_sigma(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the validated patch emigration rates
}
\description{
Check that the patch emigration rates are numeric, between zero and one,
and have length \code{nPatches}.
}
\seealso{
\link{xds_info_mosquito_demography}
}
