% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{check_time_away}
\alias{check_time_away}
\title{Check Time Spent}
\usage{
check_time_away(TA, Nh)
}
\arguments{
\item{TA}{a time spent vector}

\item{Nh}{the number of population strata}
}
\description{
Check that
\itemize{
\item there ar \eqn{N_p} columns and \eqn{N_h} rows
\item the column sums are at most 1`
}
}
\seealso{
\link{xds_info_time_spent}
}
