% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespent.R
\name{check_timespent_matrix}
\alias{check_timespent_matrix}
\title{Check Time Spent Matrix}
\usage{
check_timespent_matrix(TS, Np, Nh)
}
\arguments{
\item{TS}{a time spent matrix}

\item{Np}{the number of patches}

\item{Nh}{the number of population strata}
}
\description{
Check that
\itemize{
\item there ar \eqn{N_p} columns and \eqn{N_h} rows
\item the column sums are at most 1`
}
}
\seealso{
\link{xds_info_time_spent}
}
