% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{check_travel_eir}
\alias{check_travel_eir}
\title{Check the travel eir}
\usage{
check_travel_eir(teir, Nh)
}
\arguments{
\item{teir}{a travel eir vector}

\item{Nh}{the number of population strata}
}
\description{
Check that
\itemize{
\item the travel eir has length \code{nStrata}
\item they are all positive
}
}
\seealso{
\link{xds_info_time_spent}
}
