% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{check_vis_kappa}
\alias{check_vis_kappa}
\title{Check vis_kappa}
\usage{
check_vis_kappa(vis_kappa, nP)
}
\arguments{
\item{vis_kappa}{net infectiousness of the visitor population}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(vis_kappa) = nPatches
}
}
\seealso{
\link{xds_info_blood_feeding}
}
