% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{check_visitors}
\alias{check_visitors}
\title{Check visitors}
\usage{
check_visitors(visitors, nP)
}
\arguments{
\item{visitors}{the availability of visitors}

\item{nP}{the number of patches}
}
\description{
Check that
\itemize{
\item all elements are non-negative
\item length(visitors) = nPatches
}
}
\seealso{
\link{xds_info_blood_feeding}
}
