% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{check_XH}
\alias{check_XH}
\title{Run checks before solving (\strong{XH})}
\usage{
check_XH(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run checks before solving (\strong{XH})
}
\keyword{internal}
