% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{check_XY_interface}
\alias{check_XY_interface}
\title{Check the XY Interface}
\usage{
check_XY_interface(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Run a set of consistency checks for the \code{XY_interface}
}
\keyword{internal}
