% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{Civil.no_forcing}
\alias{Civil.no_forcing}
\title{Set up exogenous variables for civil forcing}
\usage{
\method{Civil}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Civil} for the no_forcing model of exogenous forcing (do nothing)
}
