% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_all_available_water}
\alias{compute_all_available_water}
\title{Compute and store availability of egg-laying habitats}
\usage{
compute_all_available_water(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of a variable, \eqn{Q}, that describes the availability of any habitat
or device that would attract mosquitoes and induce them to lay eggs.
}
\seealso{
\link{F_available_habitat}
}
\keyword{internal}
