% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_beta}
\alias{compute_beta}
\title{Compute beta}
\usage{
compute_beta(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This function computes the mixing matrix, beta
}
