% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_EIR_full}
\alias{compute_EIR_full}
\title{Compute EIR for each vector-host pair}
\usage{
compute_EIR_full(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
This function computes the EIR from each vector species for each stratum of each host species
}
\keyword{internal}
