% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htc.R
\name{compute_HTC}
\alias{compute_HTC}
\title{Compute the HTC Matrix}
\usage{
compute_HTC(xds_obj, i = 1, s = 1, tol = 1e-05)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{i}{the host species index}

\item{s}{the vector species index}

\item{tol}{a tolerance parameter}
}
\value{
a numeric \link{matrix}
}
\description{
Compute the \eqn{N_p \times N_p}
matrix \eqn{[D]}.
}
