% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_kappa}
\alias{compute_kappa}
\title{Compute kappa}
\usage{
compute_kappa(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
This function computes kappa for each vector species in each patch
}
\keyword{internal}
