% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{compute_O_matrix}
\alias{compute_O_matrix}
\title{Compute the Egg Laying Matrix}
\usage{
compute_O_matrix(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Compute the egg laying matrices
for all the vector species. The matrix determines
how all the eggs laid by adult mosquitoes are apportioned
among larval habitats
}
