% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-demography.R
\name{compute_Omega}
\alias{compute_Omega}
\title{Make the mosquito demography matrix for spatial RM model in continuous time}
\usage{
compute_Omega(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\description{
Make the mosquito demography matrix for spatial RM model in continuous time
}
