% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_Q}
\alias{compute_Q}
\title{Compute the total availability of egg-laying habitats, \eqn{Q}}
\usage{
compute_Q(membership_matrix, search_weights, Q_ovitraps, Q_bad_habitats)
}
\arguments{
\item{membership_matrix}{the membership matrix}

\item{search_weights}{the habitat search weights}

\item{Q_ovitraps}{the availability of ovitraps}

\item{Q_bad_habitats}{the availability of unsuitable habitats}
}
\value{
a \link{vector} of length \code{nPatches}
}
\description{
The sum of all available aquatic habitats and any other place
where a mosquito might lay eggs, including ovitraps and unsuitable habitats.
}
\details{
The availability of the habitats that we have defined in the model, denoted \eqn{Q_h}, sums search weights, \eqn{\omega}, by patch
using the habitat membership matrix, \eqn{\cal N}, and we can compute
\deqn{Q_h = {\cal N} \cdot \omega.}
If some ovitraps and bad_habitats are also available, with values \eqn{Q_o} and \eqn{Q_b} respectively, then
\deqn{Q = Q_h + Q_o + Q_b.}
The availability of habitats, ovitraps and bad_habitats are computed elsewhere and stored on \code{pars$vars}.
}
\seealso{
This function is called by \link{make_Q}

\link{create_membership_matrix} discusses \eqn{\cal N}

The availability of ovitraps and bad habitats is setup in \link{setup_EGG_LAYING}
}
