% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{compute_Qall}
\alias{compute_Qall}
\title{Compute and store availability of egg-laying habitats}
\usage{
compute_Qall(xds_obj)
}
\arguments{
\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Set the value of a variable, \eqn{Q}, that describes the availability of any habitat
or device that would attract mosquitoes and induce them to lay eggs.
}
\seealso{
\link{F_Q}
}
\keyword{internal}
