% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_RBR}
\alias{compute_RBR}
\title{Compute and attach the relative biting rates}
\usage{
compute_RBR(t, xds_obj, y)
}
\arguments{
\item{t}{the time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{y}{state vector}
}
\value{
an \code{xds} object
}
\description{
Compute the available for the pathogen's hosts for blood feeding
}
\keyword{internal}
