% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-demography.R
\name{compute_Upsilon.xde}
\alias{compute_Upsilon.xde}
\title{Make the mosquito demography matrix for spatial RM model in continuous time}
\usage{
\method{compute_Upsilon}{xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\description{
Make the mosquito demography matrix for spatial RM model in continuous time
}
