% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{compute_Upsilon_xde}
\alias{compute_Upsilon_xde}
\title{Make the mosquito demography matrix for spatial RM model in continuous time}
\usage{
compute_Upsilon_xde(eip, Omega)
}
\arguments{
\item{eip}{the extrinsic incubation period}

\item{Omega}{the demographic matrix}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
Make the mosquito demography matrix for spatial RM model in continuous time
}
\keyword{internal}
