% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity.R
\name{compute_VC}
\alias{compute_VC}
\title{Compute the VC Matrix}
\usage{
compute_VC(xds_obj, s = 1, tol = 1e-05)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}

\item{tol}{a tolerance parameter}
}
\value{
a numeric \link{matrix}
}
\description{
Compute the \eqn{N_p \times N_p}
matrix \eqn{\mathcal{V}} whose columns describe
the number of infective bites arising in each patch
from all the mosquitoes biting a single human on a
single day in each patch.

This creates a new model object to compute VC then
\itemize{
\item sets the initial condition to be the number of mosquitoes
biting a single infectious human on a single day, \eqn{fqM/W}
\item sets mean forcing to zero
\item adds a variable that tracks infective biting, \eqn{fqZ}
\item runs the model until the expected number of living mosquitoes from that
cohort is one in a million
\item saves the
}
}
