% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_W}
\alias{compute_W}
\title{Compute blood feeding availability for hosts}
\usage{
compute_W(search_weights, H, TaR)
}
\arguments{
\item{search_weights}{blood feeding search weights for the host strata}

\item{H}{host density}

\item{TaR}{the host species index}
}
\value{
host availability, a \link{vector}
}
\description{
Computes availability of the parasite's / pathogen's hosts to blood feeding
mosquitoes using the concept of search weights and time-at-risk
}
\details{
Host availability to blood searching mosquitoes in patches is
the sum of search weights of the human strata, a vector \eqn{\omega},
weighted by time at risk, defined by a matrix \eqn{\Psi}
that is \eqn{N_p \times \N_h}. The search weight is a \emph{per-capita} measure
so we weight it by human population density, \eqn{H}. Availability, \eqn{W}
is computed as \deqn{\Psi \cdot (\omega H).}
}
\references{
{This function implements equation 3
from \insertRef{WuSL2023SpatialDynamics}{ramp.xds} }
}
