% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_WB}
\alias{compute_WB}
\title{Compute available blood hosts of the i^th species}
\usage{
compute_WB(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Compute the available for the pathogen's hosts for blood feeding
}
