% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{Control.no_control}
\alias{Control.no_control}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{Control}{no_control}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Control} for the no_control model (do nothing)
}
