% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-dispersal.R
\name{create_calK_herethere}
\alias{create_calK_herethere}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
create_calK_herethere(nPatches)
}
\arguments{
\item{nPatches}{is the number of patches in the model}
}
\description{
Develop a mosquito dispersal matrix from a kernel and xy-coordinates
}
