% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{create_Linits_basicL}
\alias{create_Linits_basicL}
\title{Create Initial Values for \strong{L} Component Variables for \code{basicL}}
\usage{
create_Linits_basicL(nHabitats, Lopts = list(), L = 1)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{L}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Initial values of the variable \eqn{L} can be set
}
