% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{create_Lpar_trivial}
\alias{create_Lpar_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
create_Lpar_trivial(
  nHabitats,
  Lopts = list(),
  Lambda = 1000,
  season = F_flat,
  trend = F_flat
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{season}{a function that gives a seasonal pattern}

\item{trend}{a function that returns a temporal trend}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
