% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{create_membership_matrix}
\alias{create_membership_matrix}
\title{Create the habitat membership matrix, \eqn{\cal N}}
\usage{
create_membership_matrix(nPatches, membership)
}
\arguments{
\item{nPatches}{the number of patches}

\item{membership}{a vector describing the patch index for each habitat}
}
\value{
the membership \link{matrix}, denoted \eqn{\cal N} where \eqn{\left|\cal N\right|=}\code{nPatches} \eqn{\times} \code{nHabitats}
}
\description{
The habitat membership matrix, \eqn{\cal N}, holds
information about the patch location for each habitats.
If \eqn{w} is any vector describing a quantity in habitats (\emph{i.e.}, \eqn{\left|w\right|=} \code{nHabitats}), then
\eqn{W={\cal N}\cdot w} is a vector that has summed \eqn{w} by patch, and \eqn{\left|W\right|=} \code{nPatches}.
}
\details{
Information about the patch location of each habitat
is passed as the membership vector, an ordered list of patch locations. If
the \eqn{i^{th}} habitat is in the \eqn{j^{th}} patch, then
\eqn{{\cal N}[j,i]=1.} Otherwise, \eqn{{\cal N}[j,i]=0.}
}
\seealso{
to extract habitat membership information, see \link{view_membership_matrix}
}
