% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{create_MYZinits_basicM}
\alias{create_MYZinits_basicM}
\title{Make inits for basicM adult mosquito model}
\usage{
create_MYZinits_basicM(nPatches, MYZopts, M = 5, P = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for basicM adult mosquito model
}
