% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{create_MYZinits_RM}
\alias{create_MYZinits_RM}
\title{Make inits for RM adult mosquito model}
\usage{
create_MYZinits_RM(
  nPatches,
  Upsilon,
  MYZopts = list(),
  M = 5,
  P = 1,
  Y = 1,
  Z = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{Upsilon}{a matrix describing survival and dispersal through the EIP}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}

\item{Y}{infected mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RM adult mosquito model
}
