% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{create_MYZinits_RM_dts}
\alias{create_MYZinits_RM_dts}
\title{Make inits for RM_dts adult mosquito model}
\usage{
create_MYZinits_RM_dts(
  nPatches,
  max_eip,
  MYZopts = list(),
  M = 5,
  P = 1,
  U = 0,
  Y = 1,
  Z = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{max_eip}{the maximum number of EIP cohorts, an \link{integer}}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}

\item{U}{total uninfected mosquito density at each patch}

\item{Y}{infected mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RM_dts adult mosquito model
}
