% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-RM.R
\name{create_MYZpar_RM}
\alias{create_MYZpar_RM}
\title{Make parameters for RM ODE adult mosquito model}
\usage{
create_MYZpar_RM(
  nPatches,
  MYZopts = list(),
  eip = 12,
  g = 1/12,
  sigma = 1/8,
  mu = 0,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{eip}{the extrinsic incubation period}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{fraction lost during emigration}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a \link{list}
}
\description{
Make parameters for RM ODE adult mosquito model
}
