% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{create_residence_matrix}
\alias{create_residence_matrix}
\title{Create the Residency Matrix}
\usage{
create_residence_matrix(nPatches, residence)
}
\arguments{
\item{nPatches}{the number of patches}

\item{residence}{a vector describing the patch index for each habitat}
}
\value{
the residence \link{matrix}, denoted \eqn{J} where \eqn{\left|J\right|= n_p \times n_h}
}
\description{
The residency matrix, \eqn{J}, holds
information about residency for each human population stratum.
It is the template for the time spent and time at risk matrices, making it possible
to compute mosquito parameters describing blood feeding, the mixing matrix,
and terms describing transmission.
}
\details{
Information about residence in a patch location for each stratum
is passed as the residence vector, an ordered list of patch locations. If
the \eqn{i^{th}} stratum lives in the \eqn{j^{th}} patch, then
\eqn{{J}_{j,i}=1.} Otherwise, \eqn{{J}_{j,i}=0.}

Since \eqn{J} is a matrix, it is readily used for computation. Let:
\itemize{
\item \eqn{n_h = } \code{nStrata}, the number of population strata;
\item \eqn{n_p = } \code{nPatches}, the number of patches.
}

If \eqn{w} is any vector describing a quantity in strata (\emph{i.e.}, \eqn{\left|w\right|=n_h}), then
\deqn{W={J}\cdot w} is a vector that has summed \eqn{w} by residency for the strata, and \eqn{\left|W\right|= n_p}.
}
\seealso{
see \link{setup_BLOOD_FEEDING}

see \link{view_residence_matrix}
}
