% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{create_Xinits_hMoI}
\alias{create_Xinits_hMoI}
\title{Make inits for hybrid MoI human model}
\usage{
create_Xinits_hMoI(nStrata, Xopts = list(), m1 = 2, m2 = 1)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{Xopts}{a \link{list} that overwrites default values}

\item{m1}{mean MoI among inapparent human infections}

\item{m2}{mean MoI among patent human infections}
}
\value{
none
}
\description{
MoI stands for Multiplicity of Infection, and refers to malarial superinfection.
}
