% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{create_Xinits_SIP}
\alias{create_Xinits_SIP}
\title{Make initial values for the SIP human model, with defaults}
\usage{
create_Xinits_SIP(nStrata, H, Xopts = list(), I = 1, P = 0)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{H}{the initial human population density}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{I}{the initial values of the parameter I}

\item{P}{the initial values of the parameter P}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIP human model, with defaults
}
