% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{create_Xpar_trace}
\alias{create_Xpar_trace}
\title{Make parameters for trace human model}
\usage{
create_Xpar_trace(nPatches, Xopts, kappa = 0.1, HPop = 1)
}
\arguments{
\item{nPatches}{the number of patches}

\item{Xopts}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}
}
\value{
a \link{list}
}
\description{
Make parameters for trace human model
}
