% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{create_Xpar_trivial}
\alias{create_Xpar_trivial}
\title{Make parameters for trivial human model}
\usage{
create_Xpar_trivial(
  nPatches,
  Xopts,
  kappa = 0.1,
  HPop = 1,
  season = F_flat,
  trend = F_flat
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{Xopts}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}

\item{season}{a seasonality function}

\item{trend}{a trend function}
}
\value{
a \link{list}
}
\description{
Make parameters for trivial human model
}
