% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{d_F_eip_dt}
\alias{d_F_eip_dt}
\title{Compute the derivative of the EIP as a function of time}
\usage{
d_F_eip_dt(t, vars, eip_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{eip_par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
This method ...
}
