% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static.R
\name{d_F_eip_dt.static}
\alias{d_F_eip_dt.static}
\title{This function computes the negative derivative of the EIP as a function of time}
\usage{
\method{d_F_eip_dt}{static}(t, vars, eip_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{eip_par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{d_F_eip_dt} for the static_xde model (deip_dt=0)
}
