% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography-dHdt.R
\name{dAdt.zero}
\alias{dAdt.zero}
\title{Compute Demographic Changes}
\usage{
\method{dAdt}{zero}(t, y, Hpar)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Hpar}{a \link{list}}
}
\value{
a \link{numeric} vector of 0s
}
\description{
Return a vector of zeros
}
