% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static-xde.R
\name{dEIPdt.static_xde}
\alias{dEIPdt.static_xde}
\title{This function computes the negative derivative of the EIP as a function of time}
\usage{
\method{dEIPdt}{static_xde}(t, EIPmod)
}
\arguments{
\item{t}{current simulation time}

\item{EIPmod}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{dEIPdt} for the static_xde model (the dEIPdt=0)
}
