% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-GeM.R
\name{make_VC.GeM}
\alias{make_VC.GeM}
\title{Make the VC Matrix}
\usage{
\method{make_VC}{GeM}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a numeric \link{matrix}
}
\description{
Make the VC Matrix
}
\section{Functions}{
\itemize{
\item \code{make_VC(GeM)}: description

}}
\keyword{internal}
