% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development.no_dev}
\alias{Development.no_dev}
\title{Set up developments}
\usage{
\method{Development}{no_dev}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Development} for the no_dev model (do nothing)
}
