% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{Development.none}
\alias{Development.none}
\title{Set no development}
\usage{
\method{Development}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for development
}
