% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{dF_eip.static}
\alias{dF_eip.static}
\title{Static model for the EIP}
\usage{
\method{dF_eip}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{eip}, the extrinsic incubation period in days
}
\description{
Implements \link{F_eip} for a static model
}
\keyword{internal}
