% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-distribute-func.R
\name{DistributeBedNets.func}
\alias{DistributeBedNets.func}
\title{Set no distribute_bednets}
\usage{
\method{DistributeBedNets}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for distribute_bednets
}
