% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-no_net.R
\name{DistributeBedNets.no_nets}
\alias{DistributeBedNets.no_nets}
\title{Distribute bed nets}
\usage{
\method{DistributeBedNets}{no_nets}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{DistributeBedNets} for the no_nets model of bed nets (do nothing)
}
