% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-distribute-interface.R
\name{DistributeBedNets}
\alias{DistributeBedNets}
\title{Set the distribute_bednets}
\usage{
DistributeBedNets(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
distribute_bednets
}
