% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-none.R
\name{DistributeSugarBaits.none}
\alias{DistributeSugarBaits.none}
\title{Distribute no sugar baits}
\usage{
\method{DistributeSugarBaits}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The model for no sugar bait distribution
}
