% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-interface.R
\name{DistributeSugarBaits}
\alias{DistributeSugarBaits}
\title{Set the distribute_sugar_baits_sugar_baits}
\usage{
DistributeSugarBaits(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
distribute_sugar_baits_sugar_baits
}
