% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{dLdt.basicL}
\alias{dLdt.basicL}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dLdt}{basicL}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dLdt} for the basicL competition model.
}
