% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{dLdt.basicL_dts}
\alias{dLdt.basicL_dts}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{dLdt}{basicL_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dLdt} for the basicL_dts competition model.
}
