% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{dLdt}
\alias{dLdt}
\title{\strong{L} Component Derivatives}
\usage{
dLdt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
the derivatives, a \link{numeric} vector of length \eqn{n_q=}\code{nHabitats}
}
\description{
This method computes and returns the derivatives
for the \strong{L} Component modules
}
\note{
Dispatches on the class of \code{pars$Lpar[[s]]}
}
