% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{dLdt.trivial}
\alias{dLdt.trivial}
\title{Derivatives for \code{trivial} (\strong{L})}
\usage{
\method{dLdt}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length 0
}
\description{
Returns a numeric vector of length 0
}
\keyword{internal}
