% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{dMYdt.GeM}
\alias{dMYdt.GeM}
\title{Compute Derivatives for \strong{MY} module \code{GeM}}
\usage{
\method{dMYdt}{GeM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for the generalized, non-autonomous Macdonald model
for mosquito ecology and infection dynamics. See \link{GeM}.
}
\keyword{internal}
