% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{dMYdt.GeRM}
\alias{dMYdt.GeRM}
\title{\strong{MY} Component Derivatives for the \code{GeRM} model}
\usage{
\method{dMYdt}{GeRM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for the generalized, non-autonomous Ross-Macdonald model
for mosquito ecology and infection dynamics.
}
