% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{dMYdt.macdonald}
\alias{dMYdt.macdonald}
\title{Compute derivatives for the \strong{MY} module \code{macdonald}}
\usage{
\method{dMYdt}{macdonald}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
This implements a delay differential equation model for adult mosquito ecology and
infection dynamics that is consistent with the model published by George
Macdonald in 1952. A generalized version of this model, the \strong{MY} module \code{GeRM},
was developed to handle exogenous forcing by weather and vector control. This model
should be used only for educational purposes.

\strong{Variables:}
\itemize{
\item \eqn{M}: the density of adult mosquitoes
\item \eqn{Y}: the density of infected adult mosquitoes
\item \eqn{Z}: the density of infectious adult mosquitoes
}

\strong{Parameters and Terms:}
\itemize{
\item \eqn{\Lambda} or \code{Lambda}: the emergence rate of adult mosquitoes (from \code{F_emerge})
\item \eqn{f} or \code{f}: the blood feeding rate
\item \eqn{q} or \code{q}: maturation rate
\item \eqn{\tau} or \code{eip}: the extrinsic incubation period
\item \eqn{\Omega} or \code{Omega}: an adult mosquito demographic matrix, including mortality and migration
\item \eqn{\Upsilon} or \code{Upsilon}: survival and dispersal through the eip, \eqn{\Upsilon= e^{-\Omega \tau}}
}

\strong{Dynamics:}
In the delay equation, we use the subscript to denote the lagged value of a
variable or term: \emph{e.g.}, \eqn{M_\tau = M(t-\tau)}.

\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\
dZ/dt &= \Upsilon \cdot (fq\kappa)_\tau(M_\tau-Y_\tau) - \Omega \cdot Y \\
\end{array}}

This model was included mainly for the historical interest. It has been updated to handle
exogenous forcing by weather and vector control in the module \code{GeRM}
}
\note{
This model is not capable of being extended to
handle exogenous forcing by weather or vector control. Please
use the \code{GeRM} model.
}
