% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{dMYdt}
\alias{dMYdt}
\title{Compute Derivatives for an Adult Mosquito Model}
\usage{
dMYdt(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}

\item{s}{the species index}
}
\value{
Derivatives for an adult mosquito model, a \link{vector}
}
\description{
In the description, each method should include the equations and
a reference to relevant papers in the literature.
}
\note{
This is the \code{S3} generic. Methods dispatch on \code{MY_obj[[s]]}
}
