% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{dMYZdt.basicM}
\alias{dMYZdt.basicM}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{basicM}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the basicM xde ODE model.
}
\details{
\deqn{\begin{array}{rl}dM/dt &= \Lambda(t) - \Omega \cdot M \\ dP/dt &= f(M-P) - \Omega\cdot P\end{array}}
}
