% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ-dts.R
\name{dMYZdt.ENBRQ_dts}
\alias{dMYZdt.ENBRQ_dts}
\title{Derivatives for adult mosquitoes}
\usage{
\method{dMYZdt}{ENBRQ_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the ENBRQ_dts model.
}
