% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{dMYZdt.GeRM}
\alias{dMYZdt.GeRM}
\title{\strong{MYZ} Component Derivatives for the \code{GeRM} model}
\usage{
\method{dMYZdt}{GeRM}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for the generalized, non-autonomous Ross-Macdonald model
for mosquito ecology and infection dynamics.
}
