% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{dMYZdt.MY}
\alias{dMYZdt.MY}
\title{\strong{MYZ} Component Derivatives for the \code{MY} Mosquito Module}
\usage{
\method{dMYZdt}{MY}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a vector with the derivatives
}
\description{
The \code{MY} model for mosquito infection
dynamics has the defined \strong{variable} classes:
\itemize{
\item \eqn{M} is the density of mosquitoes in each patch;
\item \eqn{Y} is the density of infected mosquitoes in each patch.
}

The density of infectious mosquitoes in each patch is
given by a term (returned by \link{F_fqZ.MY}):
\deqn{Z = e^{-\Omega \tau} \cdot Y}

The model blood feeding \strong{parameters} are:
\itemize{
\item \eqn{f} is the overall blood feeding rate
\item \eqn{q} is the human fraction for blood feeding
}

The \strong{parameters} describing egg laying (see \link{F_eggs.MY}) are:
\itemize{
\item \eqn{\nu} is the egg laying rate
\item \eqn{\xi} is the number of eggs per batch
}

The model demographic \strong{parameters} are:
\itemize{
\item \eqn{g} is the mortality rate
\item \eqn{\sigma} is the emigration rate
\item \eqn{\mu} is the emigration loss rate
\item \eqn{\cal K} is the mosquito dispersal matrix
}

The four parameters describing mortality and migration are used to construct a
demographic matrix:
\deqn{\Omega = \mbox{diag}\left(g\right) - \left(\mbox{diag}\left(1-\mu\right) - \cal K \right) \cdot \mbox{diag}\left(\sigma\right)}

The emergence rate of adult mosquitoes, \eqn{\Lambda}, is computed by \link{F_emerge},
and the \strong{derivatives} are given by the equations:
\deqn{
\begin{array}{rr}
dM/dt = & \Lambda - \Omega \cdot M \\
dY/dt = & f q \kappa (M-Y) - \Omega \cdot Y
\end{array}.
}
}
\seealso{
\link{F_fqZ.MY} and \link{F_eggs.MY}
}
