% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{dMYZdt.mz}
\alias{dMYZdt.mz}
\title{Derivatives for a simple model for adult mosquito infection dynamics}
\usage{
\method{dMYZdt}{mz}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the mz ODE model.
}
