% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dMYZdt}
\alias{dMYZdt}
\title{\eqn{\cal MYZ} Component Derivatives for the \code{GeRM} model}
\usage{
dMYZdt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
derivatives for the \eqn{\cal MYZ} component as a \link{vector}
}
\description{
The description for each method should include the equations.
}
\note{
This is the \code{S3} generic. Methods dispatch on \code{MYZpar}
}
