% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{dMYZdt.si}
\alias{dMYZdt.si}
\title{\eqn{\cal MYZ} Component Derivatives for an SI Mosquito Model}
\usage{
\method{dMYZdt}{si}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Derivatives for a simple SI model for mosquito infection
dynamics, where \eqn{M} is the density of mosquitoes, and \eqn{Y} the
density of infected mosquitoes:
\deqn{
\begin{array}{rr}
dM/dt =& \Lambda &- \Omega \cdot M \\
dY/dt =& f q \kappa (M-Y) &- \Omega \cdot Y
\end{array}.
}
The model assumes \eqn{Z = \Upsilon Y}
}
