% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{DT_MYZt.basicM_dts}
\alias{DT_MYZt.basicM_dts}
\title{Derivatives for adult mosquitoes}
\usage{
\method{DT_MYZt}{basicM_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{DT_MYZt} for the basicM_dts model.
}
