% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{DT_MYZt.si}
\alias{DT_MYZt.si}
\title{Derivatives for adult mosquitoes}
\usage{
\method{DT_MYZt}{si}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{DT_MYZt} for the si model.
}
