% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{DT_MYZt.trace}
\alias{DT_MYZt.trace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
\method{DT_MYZt}{trace}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{DT_MYZt} for the trace (forced emergence) model.
}
