% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{DT_Xt.SEIS}
\alias{DT_Xt.SEIS}
\title{DTS updating for the SEIS model for human / vertebrate host infections}
\usage{
\method{DT_Xt}{SEIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{DT_Xt} for the SEIS model
}
