% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-dts.R
\name{DT_Xt.SIP_dts}
\alias{DT_Xt.SIP_dts}
\title{Derivatives for human population}
\usage{
\method{DT_Xt}{SIP_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{DT_Xt} for the SIP_dts model.
}
