% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{dts_make_MYZpar_si}
\alias{dts_make_MYZpar_si}
\title{Make parameters for si adult mosquito model}
\usage{
dts_make_MYZpar_si(
  nPatches,
  MYZday,
  calK,
  MYZopts = list(),
  g = 1/12,
  sigma = 1/8,
  mu = 0,
  f = 0.3,
  q = 0.95,
  eip = 12,
  nu = 1,
  eggsPerBatch = 60,
  p_mod = "static",
  sigma_mod = "static",
  mu_mod = "static",
  f_mod = "static",
  q_mod = "static",
  nu_mod = "static"
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZday}{is the runtime time step for the MYZ model}

\item{calK}{a mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{g}{daily mosquito survival}

\item{sigma}{emigration rate}

\item{mu}{emigration loss}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{eip}{the extrinsic incubation periodw}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{p_mod}{a name to dispatch F_p}

\item{sigma_mod}{a name to dispatch F_sigma}

\item{mu_mod}{a name to dispatch F_sigma}

\item{f_mod}{a name to dispatch F_f}

\item{q_mod}{a name to dispatch F_q}

\item{nu_mod}{a name to dispatch F_nu}
}
\value{
a \link{list}
}
\description{
Make parameters for si adult mosquito model
}
