% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{dts_make_Xpar_SEISd}
\alias{dts_make_Xpar_SEISd}
\title{Make parameters for SEISd human model, with defaults}
\usage{
dts_make_Xpar_SEISd(
  nStrata,
  Xday,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  nu = 1/14,
  c = 0.15
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xday}{the X component runtime time step}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{nu}{1/latent period}

\item{c}{transmission probability (efficiency) from human to mosquito}
}
\value{
a \link{list}
}
\description{
Make parameters for SEISd human model, with defaults
}
