% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_setup.R
\name{dts_setup}
\alias{dts_setup}
\title{Set up a model for dts_diffeqn}
\usage{
dts_setup(
  modelName = "unnamed",
  MYZname = "RM_dts",
  Xname = "SIS_dts",
  Lname = "trace",
  Dday = list(Xlump = 1, Xsub = 1, MYZlump = 1, MYZsub = 1, Llump = 1, Lsub = 1),
  nPatches = 1,
  nVectors = 1,
  nHosts = 1,
  HPop = 1000,
  membership = 1,
  MYZopts = list(),
  EIPopts = list(EIPname = "fixedlag_dts", eip = 12),
  calK = "herethere",
  calKopts = list(),
  Xopts = list(),
  BFopts = list(),
  residence = 1,
  searchB = 1,
  F_circadian = NULL,
  TimeSpent = "athome",
  TimeSpentOpts = list(),
  searchQ = 1,
  Lopts = list()
)
}
\arguments{
\item{modelName}{is a name for the model (arbitrary)}

\item{MYZname}{is a character string defining a MYZ model}

\item{Xname}{is a character string defining a X model}

\item{Lname}{is a character string defining a L model}

\item{Dday}{is a list of six integer values to set up mis-matched runtime time steps}

\item{nPatches}{is the number of patches}

\item{nVectors}{is the number of vector species}

\item{nHosts}{is the number of vertebrate host species}

\item{HPop}{is the number of humans in each patch}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{MYZopts}{a list to configure the MYZ model}

\item{EIPopts}{a list with options to setup EIPmod; must include EIPname}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{calKopts}{are the options to setup calK}

\item{Xopts}{a list to configure the X model}

\item{BFopts}{a list to configure the blood feeding model}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{searchB}{is a vector of search weights for blood feeding}

\item{F_circadian}{is a function describing mosquito daily activity}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{TimeSpentOpts}{are the options to setup TimeSpent}

\item{searchQ}{is a vector of search weights for egg laying}

\item{Lopts}{a list to configure the L model}
}
\value{
a \link{list}
}
\description{
Set up a model for dts_diffeqn
}
