% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{dts_setup_Lpar.basicL_dts}
\alias{dts_setup_Lpar.basicL_dts}
\title{Setup Lpar for the basicL_dts model}
\usage{
\method{dts_setup_Lpar}{basicL_dts}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_Lpar} for the basicL_dts model
}
