% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{dts_setup_Lpar.trace}
\alias{dts_setup_Lpar.trace}
\title{dts_setup Lpar for the trace model}
\usage{
\method{dts_setup_Lpar}{trace}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_Lpar} for the trace model
}
