% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{dts_setup_MYZpar.basicM_dts}
\alias{dts_setup_MYZpar.basicM_dts}
\title{Setup MYZpar for the basicM_dts model}
\usage{
\method{dts_setup_MYZpar}{basicM_dts}(MYZname, pars, s, EIPopts = list(), MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_MYZpar} for the basicM_dts model
}
