% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dts_setup_MYZpar}
\alias{dts_setup_MYZpar}
\title{A function to set up adult mosquito models}
\usage{
dts_setup_MYZpar(MYZname, pars, s, EIPopts, MYZopts = list(), calK = diag(1))
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYZname}.
}
