% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{dts_setup_MYZpar.si}
\alias{dts_setup_MYZpar.si}
\title{Setup MYZpar for the si model}
\usage{
\method{dts_setup_MYZpar}{si}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_MYZpar} for the si model
}
