% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Ztrace.R
\name{dts_setup_MYZpar.Ztrace}
\alias{dts_setup_MYZpar.Ztrace}
\title{dts_setup the Ztrace model}
\usage{
\method{dts_setup_MYZpar}{Ztrace}(MYZname, pars, s, EIPopts, MYZopts = NULL, calK = NULL)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_MYZpar} for the Ztrace model
}
