% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{dts_setup_Xpar.SEIS}
\alias{dts_setup_Xpar.SEIS}
\title{Setup Xpar for the discrete time SEIS model}
\usage{
\method{dts_setup_Xpar}{SEIS}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_Xpar} for the SEIS model
}
