% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{dts_setup_Xpar.SEISd}
\alias{dts_setup_Xpar.SEISd}
\title{Setup Xpar for the discrete time SEISd model}
\usage{
\method{dts_setup_Xpar}{SEISd}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_Xpar} for the SEISd model
}
