% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-dts.R
\name{dts_setup_Xpar.SIP_dts}
\alias{dts_setup_Xpar.SIP_dts}
\title{Setup Xpar.SIP_dts}
\usage{
\method{dts_setup_Xpar}{SIP_dts}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{dts_setup_Xpar} for the SIP_dts model
}
