% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_solving.R
\name{dts_solve}
\alias{dts_solve}
\title{Solve a discrete-time system}
\usage{
dts_solve(pars, ts = 365)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{ts}{the number of time steps}
}
\value{
a \link{list}
}
\description{
Implements \link{dts_solve} for equations
}
