% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{dts_stable_orbit}
\alias{dts_stable_orbit}
\title{Compute the stable orbit for a discrete time system}
\usage{
dts_stable_orbit(xds_obj, Tburn = 10, yr = 365)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{Tburn}{the number of steps to burn}

\item{yr}{the number of time steps in a year}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run a discrete time system for \code{Tburn} years
to get the stable orbit
}
