% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{dts_steady}
\alias{dts_steady}
\title{Compute the steady state of a discrete time system}
\usage{
dts_steady(xds_obj, Tx = 1000)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{Tx}{the number of steps to equilibrium}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run a discrete time system for \code{Tx} time steps to get the steady state
}
