% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dts_steady_state_X}
\alias{dts_steady_state_X}
\title{Compute the steady states as a function of the daily EIR}
\usage{
dts_steady_state_X(ar, H, Xpar)
}
\arguments{
\item{ar}{the daily attack rate}

\item{H}{host density}

\item{Xpar}{a list that defines a X-class model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}
}
