% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dts_steady_state_X.SIS}
\alias{dts_steady_state_X.SIS}
\title{Compute the steady states for the  dts SIS model as a function of the daily EIR}
\usage{
\method{dts_steady_state_X}{SIS}(ar, H, Xpar)
}
\arguments{
\item{ar}{the daily attack rate}

\item{H}{host density}

\item{Xpar}{a list that defines a X-class model}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the  dts SIS model as a function of the daily eir.
}
