% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{dts_update.eir}
\alias{dts_update.eir}
\title{Difference equation models for aquatic mosquito populations}
\usage{
\method{dts_update}{eir}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{vector} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
a model with only aquatic mosquitoes
}
