% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{dts_update.mosy}
\alias{dts_update.mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
\method{dts_update}{mosy}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{vector} containing the vector of all state derivatives
}
\description{
Compute and update the state variables for
a model with mosquito ecology (no transmission)
}
