% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{dVdt}
\alias{dVdt}
\title{Compute Other State Variables}
\usage{
dVdt(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the i\eqn{^{th}} set of auxiliary variables}
}
\value{
derivatives for the \eqn{\cal MYZ} component as a \link{vector}
}
\description{
The description for each method should include the equations.
}
\note{
This is the \code{S3} generic. Methods dispatch on \code{V_obj}
}
\keyword{internal}
