% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{dXdt.SEIS}
\alias{dXdt.SEIS}
\title{Derivatives for the SEIS model for human / vertebrate host infections}
\usage{
\method{dXdt}{SEIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SEIS model
}
