% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{dXdt.SEISd}
\alias{dXdt.SEISd}
\title{Derivatives for the SEISd model for human / vertebrate host infections}
\usage{
\method{dXdt}{SEISd}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SEISd model
}
