% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{dXdt.SIS_dts}
\alias{dXdt.SIS_dts}
\title{Derivatives for human population}
\usage{
\method{dXdt}{SIS_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIS_dts model, no demography.
}
