% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{dXdt.trivial}
\alias{dXdt.trivial}
\title{Derivatives for human population}
\usage{
\method{dXdt}{trivial}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the trivial model.
}
