% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{dXHdt.hMoI}
\alias{dXHdt.hMoI}
\title{Compute Derivatives for \code{hMoI} (\strong{X} Module)}
\usage{
\method{dXHdt}{hMoI}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXHdt} for the hybrid MoI model.
}
