% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{dXHdt}
\alias{dXHdt}
\title{Compute \strong{XH} Component Derivatives}
\usage{
dXHdt(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the derivatives as a \link{numeric} vector
}
\description{
Using the stored values
of the daily FoI, compute the derivatives and
return the derivatives as a numeric vector.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
