% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{dXHdt.SIS}
\alias{dXHdt.SIS}
\title{\code{SIS} Derivatives Function (an \strong{XH} Module)}
\usage{
\method{dXHdt}{SIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the derivatives, as a vector
}
\description{
Computes the derivatives for SIS compartmental model with variables:
\itemize{
\item \eqn{I} - the density of infected humans (or hosts)
\item \eqn{S} - the density of susceptible humans (or hosts)
\item \eqn{H} - human (or host) population density
}

The model assumes \eqn{S+I=H.} This function computes
computes \eqn{dH/dt} and \eqn{dI/dt}.
For convenience, \eqn{S} is also computed
by \link{get_XH_vars.SIS}
and \link{parse_XH_orbits.SIS}.

The force of infection or "happenings" rate is \eqn{h}. It is computed
upstream and used here.

The clearance rate for infections is \eqn{r}, and by assumption,
individuals are assumed to be susceptible to infection after clearing
infections.

The module includes a port to model mass treatment, \eqn{\xi(t)}.

\strong{Human Demographic}  change is modeled with two functions:
\itemize{
\item \eqn{B(t, H)} is the time-dependent population birth rate;
\item \eqn{D} is a linear operator, a matrix describing mortality, migration, aging,
and dynamical transfers among population strata.
}

The
derivatives computed are:

\deqn{
\begin{array}{rl}
dH/dt = & B(t,H)  + D \cdot H \\
dI/dt = & h (H-I) - r I - \xi(t) +  D \cdot I
\end{array}
}
}
\seealso{
\href{https://dd-harp.github.io/ramp.xds/articles/human_sis.html}{SIS Dynamics}
}
