% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{dXHdt.trivial}
\alias{dXHdt.trivial}
\title{Compute derivatives for \code{trivial} (\strong{XH})}
\usage{
\method{dXHdt}{trivial}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
The trivial model has no state variables so it returns
a numeric vector of length 0
}
\keyword{internal}
