% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{dynamic_forcing.setup}
\alias{dynamic_forcing.setup}
\title{Set up dynamic forcing}
\usage{
\method{dynamic_forcing}{setup}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
