% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{dynamic_health}
\alias{dynamic_health}
\title{Set up dynamic health}
\usage{
dynamic_health(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic health has not
already been set up, then turn on dynamic
health and set all the
}
