% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{dynamic_MYZpar}
\alias{dynamic_MYZpar}
\title{Call a function to set baseline bionomic parameters}
\usage{
dynamic_MYZpar(t, vars, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{current simulation time}

\item{MYZpar}{the mosquito model}
}
\value{
baseline bionomics as a named \link{list}
}
\description{
This method dispatches on the type of \code{baseline}.
}
