% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{dynamic_MYZpar.RMfunc}
\alias{dynamic_MYZpar.RMfunc}
\title{Update adult bionomic parameters}
\usage{
\method{dynamic_MYZpar}{RMfunc}(t, vars, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{current simulation time}

\item{MYZpar}{the mosquito model}
}
\value{
baseline bionomics as a named \link{list}
}
\description{
This method dispatches on the type of \code{baseline}.
}
