% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control.dynamic}
\alias{dynamic_vector_control.dynamic}
\title{Vector Control is Turned On}
\usage{
\method{dynamic_vector_control}{dynamic}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If
\code{class(pars$vector_control) == 'dynamic'}
then dynamic vector control has been turned
on. The unmodified \strong{\code{xds}} object is returned.
}
