% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control.none}
\alias{dynamic_vector_control.none}
\title{Turn On Vector Control}
\usage{
\method{dynamic_vector_control}{none}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If
\code{class(pars$vector_control) == 'none'}
then dynamic vector_control has not
been set up by any other function.
This sets \code{class(vector_control) <- 'dynamic'} and
then sets up a  trivial module for every mode of
vector control.
}
