% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{dynamic_weather.dynamic}
\alias{dynamic_weather.dynamic}
\title{Set up dynamic weather}
\usage{
\method{dynamic_weather}{dynamic}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic weather has not
already been set up, then turn on dynamic
weather and set all the
}
