% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamical_components.R
\name{dynamical_components}
\alias{dynamical_components}
\title{Dynamical Components}
\description{
Dynamical systems for malaria and other mosquito-borne diseases
are made up of three core \strong{dynamical components:}
\itemize{
\item \strong{XH} - a system of equations describing the dynamics of host infection and immunity (\strong{X}) and demography (\strong{H})
\item \strong{MY} - a system of equations describing adult mosquito ecology (\strong{M}) and infection dynamics (\strong{Y})
\item \strong{L} - a system of equations describing immature mosquito ecology
}

During setup, modules are selected by passing their names:

\describe{
\item{\code{Xname}}{a character string: module name for \strong{X} Component module}
\item{\code{MYname}}{a character string: module name for \strong{MY} Component module}
\item{\code{Lname}}{a character string: module name for \strong{L} Component module}
}

Other variables can also be added to the dynamical system (see \link{Other_State_Variables}).
}
\seealso{
\link{xds_object_frame}
}
