% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{egg_laying_dynamics}
\alias{egg_laying_dynamics}
\title{Compute eggs laid}
\usage{
egg_laying_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Computes eggs laid with exogenous forcing
on parameter affecting host availability.
}
\keyword{internal}
