% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{EggLaying.dynamic}
\alias{EggLaying.dynamic}
\title{Compute eggs laid}
\usage{
\method{EggLaying}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Computes eggs laid with exogenous forcing
on parameter affecting host availability.
}
