% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{EggLaying.setup}
\alias{EggLaying.setup}
\title{Compute eggs laid, the first time}
\usage{
\method{EggLaying}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
In autonomous models, this
function gets called after the \code{ML_interface}
is set up, or after any structural element in
the ML_interface changes

describing host availability (\eqn{Q}) and
the egg distribution matrix (\eqn{O}).
}
\details{
If conditions are time, invariant,
then \link{EggLaying.static} computes eggs laid. The functions that
compute \eqn{N} and \eqn{O} are called \emph{once} after setup.
The class of \code{ML_interface} is set to \code{static} and the function
is never called again. If any parameters are changed that would
affect egg laying, then the class of \code{ML_interface} should
get reset to \code{setup} to reconfigure \eqn{N} and \eqn{O}.
}
\seealso{
For \eqn{Q}, see \link{compute_Qall}

For \eqn{N}, see \link{make_habitat_matrix}
}
\keyword{internal}
