% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{EggLaying.static}
\alias{EggLaying.static}
\title{Compute eggs laid}
\usage{
\method{EggLaying}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Computes eggs laid for an autonomous model
}
