% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-fixedlag-dts.R
\name{EIP.fixedlag_dts}
\alias{EIP.fixedlag_dts}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{fixedlag_dts}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
the EIP maturation vector G \link{numeric}
}
\description{
Implements \link{EIP} for the fixedlag_dts model (the EIP is constant)
}
