% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static-xde.R
\name{EIP.static_xde}
\alias{EIP.static_xde}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{EIP}{static_xde}(t, MYZpar)
}
\arguments{
\item{t}{current simulation time}

\item{MYZpar}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{EIP} for the static_xde model (the EIP is constant)
}
