% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mectric-conversion.R
\name{eir2fqZ}
\alias{eir2fqZ}
\title{Convert the EIR into a vector describing infective biting density}
\usage{
eir2fqZ(eir, beta)
}
\arguments{
\item{eir}{a vector describing the EIR in several strata}

\item{beta}{the mixing matrix}
}
\value{
a numeric \link{vector}
}
\description{
Computes a vector of length \code{nStrata} describing the daily eir, per patch from
a vector of length \code{nPatches} describing the daily infective biting density.
\deqn{fqZ = \left(\beta^T \cdot \beta \right)^{-1} \cdot \beta^T \cdot E}
}
