% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emergence.R
\name{Emergence}
\alias{Emergence}
\title{Compute emerging adults}
\usage{
Emergence(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{a \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Sums up adults emerging from all aquatic habitats using the habitat membership matrix
}
