% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous.R
\name{Exogenous.human}
\alias{Exogenous.human}
\title{Differential equations isolating the humans, forced with Ztrace}
\usage{
\method{Exogenous}{human}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
