% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous.R
\name{Exogenous.mosy}
\alias{Exogenous.mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
\method{Exogenous}{mosy}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Mirrors \link{Exogenous} but only includes the adult and aquatic
mosquito components.
}
