% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exogenous.R
\name{Exogenous}
\alias{Exogenous}
\title{Compute Derivatives}
\usage{
Exogenous(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The function to compute the derivatives dispatches on \code{class(frame)}
}
\seealso{
\link{xds_solve}
}
