% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.dts}
\alias{Exposure.dts}
\title{Compute the attack rate (AR) given the local EIR and a travel model}
\usage{
\method{Exposure}{dts}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object; ARs are stored as \code{pars$AR}
}
\description{
For dts models, compute the attack rate (AR) for
all the population strata. The AR is a function of local daily entomological
inoculation rates (dEIR), and a model for exposure to malaria while traveling.
}
