% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.xde}
\alias{Exposure.xde}
\title{Compute the FoI}
\usage{
\method{Exposure}{xde}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For \strong{\code{xde}} models, compute the FoI
}
\details{
The local force of infection (FoI or \eqn{h}) is
a function of local daily entomological inoculation
rate (dEIR or \eqn{E}) under
a model for the probability of infection
per infectious bite, \eqn{b} (called from \eqn{F_b},
as defined by an \eqn{\cal X} model).
The total FoI is a weighted sum of the local FoI and
exposure to malaria while traveling, (\eqn{T_h}):
\deqn{h = (1-\delta) \; F_h(E, b) + \delta\; T_h(b,t)}
}
\seealso{
Related: \link{Exposure} & \link{F_foi.pois} & \link{F_foi.nb} & \link{travel_malaria}
}
