% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{F_all_available_water}
\alias{F_all_available_water}
\title{Compute Available Laying Sites, \eqn{O}}
\usage{
F_all_available_water(Q, Q_traps, Q_bad)
}
\arguments{
\item{Q}{the availability of habitats}

\item{Q_traps}{the availability of traps}

\item{Q_bad}{the availability of unsuitable habitats}
}
\value{
Availability laying sites, \eqn{O}
}
\description{
Mosquitoes lay eggs in aquatic habitats, in
water bodies that are unsuitable for larval development
(bad habitats), and in ovitraps.
}
\details{
Habitat availability denoted \eqn{Q},
sums habitat search weights, \eqn{\omega}, by patch
using the habitat membership matrix, \eqn{N}:
\deqn{Q = {N} \cdot \omega_h.}

If some ovitraps (\eqn{Q_t}) and bad_habitats (\eqn{Q_b}) are
also available, then total availability of laying sites is
\deqn{O = Q + Q_t + Q_b.}

The availability of habitats, ovitraps and bad_habitats are computed elsewhere and stored on \code{xds_obj$vars}.
}
\seealso{
This function is called by \link{compute_all_available_water}

\link{make_habitat_matrix} discusses \eqn{N}

The availability of traps and bad habitats is setup in \link{setup_egg_laying}
}
\keyword{internal}
