% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{F_alpha.basicL_dts}
\alias{F_alpha.basicL_dts}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_alpha}{basicL_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_alpha} for the basicL_dts competition model.
}
