% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{F_ar.nb}
\alias{F_ar.nb}
\title{A negative binomial model for the attack rate as a function of the daily EIR.}
\usage{
\method{F_ar}{nb}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \code{xds} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ar} for a negative binomial model
}
