% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{F_ar.pois}
\alias{F_ar.pois}
\title{A Poisson model for the daily attack rate as a function of the daily EIR.}
\usage{
\method{F_ar}{pois}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{the model object as a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ar} for a Poisson model
}
