% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{F_ar}
\alias{F_ar}
\title{Compute Local Attack Rates}
\usage{
F_ar(eir, b, env_het_obj)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{env_het_obj}{an environmental heterogeneity model object}
}
\value{
a \link{numeric} vector: local attack rates for the strata
}
\description{
Compute the local attack rates as a function of the daily EIR and immunity.
}
\details{
This function computes the daily
local attack rates (ARs)
as a function of the daily EIR under
a probabilistic model for environmentally
heterogeneous exposure --
the distribution of bites per person.

The model for human / host infection (\eqn{\cal X}) provide
a term \eqn{b}, from \link{F_infectivity} that describe the probability of infection
per infectious bite, possibly affected by
pre-erythrocytic immunity
}
\seealso{
Cases: \link{F_ar.nb} and \link{F_ar.pois}. Related: \link{ar2eir} & \link{Exposure.dts}
}
