% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{F_b.SIS_xde}
\alias{F_b.SIS_xde}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_b}{SIS_xde}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_b} for the SIS_xde model.
}
