% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{F_B_available}
\alias{F_B_available}
\title{Compute Vertebrate Host available for Blood Feeding}
\usage{
F_B_available(W, visitors, other_blood, Btraps)
}
\arguments{
\item{W}{available of the parasite's / pathogen' local hosts}

\item{visitors}{available of \emph{visitors,} or non-resident host populations}

\item{other_blood}{available of other vertebrate hosts}

\item{Btraps}{blood feeding trap availability}
}
\value{
host available, a \link{vector}
}
\description{
Compute the available all vertebrate hosts for blood feeding by mosquitoes
}
\details{
The available of other vertebrate hosts is a sum of available local
hosts \eqn{W}, and visitors \eqn{W_\delta}, and other available vertebrate hosts, \eqn{O}.
Total available is a simple sum: \deqn{B = W + W_\delta + O}
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\keyword{internal}
