% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_batch_rate.Q2}
\alias{F_batch_rate.Q2}
\title{Type 2 Functional Response for Egg Laying}
\usage{
\method{F_batch_rate}{Q2}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_batch_rate} for a static model
}
