% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_batch_rate}
\alias{F_batch_rate}
\title{Compute the Mosquito Patch Emigration Rate}
\usage{
F_batch_rate(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector onu length \code{nPatches}
}
\description{
This method dispatches on the type of \code{nu_obj}. It should
set the values the patch emigration rate, \eqn{\nu}
}
