% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_batch_rate.static}
\alias{F_batch_rate.static}
\title{Static model patch emigration}
\usage{
\method{F_batch_rate}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{nu}, the patch emigration rate
}
\description{
Implements \link{F_batch_rate} for a static model
}
