% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-functions.R
\name{F_calK}
\alias{F_calK}
\title{Compute the egg laying rate}
\usage{
F_calK(t, vars, calK_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{calK_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method ...
}
