% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-GeRM.R
\name{F_calK.static}
\alias{F_calK.static}
\title{Static model for mosquito emigration}
\usage{
\method{F_calK}{static}(t, vars, calK_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{calK_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_calK} for a static model
}
