% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_dispersal_loss}
\alias{F_dispersal_loss}
\title{Compute the emigration loss fraction}
\usage{
F_dispersal_loss(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector omu length \code{nPatches}
}
\description{
This method dispatches on the type of \code{mu_obj}. It should
set the values omu the bionomic parameters to baseline values
}
