% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_dispersal_loss.static}
\alias{F_dispersal_loss.static}
\title{Static model for the blood feeding rate}
\usage{
\method{F_dispersal_loss}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{mu}, the baseline human fraction
}
\description{
Implements \link{F_dispersal_loss} for a static model
}
