% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_dlay_maturation}
\alias{F_dlay_maturation}
\title{Compute the blood xieeding rate, xi}
\usage{
F_dlay_maturation(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector oxi length \code{nPatches}
}
\description{
This method dispatches on the type of \code{xi_obj}. It should
set the values oxi the bionomic parameters to baseline values
}
