% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{F_eggs.RMdts}
\alias{F_eggs.RMdts}
\title{Number of eggs laid by adult mosquitoes}
\usage{
\method{F_eggs}{RMdts}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_eggs} for the RMdts model.
}
