% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{F_eggs.si}
\alias{F_eggs.si}
\title{\eqn{\cal MYZ} Component Egg Laying for the \code{si} Mosquito Model}
\usage{
\method{F_eggs}{si}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
The density of adult mosquitoes is \eqn{M}.
The \strong{parameters} describing egg laying by adult mosquitoes are:
\itemize{
\item \eqn{\nu} or \code{nu} is the egg laying rate
\item \eqn{\xi} or \code{eggsPerBatch} is the number of eggs per batch
}

The egg laying rate, per patch, is \deqn{\nu \xi M}
}
