% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-fixedlag-dts.R
\name{F_eip.fixedlag_dts}
\alias{F_eip.fixedlag_dts}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{F_eip}{fixedlag_dts}(t, vars, eip_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{eip_par}{a \link{list}}
}
\value{
the EIP maturation vector G \link{numeric}
}
\description{
Implements \link{F_eip} for the fixedlag_dts model (the EIP is constant)
}
