% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{F_eip.random}
\alias{F_eip.random}
\title{Compute random extrinsic incubation periods}
\usage{
\method{F_eip}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch-level EIP values with the configured positive
random number generator.
}
\keyword{internal}
