% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{F_eip}
\alias{F_eip}
\title{Compute the EIP}
\usage{
F_eip(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{eip_obj} and computes
the extrinsic incubation period.
}
\keyword{internal}
