% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_eir}
\alias{F_eir}
\title{Compute the daily Entomological Inoculation Rate (EIR)}
\usage{
F_eir(fqZ, beta, local_frac)
}
\arguments{
\item{fqZ}{the infective biting density}

\item{beta}{the mixing matrix}

\item{local_frac}{is the fraction of bites occurring on residents}
}
\value{
\link{numeric} vector of length \code{nStrata}
}
\description{
Compute the daily EIR for a set of human population
strata given the biting density of infective mosquitoes in each patch
}
