% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{F_emerge.basicL}
\alias{F_emerge.basicL}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_emerge}{basicL}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_emerge} for the basicL competition model.
}
