% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{F_emerge.trivial}
\alias{F_emerge.trivial}
\title{Compute emergent adults for \code{trivial} (\strong{L} component)}
\usage{
\method{F_emerge}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
The number of emerging adults is a function \deqn{\Lambda F_S(t) F_T(t) F_K(t)} where
\itemize{
\item \eqn{\Lambda} or \code{Lambda} is the mean number of adult female mosquitoes emerging per day
\item \eqn{F_S(t)} or \code{F_season} is a seasonal signal (ideally, with an average annual mean of 1)
\item \eqn{F_T(t)} or \code{F_trend} is a function returning a trend (ideally, with an average value of 1)
\item \eqn{F_K(t)} or \code{F_shock} is a function describing a perturbation (by default, set to 1)
}
}
\keyword{internal}
