% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_emigrate.BQS}
\alias{F_emigrate.BQS}
\title{Model for mosquito emigration based on resource availability}
\usage{
\method{F_emigrate}{BQS}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_emigrate} for a static model
}
