% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_emigrate.static}
\alias{F_emigrate.static}
\title{Static model patch emigration}
\usage{
\method{F_emigrate}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{sigma}, the patch emigration rate
}
\description{
Implements \link{F_emigrate} for a static model
}
