% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{F_eta}
\alias{F_eta}
\title{Eggs Laying in Habitats}
\usage{
F_eta(eggs_laid, U_matrix, Q, Qall)
}
\arguments{
\item{eggs_laid}{the number of eggs laid in each patch, a vector of length \code{nPatches}}

\item{U_matrix}{the egg laying matrix}

\item{Q}{larval habitat availability}

\item{Qall}{total availability}
}
\value{
a \link{vector}, \eqn{\eta} where \eqn{\left|\eta\right|=}\code{nHabitats}
}
\description{
Computes egg distribution for the aquatic habitats
}
\seealso{
\link{compute_U_matrix}
}
\keyword{internal}
