% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{F_f.random}
\alias{F_f.random}
\title{Compute random blood feeding rates}
\usage{
\method{F_f}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch-level blood feeding rates using the configured
random number generator.
}
\keyword{internal}
