% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{F_f}
\alias{F_f}
\title{Compute the blood feeding rate, f}
\usage{
F_f(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Set the baseline value of the feeding rate, \eqn{f}.
}
\note{
This method dispatches on the type of \code{f_obj} attached to the \code{MY_obj}.
}
\keyword{internal}
