% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{F_f.type2}
\alias{F_f.type2}
\title{Compute the blood feeding rate, f}
\usage{
\method{F_f}{type2}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{f}, the baseline blood feeding rate
}
\description{
Implements a type2 functional response to compute
blood feeding rates as a functional response to resource availability:
\deqn{F_f(B)= f_x \frac{s_f B}{1+s_f B}}.
}
\keyword{internal}
