% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_feeding_rate.B2}
\alias{F_feeding_rate.B2}
\title{Type 2 functional response for the blood feeding rate}
\usage{
\method{F_feeding_rate}{B2}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_feeding_rate} for a static model
}
