% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{F_foi.nb}
\alias{F_foi.nb}
\title{A negative binomial model for Exposure}
\usage{
\method{F_foi}{nb}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the daily FoI, \eqn{h},
given the daily EIR, \eqn{E},
under a negative binomial model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{h = \log{1 + b E}/\phi}
The model is parameterized using the \strong{mu} \eqn{= E} and \strong{sz} \eqn{= 1/\phi}
parameterization.
}
