% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{F_fqM.basicM}
\alias{F_fqM.basicM}
\title{The net blood feeding rate of the mosquito population in a patch}
\usage{
\method{F_fqM}{basicM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqM} for the basicM xde model.
}
