% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM-xde.R
\name{F_fqM.GeRM_xde}
\alias{F_fqM.GeRM_xde}
\title{Blood feeding rate of the infective mosquito population}
\usage{
\method{F_fqM}{GeRM_xde}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqM} for the GeRM_xde model.
}
