% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-sei.R
\name{F_fqM.sei}
\alias{F_fqM.sei}
\title{The net blood feeding rate of the infective mosquito population in a patch}
\usage{
\method{F_fqM}{sei}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqM} for the sei model.
}
