% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{F_fqZ.basicM_xde}
\alias{F_fqZ.basicM_xde}
\title{The net blood feeding rate of the infective mosquito population in a patch}
\usage{
\method{F_fqZ}{basicM_xde}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqZ} for the basicM xde model.
}
