% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{F_fqZ.macdonald}
\alias{F_fqZ.macdonald}
\title{The net blood feeding rate of the infective mosquito population in a patch}
\usage{
\method{F_fqZ}{macdonald}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_fqZ} for the macdonald model.
}
