% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{F_fqZ.SI}
\alias{F_fqZ.SI}
\title{Net blood feeding by infectious mosquitoes - \code{SI} mosquito model}
\usage{
\method{F_fqZ}{SI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
The variable \eqn{Y} is the density of \emph{infected} mosquitoes.
The model blood feeding \strong{parameters} are:
\itemize{
\item \eqn{f} is the overall blood feeding rate
\item \eqn{q} is the human fraction for blood feeding
\item \eqn{\Omega} is the demographic matrix
\item \eqn{\tau} is the EIP
}

The net blood feeding rate by infectious mosquitoes is computed by accounting
for survival and dispersal that would occur in a delay differential equation:
\deqn{\Upsilon = e^{-\Omega \tau}}
so
\deqn{fqZ = fq (\Upsilon \cdot Y)}
The daily EIR for the population strata is \eqn{\beta \cdot fqZ}
}
\keyword{internal}
