% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{F_g.random}
\alias{F_g.random}
\title{Compute random mosquito mortality rates}
\usage{
\method{F_g}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch-level mosquito mortality rates using the
configured random number generator.
}
\keyword{internal}
