% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{F_g.static}
\alias{F_g.static}
\title{Static model mosquito mortality rate}
\usage{
\method{F_g}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{g}, the baseline mosquito mortality rate
}
\description{
Implements \link{F_g} for a static model
}
\keyword{internal}
