% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI-xde.R
\name{F_H.hMoI_xde}
\alias{F_H.hMoI_xde}
\title{Size of the human population}
\usage{
\method{F_H}{hMoI_xde}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the hybrid MoI model.
}
