% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{F_H}
\alias{F_H}
\title{Size of human population denominators}
\usage{
F_H(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
