% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_H.SEISd}
\alias{F_H.SEISd}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SEISd}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the SEISd model.
}
