% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{F_H.SIP}
\alias{F_H.SIP}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SIP}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SIP model.
}
