% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-dts.R
\name{F_H.SIP_dts}
\alias{F_H.SIP_dts}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SIP_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SIP_dts model.
}
