% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{F_H.SIPd}
\alias{F_H.SIPd}
\title{Size of effective infectious human population}
\usage{
\method{F_H}{SIPd}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SIPd model.
}
