% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{F_H.SIS}
\alias{F_H.SIS}
\title{Compute Population Density}
\usage{
\method{F_H}{SIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
Human (host) population density, \eqn{H}
}
\description{
Implements \link{F_H} for the \emph{SIS} module
}
\keyword{internal}
