% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{F_H.trivial}
\alias{F_H.trivial}
\title{Size of the human population}
\usage{
\method{F_H}{trivial}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_H} for the trivial model.
}
