% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-object.R
\name{F_H.xds_obj}
\alias{F_H.xds_obj}
\title{Get the human population size}
\usage{
\method{F_H}{xds_obj}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This makes the initial human population sizes available
during \link{make_xds_object}
}
\note{
This method dispatches on the type of \code{pars$Xpar[[i]]}
}
