% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{F_human_frac.static}
\alias{F_human_frac.static}
\title{Static model for the blood feeding rate}
\usage{
\method{F_human_frac}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{q}, the baseline human fraction
}
\description{
Implements \link{F_human_frac} for a static model
}
