% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{F_I.trivial}
\alias{F_I.trivial}
\title{Size of effective infectious human population}
\usage{
\method{F_I}{trivial}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
The trace function for \eqn{\kappa} follows
the \strong{\code{ramp.xds}} standard for time series forcing. The
trace function for \eqn{H} can be modified to follow a
trend. Here, \eqn{I = \kappa H}.
}
