% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{F_infectivity.hMoI}
\alias{F_infectivity.hMoI}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
\method{F_infectivity}{hMoI}(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_infectivity} for the hMoI model.
}
