% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{F_infectivity}
\alias{F_infectivity}
\title{Infection blocking pre-erythrocytic immunity}
\usage{
F_infectivity(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
