% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{F_infectivity.SIS}
\alias{F_infectivity.SIS}
\title{Compute Infectivity}
\usage{
\method{F_infectivity}{SIS}(y, xds_obj, i)
}
\arguments{
\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
The constant \eqn{b}
}
\description{
This function computes the fraction of infectious
bites that cause an infection. The SIS model,
assumes a constant fraction causes infection, \eqn{b}.
}
