% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{F_K_matrix}
\alias{F_K_matrix}
\title{Dynamically update a K Matrix}
\usage{
F_K_matrix(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
A port function to
updates the mosquito dispersal
matrix dynamically.
}
\note{
The port object is called \code{K_obj}
}
\keyword{internal}
