% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{F_K_matrix.static}
\alias{F_K_matrix.static}
\title{Dynamically update a K Matrix}
\usage{
\method{F_K_matrix}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{F_K_matrix} for a static model
}
\keyword{internal}
