% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_kappa}
\alias{F_kappa}
\title{Net infectiousness of human population to mosquitoes}
\usage{
F_kappa(Wi, W, beta, X)
}
\arguments{
\item{Wi}{availability of this host}

\item{W}{availability of all hosts}

\item{beta}{the mixing matrix}

\item{X}{the infectious density of the strata}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the net infectiousness of humans in each patch at time \code{t}
}
