% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_larval_dd_mort.static}
\alias{F_larval_dd_mort.static}
\title{Static model for the blood feeding rate}
\usage{
\method{F_larval_dd_mort}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{theta}, the baseline human fraction
}
\description{
Implements \link{F_larval_dd_mort} for a static model
}
