% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_larval_mort}
\alias{F_larval_mort}
\title{Compute the blood phieeding rate, phi}
\usage{
F_larval_mort(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector ophi length \code{nPatches}
}
\description{
This method dispatches on the type of \code{phi_obj}. It should
set the values ophi the bionomic parameters to baseline values
}
