% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{F_local_frac}
\alias{F_local_frac}
\title{Compute the local fraction}
\usage{
F_local_frac(W, Visitors)
}
\arguments{
\item{W}{availability of blood hosts}

\item{Visitors}{availability of visitors}
}
\value{
the fraction of bites
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
\keyword{internal}
