% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_maturation}
\alias{F_maturation}
\title{Compute the blood psieeding rate, psi}
\usage{
F_maturation(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector opsi length \code{nPatches}
}
\description{
This method dispatches on the type of \code{psi_obj}. It should
set the values opsi the bionomic parameters to baseline values
}
