% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_maturation.static}
\alias{F_maturation.static}
\title{Static model for the blood feeding rate}
\usage{
\method{F_maturation}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{a \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{psi}, the baseline human fraction
}
\description{
Implements \link{F_maturation} for a static model
}
