% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{F_mu.random}
\alias{F_mu.random}
\title{Compute random emigration-loss fractions}
\usage{
\method{F_mu}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch-level emigration-loss fractions using the
configured random number generator.
}
\keyword{internal}
