% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{F_mu}
\alias{F_mu}
\title{Compute the emigration loss fraction}
\usage{
F_mu(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{mu_obj}. It should
compute the baseline emigration-loss fraction, \eqn{\mu}
}
\keyword{internal}
