% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{F_mu.static}
\alias{F_mu.static}
\title{Static model emigration-loss fraction}
\usage{
\method{F_mu}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{mu}, the baseline emigration-loss fraction
}
\description{
Implements \link{F_mu} for a static model
}
\keyword{internal}
