% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_ni.hMoI}
\alias{F_ni.hMoI}
\title{Compute the "true" nievalence of infection / parasite rate}
\usage{
\method{F_ni}{hMoI}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ni} for the hMoI model.
}
