% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{F_ni.trivial}
\alias{F_ni.trivial}
\title{Compute the NI}
\usage{
\method{F_ni}{trivial}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_ni} for the trivial model.
}
