% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{F_no_trend}
\alias{F_no_trend}
\title{A trivial function F_no_trend}
\usage{
F_no_trend(t, trend_opts = list())
}
\arguments{
\item{t}{the time}

\item{trend_opts}{a list of options}
}
\value{
a vector of ones of length x
}
\description{
A function that returns 1
}
