% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{F_nu.random}
\alias{F_nu.random}
\title{Compute random egg laying rates}
\usage{
\method{F_nu}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{nu}, the egg laying rate
}
\description{
Compute random egg laying rates
}
\keyword{internal}
