% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{F_nu}
\alias{F_nu}
\title{Compute the Egg Laying Rate}
\usage{
F_nu(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{nu_obj}. It should
set the values of the egg laying rate, \eqn{\nu}
}
\keyword{internal}
