% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{F_nu.static}
\alias{F_nu.static}
\title{Static model egg laying rate}
\usage{
\method{F_nu}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{nu}, the egg laying rate
}
\description{
Implements \link{F_nu} for a static model
}
\keyword{internal}
