% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{F_nu.type2}
\alias{F_nu.type2}
\title{Egg laying rate}
\usage{
\method{F_nu}{type2}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{\nu}, the baseline egg laying rate
}
\description{
Egg laying rates are a type2 functional response
to available water.

Letting \eqn{Q} be the availability of all water. The egg laying rate is
\deqn{\nu = \nu_x \frac{s_\nu Q}{1+s_\nu Q}}
}
\keyword{internal}
