% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{F_Omega_dts}
\alias{F_Omega_dts}
\title{F_Omega for dts}
\usage{
F_Omega_dts(p, ssigma, mu, K_matrix)
}
\arguments{
\item{p}{mosquito daily survival, a vector of length \code{nPatches}}

\item{ssigma}{mosquito daily fraction emigrating}

\item{mu}{emigration survival, a vector of length \code{nPatches}}

\item{K_matrix}{a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
F_Omega for dts
}
\keyword{internal}
