% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{F_Omega_xde}
\alias{F_Omega_xde}
\title{F_Omega for xde}
\usage{
F_Omega_xde(g, sigma, mu, K_matrix)
}
\arguments{
\item{g}{mosquito death rate, a vector of length \code{nPatches}}

\item{sigma}{mosquito emigration rate, a vector of length \code{nPatches}}

\item{mu}{emigration loss, a vector of length \code{nPatches}}

\item{K_matrix}{a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
Computes the mosquito demographic matrix:
\deqn{
 \Omega = \mbox{diag} \left( g + \sigma \mu \right) - K \cdot \mbox{diag} \left( \sigma \left(1-\mu\right) \right)
}
}
\seealso{
\link{xds_info_mosquito_demography}
}
\keyword{internal}
