% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_p}
\alias{F_p}
\title{Compute the human blood fraction}
\usage{
F_p(t, vars, p_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{p_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{p_par}. It should
set the values of the bionomic parameters to baseline values.
}
