% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{F_pfpr_by_pcr.trivial}
\alias{F_pfpr_by_pcr.trivial}
\title{Compute the prevalence of infection by PCR}
\usage{
\method{F_pfpr_by_pcr}{trivial}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_prevalence} for the trivial model.
}
