% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_pfpr_by_rdt.hMoI}
\alias{F_pfpr_by_rdt.hMoI}
\title{Compute the prevalence of infection by RDT}
\usage{
\method{F_pfpr_by_rdt}{hMoI}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the hMoI model.
}
