% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_phi}
\alias{F_phi}
\title{Compute the blood phieeding rate, phi}
\usage{
F_phi(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector ophi length \code{nHabitats}
}
\description{
This method dispatches on the type of \code{phi_obj}. It should
set the values ophi the bionomic parameters to baseline values
}
\keyword{internal}
