% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_phi.static}
\alias{F_phi.static}
\title{Compute phi, density independent mortality rate}
\usage{
\method{F_phi}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{phi}, the baseline human fraction
}
\description{
Implements \link{F_phi} for a static model
}
\keyword{internal}
