% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{F_pr.hMoI}
\alias{F_pr.hMoI}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{hMoI}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human infections}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the hMoI model.
}
