% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI-xde.R
\name{F_pr.hMoI_xde}
\alias{F_pr.hMoI_xde}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{hMoI_xde}(varslist, pars, i)
}
\arguments{
\item{varslist}{a list with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the hMoI_xde model.
}
