% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_pr}
\alias{F_pr}
\title{Compute the \emph{true} prevalence of infection / parasite rate}
\usage{
F_pr(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
