% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{F_pr.trivial}
\alias{F_pr.trivial}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_pr}{trivial}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_pr} for the trivial model.
}
