% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr_by_pcr.SEISd}
\alias{F_pr_by_pcr.SEISd}
\title{Compute the prevalence of infection by pcr}
\usage{
\method{F_pr_by_pcr}{SEISd}(varslist, pars, i)
}
\arguments{
\item{varslist}{a parsed list of outputs with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEISd model.
}
