% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr_by_rdt.SEISd}
\alias{F_pr_by_rdt.SEISd}
\title{Compute the prevalence of infection by RDT}
\usage{
\method{F_pr_by_rdt}{SEISd}(varslist, pars, i)
}
\arguments{
\item{varslist}{a parsed list of outputs with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEISd model.
}
