% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{F_pr_by_rdt.SIPd}
\alias{F_pr_by_rdt.SIPd}
\title{Compute the prevalence of infection by RDT}
\usage{
\method{F_pr_by_rdt}{SIPd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SIPd model.
}
