% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{F_pr_by_rdt.trace}
\alias{F_pr_by_rdt.trace}
\title{Compute the prevalence of infection by RDT}
\usage{
\method{F_pr_by_rdt}{trace}(varslist, pars, i)
}
\arguments{
\item{varslist}{a list with variables attached by name}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_pr} for the trace model.
}
