% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{F_prevalence}
\alias{F_prevalence}
\title{Compute the \emph{true} prevalence of infection / parasite rate}
\usage{
F_prevalence(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A function that translates the state variables into
the "true" \emph{Pf}PR
}
