% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{F_q.random}
\alias{F_q.random}
\title{Compute random human feeding fractions}
\usage{
\method{F_q}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch-level human feeding fractions using the
configured random number generator.
}
\keyword{internal}
