% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{F_q.static}
\alias{F_q.static}
\title{Static model human feeding fraction}
\usage{
\method{F_q}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{q}, the baseline human feeding fraction
}
\description{
Implements \link{F_q} for a static model
}
\keyword{internal}
