% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{F_q.WVB}
\alias{F_q.WVB}
\title{WVB human feeding fraction}
\usage{
\method{F_q}{WVB}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{q}, the human feeding fraction
}
\description{
Computes the human feeding fraction from the availability of
resident hosts (\eqn{W}), visitors (\eqn{V}), and all available blood hosts
(\eqn{B}).

\deqn{q = \frac{W+V}{B}}
}
\keyword{internal}
