% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{F_rbr}
\alias{F_rbr}
\title{Compute relative biting rates}
\usage{
F_rbr(search_weights, H)
}
\arguments{
\item{search_weights}{blood feeding search weights for the host strata}

\item{H}{host density}
}
\value{
host available, a \link{vector}
}
\description{
Relative biting rates translate biting weights into a
frailty parameter -- a multiplicative term.
}
