% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{F_season.MY}
\alias{F_season.MY}
\title{Compute the seasonal pattern}
\usage{
\method{F_season}{MY}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the seasonal pattern, invisibly
}
\description{
Evaluate the function \code{F_season} for a trivial \strong{MY} model
}
\keyword{internal}
