% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{F_season}
\alias{F_season}
\title{Compute the seasonal pattern}
\usage{
F_season(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the seasonal pattern, invisibly
}
\description{
Evaluate the function \code{F_season} for a forced model.
}
\keyword{internal}
