% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_shock.R
\name{F_shock.kappa}
\alias{F_shock.kappa}
\title{Compute the temporal shock}
\usage{
\method{F_shock}{kappa}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal shock
}
\description{
Evaluate the function \code{F_shock} for a trivial \strong{XH} model
}
\keyword{internal}
