% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_shock.R
\name{F_shock}
\alias{F_shock}
\title{Compute the temporal shock}
\usage{
F_shock(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal shock, invisibly
}
\description{
Evaluate the function \code{F_shock} for a forced model.
}
\keyword{internal}
