% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{F_sigma.random}
\alias{F_sigma.random}
\title{Random patch emigration rate}
\usage{
\method{F_sigma}{random}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Generate patch emigration rates using the configured random
number generator.
}
\keyword{internal}
