% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{F_sigma}
\alias{F_sigma}
\title{Compute the mosquito patch emigration rate}
\usage{
F_sigma(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{sigma_obj} and computes
the patch emigration rate, \eqn{\sigma}.
}
\keyword{internal}
