% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{F_sigma.type2}
\alias{F_sigma.type2}
\title{Patch emigration rate}
\usage{
\method{F_sigma}{type2}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{\sigma}, the patch emigration rate
}
\description{
Implements a type 2 functional response to the availability of three resources:
blood hosts (\eqn{B}), aquatic habitats (\eqn{Q}), and sugar (\eqn{S}).

\deqn{\sigma = \sigma_X \left( \frac{\sigma_Q Q}{1+\sigma_Q Q} 
+ \frac{\sigma_B B}{1+\sigma_B B}
+ \frac{\sigma_S S}{1+\sigma_S S}
\right)}
}
\keyword{internal}
