% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{F_sin_season}
\alias{F_sin_season}
\title{A trivial function F_no_season}
\usage{
F_sin_season(t, phase, season_opts = list(pwr = 1))
}
\arguments{
\item{t}{the time}

\item{phase}{the phase}

\item{season_opts}{a list of options}
}
\value{
a vector of ones of length x
}
\description{
A function that returns 1
}
