% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{F_TaR}
\alias{F_TaR}
\title{Compute blood feeding available of all vertebrate hosts}
\usage{
F_TaR(t, TiSp, F_circadian, time_at_home)
}
\arguments{
\item{t}{the time}

\item{TiSp}{a time spent matrix}

\item{F_circadian}{a function to compute relative activity rates by time of day}

\item{time_at_home}{the fraction of time spent at home}
}
\value{
a TaR \link{matrix}
}
\description{
Computes available all vertebrate hosts to blood feeding
mosquitoes
}
\details{
The available of other vertebrate hosts is a sum of available local
hosts \eqn{W}, and visitors \eqn{W_\delta}, and other available vertebrate hosts, \eqn{O}.
Total available is a simple sum: \deqn{B = W + W_\delta + O}
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
