% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_theta}
\alias{F_theta}
\title{Compute theta (density dependent mortality)}
\usage{
F_theta(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
This method dispatches on the type of \code{theta_obj}. It should
set the values otheta the bionomic parameters to baseline values
}
\keyword{internal}
