% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_theta.static}
\alias{F_theta.static}
\title{Compute theta (density dependent mortality)}
\usage{
\method{F_theta}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{theta}, slope of density-dependent mortality
}
\description{
Implements \link{F_theta} for a static model
}
\keyword{internal}
