% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend.Lambda}
\alias{F_trend.Lambda}
\title{Compute the temporal trend}
\usage{
\method{F_trend}{Lambda}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend
}
\description{
Evaluate the function \code{F_trend} for a forced model when \code{forced_by} is "Lambda"
}
\keyword{internal}
