% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend.MY}
\alias{F_trend.MY}
\title{Compute the temporal trend}
\usage{
\method{F_trend}{MY}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend
}
\description{
Evaluate the function \code{F_trend} for a trivial \strong{MY} model
}
\keyword{internal}
