% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend.none}
\alias{F_trend.none}
\title{Compute the temporal trend}
\usage{
\method{F_trend}{none}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an empty vector
}
\description{
Evaluate the function \code{F_trend} for a forced model when \code{forced_by} is "none"
}
\keyword{internal}
