% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend}
\alias{F_trend}
\title{Compute the temporal trend}
\usage{
F_trend(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend, invisibly
}
\description{
Evaluate the function \code{F_trend} for a forced model.
}
\keyword{internal}
