% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{F_W_available}
\alias{F_W_available}
\title{Compute Host available for Blood Feeding}
\usage{
F_W_available(search_weights, H, TaR)
}
\arguments{
\item{search_weights}{blood feeding search weights for the host strata}

\item{H}{host density}

\item{TaR}{the host species index}
}
\value{
host available, a \link{vector}
}
\description{
Compute the available of the population strata defined
in the model as hosts for blood feeding by mosquitoes
}
\details{
Host available to blood searching mosquitoes in patches is
the sum of search weights of the human strata, a vector \eqn{\omega},
weighted by time at risk, defined by a matrix \eqn{\Psi}
that is \eqn{N_p \times N_h}. The search weight is a \emph{per-capita} measure
so we weight it by human population density, \eqn{H}. available, \eqn{W}
is computed as \deqn{\Psi \cdot (\omega H).}
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\seealso{
available of all vertebrate hosts for blood feeding is computed by \code{\link[=F_B_available]{F_B_available()}}
}
