% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_X.SEISd}
\alias{F_X.SEISd}
\title{Size of effective infectious human population}
\usage{
\method{F_X}{SEISd}(y, pars, i)
}
\arguments{
\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SEISd model.
}
