% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{F_X.SIS}
\alias{F_X.SIS}
\title{Compute Infectious Density}
\usage{
\method{F_X}{SIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
Infectious density
}
\description{
In the \emph{SIS} model family, infectious
density is \eqn{cI}.
}
\keyword{internal}
