% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{F_X.SIS_dts}
\alias{F_X.SIS_dts}
\title{Size of effective infectious human population}
\usage{
\method{F_X}{SIS_dts}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a list}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_X} for the SIS_dts model.
}
