% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_xi}
\alias{F_xi}
\title{Compute xi, the maturation delay parameter}
\usage{
F_xi(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
This method dispatches on the type of \code{xi_obj}. It should
set the values oxi the bionomic parameters to baseline values
}
\keyword{internal}
