% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_xi.static}
\alias{F_xi.static}
\title{Static model for the maturation delay parameter}
\usage{
\method{F_xi}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{xi}, the baseline human fraction
}
\description{
Implements \link{F_xi} for a static model
}
\keyword{internal}
