% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{foi2eir.nb}
\alias{foi2eir.nb}
\title{A negative binomial model for the daily FoI as a function of the daily EIR.}
\usage{
\method{foi2eir}{nb}(foi, b, pars)
}
\arguments{
\item{foi}{the attack rate}

\item{b}{the probability of infection, per bite}

\item{pars}{an \code{xds} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{foi2eir} for a negative binomial model
}
