% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.none}
\alias{Forcing.none}
\title{Exogenous Forcing}
\usage{
\method{Forcing}{none}(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After basic setup, no exogenous
variables are set up. \code{Forcing} returns
the unmodified \strong{\code{xds}} object.
}
