% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.setup}
\alias{Forcing.setup}
\title{Set the values of exogenous variables}
\usage{
\method{Forcing}{setup}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
none
}
\description{
This method...
}
