% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mectric-conversion.R
\name{fqZ2eir}
\alias{fqZ2eir}
\title{Convert a vector describing infective biting density into the EIR}
\usage{
fqZ2eir(fqZ, beta)
}
\arguments{
\item{fqZ}{a vector describing infective biting density}

\item{beta}{the mixing matrix}
}
\value{
a numeric \link{vector}
}
\description{
Computes a vector of length \code{nPatches} describing infective biting density, per patch, from
a vector of length \code{n} describing the daily EIR per stratum:
\deqn{eir = \beta \cdot fqZ}
}
