% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{GeM}
\alias{GeM}
\title{The \code{GeM} module for the MY component}
\description{
\code{GeM} is a system of delay differential equation model for mosquito ecology
and infection dynamics. The core dynamics trace back to Macdonald's
mosquito model (1952). Macdonald's model was extended by Aron & May (1982) to
include a variable for mosquito population density and a term for
emergence (\eqn{\Lambda}). Aron & May's model was extended by Wu \emph{et al.} (2023) to include
spatial dynamics. In this version, a new method is introduced to non-autonomous
dynamics: an accessory variable computes time-varying survival and dispersal
through a time varying EIP.
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{P}}{density of parous adult mosquitoes}
\item{\code{Y}}{density of infected adult mosquitoes}
\item{\code{Z}}{density of infectious adult mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{nu}}{egg batch laying rate}
\item{\code{eip}}{extrinsic incubation period (\eqn{\tau})}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma}}{emigration rate (\eqn{\sigma})}
\item{\code{mu}}{emigration loss rate (\eqn{\mu})}
\item{\code{K}}{mosquito dispersal matrix}
\item{\code{eggs_per_patch}}{# eggs per batch}
}
}

\section{Mosquito Demography}{


The \emph{demographic matrix} \eqn{\Omega} is:
\deqn{
 \Omega(t) = \mbox{diag} \left( g(t) + \sigma(t) \mu(t) \right) - K \cdot \mbox{diag} \left( \sigma(t) \left(1-\mu(t)\right) \right)
}
}

\section{EIP}{

Let \eqn{\tau(t)} denote the EIP for a mosquito, if infected at time \eqn{t}:
the mosquito would become infectious at time \eqn{t+\tau(t).}
Similarly, let \eqn{\tau'(t)} denote the lag for a mosquito that became \emph{infectious} at time \eqn{t}.
It was infected at time \eqn{t-\tau'(t).}  The two are related by the identities
\deqn{\tau(t) = \tau'(t+\tau(t))} and
\deqn{\tau'(t) = \tau(t-\tau'(t)).}
In the implementation, a function \code{F_eip} returns \eqn{\tau'(t)}. Implementation of dynamically
changing EIP also requires a function to compute
the derivative \code{dF_eip} (below).
}

\section{The Accessory Variables, \eqn{\Upsilon}}{


To compute time-varying survival and dispersal through a time-varying EIP,
we use a set of accessory variables, called \eqn{\Upsilon}. To motivate the
algorithm used to compute it, we introduce a new variable,
\eqn{U} that integrates \eqn{\Omega} over time:
\deqn{dU/dt = \Omega(t).}

Cumulative mortality from time \eqn{t} to time \eqn{t+s} is:
\deqn{\gamma(t,s)  = U(t+s) - U(t).}
Survival and dispersal for the mosquitoes that became infectious at time \eqn{t} is:
\deqn{\Upsilon(t) = e^{-\gamma\left(t_\tau, t\right)} = e^{-\left(U\left(t\right) - U\left(t_\tau \right)\right)}.}
where \eqn{t_\tau = t-\tau'(t)}.

In \code{GeM}, \eqn{\Upsilon} is computed as an accessory variable, with derivatives:
\deqn{d\Upsilon/dt = \left(\Omega\left(t_\tau\right)\left(1-\frac{d\tau'(t)}{dt}\right)-\Omega(t)\right) \cdot \Upsilon(t)}
Initial conditions are set to:
\deqn{\Upsilon(t_0) = e^{-\Omega(t_0)}}
}

\section{Inputs}{


\strong{Emergence} -- \code{Lambda} or \eqn{\Lambda}, is computed
by the \strong{ML}-Interface using outputs of the \strong{L} Component

\strong{Net Infectiousness} -- \code{kappa} or \eqn{\kappa}, is computed
by the \strong{XY}-Interface using outputs of the \strong{XH} Component
}

\section{Dynamics}{

In the following we use the subscript \eqn{\tau} to denote the value of
a parameter, term, or variable at time \eqn{t-\tau'(t)}: \deqn{X_\tau = X(t-\tau'(t))}.

The derivatives of the state variables are:
\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dP/dt &=  f(M-P) - \Omega \cdot P \\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\
dZ/dt &= \Upsilon \cdot (fq\kappa)_\tau(M_\tau-Y_\tau) - \Omega \cdot Z \\
\end{array}}

The function \link{dMYdt.GeM} also computes the variables: \eqn{\Upsilon}; and an accessory variable that internalizes
computation of \eqn{(fq\kappa)_\tau}.
}

\section{Egg Laying}{


The number of egg batches laid, per patch, per day is \eqn{\nu M}. The total
of eggs laid is \eqn{\nu M \times} \code{eggs_per_batch}
}

\section{Infectious Biting}{


The number of bites on humans, per patch, per day is \eqn{fqM}, and the number of
infectious bites, per patch, per day is \eqn{fqZ}.
}

\references{
\itemize{
\item Macdonald G (1952) The analysis of the sporozoite rate. Tropical Diseases Bulletin 49:569-586.
\item Aron JL, May RM (1982) The population dynamics of malaria. Chapter 5 in \emph{The Population Dynamics of Infectious Diseases: Theory and Applications,} Springer, Boston, MA.
\item Wu SL, \emph{et al.} (2023) Spatial dynamics of malaria transmission. PLoS Computational Biology 19(6): e1010684. \url{https://doi.org/10.1371/journal.pcbi.1010684}
}
}
