% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{GeRM}
\alias{GeRM}
\title{The \code{GeRM} Module for the MY Component}
\description{
Implements the \strong{MY} component using a generalized, non-autonomous
Ross-Macdonald model of adult mosquito ecology and infection dynamics,
capable of handling exogenous forcing by weather and vector control.
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{Y}}{density of infected adult mosquitoes}
\item{\code{Z}}{density of infectious adult mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{eip}}{extrinsic incubation period (\eqn{\tau})}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma}}{emigration rate}
\item{\code{mu}}{emigration loss rate}
\item{\code{K}}{mosquito dispersal matrix}
\item{\code{Omega}}{adult mosquito demographic matrix (mortality + migration)}
\item{\code{Upsilon}}{survival and dispersal through the eip: \eqn{\Upsilon = e^{-\Omega\tau}}}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\
dZ/dt &= \Upsilon \cdot (fq\kappa)_\tau(M_\tau-Y_\tau) - \Omega \cdot Z \\
\end{array}}
}

