% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{get_bad_habitats}
\alias{get_bad_habitats}
\title{Get availability of other blood hosts}
\usage{
get_bad_habitats(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of other blood hosts
}
