% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_bionomics}
\alias{get_bionomics}
\title{Compute dynamical terms}
\usage{
get_bionomics(tm, de_vars, pars)
}
\arguments{
\item{tm}{the time}

\item{de_vars}{the output of \link{deSolve}}

\item{pars}{an \code{xds} object}
}
\value{
\link{list}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve.ode} or \code{xde_solve.dde}
}
