% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-get.R
\name{get_bionomics_s}
\alias{get_bionomics_s}
\title{Compute dynamical terms}
\usage{
get_bionomics_s(tm, de_vars, pars, s)
}
\arguments{
\item{tm}{the time}

\item{de_vars}{the output of \link{deSolve}}

\item{pars}{an \code{xds} object}

\item{s}{a vector species index}
}
\value{
\link{list}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve.ode} or \code{xde_solve.dde}
}
