% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{get_blood_search_weights}
\alias{get_blood_search_weights}
\title{Get availability of blood search weights}
\usage{
get_blood_search_weights(xds_obj, i = 1, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{s}{the vector species index}
}
\value{
a vector
}
\description{
Get availability of blood search weights
}
