% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{get_blood_traps}
\alias{get_blood_traps}
\title{Get availability of other blood hosts}
\usage{
get_blood_traps(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a numeric vector
}
\description{
Get availability of other blood hosts
}
