% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{get_f.GeM}
\alias{get_f.GeM}
\title{Get the feeding rates}
\usage{
\method{get_f}{GeM}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the feeding rate, as a vector
}
\description{
Get the feeding rates
}
\keyword{internal}
