% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_f}
\alias{get_f}
\title{Get the feeding rate(s)}
\usage{
get_f(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the feeding rate, as a vector
}
\description{
Get the feeding rate(s)
}
