% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{get_f.SEI}
\alias{get_f.SEI}
\title{Get the feeding rate}
\usage{
\method{get_f}{SEI}(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
