% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_ft}
\alias{get_ft}
\title{Get the feeding rate}
\usage{
get_ft(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
\seealso{
\link{get_bionomics}
}
