% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{get_g.macdonald}
\alias{get_g.macdonald}
\title{Get the mortality rates}
\usage{
\method{get_g}{macdonald}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the mortality rate, as a vector
}
\description{
Get the mortality rates
}
\keyword{internal}
