% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skillsets.R
\name{get_get_skills}
\alias{get_get_skills}
\title{: Get the skill set}
\usage{
get_get_skills(module, xds_obj, ix)
}
\arguments{
\item{module}{"all" or "XY" or "MY" or "L"}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{ix}{the species index}
}
\value{
the skill set, a list
}
\description{
: Get the skill set
}
