% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mosquito.R
\name{get_gt}
\alias{get_gt}
\title{Get the mosquito mortality rate}
\usage{
get_gt(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a vector
}
\description{
Get the mosquito mortality rate
}
