% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_H}
\alias{get_H}
\title{Get the initial values as a vector}
\usage{
get_H(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the human species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the initial values as a vector
}
