% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{get_habitat_search_weights}
\alias{get_habitat_search_weights}
\title{Get availability of habitat search weights}
\usage{
get_habitat_search_weights(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a vector
}
\description{
Get availability of habitat search weights
}
