% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{get_habitats}
\alias{get_habitats}
\title{Get the habitat membership vector}
\usage{
get_habitats(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the patch membership for aquatic habitats, a numeric vector
}
\description{
Output the habitat membership information as a list
}
\seealso{
\code{\link[=make_habitat_matrix]{make_habitat_matrix()}}
}
