% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{get_HTC.hMoI}
\alias{get_HTC.hMoI}
\title{Compute the get_HTC for the hMoI model}
\usage{
\method{get_HTC}{hMoI}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \code{get_HTC} for the hMoI model with demography.
}
\keyword{internal}
