% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{get_HTC}
\alias{get_HTC}
\title{Compute the human transmitting capacity}
\usage{
get_HTC(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
\keyword{internal}
