% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{get_HTC.SIS}
\alias{get_HTC.SIS}
\title{Compute the get_HTC for the SIS model}
\usage{
\method{get_HTC}{SIS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \code{get_HTC} for the SIS model with demography.
}
\keyword{internal}
