% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{get_inits_MYZ.basicM_xde}
\alias{get_inits_MYZ.basicM_xde}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{basicM_xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_inits_MYZ} for the basicM_xde model.
}
