% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{get_inits_MYZ.mz}
\alias{get_inits_MYZ.mz}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{mz}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_inits_MYZ} for the mz model.
}
