% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{get_inits_MYZ.RM_dts}
\alias{get_inits_MYZ.RM_dts}
\title{Return initial values as a vector}
\usage{
\method{get_inits_MYZ}{RM_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_inits_MYZ} for the RM_dts model.
}
