% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{get_K_matrix.0}
\alias{get_K_matrix.0}
\title{Get the Mosquito Dispersal Matrix}
\usage{
\method{get_K_matrix}{`0`}(xds_obj, behave = "0", s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{behave}{the behavioral state}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get the Mosquito Dispersal Matrix
}
\keyword{internal}
