% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{get_K_matrix}
\alias{get_K_matrix}
\title{Get the Mosquito Dispersal Matrix}
\usage{
get_K_matrix(xds_obj, which_K = "K", s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{which_K}{which K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get (inspect) a mosquito dispersal matrix.
}
\note{
The argument \code{K} is used to retrieve state-dependent
mosquito dispersal arguments for behavioral state models.
The default "K" returns \code{K_matrix}
}
