% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_inits}
\alias{get_L_inits}
\title{Get initial values (\strong{L})}
\usage{
get_L_inits(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
The initial values are stored on the \strong{L} object as a
named list called \code{inits}; \code{get_L_inits} returns that list.
}
\keyword{internal}
