% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_ix}
\alias{get_L_ix}
\title{Get L indices}
\usage{
get_L_ix(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list} of indices for the \strong{L} variables
}
\description{
The indices are stored on the \strong{L} object as a
named list called \code{ix}; \code{get_L_ix} returns that list.
}
\keyword{internal}
