% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_out}
\alias{get_L_out}
\title{Get \strong{L} outputs}
\usage{
get_L_out(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the orbits for the \strong{L} component
}
\description{
Pull the saved, parsed orbits for the \strong{L}
component from an \strong{\code{xds}} object
}
