% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{get_L_pars.basicL}
\alias{get_L_pars.basicL}
\title{Get \strong{L} Component Parameters for \code{basicL}}
\usage{
\method{get_L_pars}{basicL}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get the \strong{L} component parameters
}
\seealso{
\link{dLdt.basicL} or \link{change_L_pars.basicL}
}
