% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{get_L_pars.trivial}
\alias{get_L_pars.trivial}
\title{Get \strong{L} component parameters for \code{trivial}}
\usage{
\method{get_L_pars}{trivial}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get \eqn{\Lambda} and parameters that construct
the forcing functions
}
\keyword{internal}
