% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{get_L_vars.basicL}
\alias{get_L_vars.basicL}
\title{List \strong{L} Component Variables for \code{basicL}}
\usage{
\method{get_L_vars}{basicL}(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Extract the \strong{L} component variables from the
vector of state variables (\code{y}) and return them as a named list
}
