% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_vars}
\alias{get_L_vars}
\title{List \strong{L} Component Variables}
\usage{
get_L_vars(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a named \link{list}: the variables of \eqn{\cal L} by name
}
\description{
Extract the variables describing \strong{L} Component
states for the \eqn{s^{th}} species and return them as a named list
}
\note{
This method dispatches on the class of \code{xds_obj$L_obj[[s]]}.
}
