% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{get_L_vars.trivial}
\alias{get_L_vars.trivial}
\title{List \strong{L} Component Variables for \code{trivial}}
\usage{
\method{get_L_vars}{trivial}(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
an empty \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$L_obj[[s]]}
}
