% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_last}
\alias{get_last}
\title{Get the last state}
\usage{
get_last(xds_obj, parse = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{parse}{if TRUE return a named list}
}
\value{
a \link{numeric} vector
}
\description{
Get the last state
}
