% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{get_Linits.basicL}
\alias{get_Linits.basicL}
\title{Return initial values as a vector}
\usage{
\method{get_Linits}{basicL}(pars, s = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_Linits} for the GeRM model.
}
