% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{get_Linits.trace}
\alias{get_Linits.trace}
\title{Return initial values as a vector}
\usage{
\method{get_Linits}{trace}(pars, s = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_Linits} for the GeRM model.
}
