% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{get_Lpars}
\alias{get_Lpars}
\title{Get parameters for the \strong{L} Component module}
\usage{
get_Lpars(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get parameters for the \strong{L} Component module
}
