% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{get_Lpars.trivial}
\alias{get_Lpars.trivial}
\title{Get \strong{L} Component Parameters for \code{trivial}}
\usage{
\method{get_Lpars}{trivial}(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Show the trace function
}
