% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{get_M.SI}
\alias{get_M.SI}
\title{Get the initial mosquito population density}
\usage{
\method{get_M}{SI}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the initial mosquito population density
}
\description{
Get the initial mosquito population density
}
\keyword{internal}
