% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.cohort}
\alias{get_mean_forcing.cohort}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{cohort}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the mean daily EIR for
a \code{cohort} model
}
