% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.eggs}
\alias{get_mean_forcing.eggs}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{eggs}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get mean egg deposition rates
for a trivial \strong{MY} model
}
\keyword{internal}
