% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.eir}
\alias{get_mean_forcing.eir}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{eir}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get the mean daily EIR for an \code{eir} model
}
