% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.fqZ}
\alias{get_mean_forcing.fqZ}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{fqZ}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get the mean density of infectious mosquitoes,
for a trivial \strong{MY}  model
}
\keyword{internal}
