% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.kappa}
\alias{get_mean_forcing.kappa}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{kappa}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get net infectiousness
}
\keyword{internal}
