% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.Lambda}
\alias{get_mean_forcing.Lambda}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get \code{Lambda}, the mean emergence
rate(s) \eqn{\Lambda}, for a model
forced by adult mosquito emergence
}
