% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.MY}
\alias{get_mean_forcing.MY}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{MY}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get \code{Z}, the mean density of infectious mosquitoes,
for an \strong{MY} trivial model
}
\keyword{internal}
