% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing.none}
\alias{get_mean_forcing.none}
\title{Get mean forcing}
\usage{
\method{get_mean_forcing}{none}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Return null for models without forcing
}
