% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_forcing.R
\name{get_mean_forcing}
\alias{get_mean_forcing}
\title{Get mean forcing}
\usage{
get_mean_forcing(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the mean parameter for forcing
}
\description{
Get the parameter(s) that set mean forcing
}
