% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_MY_inits}
\alias{get_MY_inits}
\title{Return initial values as a vector}
\usage{
get_MY_inits(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj}.
}
