% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{get_MY_inits.trivial}
\alias{get_MY_inits.trivial}
\title{Get Inits for \code{trivial} (MY)}
\usage{
get_MY_inits.trivial(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length 0
}
\description{
Return a \link{numeric} vector of length 0
}
\keyword{internal}
