% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_MY_ix}
\alias{get_MY_ix}
\title{Get \strong{MY} variable indices}
\usage{
get_MY_ix(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj}.
}
\keyword{internal}
