% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-get.R
\name{get_MY_orbits}
\alias{get_MY_orbits}
\title{Get \strong{MY} outputs}
\usage{
get_MY_orbits(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the orbits for the \strong{MY} component
}
\description{
Get \strong{MY} outputs
}
