% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{get_MY_pars.basicM}
\alias{get_MY_pars.basicM}
\title{Return the parameters as a list}
\usage{
\method{get_MY_pars}{basicM}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
