% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_MY_pars}
\alias{get_MY_pars}
\title{Return the parameters as a list}
\usage{
get_MY_pars(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
