% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{get_MY_pars.trivial}
\alias{get_MY_pars.trivial}
\title{Return the parameters as a list}
\usage{
\method{get_MY_pars}{trivial}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the parameters as a named list
}
\description{
Return the parameters
in the trivial \strong{MY}-Component
model as a named list
}
