% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{get_MY_vars.basicM}
\alias{get_MY_vars.basicM}
\title{Return the variables as a list}
\usage{
\method{get_MY_vars}{basicM}(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}
}
