% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_MY_vars}
\alias{get_MY_vars}
\title{Return the variables as a list}
\usage{
get_MY_vars(y, xds_obj, s)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
