% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-get.R
\name{get_MYZ}
\alias{get_MYZ}
\title{Get \strong{MYZ} outputs}
\usage{
get_MYZ(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\description{
Get \strong{MYZ} outputs
}
