% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{get_MYZinits.basicM}
\alias{get_MYZinits.basicM}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{basicM}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_MYZinits} for the basicM model.
}
