% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{get_MYZinits.GeRM}
\alias{get_MYZinits.GeRM}
\title{Return initial values as a list}
\usage{
\method{get_MYZinits}{GeRM}(pars, s = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{get_MYZinits} for the GeRM model.
}
