% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-macdonald.R
\name{get_MYZinits.macdonald}
\alias{get_MYZinits.macdonald}
\title{Return initial values as a list}
\usage{
\method{get_MYZinits}{macdonald}(pars, s = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{get_MYZinits} for the macdonald model.
}
