% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{get_MYZinits.MY}
\alias{get_MYZinits.MY}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{MY}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_MYZinits} for the \code{MY} model.
}
