% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-sei.R
\name{get_MYZinits.sei}
\alias{get_MYZinits.sei}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{sei}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_MYZinits} for the sei model.
}
