% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{get_MYZinits.si_old}
\alias{get_MYZinits.si_old}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{si_old}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{numeric}
}
\description{
Implements \link{get_MYZinits} for the \code{si_old} model.
}
