% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{get_MYZinits.trace}
\alias{get_MYZinits.trace}
\title{Return initial values as a vector}
\usage{
\method{get_MYZinits}{trace}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_MYZinits} for the GeRM model.
}
