% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-get.R
\name{get_Omega}
\alias{get_Omega}
\title{Get Omega}
\usage{
get_Omega(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the demographic matrix
}
\description{
Return the demographic
matrix, called \eqn{\Omega}.
}
\seealso{
\link{xds_info_mosquito_demography}
}
