% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{get_other_blood_hosts}
\alias{get_other_blood_hosts}
\title{Get availability of other blood hosts}
\usage{
get_other_blood_hosts(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of other blood hosts
}
