% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{get_ovitraps}
\alias{get_ovitraps}
\title{Get availability of ovitraps}
\usage{
get_ovitraps(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of ovitraps
}
