% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.lm}
\alias{get_PR.lm}
\title{Get the \emph{Pf}PR from a malaria model}
\usage{
\method{get_PR}{lm}(xds_obj, method = "lm", i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{method}{the method used for computing \emph{Pf}PR}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Return the PR by light microscopy
}
\keyword{internal}
