% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.pcr}
\alias{get_PR.pcr}
\title{Get the \emph{Pf}PR from a malaria model}
\usage{
\method{get_PR}{pcr}(xds_obj, method = "pcr", i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{method}{the method used for computing \emph{Pf}PR}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Return the \emph{Pf}PR by PCR
}
\keyword{internal}
