% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR}
\alias{get_PR}
\title{Get the \emph{Pf}PR from a Malaria Model}
\usage{
get_PR(xds_obj, i = 1, method = "true")
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{method}{the method used for computing \emph{Pf}PR}
}
\value{
none
}
\description{
This method assigns \code{class(method)=method} and then
dispatches on "type." Options are:
\itemize{
\item \code{true} for the true \emph{Pf}PR
\item \code{lm} for the \emph{Pf}PR by light microscopy
\item \code{rdt} for the \emph{Pf}PR by RDT
\item \code{pcr} for the \emph{Pf}PR by PCR
}
}
