% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.true}
\alias{get_PR.true}
\title{Get the \emph{Pf}PR from a Malaria Model}
\usage{
\method{get_PR}{true}(xds_obj, i = 1, method = "true")
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{method}{the method used for computing \emph{Pf}PR}
}
\value{
the true PR
}
\description{
Return the true \emph{Pf}PR
}
