% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-get.R
\name{get_psi}
\alias{get_psi}
\title{Get psi, the maturation rate}
\usage{
get_psi(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get psi, the maturation rate
}
