% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{get_q.trivial}
\alias{get_q.trivial}
\title{Get the human fractions}
\usage{
\method{get_q}{trivial}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the human fraction, as a vector
}
\description{
Get the human fractions
}
\keyword{internal}
