% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-get.R
\name{get_qt}
\alias{get_qt}
\title{Get the human fraction}
\usage{
get_qt(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a vector
}
\description{
Get the human fraction
}
