% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{get_residence}
\alias{get_residence}
\title{Get residence}
\usage{
get_residence(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a named \link{list}
}
\description{
Get the residence vector
}
