% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_search_weights_blood}
\alias{get_search_weights_blood}
\title{Get Blood Feeding Search Weights}
\usage{
get_search_weights_blood(xds_obj, s = 1, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Get Blood Feeding Search Weights
}
