% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season.cohort}
\alias{get_season.cohort}
\title{Get seasonal pattern}
\usage{
\method{get_season}{cohort}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\description{
Get the parameters set the seasonal pattern
}
