% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season.MY}
\alias{get_season.MY}
\title{Get seasonality parameters}
\usage{
\method{get_season}{MY}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a list, the value of \code{season_par}
}
\description{
Get seasonality parameters when \code{forced_by} = "MY"
}
\keyword{internal}
