% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season.none}
\alias{get_season.none}
\title{Get seasonality parameters}
\usage{
\method{get_season}{none}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an empty vector
}
\description{
Get seasonality parameters when \code{forced_by} = "none"
}
\keyword{internal}
