% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_bottom.R
\name{get_season_bottom.cohort}
\alias{get_season_bottom.cohort}
\title{Get the seasonal bottom}
\usage{
\method{get_season_bottom}{cohort}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value that sets bottom
for a seasonally forced cohort model
}
