% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_bottom.R
\name{get_season_bottom.eir}
\alias{get_season_bottom.eir}
\title{Get the seasonal bottom}
\usage{
\method{get_season_bottom}{eir}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value that sets the bottom
for a seasonally forced eir model
}
