% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_bottom.R
\name{get_season_bottom.Lambda}
\alias{get_season_bottom.Lambda}
\title{Get the seasonal bottom}
\usage{
\method{get_season_bottom}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value(s) that set the bottom
for a model that is forced by mosquito emergence
}
