% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_bottom.R
\name{get_season_bottom.none}
\alias{get_season_bottom.none}
\title{Get the seasonal bottom}
\usage{
\method{get_season_bottom}{none}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
a null value
}
\description{
The null case for \code{get_season_bottom}
}
