% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season_par.Lambda}
\alias{get_season_par.Lambda}
\title{Get seasonal pattern}
\usage{
\method{get_season_par}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the parameters set the seasonal pattern
}
