% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season_par}
\alias{get_season_par}
\title{Get seasonal pattern}
\usage{
get_season_par(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the parameters set the seasonal pattern
}
