% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_phase.R
\name{get_season_phase.cohort}
\alias{get_season_phase.cohort}
\title{Get phase}
\usage{
\method{get_season_phase}{cohort}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value that sets phase
for a seasonally forced cohort model
}
