% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_phase.R
\name{get_season_phase.Lambda}
\alias{get_season_phase.Lambda}
\title{Get phase}
\usage{
\method{get_season_phase}{Lambda}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value(s) that set the phase
for a model that is forced by mosquito emergence
}
