% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_phase.R
\name{get_season_phase}
\alias{get_season_phase}
\title{Get phase}
\usage{
get_season_phase(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
the phase parameter(s)
}
\description{
Get the parameter that sets the phase
for a seasonally forced model
}
