% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_pw.R
\name{get_season_pw.cohort}
\alias{get_season_pw.cohort}
\title{Get the seasonal pw}
\usage{
\method{get_season_pw}{cohort}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the value that sets pw
for seasonally forcing on the daily
EIR for a cohort model
}
