% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_pw.R
\name{get_season_pw.eir}
\alias{get_season_pw.eir}
\title{Get the seasonal pw}
\usage{
\method{get_season_pw}{eir}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the pw parameter(s)
for a seasonally forced eir model
}
