% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_pw.R
\name{get_season_pw.none}
\alias{get_season_pw.none}
\title{Get the seasonal pw}
\usage{
\method{get_season_pw}{none}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
a null value
}
\description{
The null case for \code{get_season_pw}
}
