% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season_pw}
\alias{get_season_pw}
\title{Get seasonality parameter \code{pw}}
\usage{
get_season_pw(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the pw parameter(s)
}
\description{
Get the \code{pw} parameter for the seasonality function
}
