% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shock.R
\name{get_shock.eir}
\alias{get_shock.eir}
\title{Get the shock parameters}
\usage{
\method{get_shock}{eir}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a list, the value of \code{shock_par}
}
\description{
Get the shock parameters for an EIR forced model
}
\keyword{internal}
