% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shock.R
\name{get_shock.MY}
\alias{get_shock.MY}
\title{Get the shock parameters}
\usage{
\method{get_shock}{MY}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
shock parameters
}
\description{
Get the shock parameters for a model forced by infectious biting
}
\keyword{internal}
