% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shock.R
\name{get_shock.none}
\alias{get_shock.none}
\title{Get the shock parameters}
\usage{
\method{get_shock}{none}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
shock parameters
}
\description{
Get the parameters for the perturbation function
}
\keyword{internal}
