% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shock.R
\name{get_shock}
\alias{get_shock}
\title{Get the shock parameters}
\usage{
get_shock(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
shock parameters
}
\description{
Get the parameters for the perturbation function
forcing the model
}
\note{
Dispatches on \code{forced_by}
}
