% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{get_sigma}
\alias{get_sigma}
\title{Get the feeding rate}
\usage{
get_sigma(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
