% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_sigma.sei}
\alias{get_sigma.sei}
\title{Get the feeding rate}
\usage{
\method{get_sigma}{sei}(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the feeding rate
}
