% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-get.R
\name{get_sigmat}
\alias{get_sigmat}
\title{Get the mosquito emigration rate}
\usage{
get_sigmat(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a vector
}
\description{
Get the mosquito emigration rate
}
