% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skillsets.R
\name{get_skills.L}
\alias{get_skills.L}
\title{Get the skill set}
\usage{
\method{get_skills}{L}(module, xds_obj, ix = 1)
}
\arguments{
\item{module}{"all" or "XY" or "MY" or "L"}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{ix}{the species index}
}
\value{
the skill set, a list
}
\description{
Get the skill set
}
\keyword{internal}
