% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_spline}
\alias{get_spline}
\title{Get spline interpolation points}
\usage{
get_spline(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\description{
Get the interpolating points
for a spline function.
}
