% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spline.R
\name{get_spline_s.cohort}
\alias{get_spline_s.cohort}
\title{Get spline interpolation points}
\usage{
\method{get_spline_s}{cohort}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\description{
Return the
}
