% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_spline_s.Lambda}
\alias{get_spline_s.Lambda}
\title{Get spline interpolation points}
\usage{
\method{get_spline_s}{Lambda}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a list, the value of \code{trend_par}
}
\description{
Get spline interpolation points when \code{forced_by} = "Lambda"
}
\keyword{internal}
