% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_spline_s}
\alias{get_spline_s}
\title{Get spline interpolation points}
\usage{
get_spline_s(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\description{
Get the interpolating points
for a spline function.
}
