% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-get.R
\name{get_terms}
\alias{get_terms}
\title{Compute dynamical terms}
\usage{
get_terms(tm, de_vars, pars)
}
\arguments{
\item{tm}{a vector of times}

\item{de_vars}{a matrix with the values of the variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
Using the output of deSolve
compute the dynamical terms for the output of \code{xde_solve.ode} or \code{xde_solve.dde}
}
