% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-get.R
\name{get_theta}
\alias{get_theta}
\title{Get theta, density dependent mortality}
\usage{
get_theta(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get theta, density dependent mortality
}
