% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R, R/travel-time_away.R
\name{get_time_away}
\alias{get_time_away}
\title{Get the Time Away Vector}
\usage{
get_time_away(xds_obj, i = 1)

get_time_away(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector of length \code{nStrata}

a []
}
\description{
Get the Time Away Vector

Get the Time Spent
}
