% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespent.R
\name{get_timespent_matrix}
\alias{get_timespent_matrix}
\title{Get the Time Spent Matrix}
\usage{
get_timespent_matrix(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{matrix}
}
\description{
Get the Time Spent Matrix
}
