% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{get_travel_eir}
\alias{get_travel_eir}
\title{Get the travel eir}
\usage{
get_travel_eir(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the travel eir, a vector of length nStrata
}
\description{
Get the travel eir
}
