% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend.cohort}
\alias{get_trend.cohort}
\title{Get the trend parameters}
\usage{
\method{get_trend}{cohort}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\description{
Get the interpolating points
for a spline function.
}
