% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend.MY}
\alias{get_trend.MY}
\title{Get the trend parameters}
\usage{
\method{get_trend}{MY}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
trend parameters
}
\description{
Get the trend parameters for an \strong{MY} trivial model
}
\keyword{internal}
