% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trend.R
\name{get_trend}
\alias{get_trend}
\title{Get the trend parameters}
\usage{
get_trend(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\description{
Get the interpolating points
for a spline function.
}
