% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-get.R
\name{get_Upsilon}
\alias{get_Upsilon}
\title{Get Upsilon}
\usage{
get_Upsilon(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the demographic matrix
}
\description{
Return \eqn{\Upsilon},
a matrix describing survival and dispersal
through the EIP.
}
\seealso{
\link{xds_info_mosquito_demography}
}
