% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{get_V_inits}
\alias{get_V_inits}
\title{A function to set up V_obj}
\usage{
get_V_inits(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the initial values, as a named list
}
\description{
Get and display the stored
set of initial values.
}
