% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{get_V_pars}
\alias{get_V_pars}
\title{Return the parameters as a list}
\usage{
get_V_pars(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$V_obj[[i]]}.
}
