% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_variables.na}
\alias{get_variables.na}
\title{Get Variables}
\usage{
\method{get_variables}{na}(t, y, func, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{func}{a function that dispatches \link{get_variables} and pulls variables it needs}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a set of variables
}
\description{
The null case for \link{get_variables}
}
