% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variables.R
\name{get_variables}
\alias{get_variables}
\title{Get Variables}
\usage{
get_variables(opts, t, y, xds_obj, ix)
}
\arguments{
\item{opts}{an options list and dispatching}

\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
a list of named variables
}
\description{
A generic function to retrieve variables from the \strong{\code{xds}} model object
}
\keyword{internal}
