% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity.R
\name{get_VC}
\alias{get_VC}
\title{Get VC}
\usage{
get_VC(xds_obj, s = 1, computed = TRUE)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}

\item{computed}{if true, get the VC matrix that was computed; if false, by formula}
}
\value{
a numeric \link{matrix}
}
\description{
Get the VC matrix
}
