% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{get_vis_kappa}
\alias{get_vis_kappa}
\title{Get availability of vis_kappa}
\usage{
get_vis_kappa(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of vis_kappa
}
