% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{get_visitors}
\alias{get_visitors}
\title{Get availability of visitors}
\usage{
get_visitors(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of visitors
}
