% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_XH}
\alias{get_XH}
\title{Get \strong{XH} outputs}
\usage{
get_XH(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\description{
Get \strong{XH} outputs
}
