% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{get_XH_ix}
\alias{get_XH_ix}
\title{Add indices for human population to parameter list}
\usage{
get_XH_ix(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get and display the values of the indices
for the variables
}
