% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_XH_orbits}
\alias{get_XH_orbits}
\title{Get orbits (\strong{XH})}
\usage{
get_XH_orbits(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the orbits for the \strong{XH} component
}
\description{
Get orbits (\strong{XH})
}
