% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_XH_out}
\alias{get_XH_out}
\title{Get \strong{XH} outputs}
\usage{
get_XH_out(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
the orbits for the \strong{XH} component
}
\description{
Get \strong{XH} outputs
}
