% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{get_XH_pars.hMoI}
\alias{get_XH_pars.hMoI}
\title{Return the parameters as a list}
\usage{
\method{get_XH_pars}{hMoI}(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Parameter values for the \eqn{i^{th}} host are
stored as \code{xds_obj$XH_obj[[i]]}. This returns the stored parameter
values as a list.
}
\seealso{
\link{make_XH_obj_hMoI}
}
