% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{get_XH_pars.SIS}
\alias{get_XH_pars.SIS}
\title{Get \emph{SIS} model parameters}
\usage{
\method{get_XH_pars}{SIS}(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
SIS model parameters
}
\description{
Parameter values for the \eqn{i^{th}} host are
stored as \code{xds_obj$XH_obj[[i]]}.
This returns the stored parameter
values as a list.
}
