% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{get_XH_vars}
\alias{get_XH_vars}
\title{Get Variables by Name}
\usage{
get_XH_vars(y, xds_obj, i = 1)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
Variables as a named list
}
\description{
Get the the values
of variables from the flat state
variable vector \eqn{y}, and return
the values as a named list
}
