% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{get_XH_vars.SIS}
\alias{get_XH_vars.SIS}
\title{Get Variables by Name}
\usage{
\method{get_XH_vars}{SIS}(y, xds_obj, i)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Get the the value of variables
from the flat state variable
vector \eqn{y}, and return
the values as a named list
}
\keyword{internal}
