% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{get_XH_vars.trivial}
\alias{get_XH_vars.trivial}
\title{Get Variables by Name}
\usage{
\method{get_XH_vars}{trivial}(y, xds_obj, i)
}
\arguments{
\item{y}{the variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Return an empty list
}
