% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-get.R
\name{get_xi}
\alias{get_xi}
\title{Get xi, delayed maturation}
\usage{
get_xi(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get xi, delayed maturation
}
