% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{get_Xinits.SIP}
\alias{get_Xinits.SIP}
\title{Return initial values as a vector}
\usage{
\method{get_Xinits}{SIP}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a named \link{vector}
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
