% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{get_Xinits.trivial}
\alias{get_Xinits.trivial}
\title{Return initial values as a vector}
\usage{
\method{get_Xinits}{trivial}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar}.
}
