% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{get_Xpars}
\alias{get_Xpars}
\title{Return the parameters as a list}
\usage{
get_Xpars(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
