% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{get_Xpars.trivial}
\alias{get_Xpars.trivial}
\title{Return the parameters as a list}
\usage{
\method{get_Xpars}{trivial}(pars, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[s]]}.
}
