% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{get_Z.GeM}
\alias{get_Z.GeM}
\title{Get the initial infective mosquito density}
\usage{
\method{get_Z}{GeM}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the initial infective mosquito population density
}
\description{
Get the initial infective mosquito density
}
\keyword{internal}
