% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_Z}
\alias{get_Z}
\title{Get the density of infectious, blood feeding mosquitoes}
\usage{
get_Z(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the density of infectious, blood feeding mosquitoes
}
