% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{habitat_dynamics}
\alias{habitat_dynamics}
\title{Habitats Dynamics}
\usage{
habitat_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
A function \code{F_habitat_weights} computes \code{TiSp_frac_searching} and
this function sets \code{TiSp_frac_here.}
}
\note{
The quantity \code{TiSp_frac_here} is part of the \code{ML_interface}
It is used to compute the Time at Risk matrix, and it is also
used to weight local exposure \emph{vs.}  searching exposure.
}
