% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{habitat_search_weights}
\alias{habitat_search_weights}
\title{habitat_search_weights}
\usage{
habitat_search_weights(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{S3} definition for the
function that computes habitat_search_weights
}
\keyword{internal}
