% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{habitat_search_weights.static}
\alias{habitat_search_weights.static}
\title{Compute availability of habitat search weights}
\usage{
\method{habitat_search_weights}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up...
}
\keyword{internal}
