% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics}
\alias{HabitatDynamics}
\title{Habitat Dynamics}
\usage{
HabitatDynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function modifies search weights for
habitats and availability of bad habitat
}
\keyword{internal}
