% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{HabitatDynamics.setup}
\alias{HabitatDynamics.setup}
\title{Habitat Dynamics}
\usage{
\method{HabitatDynamics}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function modifies search weights for
habitats and availability of bad habitat
}
