% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health.dynamic}
\alias{Health.dynamic}
\title{Set the values of exogenous variables}
\usage{
\method{Health}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
With dynamic health, exogenous variables
can be set in one of four function calls:
\itemize{
\item Clinic
\item School
\item MassHealth
\item Behavior
\item ActiveCaseDetection
}
}
\seealso{
\link{dynamic_health}
}
