% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health.none}
\alias{Health.none}
\title{Set no exogenous health variables}
\usage{
\method{Health}{none}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After none setup, no exogenous
variables are configured so \code{Health} returns
the \strong{\code{xds}} object without modification
}
