% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health}
\alias{Health}
\title{Set the values variables for health-based malaria control}
\usage{
Health(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the values of
variables for health interventions.
}
