% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health.setup}
\alias{Health.setup}
\title{The \code{setup} case for health variables}
\usage{
\method{Health}{setup}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Call all the functions to set the
values of health variables and then revert
the \code{none} case
}
