% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{hMoI}
\alias{hMoI}
\title{The \code{hMoI} Module for the XH Component}
\description{
Implements the \strong{XH} component using a hybrid model that tracks the
mean Multiplicity of Infection (MoI) for total and apparent infections.
This model is not extensible to dynamic human population density or
mass treatment.
}
\section{State Variables}{

\describe{
\item{\code{m1}}{mean MoI for all infections}
\item{\code{m2}}{mean MoI for apparent (detectable) infections}
}
}

\section{Parameters}{

\describe{
\item{\code{r1}}{clearance rate for all infections}
\item{\code{r2}}{clearance rate for apparent infections}
\item{\code{b}}{probability of infection per infectious bite}
}
}

\section{Dynamics}{

\deqn{
\begin{array}{rl}
dm_1/dt &= h - r_1 m_1 \\
dm_2/dt &= h e^{-m_1} (e^{m_2} - 1) - r_2 m_2 \\
\end{array}}
where \eqn{h} is the force of infection computed upstream.
}

