% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{HTC.hMoI}
\alias{HTC.hMoI}
\title{Compute the HTC for the hMoI model}
\usage{
\method{HTC}{hMoI}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the hMoI model with demography.
}
