% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI-xde.R
\name{HTC.hMoI_xde}
\alias{HTC.hMoI_xde}
\title{Compute the HTC for the hMoI_xde model}
\usage{
\method{HTC}{hMoI_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the hMoI_xde model with demography.
}
