% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{HTC}
\alias{HTC}
\title{Compute the human transmitting capacity}
\usage{
HTC(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
