% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{HTC.SEISd}
\alias{HTC.SEISd}
\title{Compute the HTC for the SEISd model}
\usage{
\method{HTC}{SEISd}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SEISd model with demography.
}
