% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-xde.R
\name{HTC.SIP_xde}
\alias{HTC.SIP_xde}
\title{Compute the HTC for the SIP_xde model}
\usage{
\method{HTC}{SIP_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIP_xde model with demography.
}
