% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{HTC.SIPd}
\alias{HTC.SIPd}
\title{Compute the HTC for the SIPd model}
\usage{
\method{HTC}{SIPd}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIPd model with demography.
}
