% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{HTC.SIS}
\alias{HTC.SIS}
\title{Compute the HTC for the SIS model}
\usage{
\method{HTC}{SIS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIS model with demography.
}
\keyword{internal}
