% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{HTC.SIS_dts}
\alias{HTC.SIS_dts}
\title{Compute the HTC for the SIS_dts model}
\usage{
\method{HTC}{SIS_dts}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIS_dts model with demography.
}
