% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{HTC.SIS_xde}
\alias{HTC.SIS_xde}
\title{Compute the HTC for the SIS_xde model}
\usage{
\method{HTC}{SIS_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{HTC} for the SIS_xde model with demography.
}
