% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{Hydrology.no_forcing}
\alias{Hydrology.no_forcing}
\title{Set the values of exogenous variables describing hydrology}
\usage{
\method{Hydrology}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Hydrology} for the no_forcing model of hydrology (do nothing)
}
