% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{Importation.dynamic}
\alias{Importation.dynamic}
\title{Importation}
\usage{
\method{Importation}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
No-op method for the \code{Importation} junction when no
dynamic importation has been configured. The static default
parameters set by \link{setup_importation} are used as-is.
}
\keyword{internal}
