% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{Importation.static}
\alias{Importation.static}
\title{Importation}
\usage{
\method{Importation}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Importation for static models
}
\keyword{internal}
