% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{importation_dynamics}
\alias{importation_dynamics}
\title{Importation terms}
\usage{
importation_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute terms associated with importation
}
\keyword{internal}
