% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{IRS.dynamic}
\alias{IRS.dynamic}
\title{Set no irs}
\usage{
\method{IRS}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for irs
}
