% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{IRS.none}
\alias{IRS.none}
\title{Set no irs}
\usage{
\method{IRS}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for irs
}
