% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-func.R
\name{IRSCoverage.func}
\alias{IRSCoverage.func}
\title{Set no irs_coverage}
\usage{
\method{IRSCoverage}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for irs_coverage
}
