% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-coverage-interface.R
\name{IRSCoverage}
\alias{IRSCoverage}
\title{Set the irs_coverage}
\usage{
IRSCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
irs_coverage
}
