% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects-none.R
\name{IRSEffects.none}
\alias{IRSEffects.none}
\title{Set no irs_effects}
\usage{
\method{IRSEffects}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for irs_effects
}
