% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects-interface.R
\name{IRSEffects}
\alias{IRSEffects}
\title{Set the irs_effects_irs}
\usage{
IRSEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
irs_effects_irs
}
