% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-none.R
\name{IRSEffectSizes.none}
\alias{IRSEffectSizes.none}
\title{Set no irs_effectsizes}
\usage{
\method{IRSEffectSizes}{none}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for irs_effectsizes
}
