% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{IRSEffectSizes}
\alias{IRSEffectSizes}
\title{Set the irs_effectsizes}
\usage{
IRSEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
irs_effectsizes
}
