% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{junction}
\alias{junction}
\title{Junctions}
\description{
In \strong{\verb{ramp.xds,}} clusters of \link{port}s are bundled
into a \strong{junction.} A junction is a structural element,
a function that is designed to handle one or more \link{port}s.
\itemize{
\item \link{Other_State_Variables} -- a junction to add new state variables, effectively extending the dynamical system
\item \link{Forcing} -- a junction to handle exogenous forcing by weather, \emph{etc.}
\item \link{Vector_Control} -- a junction to handle vector control
}
}
