% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-help.R
\name{L_get}
\alias{L_get}
\title{Get Functions for the L Component}
\description{
write me.

\itemize{
\item \code{\link{get_L_vars}} -- get state variables from \eqn{y}
\item \code{\link{get_L_ix}} -- get variable indices
\item \code{\link{get_L_inits}} -- get initial values
\item \code{\link{get_L_pars}} -- get parameters
\item \code{\link{get_L_orbits}} -- get saved orbits
}
}
