% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-modules.R
\name{L_module_list}
\alias{L_module_list}
\title{L Modules}
\description{
A List of Modules for the \strong{L}-Component
}
\section{In \strong{\code{ramp.xds}}}{

\itemize{
\item \code{trivial} - the trivial module: see \link{trivial_L}
\item \code{basicL} - a simple model for aquatic dynamics: see \link{basicL}
}
}

\section{In \strong{\code{ramp.library}}}{


None (so far)

**\code{ramp.library} -- ** \href{https://github.com/dd-harp/ramp.library}{source code} & \href{https://github.com/dd-harp/ramp.library}{github.io}
}

