% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-help.R
\name{L_setup}
\alias{L_setup}
\title{Setup functions (\strong{L})}
\description{
A list of setup functions for aquatic mosquito modules.

\itemize{
\item \code{\link{setup_L_obj}} -- set up the L module object
\item \code{\link{setup_L_ix}} -- set up variable indices
\item \code{\link{setup_L_inits}} -- set up initial values
}
}
