% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_to_inits}
\alias{last_to_inits}
\title{Set the initial values to the last values of the last simulation}
\usage{
last_to_inits(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set the initial values to the last values of the last simulation
}
