% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{LBaseline.trivial}
\alias{LBaseline.trivial}
\title{Baseline Bionomics for \code{trivial} (\strong{L} Component)}
\usage{
\method{LBaseline}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{LBaseline} for the RM model
}
