% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{LBionomics.basicL}
\alias{LBionomics.basicL}
\title{Mosquito bionomics for \code{basicL} (\strong{L})}
\usage{
\method{LBionomics}{basicL}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the bionomic parameter values for \code{basicL}
by calling the forcing functions for each parameter. Also resets
all effect sizes (\code{es_psi}, \code{es_phi}, \code{es_xi}, \code{es_theta}) to 1.
}
\keyword{internal}
