% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{LBionomics.stages_dts}
\alias{LBionomics.stages_dts}
\title{Reset aquatic parameters to baseline}
\usage{
\method{LBionomics}{stages_dts}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{LBionomics} for the stages_dts model
}
