% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{LBionomics.static}
\alias{LBionomics.static}
\title{Set mosquito bionomics to baseline}
\usage{
\method{LBionomics}{static}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
the model as a \link{list}
}
\description{
Implements \link{LBionomics} for models with no forcing
}
