% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{LBionomics.trivial}
\alias{LBionomics.trivial}
\title{Mosquito bionomics for \code{trivial} (\strong{L})}
\usage{
\method{LBionomics}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the \strong{\code{xds}} object unmodified
}
\keyword{internal}
