% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{LEffectSizes}
\alias{LEffectSizes}
\title{Immature Mosquito Bionomics - Vector Control Effect Sizes}
\usage{
LEffectSizes(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Apply vector control effect sizes to immature mosquito bionomic parameters
}
\keyword{internal}
