% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{LEffectSizes.trivial}
\alias{LEffectSizes.trivial}
\title{Apply effect sizes for \code{trivial} (\strong{L})}
\usage{
\method{LEffectSizes}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{LEffectSizes} for the \code{trivial} module
}
\keyword{internal}
