% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{list_Lvars.basicL}
\alias{list_Lvars.basicL}
\title{List \strong{L} Component Variables for \code{basicL}}
\usage{
\method{list_Lvars}{basicL}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a named \link{list}
}
\description{
Extract the \strong{L} component variables from the
vector of state variables (\code{y}) and return them as a named list
}
