% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{list_Lvars}
\alias{list_Lvars}
\title{Return \strong{L} Component Variables as a list}
\usage{
list_Lvars(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a named \link{list}: the variables of \eqn{\cal L} by name
}
\description{
Extract the variables describing \strong{L} Component
states for the \eqn{s^{th}} species and return them as a named list
}
\note{
This method dispatches on the class of \code{pars$Lpar[[s]]}.
}
