% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{list_Lvars.trace}
\alias{list_Lvars.trace}
\title{Return the variables as a list}
\usage{
\method{list_Lvars}{trace}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
an empty \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}
}
