% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{list_Lvars.trivial}
\alias{list_Lvars.trivial}
\title{List \strong{L} Component Variables for \code{trivial}}
\usage{
\method{list_Lvars}{trivial}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
an empty \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}
}
