% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_vars}
\alias{list_vars}
\title{Set the initial values to the last values of the last simulation}
\usage{
list_vars(xds_obj, y0 = NULL, s = 1, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{y0}{a \link{vector} of initial values}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\description{
Set the initial values to the last values of the last simulation
}
