% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{LSM.dynamic}
\alias{LSM.dynamic}
\title{Set no LSM}
\usage{
\method{LSM}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for LSM
}
