% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{LSM.none}
\alias{LSM.none}
\title{Set no LSM}
\usage{
\method{LSM}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for LSM
}
