% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{LSM}
\alias{LSM}
\title{Set the LSM}
\usage{
LSM(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
LSM
}
