% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-func.R
\name{LSMCoverage.func}
\alias{LSMCoverage.func}
\title{Set no lsm_coverage}
\usage{
\method{LSMCoverage}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for lsm_coverage
}
