% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-interface.R
\name{LSMCoverage}
\alias{LSMCoverage}
\title{Set the lsm_coverage}
\usage{
LSMCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
lsm_coverage
}
