% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-none.R
\name{LSMEffects.none}
\alias{LSMEffects.none}
\title{Set no lsm_effects}
\usage{
\method{LSMEffects}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for lsm_effects
}
