% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effects-interface.R
\name{LSMEffects}
\alias{LSMEffects}
\title{Set the lsm_effects_lsm}
\usage{
LSMEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
lsm_effects_lsm
}
