% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{LSMEffectSizes.none}
\alias{LSMEffectSizes.none}
\title{Set no LSMEffectSizes}
\usage{
\method{LSMEffectSizes}{none}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The null model for LSMEffectSizes
}
