% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{make_beta}
\alias{make_beta}
\title{Compute beta}
\usage{
make_beta(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
This function computes the mixing matrix, beta
}
