% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-dispersal.R
\name{make_calK.as_matrix}
\alias{make_calK.as_matrix}
\title{Pass a pre-configured calK}
\usage{
\method{make_calK}{as_matrix}(calK, s, pars, opts = list())
}
\arguments{
\item{calK}{a matrix or a string}

\item{s}{the vector species index}

\item{pars}{an \code{xds} object}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_calK} for as_matrix
}
