% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-dispersal.R
\name{make_calK.herethere}
\alias{make_calK.herethere}
\title{Dispersal to every other patch, with equal probability}
\usage{
\method{make_calK}{herethere}(calK, s, pars, opts = list())
}
\arguments{
\item{calK}{a matrix or a string}

\item{s}{the vector species index}

\item{pars}{an \code{xds} object}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_calK} for the herethere model
}
