% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-dispersal.R
\name{make_calK}
\alias{make_calK}
\title{Make a mosquito dispersal matrix, called calK}
\usage{
make_calK(calK, s, pars, opts = list())
}
\arguments{
\item{calK}{a matrix or a string}

\item{s}{the vector species index}

\item{pars}{an \code{xds} object}

\item{opts}{a list of options to configure calK}
}
\value{
a modified \code{xds} object
}
\description{
Make a mosquito dispersal matrix, called calK
}
