% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-dispersal.R
\name{make_calK.xy}
\alias{make_calK.xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
\method{make_calK}{xy}(calK = "xy", s, pars, opts = list())
}
\arguments{
\item{calK}{a matrix or a string}

\item{s}{the vector species index}

\item{pars}{an \code{xds} object}

\item{opts}{a list of options to configure calK}
}
\value{
a \link{matrix}
}
\description{
Implements \link{make_calK} for kernels
}
