% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{make_calU}
\alias{make_calU}
\title{Compute and store \eqn{\cal U}}
\usage{
make_calU(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Compute and store the egg distribution matrices
for all the vector species
}
