% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{make_EIR}
\alias{make_EIR}
\title{Compute EIR}
\usage{
make_EIR(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This function computes the EIR for each stratum of each host species
}
