% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{make_EIR_full}
\alias{make_EIR_full}
\title{Compute EIR for each vector-host pair}
\usage{
make_EIR_full(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This function computes the EIR from each vector species for each stratum of each host species
}
