% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{make_exposure_nb}
\alias{make_exposure_nb}
\title{Make a nbson Exposure Model Object}
\usage{
make_exposure_nb(nStrata, options, sz = 1)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{options}{options to configure negative binomial exposure}

\item{sz}{the \code{size} parameter for a negative binomial distribution}
}
\value{
a local exposure model object
}
\description{
Set up the nbson model for
exposure for continuous time models
}
\seealso{
Also, see \link{F_foi.nb}
}
