% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{make_exposure_pois}
\alias{make_exposure_pois}
\title{Make a Poisson Exposure Model Object}
\usage{
make_exposure_pois()
}
\value{
a local exposure model object
}
\description{
Set up the Poisson model for
exposure for continuous time models
}
\seealso{
Also, see \link{F_foi.pois}
}
