% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{make_F_eip_random}
\alias{make_F_eip_random}
\title{Make random EIP configuration}
\usage{
make_F_eip_random(
  nPatches,
  options = list(),
  f_rand = stats::rgamma,
  p1 = 12,
  p2 = 1
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a positive random number generator}

\item{p1}{first distribution parameter (Gamma shape by default)}

\item{p2}{second distribution parameter (Gamma rate by default)}
}
\value{
a configuration object for \link{F_eip.random}
}
\description{
Build a configuration object for random EIP values.
}
