% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{make_F_f_random}
\alias{make_F_f_random}
\title{Make random blood feeding rate object}
\usage{
make_F_f_random(
  nPatches,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 280,
  p2 = 1000
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for \code{f_rand}}

\item{p2}{argument for f_rand}
}
\value{
a configuration object for \link{F_f.random}
}
\description{
Build a configuration object for random blood feeding rates.
}
