% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{make_F_f_type2}
\alias{make_F_f_type2}
\title{Make type 2 blood feeding rate configuration}
\usage{
make_F_f_type2(nPatches, options = list(), fx = 0.3, sf = 0.1)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{fx}{the maximum blood feeding rate, \eqn{f_x}}

\item{sf}{the shape parameter, \eqn{s_f}}
}
\value{
a configuration object for \link{F_f.type2}
}
\description{
Build a configuration object for the type 2 blood feeding
response in response to the availability of blood hosts:
\deqn{F_f(B)= f_x \frac{s_f B}{1+s_f B}}.
}
