% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{make_F_g_random}
\alias{make_F_g_random}
\title{Make random mosquito mortality rates}
\usage{
make_F_g_random(
  nPatches,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 80,
  p2 = 1000
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a configuration object for \link{F_g.random}
}
\description{
Build a configuration object for random mosquito mortality.
}
