% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{make_F_nu_random}
\alias{make_F_nu_random}
\title{Make random laying rates}
\usage{
make_F_nu_random(
  nPatches,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 980,
  p2 = 1000
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument 1 for f_rand}

\item{p2}{argument 2 for f_rand}
}
\value{
an object to set up egg laying
}
\description{
Set up a random vector
}
