% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{make_F_nu_type2}
\alias{make_F_nu_type2}
\title{Make type2 egg laying rates}
\usage{
make_F_nu_type2(nPatches, options = list(), vx = 0.3, sv = 0.1)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{vx}{the maximum laying}

\item{sv}{the shape parameter}
}
\value{
an object to set up egg laying
}
\description{
Set up a type2 vector
}
