% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{make_F_q_WVB}
\alias{make_F_q_WVB}
\title{Make a WVB human feeding fraction configuration}
\usage{
make_F_q_WVB(nPatches, options = list())
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}
}
\value{
a configuration object for \link{F_q.WVB}
}
\description{
Create a configuration object for the WVB human feeding
fraction method.
}
\seealso{
\link{F_q.WVB}
}
