% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{make_F_sigma_random}
\alias{make_F_sigma_random}
\title{Make random patch emigration rates}
\usage{
make_F_sigma_random(
  nPatches,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 100,
  p2 = 1000
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a configuration object for \link{F_sigma.random}
}
\description{
Set up a random vector
}
