% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{make_F_sigma_type2}
\alias{make_F_sigma_type2}
\title{Make a patch emigration configuration}
\usage{
make_F_sigma_type2(
  nPatches,
  options = list(),
  sigX = 0.3,
  sigB = 0.1,
  sigQ = 0.1,
  sigS = 0.1
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{sigX}{the emigration rate from a patch with no resources}

\item{sigB}{the shape parameter for blood hosts}

\item{sigQ}{the shape parameter for aquatic habitats}

\item{sigS}{the shape parameter for sugar}
}
\value{
a configuration object for \link{F_sigma.type2}
}
\description{
Make an object to model patch emigration as a
type 2 functional response
}
\seealso{
\link{F_sigma.type2}
}
