% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.product}
\alias{make_function.product}
\title{Make a Sinusoidal Function}
\usage{
\method{make_function}{product}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function that is the product of two other functions
}
\description{
Build a function that is the
product of two other functions
}
