% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.sharkfin}
\alias{make_function.sharkfin}
\title{Make a Sharkfin Function}
\usage{
\method{make_function}{sharkfin}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function
}
\description{
A sharkfin function is built in steps:
\enumerate{
\item take the product of two sigmoidal functions
\itemize{
\item the first one rises around day \eqn{D} with rate parameter \eqn{uk}
\item the second one decays around day \eqn{D+L} with rate \eqn{-dk}
}
\item the product is raised a power \eqn{pw}
\item the result is scaled so that the maximum is \eqn{mx}
For the default values, the function looks like a shark fin.
}
}
