% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.sigmoid}
\alias{make_function.sigmoid}
\title{Make a Sigmoidal Function}
\usage{
\method{make_function}{sigmoid}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function
}
\description{
Build a function to model a
forced seasonal signal. The shape parameters determine
the timing, frequency, and relative intensity
over the season. The function is normalized
to have an annual value set by \code{norm}
}
