% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.splinef}
\alias{make_function.splinef}
\title{Make a spline function}
\usage{
\method{make_function}{splinef}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function
}
\description{
A spline function passes time points \code{tt} and
associated values \code{yy} and returns a spline function
}
