% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.sum}
\alias{make_function.sum}
\title{Make a Function that is the sum of Two other Functions}
\usage{
\method{make_function}{sum}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function that is the sum of two other functions
}
\description{
Build a function that is the sum of two
other functions.
}
