% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.val}
\alias{make_function.val}
\title{Make a Function that is the sum of Two other Functions}
\usage{
\method{make_function}{val}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function that returns a constant value
}
\description{
Build a function that returns a constant value
}
