% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_function.zero}
\alias{make_function.zero}
\title{Make a Function that is the sum of Two other Functions}
\usage{
\method{make_function}{zero}(opts)
}
\arguments{
\item{opts}{a named list}
}
\value{
a function that is the sum of two other functions
}
\description{
Return \link{F_zero}
}
