% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{make_indices_L.basicL}
\alias{make_indices_L.basicL}
\title{Make Indices for \strong{L} Component Variables for \code{basicL}}
\usage{
\method{make_indices_L}{basicL}(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of the indices for the \strong{L} component variables
for the \code{basicL} module
}
