% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{make_indices_L.basicL_dts}
\alias{make_indices_L.basicL_dts}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_L}{basicL_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_L} for basicL_dts competition model.
}
