% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{make_indices_L.basicL_xde}
\alias{make_indices_L.basicL_xde}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_L}{basicL_xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_L} for basicL_xde competition model.
}
