% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_indices_L}
\alias{make_indices_L}
\title{Set the Values of the Indices for \strong{L} Component Modules}
\usage{
make_indices_L(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the Values of the Indices for \strong{L} Component Modules
}
