% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{make_indices_L.trivial}
\alias{make_indices_L.trivial}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_L}{trivial}(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_L} for trivial (forced emergence) model.
}
