% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{make_indices_MYZ.basicM_dts}
\alias{make_indices_MYZ.basicM_dts}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{basicM_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{make_indices_MYZ} for the basicM_dts model.
}
