% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ-dts.R
\name{make_indices_MYZ.ENBRQ_dts}
\alias{make_indices_MYZ.ENBRQ_dts}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{ENBRQ_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{make_indices_MYZ} for the ENBRQ_dts model.
}
