% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{make_indices_MYZ.trivial}
\alias{make_indices_MYZ.trivial}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{trivial}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_MYZ} for trivial (forced emergence) model.
}
