% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI-xde.R
\name{make_indices_X.hMoI_xde}
\alias{make_indices_X.hMoI_xde}
\title{Add indices for human population to parameter list}
\usage{
\method{make_indices_X}{hMoI_xde}(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{make_indices_X} for the hybrid MoI model.
}
