% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{make_inits_L_basicL_xde}
\alias{make_inits_L_basicL_xde}
\title{Make inits for basicL_xde competition aquatic mosquito model}
\usage{
make_inits_L_basicL_xde(pars, L0)
}
\arguments{
\item{pars}{a \link{list}}

\item{L0}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for basicL_xde competition aquatic mosquito model
}
