% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{make_inits_L_stages_dts}
\alias{make_inits_L_stages_dts}
\title{Make inits for stages_dts competition aquatic mosquito model}
\usage{
make_inits_L_stages_dts(pars, L10, L20, L30, L40, P0)
}
\arguments{
\item{pars}{a \link{list}}

\item{L10}{initial conditions}

\item{L20}{initial conditions}

\item{L30}{initial conditions}

\item{L40}{initial conditions}

\item{P0}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Make inits for stages_dts competition aquatic mosquito model
}
