% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{make_inits_MYZ_basicM_dts}
\alias{make_inits_MYZ_basicM_dts}
\title{Make inits for basicM_dts adult mosquito model}
\usage{
make_inits_MYZ_basicM_dts(pars, M0, P0)
}
\arguments{
\item{pars}{a \link{list}}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for basicM_dts adult mosquito model
}
