% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{make_inits_MYZ_basicM_xde}
\alias{make_inits_MYZ_basicM_xde}
\title{Make inits for basicM_xde adult mosquito model}
\usage{
make_inits_MYZ_basicM_xde(pars, M0, P0)
}
\arguments{
\item{pars}{a \link{list}}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for basicM_xde adult mosquito model
}
