% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{make_inits_MYZ_mz}
\alias{make_inits_MYZ_mz}
\title{Make inits for mz adult mosquito model}
\usage{
make_inits_MYZ_mz(pars, M0, Z0)
}
\arguments{
\item{pars}{a \link{list}}

\item{M0}{total mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for mz adult mosquito model
}
