% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{make_inits_MYZ_trace}
\alias{make_inits_MYZ_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_inits_MYZ_trace(pars, MYZ0 = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{MYZ0}{is set to NULL for the trace model}
}
\value{
none
}
\description{
Make parameters for trace aquatic mosquito model
}
