% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-dts.R
\name{make_inits_X_SIP_dts}
\alias{make_inits_X_SIP_dts}
\title{Make inits for SIP_dts human model}
\usage{
make_inits_X_SIP_dts(pars, S0, I0, P0)
}
\arguments{
\item{pars}{a \link{list}}

\item{S0}{size of infected population in each strata}

\item{I0}{size of infected population in each strata}

\item{P0}{size of population protected by prophylaxis in each strata}
}
\value{
none
}
\description{
Make inits for SIP_dts human model
}
