% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{make_inits_X_SIS_dts}
\alias{make_inits_X_SIS_dts}
\title{Make inits for SIS_dts human model}
\usage{
make_inits_X_SIS_dts(pars, S0, I0)
}
\arguments{
\item{pars}{a \link{list}}

\item{S0}{size of infected population in each strata}

\item{I0}{size of infected population in each strata}
}
\value{
none
}
\description{
Make inits for SIS_dts human model
}
