% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{make_inits_X_SIS_xde}
\alias{make_inits_X_SIS_xde}
\title{Make inits for SIS xde human model}
\usage{
make_inits_X_SIS_xde(pars, S0, I0)
}
\arguments{
\item{pars}{a \link{list}}

\item{S0}{size of infected population in each strata}

\item{I0}{size of infected population in each strata}
}
\value{
none
}
\description{
Make inits for SIS xde human model
}
