% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{make_K_matrix_herethere}
\alias{make_K_matrix_herethere}
\title{Make a Here-There Dispersal Matrix}
\usage{
make_K_matrix_herethere(nPatches)
}
\arguments{
\item{nPatches}{is the number of patches in the model}
}
\description{
Make a Here-There Dispersal Matrix
}
