% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_K_matrix_xy}
\alias{make_K_matrix_xy}
\title{make a Kernel-Based Mosquito Dispersal Matrix}
\usage{
make_K_matrix_xy(xy, F_K = F_exp, V = list())
}
\arguments{
\item{xy}{is a vector of the xy-coordinates of patch locations}

\item{F_K}{is a function that weights putative locations by distance}

\item{V}{variables}
}
\description{
make a Kernel-Based Mosquito Dispersal Matrix
}
