% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{make_kappa}
\alias{make_kappa}
\title{Compute kappa}
\usage{
make_kappa(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This function computes kappa for each vector species in each patch
}
