% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{make_L_inits_basicL}
\alias{make_L_inits_basicL}
\title{Make Initial Values for \code{basicL} (\strong{L} Component)}
\usage{
make_L_inits_basicL(nHabitats, options = list(), L = 1)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{options}{a \link{list} that overwrites default values}

\item{L}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Initial values of the variable \eqn{L} can be set
}
