% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{make_L_obj_basicL}
\alias{make_L_obj_basicL}
\title{Make \code{L_obj} for \code{basicL} (\strong{L} Component)}
\usage{
make_L_obj_basicL(
  nHabitats,
  options = list(),
  psi = 1/8,
  xi = 0,
  phi = 1/8,
  theta = 1/100
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{options}{a named \link{list}}

\item{psi}{maturation rates for each aquatic habitat}

\item{xi}{delayed maturation in response to mean crowding}

\item{phi}{density-independent mortality rates for each aquatic habitat}

\item{theta}{density-dependent mortality terms for each aquatic habitat}
}
\value{
\strong{\code{L_obj}} an \strong{L} component object
}
\description{
The following parameters will be set to the values in
\code{options.} If they are not found, default values will be used.
\itemize{
\item \eqn{\psi} or \code{psi}: maturation rate
\item \eqn{\xi} or \code{xi}: delayed maturation response due to mean crowding
\item \eqn{\phi} or \code{phi}: density-independent death rate
\item \eqn{\theta} or \code{theta}: density dependence in mortality: the slope of the response to mean crowding
}
}
\seealso{
Called by: \link{setup_L_obj.basicL}. Related: \link{dLdt.basicL} & \link{Update_Lt.basicL}
}
