% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{make_Linits.basicL}
\alias{make_Linits.basicL}
\title{Setup the Initial Values for \strong{L} Component Variables for \code{basicL}}
\usage{
\method{make_Linits}{basicL}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This sets initial values of the variable \eqn{L} by
calling \link{create_Linits_basicL}. Default values are used unless other values
are passed in \code{Lopts} by name (\emph{i.e.} \code{Lopts$L})
}
\seealso{
\link{create_Linits_basicL}
}
