% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_Linits}
\alias{make_Linits}
\title{Set up Initial Values of Variables for \strong{L} Component modules}
\usage{
make_Linits(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{Lname}.
}
