% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{make_local_frac}
\alias{make_local_frac}
\title{Compute the local fraction}
\usage{
make_local_frac(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
