% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{make_Lpar.basicL}
\alias{make_Lpar.basicL}
\title{xde_setup Lpar for the basicL model}
\usage{
\method{make_Lpar}{basicL}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a named \link{list} to configure \strong{\code{Lpar}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{make_Lpar} for the basicL model
}
