% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_Lpar}
\alias{make_Lpar}
\title{Set up \code{Lpar} for \strong{L} Component modules}
\usage{
make_Lpar(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a named \link{list} to configure \strong{\code{Lpar}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Each instance of \code{make_Lpar} calls a function (usually \verb{create_Lpar_*}) that
creates an object \strong{\code{Lpar}.}
It is attached the \strong{\code{xds}} object as \verb{pars$Lpar[[s]].} Each instance of \verb{create_Lpar_*}
should assign default parameter values. These will be over-written by \code{Lopts}
}
\note{
This method assigns \code{Lname} to class(\code{Lname}) and dispatches on \code{Lname}.
}
