% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{make_Lpar.trivial}
\alias{make_Lpar.trivial}
\title{Setup \code{Lpar} for the \code{trivial} module}
\usage{
\method{make_Lpar}{trivial}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a named \link{list} to configure \strong{\code{Lpar}}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Lpar} for the trivial model
}
