% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{make_Lpar_stages_dts}
\alias{make_Lpar_stages_dts}
\title{Make parameters for stages_dts competition aquatic mosquito model}
\usage{
make_Lpar_stages_dts(
  nHabitats,
  Lopts = list(),
  psi1 = 1/8,
  psi2 = 1/8,
  psi3 = 1/8,
  psi4 = 1/8,
  xi1 = 0,
  xi2 = 0,
  xi3 = 0,
  xi4 = 0,
  phi1 = 1/8,
  phi2 = 1/8,
  phi3 = 1/8,
  phi4 = 1/8,
  theta1 = 1/100,
  theta2 = 1/100,
  theta3 = 1/100,
  theta4 = 1/100
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{psi1}{maturation rates for L1 in each aquatic habitat}

\item{psi2}{maturation rates for L2 in each aquatic habitat}

\item{psi3}{maturation rates for L3 in each aquatic habitat}

\item{psi4}{maturation rates for L1 in each aquatic habitat}

\item{xi1}{delayed maturation for L1 due to mean crowding}

\item{xi2}{delayed maturation for L2 due to mean crowding}

\item{xi3}{delayed maturation for L3 due to mean crowding}

\item{xi4}{delayed maturation for L4 due to mean crowding}

\item{phi1}{density-independent mortality rates for L1 for each aquatic habitat}

\item{phi2}{density-independent mortality rates for L2 for each aquatic habitat}

\item{phi3}{density-independent mortality rates for L3 for each aquatic habitat}

\item{phi4}{density-independent mortality rates for L4 for each aquatic habitat}

\item{theta1}{density-dependent mortality terms for L1 for each aquatic habitat}

\item{theta2}{density-dependent mortality terms for L2 for each aquatic habitat}

\item{theta3}{density-dependent mortality terms for L3 for each aquatic habitat}

\item{theta4}{density-dependent mortality terms for L4 for each aquatic habitat}
}
\value{
a \link{list} with Lpar added
}
\description{
Make parameters for stages_dts competition aquatic mosquito model
}
