% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{make_Lpar_trivial}
\alias{make_Lpar_trivial}
\title{Make \code{Lpar} for \code{trivial} (\strong{L} Component)}
\usage{
make_Lpar_trivial(
  nHabitats,
  Lopts = list(),
  Lambda = 1000,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{F_season}{a function describing a seasonal pattern over time}

\item{season_par}{an object to configure a seasonality function using \link{make_function}}

\item{F_trend}{a function describing a temporal trend over time}

\item{trend_par}{an object to configure a trends function using \link{make_function}}
}
\value{
none
}
\description{
The number of emerging adults is a function \deqn{\Lambda S(t) T(t)} where
\itemize{
\item \eqn{\Lambda} or \code{Lambda} is the mean number of adult female mosquitoes emerging per day
\item \eqn{S(t)} or \code{F_season} is a seasonal signal (ideally, with an average annual mean of 1)
\item \eqn{T(t)} or \code{F_trend} is a function returning a trend (ideally, with an average value of 1)
}
}
