% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{make_M_obj_basicM}
\alias{make_M_obj_basicM}
\title{Make parameters for GeRM ODE adult mosquito model}
\usage{
make_M_obj_basicM(
  nPatches,
  options = list(),
  g = 1/12,
  sigma = 1/8,
  mu = 0,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{options}{a named list: named values overwrite defaults}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{emigration loss}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a \link{list}
}
\description{
Make parameters for GeRM ODE adult mosquito model
}
