% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{make_MY_inits_basicM}
\alias{make_MY_inits_basicM}
\title{Make inits for basicM adult mosquito model}
\usage{
make_MY_inits_basicM(nPatches, options, M = 5, P = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}
}
\value{
none
}
\description{
Make inits for basicM adult mosquito model
}
