% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{make_MY_inits_GeRM}
\alias{make_MY_inits_GeRM}
\title{Make inits for GeRM adult mosquito model}
\usage{
make_MY_inits_GeRM(
  nPatches,
  Upsilon,
  options = list(),
  M = 5,
  P = 1,
  Y = 1,
  Z = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{Upsilon}{a matrix describing survival and dispersal through the EIP}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}

\item{Y}{infected mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for GeRM adult mosquito model
}
