% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{make_MY_inits_RMdts}
\alias{make_MY_inits_RMdts}
\title{Make inits for RMdts adult mosquito model}
\usage{
make_MY_inits_RMdts(
  nPatches,
  max_eip,
  options = list(),
  M = 5,
  P = 1,
  U = 0,
  Y = 1,
  Z = 1
)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{max_eip}{the maximum number of EIP cohorts, an \link{integer}}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{P}{total parous mosquito density at each patch}

\item{U}{total uninfected mosquito density at each patch}

\item{Y}{infected mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for RMdts adult mosquito model
}
