% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{make_MY_inits_SI}
\alias{make_MY_inits_SI}
\title{Make inits for \code{SI} adult mosquito model}
\usage{
make_MY_inits_SI(nPatches, options = list(), M = 5, Y = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{Y}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for \code{SI} adult mosquito model
}
