% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{make_MY_obj_GeM}
\alias{make_MY_obj_GeM}
\title{Make parameters for GeM ODE adult mosquito model}
\usage{
make_MY_obj_GeM(
  nPatches,
  options = list(),
  eip = 12,
  g = 1/12,
  sigma = 1/8,
  mu = 0,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggs_per_batch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{eip}{the extrinsic incubation period}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{emigration loss}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggs_per_batch}{eggs laid per oviposition}
}
\value{
a \link{list}
}
\description{
Make parameters for GeM ODE adult mosquito model
}
\keyword{internal}
