% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{make_MY_obj_SI}
\alias{make_MY_obj_SI}
\title{Make parameters for SI ODE adult mosquito model}
\usage{
make_MY_obj_SI(
  nPatches,
  options = list(),
  eip = 12,
  g = 1/12,
  sigma = 1/8,
  mu = 0,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{options}{a named list: named values overwrite defaults}

\item{eip}{the extrinSIc incubation period}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{fraction lost during emigration}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{ovipoSItion rate, per mosquito}

\item{eggsPerBatch}{eggs laid per ovipoSItion}
}
\value{
a \link{list}
}
\description{
Make parameters for SI ODE adult mosquito model
}
