% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{make_MY_obj_trivial}
\alias{make_MY_obj_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
make_MY_obj_trivial(
  nPatches,
  options,
  f = 1,
  q = 1,
  Z = 1,
  eggs = 1,
  F_season = F_one,
  F_trend = F_one,
  F_shock = F_one,
  season_par = list(name = "F_one"),
  trend_par = list(name = "F_one"),
  shock_par = list(name = "F_one")
)
}
\arguments{
\item{nPatches}{an integer}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Z}{the density of infectious mosquitoes}

\item{eggs}{the mean egg laying rate}

\item{F_season}{the seasonal pattern function}

\item{F_trend}{the trend function}

\item{F_shock}{the shock function}

\item{season_par}{a list of options for F_season}

\item{trend_par}{a list of options for F_trend}

\item{shock_par}{a list of options for F_shock}
}
\value{
a \link{list}
}
\description{
Make parameters for trivial aquatic mosquito model
}
\keyword{internal}
