% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_MYZinits.GeRM}
\alias{make_MYZinits.GeRM}
\title{Setup initial values for the GeRM model}
\usage{
\method{make_MYZinits}{GeRM}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_MYZinits} for the GeRM model
}
