% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{make_MYZinits}
\alias{make_MYZinits}
\title{A function to set up adult mosquito models}
\usage{
make_MYZinits(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYZname}.
}
