% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_MYZinits.RM}
\alias{make_MYZinits.RM}
\title{Setup initial values for the RM model}
\usage{
\method{make_MYZinits}{RM}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_MYZinits} for the RM model
}
