% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{make_MYZinits.SI}
\alias{make_MYZinits.SI}
\title{Setup initial values for the \code{SI} model}
\usage{
\method{make_MYZinits}{SI}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_MYZinits} for the \code{SI} model
}
