% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-dts.R
\name{make_MYZinits_basicM_dts}
\alias{make_MYZinits_basicM_dts}
\title{Make inits for basicM_dts adult mosquito model}
\usage{
make_MYZinits_basicM_dts(nPatches, MYZopts = list(), M0 = 5, P0 = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{P0}{total parous mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for basicM_dts adult mosquito model
}
