% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{make_MYZinits_si}
\alias{make_MYZinits_si}
\title{Make inits for si adult mosquito model}
\usage{
make_MYZinits_si(nPatches, MYZopts = list(), M0 = 5, Z0 = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M0}{total mosquito density at each patch}

\item{Z0}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for si adult mosquito model
}
