% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{make_MYZinits_si_old}
\alias{make_MYZinits_si_old}
\title{Make inits for \code{si_old} adult mosquito model}
\usage{
make_MYZinits_si_old(nPatches, MYZopts = list(), M = 5, Y = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{Y}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for \code{si_old} adult mosquito model
}
