% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{make_MYZpar.basicM}
\alias{make_MYZpar.basicM}
\title{Setup MYZpar for the basicM xde model}
\usage{
\method{make_MYZpar}{basicM}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_MYZpar} for the basicM xde model
}
