% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{make_MYZpar.GeRM}
\alias{make_MYZpar.GeRM}
\title{Setup MYZpar for the GeRM model}
\usage{
\method{make_MYZpar}{GeRM}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_MYZpar} for the GeRM model
}
