% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{make_MYZpar}
\alias{make_MYZpar}
\title{A function to set up adult mosquito models}
\usage{
make_MYZpar(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYZname}.
}
