% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{make_MYZpar.RM}
\alias{make_MYZpar.RM}
\title{Setup MYZpar for the RM model}
\usage{
\method{make_MYZpar}{RM}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_MYZpar} for the RM model
}
