% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-sei.R
\name{make_MYZpar.sei}
\alias{make_MYZpar.sei}
\title{Setup MYZpar for the sei model}
\usage{
\method{make_MYZpar}{sei}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_MYZpar} for the sei model
}
