% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{make_MYZpar_trivial}
\alias{make_MYZpar_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
make_MYZpar_trivial(
  nPatches,
  MYZopts,
  f = 1,
  q = 1,
  Z = 1,
  eggs = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list()
)
}
\arguments{
\item{nPatches}{an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Z}{the human fraction}

\item{eggs}{the human fraction}

\item{F_season}{a function describing a seasonal pattern}

\item{season_par}{parameters to configure a \code{F_season} using \link{make_function}}

\item{F_trend}{a function describing a temporal trend}

\item{trend_par}{parameters to configure \code{F_trend} using \link{make_function}}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
