% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{make_MYZprobs.dts}
\alias{make_MYZprobs.dts}
\title{Compute probabilities from rates for \code{dts} models with runtime support}
\usage{
\method{make_MYZprobs}{dts}(pars, s)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
The \code{dts} case calls \code{MYZ_rates2probs} to compute
}
