% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{make_MYZprobs}
\alias{make_MYZprobs}
\title{Adult bionomics: Compute probabilities from rates}
\usage{
make_MYZprobs(pars, s)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Where homologous \code{xde} and \code{dts} models exist,
the rates can be translated into probabilities with runtime
support.
}
\seealso{
\link{dynamic_MYZpar}
}
