% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{make_MYZprobs.xde}
\alias{make_MYZprobs.xde}
\title{Adult bionomics: Compute probabilities from rates}
\usage{
\method{make_MYZprobs}{xde}(pars, s)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
unmodified \code{xds} object
}
\description{
The \code{xde} case returns the unmodified \code{xds} object.
}
