% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-demography.R
\name{make_Omega.dts}
\alias{make_Omega.dts}
\title{Make the mosquito demography matrix for spatial RM model in continuous time}
\usage{
\method{make_Omega}{dts}(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\description{
Make the mosquito demography matrix for spatial RM model in continuous time
}
