% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-demography.R
\name{make_Omega.xde}
\alias{make_Omega.xde}
\title{Make the mosquito demography matrix for spatial RM model in continuous time}
\usage{
\method{make_Omega}{xde}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\description{
Make the mosquito demography matrix for spatial RM model in continuous time
}
