% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-demography.R
\name{make_Omega_dts}
\alias{make_Omega_dts}
\title{Make the mosquito demography matrix for spatial RM model in discrete time}
\usage{
make_Omega_dts(p, sigma, mu, calK)
}
\arguments{
\item{p}{mosquito daily survival, a vector of length \code{nPatches}}

\item{sigma}{mosquito daily emigration fraction, a vector of length \code{nPatches}}

\item{mu}{emigration survival, a vector of length \code{nPatches}}

\item{calK}{a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
Make the mosquito demography matrix for spatial RM model in discrete time
}
