% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{make_outputs.cohort}
\alias{make_outputs.cohort}
\title{Compute}
\usage{
\method{make_outputs}{cohort}(pars, de_vars, tm)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{de_vars}{the output of \link{deSolve}}

\item{tm}{the time}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute
}
