% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs-make.R
\name{make_outputs}
\alias{make_outputs}
\title{Make Outputs}
\usage{
make_outputs(pars, de_vars, tm)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{de_vars}{the output of \link{deSolve}}

\item{tm}{the time}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make Outputs
}
