% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{make_ovitraps_random}
\alias{make_ovitraps_random}
\title{Random ovitrap availability}
\usage{
make_ovitraps_random(
  nPatches,
  options = list(),
  f_rand = stats::rlnorm,
  p1 = 0,
  p2 = 0.5
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a numeric vector
}
\description{
Set up a random vector describing ovitrap availability
}
