% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{make_parameters_L_basicL_dts}
\alias{make_parameters_L_basicL_dts}
\title{Make parameters for basicL_dts competition aquatic mosquito model}
\usage{
make_parameters_L_basicL_dts(pars, psi, xi, phi, theta)
}
\arguments{
\item{pars}{a \link{list}}

\item{psi}{maturation rates for each aquatic habitat}

\item{xi}{delayed maturation due to mean crowding}

\item{phi}{density-independent mortality rates for each aquatic habitat}

\item{theta}{density-dependent mortality terms for each aquatic habitat}
}
\value{
a \link{list} with Lpar added
}
\description{
Make parameters for basicL_dts competition aquatic mosquito model
}
