% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{make_parameters_L_basicL_xde}
\alias{make_parameters_L_basicL_xde}
\title{Make parameters for basicL_xde competition aquatic mosquito model}
\usage{
make_parameters_L_basicL_xde(pars, psi, phi, theta)
}
\arguments{
\item{pars}{a \link{list}}

\item{psi}{maturation rates for each aquatic habitat}

\item{phi}{density-independent mortality rates for each aquatic habitat}

\item{theta}{density-dependent mortality terms for each aquatic habitat}
}
\value{
a \link{list} with Lpar added
}
\description{
Make parameters for basicL_xde competition aquatic mosquito model
}
