% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{make_parameters_L_stages_dts}
\alias{make_parameters_L_stages_dts}
\title{Make parameters for stages_dts competition aquatic mosquito model}
\usage{
make_parameters_L_stages_dts(
  pars,
  psi1,
  psi2,
  psi3,
  psi4,
  xi1,
  xi2,
  xi3,
  xi4,
  phi1,
  phi2,
  phi3,
  phi4,
  theta1,
  theta2,
  theta3,
  theta4
)
}
\arguments{
\item{pars}{a \link{list}}

\item{psi1}{maturation rates for L1 in each aquatic habitat}

\item{psi2}{maturation rates for L2 in each aquatic habitat}

\item{psi3}{maturation rates for L3 in each aquatic habitat}

\item{psi4}{maturation rates for L1 in each aquatic habitat}

\item{xi1}{delayed maturation for L1 due to mean crowding}

\item{xi2}{delayed maturation for L2 due to mean crowding}

\item{xi3}{delayed maturation for L3 due to mean crowding}

\item{xi4}{delayed maturation for L4 due to mean crowding}

\item{phi1}{density-independent mortality rates for L1 for each aquatic habitat}

\item{phi2}{density-independent mortality rates for L2 for each aquatic habitat}

\item{phi3}{density-independent mortality rates for L3 for each aquatic habitat}

\item{phi4}{density-independent mortality rates for L4 for each aquatic habitat}

\item{theta1}{density-dependent mortality terms for L1 for each aquatic habitat}

\item{theta2}{density-dependent mortality terms for L2 for each aquatic habitat}

\item{theta3}{density-dependent mortality terms for L3 for each aquatic habitat}

\item{theta4}{density-dependent mortality terms for L4 for each aquatic habitat}
}
\value{
a \link{list} with Lpar added
}
\description{
Make parameters for stages_dts competition aquatic mosquito model
}
