% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{make_parameters_MYZ_basicM_xde}
\alias{make_parameters_MYZ_basicM_xde}
\title{Make parameters for a basic adult mosquito model}
\usage{
make_parameters_MYZ_basicM_xde(
  pars,
  g,
  sigma,
  mu,
  f,
  q,
  nu,
  eggsPerBatch,
  calK
)
}
\arguments{
\item{pars}{a \link{list}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{emigration loss}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
none
}
\description{
Make parameters for a basic adult mosquito model
}
