% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{make_parameters_MYZ_si}
\alias{make_parameters_MYZ_si}
\title{Make parameters for si ODE adult mosquito model}
\usage{
make_parameters_MYZ_si(pars, g, eip, sigma, mu, f, q, nu, eggsPerBatch, calK)
}
\arguments{
\item{pars}{a \link{list}}

\item{g}{mosquito mortality rate}

\item{eip}{the extrinsic incubation period (in days)}

\item{sigma}{emigration rate}

\item{mu}{emigration loss}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}

\item{calK}{mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}
}
\value{
a \link{list}
}
\description{
Make parameters for si ODE adult mosquito model
}
