% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{make_parameters_MYZ_trace}
\alias{make_parameters_MYZ_trace}
\title{Make parameters for trace aquatic mosquito model}
\usage{
make_parameters_MYZ_trace(pars, f = 0.3, q = 0.95, MYZm = 1, MYZf = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{MYZm}{a vector of mean mosquito densities}

\item{MYZf}{a \link{function} of the form MYZf(t, pars) that computes temporal fluctuations}
}
\value{
none
}
\description{
Make parameters for trace aquatic mosquito model
}
