% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-xde.R
\name{make_parameters_X_SIP_xde}
\alias{make_parameters_X_SIP_xde}
\title{Make parameters for SIP_xde human model}
\usage{
make_parameters_X_SIP_xde(pars, b, c, r, rho, eta, xi)
}
\arguments{
\item{pars}{a \link{list}}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}

\item{rho}{probability of successful treatment upon infection}

\item{eta}{prophylaxis waning rate}

\item{xi}{background treatment rate}
}
\value{
none
}
\description{
Make parameters for SIP_xde human model
}
