% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{make_parameters_X_SIS_dts}
\alias{make_parameters_X_SIS_dts}
\title{Make parameters for SIS_dts human model}
\usage{
make_parameters_X_SIS_dts(pars, b, c, r)
}
\arguments{
\item{pars}{a \link{list}}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}
}
\value{
a \link{list}
}
\description{
Make parameters for SIS_dts human model
}
