% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{make_parameters_X_SIS_xde}
\alias{make_parameters_X_SIS_xde}
\title{Make parameters for SIS xde human model}
\usage{
make_parameters_X_SIS_xde(pars, b, c, r)
}
\arguments{
\item{pars}{a \link{list}}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{c}{transmission probability (efficiency) from human to mosquito}

\item{r}{recovery rate}
}
\value{
a \link{list}
}
\description{
Make parameters for SIS xde human model
}
