% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{make_Q}
\alias{make_Q}
\title{Compute and store availability of egg-laying habitats}
\usage{
make_Q(pars)
}
\value{
a \link{list}  -- the modified \code{xds} object
}
\description{
Set the value of a variable, \eqn{Q}, that describes the availability of any habitat
or device that would attract mosquitoes and induce them to lay eggs.
}
\seealso{
\link{compute_Q}
}
