% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{make_RBR}
\alias{make_RBR}
\title{Compute and attach the relative biting rates}
\usage{
make_RBR(pars, y)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{y}{state vector}
}
\value{
an \code{xds} object
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
