% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{make_runtime.xde}
\alias{make_runtime.xde}
\title{Set up run-time time step support for \code{xde} models}
\usage{
\method{make_runtime}{xde}(pars, Xday, MYZday, Lday, Lname)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{Lname}{is the S3 class of the L model}
}
\value{
the unmodified \code{xds} object
}
\description{
Continuous time models don't use run-time support,
so this returns the \code{xds} object without modification
}
