% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_static_obj}
\alias{make_static_obj}
\title{Set up a static object}
\usage{
make_static_obj()
}
\value{
an empty list with \code{class} set to \code{static}
}
\description{
Set up a static object
}
\keyword{internal}
