% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{make_TaR}
\alias{make_TaR}
\title{Make TaR}
\usage{
make_TaR(pars, t = 0)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{t}{the time}
}
\value{
the modified \code{xds} object
}
\description{
Make a time at risk matrix (TaR) from a time spent matrix and a circadian function
}
