% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{make_time_away_random}
\alias{make_time_away_random}
\title{Make other blood hosts}
\usage{
make_time_away_random(
  nStrata,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 50,
  p2 = 1000
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a numeric vector
}
\description{
Set up a random vector
}
