% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{make_times_dts}
\alias{make_times_dts}
\title{Times Utility for Discrete Time Systems}
\usage{
make_times_dts(Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method sets up the time points when output is wanted.
For discrete time systems, the system must compute values at every point,
but it is possible

If \code{times} is not null, it sets \code{maxT} to the maximum value in times.
Otherwise it sets \code{maxT=Tmax.}

It returns a sequence from \code{0} to \code{maxT} by \code{dt.}

A check
values of \code{times} should be some integer multiple of \code{dt}
}
