% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{make_times_xde}
\alias{make_times_xde}
\title{Times Utility for Differential Equations}
\usage{
make_times_xde(Tmax, dt, times = NULL)
}
\arguments{
\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method sets up the time points for the
independent variable to return solutions for the dependent variables.

If \code{times} is not \verb{NULL,} then it returns solutions at those values.

Otherwise, it returns a sequence from \code{0} to \code{Tmax} by \code{dt.}

If values in \code{times} are not a sequence, the function checks that
all its values are in the sequence of time points when solutions are returned.
}
