% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_timespent_athome}
\alias{make_timespent_athome}
\title{Make a time spent matrix, called timespent}
\usage{
make_timespent_athome(
  nPatches,
  residence,
  options = list(),
  atHome = 1,
  travel = 0
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the home patch for each stratum}

\item{options}{is a set of options that overwrites the defaults}

\item{atHome}{is the fraction of time spent at home}

\item{travel}{is the fraction of time spent traveling}
}
\value{
a \link{matrix}
}
\description{
Make a time spent matrix, called timespent
}
