% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{make_travel_eir_random}
\alias{make_travel_eir_random}
\title{Make other blood hosts}
\usage{
make_travel_eir_random(
  nStrata,
  options = list(),
  f_rand = stats::rlnorm,
  p1 = log(1/3650),
  p2 = 0.2
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a numeric vector
}
\description{
Set up a random vector
}
