% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{make_Umatrix}
\alias{make_Umatrix}
\title{Compute and store \eqn{U}}
\usage{
make_Umatrix(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Compute and store the egg distribution matrices
for all the vector species
}
