% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-demography.R
\name{make_Upsilon}
\alias{make_Upsilon}
\title{Make Upsilon for delay differential equations}
\usage{
make_Upsilon(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
This method dispatches on the type of \code{pars$MYZpar[[s]]}.
}
