% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-demography.R
\name{make_Upsilon.static}
\alias{make_Upsilon.static}
\title{Handle the mosquito demography matrix for static models}
\usage{
\method{make_Upsilon}{static}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
Handle the mosquito demography matrix for static models
}
