% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-SI.R
\name{make_VC.SI}
\alias{make_VC.SI}
\title{Make the VC Matrix}
\usage{
\method{make_VC}{SI}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a numeric \link{matrix}
}
\description{
Compute the \eqn{N_p \times N_p}
matrix \eqn{\mathcal{V}} whose columns describe
the number of infective bites arising in each patch
from all the mosquitoes biting a single human on a
single day in each patch.
}
\keyword{internal}
