% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-macdonald.R,
%   R/vectorial_capacity.R
\name{make_VC.macdonald}
\alias{make_VC.macdonald}
\alias{make_VC_gem}
\title{Make the VC Matrix}
\usage{
\method{make_VC}{macdonald}(xds_obj, s = 1)

make_VC_gem(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a numeric \link{matrix}

a numeric \link{matrix}
}
\description{
Compute the \eqn{N_p \times N_p}
matrix \eqn{\mathcal{V}} whose columns describe
the number of infective bites arising in each patch
from all the mosquitoes biting a single human on a
single day in each patch.

Compute the VC matrix
for models that use the parameter set for the
generalized Macdonald model
}
\note{
This is called for the models \code{macdonald},
\code{GeM}, and \code{SI}
}
