% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{make_vis_kappa_random}
\alias{make_vis_kappa_random}
\title{Make random values for vis_kappa}
\usage{
make_vis_kappa_random(
  nPatches,
  options = list(),
  f_rand = stats::rbeta,
  p1 = 20,
  p2 = 1000
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{options}{is a set of options that overwrites the defaults}

\item{f_rand}{a random number generator}

\item{p1}{argument for f_rand}

\item{p2}{argument for f_rand}
}
\value{
a \link{matrix}
}
\description{
Set up a random vector
}
