% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{make_WB}
\alias{make_WB}
\title{Compute availability blood hosts of the i^th species}
\usage{
make_WB(t, pars, y = 0)
}
\arguments{
\item{t}{the time}

\item{pars}{an \code{xds} object}

\item{y}{state vector}
}
\value{
an \code{xds} object
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
