% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{make_X_indices}
\alias{make_X_indices}
\title{Add indices for human population to parameter list}
\usage{
make_X_indices(pars, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
