% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{make_X_indices.SIPd}
\alias{make_X_indices.SIPd}
\title{Add indices for human population to parameter list}
\usage{
\method{make_X_indices}{SIPd}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{make_X_indices} for the SIPd model.
}
