% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-object.R
\name{make_xds_object}
\alias{make_xds_object}
\title{Make \strong{\code{xds}} Object}
\usage{
make_xds_object(
  xds = "xde",
  frame = "full",
  dlay = "ode",
  nPatches = 1,
  membership = 1,
  residence = 1,
  HPop = 1000
)
}
\arguments{
\item{xds}{is either "xde" or "dts" for differential / difference equations}

\item{frame}{model component subset}

\item{dlay}{is either "ode" for ordinary and "dde" for delay differential equations}

\item{nPatches}{is the number of patches}

\item{membership}{is the habitat membership vector}

\item{residence}{is the strata residence vector}

\item{HPop}{is the initial human population density}
}
\value{
an \code{xds} template as a compound \link{list}
}
\description{
Create the template for a fully defined \strong{\code{xds}} object
}
\details{
This function sets up everything that is required to configure and run a basic model.

First, the function sets up some short text strings (assigned to the same \code{S3} class)
to dispatch various \strong{cases} of of various \code{S3} functions:
\itemize{
\item \emph{\code{xds}} is either "xde" for differential equations or "dts" for difference equations;
\item \emph{\code{frame}} is one of several cases:
\itemize{
\item "full" includes all three dynamical components: a human/host dynamical component, \eqn{\cal XH}; and adult mosquito dynamical component, \eqn{\cal MYZ}; and an aquatic mosquito dynamical component, and \eqn{\cal L}.
in some form (possible the trace case) (see \code{\link[=xds_setup]{xds_setup()}})
\item "mozy" is for mosquito ecology models (see \code{\link[=xds_setup_mosy]{xds_setup_mosy()}}), including models without pathogen infection dynamics in mosquitoes
\item "aquatic" is for aquatic mosquito ecology models (see \code{\link[=xds_setup_mosy]{xds_setup_mosy()}}), forced by a function describing egg laying
\item "human" is for human/host infection dynamics(see \code{\link[=xds_setup_human]{xds_setup_human()}}), forced by the infective density of adult mosquitoes, \eqn{fqZ}
\item "cohort" is for human/host cohort infection dynamics (see \code{\link[=xds_setup_cohort]{xds_setup_cohort()}}), forced by a function \code{F_eir}
}
\item \emph{\code{dlay}} is either "ode" or "dde" and it only affects dispatching for differential equations
\item \emph{\code{forcing}} is set to "static"
\item \emph{\code{compute}} dispatches various cases of \code{\link[=compute_terms]{compute_terms()}} linked to \code{frame}
}

Next, the function sets the basic \strong{structural parameters}:
\itemize{
\item \emph{\code{nVectors}} or \eqn{n_s}, the number of vector species is set to 1;
\item \emph{\code{nHosts}} or \eqn{n_i}, the number of host species is set to 1;
\item \emph{\code{nPatches}} or \eqn{n_p} is the number of patches
\item \emph{\code{nHabitats}} or \eqn{n_q} is the number of aquatic habitats is \code{length(membership)}
\item \emph{\code{nStrata}} or \eqn{n_q} is the number of human population strata is \code{length(HPop)}
}

Next, the model sets up egg laying, blood feeding, and transmission:
\itemize{
\item \strong{Egg Laying} is set up by calling \code{\link[=create_habitat_matrix]{create_habitat_matrix()}} and then \code{\link[=setup_EGG_LAYING]{setup_EGG_LAYING()}}
\item \strong{Blood Feeding} is set up by calling \code{\link[=create_residence_matrix]{create_residence_matrix()}}and then calls \code{\link[=setup_BLOOD_FEEING]{setup_BLOOD_FEEING()}}:
\item \strong{Transmission}  is set up by calling \code{\link[=setup_TRANSMISSION]{setup_TRANSMISSION()}}
}

Finally, the function sets up the default cases for other options:
\itemize{
\item \link{Exposure} is called \emph{after} \link{Transmission} to implement some model of environmental exposure, including malaria importation throught travel:
\itemize{
\item \strong{Poisson} exposure is set up by \code{\link[=setup_exposure_pois]{setup_exposure_pois()}}
\item \strong{Travel Malaria} is set up as a static model with no malaria importation by \code{\link[=setup_travel_static]{setup_travel_static()}}
}
\item \link{Transmission} calls \code{\link[=compute_local_frac]{compute_local_frac()}}. The basic static model is set up by
\code{\link[=setup_visitors_static]{setup_visitors_static()}} and the availabity of visitors is set to zero.
}
}
\note{
setup requires that \code{length(residence) == length(HPop)}
}
\seealso{
\code{\link[=xds_setup]{xds_setup()}}
}
