% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{make_XH_inits_hMoI}
\alias{make_XH_inits_hMoI}
\title{Make inits for hybrid MoI human model}
\usage{
make_XH_inits_hMoI(nStrata, options = list(), m1 = 2, m2 = 1)
}
\arguments{
\item{nStrata}{the number of population strata}

\item{options}{a \link{list} that overwrites default values}

\item{m1}{mean MoI among inapparent human infections}

\item{m2}{mean MoI among patent human infections}
}
\value{
none
}
\description{
MoI stands for Multiplicity of Infection, and refers to malarial superinfection.
}
