% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{make_XH_inits_SIS}
\alias{make_XH_inits_SIS}
\title{Make initial values for the SIS xde human model, with defaults}
\usage{
make_XH_inits_SIS(nStrata, H, options = list(), I = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial human population density}

\item{options}{a \link{list} to overwrite defaults}

\item{I}{the initial values of the parameter I}
}
\value{
a \link{list}
}
\description{
Set the initial values
}
\keyword{internal}
