% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{make_Xinits}
\alias{make_Xinits}
\title{A function to set up Xpar}
\usage{
make_Xinits(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This method dispatches on \code{pars$Xpar[[i]]}.
}
