% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{make_Xinits.SEISd}
\alias{make_Xinits.SEISd}
\title{Setup Xinits.SEISd}
\usage{
\method{make_Xinits}{SEISd}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Xinits} for the SEISd model
}
