% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_Xinits.SIP}
\alias{make_Xinits.SIP}
\title{Setup Xinits.SIP}
\usage{
\method{make_Xinits}{SIP}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Xinits} for the SIP model
}
