% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_Xinits.SIS}
\alias{make_Xinits.SIS}
\title{Setup Xinits.SIS}
\usage{
\method{make_Xinits}{SIS}(pars, H, i, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Xinits} for the SIS model
}
