% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{make_Xinits_SIS}
\alias{make_Xinits_SIS}
\title{Make initial values for the SIS xde human model, with defaults}
\usage{
make_Xinits_SIS(nStrata, H, Xopts = list(), I = 1)
}
\arguments{
\item{nStrata}{the number of strata in the model}

\item{H}{the initial human population density}

\item{Xopts}{a \link{list} to overwrite defaults}

\item{I}{the initial values of the parameter I}
}
\value{
a \link{list}
}
\description{
Make initial values for the SIS xde human model, with defaults
}
