% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{make_Xpar}
\alias{make_Xpar}
\title{Make a \eqn{cal X} Model}
\usage{
make_Xpar(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make a model
}
