% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{make_Xpar.SIP}
\alias{make_Xpar.SIP}
\title{Setup Xpar.SIP}
\usage{
\method{make_Xpar}{SIP}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Xpar} for the SIP model
}
