% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{make_Xpar.trace}
\alias{make_Xpar.trace}
\title{xde_setup Xpar.trace}
\usage{
\method{make_Xpar}{trace}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vectord
}
\description{
Implements \link{make_Xpar} for the trace model
}
