% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{make_Xpar.trivial}
\alias{make_Xpar.trivial}
\title{xde_setup Xpar.trivial}
\usage{
\method{make_Xpar}{trivial}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vectord
}
\description{
Implements \link{make_Xpar} for the trivial model
}
