% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{make_Xpar_trivial}
\alias{make_Xpar_trivial}
\title{Make parameters for trivial human model}
\usage{
make_Xpar_trivial(
  nPatches,
  Xopts,
  kappa = 0.1,
  HPop = 1,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{Xopts}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}

\item{F_season}{a function returning a seasonal signal}

\item{F_trend}{a function returning a trend}
}
\value{
a \link{list}
}
\description{
Make parameters for trivial human model
}
