% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_one}
\alias{makepar_F_one}
\title{parameters for make_function}
\usage{
makepar_F_one()
}
\value{
a function for seasonality
}
\description{
Return an object to return \link{F_one}
}
\seealso{
\link{make_function.one}
}
