% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_sharkfin}
\alias{makepar_F_sharkfin}
\title{Make Parameters for a Sharkfin Function}
\usage{
makepar_F_sharkfin(
  D = 100,
  L = 180,
  uk = 1/7,
  dk = 1/40,
  pw = 1,
  mx = 1,
  N = 1
)
}
\arguments{
\item{D}{the half-saturation day}

\item{L}{shape parameter}

\item{uk}{the rate parameter for}

\item{dk}{the length of the vector to return}

\item{pw}{shape parameter, power}

\item{mx}{a maximum value}

\item{N}{the length of the vector to return}
}
\value{
a function F_season
}
\description{
Return an object for \link{make_function.sharkfin}
}
