% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_spline}
\alias{makepar_F_spline}
\title{Make Parameters for a Spline}
\usage{
makepar_F_spline(tt, yy, X = FALSE)
}
\arguments{
\item{tt}{the nodes}

\item{yy}{the y values}

\item{X}{a switch to configure for splinef or splineX}
}
\value{
parameters to configure the \code{splinef} or \code{splineX} case of \code{make_function}
}
\description{
Return an object for \link{make_function.splinef} or \link{make_function.splineX}
}
