% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_splinef}
\alias{makepar_F_splinef}
\title{Make Parameters for a Spline}
\usage{
makepar_F_splinef(tt, yy)
}
\arguments{
\item{tt}{the nodes}

\item{yy}{the y values}
}
\value{
parameters to configure the \code{splinef} case of \code{make_function}
}
\description{
Return an object for \link{make_function.splinef}
}
