% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_sum}
\alias{makepar_F_sum}
\title{parameters for make_function}
\usage{
makepar_F_sum(opts1, opts2)
}
\arguments{
\item{opts1}{options for first function}

\item{opts2}{options for the second function}
}
\value{
a function
}
\description{
Return an object to configure
a function \link{make_function.sum}
}
