% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_type2}
\alias{makepar_F_type2}
\title{parameters for make_function}
\usage{
makepar_F_type2(shift = 30, A = 1.8, B = 5, N = 1)
}
\arguments{
\item{shift}{a shift to set F(0)>0}

\item{A}{asymptote}

\item{B}{slope}

\item{N}{the length of the vector to return}
}
\value{
a type 2 functional response
}
\description{
Return an object to configure
a function \link{make_function.type2}
}
\seealso{
\link{make_function.type2}
}
