% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_val}
\alias{makepar_F_val}
\title{parameters for make_function}
\usage{
makepar_F_val(val)
}
\arguments{
\item{val}{the value to return}
}
\value{
a function for seasonality
}
\description{
Return an object to return a value
}
\seealso{
\link{make_function.val}
}
