% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makepar_F_zero}
\alias{makepar_F_zero}
\title{parameters for make_function}
\usage{
makepar_F_zero()
}
\value{
a function to return
}
\description{
Return an object to configure
a function \link{make_function.zero}
}
\seealso{
\link{make_function.zero}
}
