% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_health.R
\name{MassHealth.none}
\alias{MassHealth.none}
\title{Methods for distributing interventions during mass_healthal visits}
\usage{
\method{MassHealth}{none}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{MassHealth} for the none model (do nothing)
}
