% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_health.R
\name{MassHealth}
\alias{MassHealth}
\title{Methods for mass distributing health interventions}
\usage{
MassHealth(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$mass_health}.
}
