% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{MassMedical.no_control}
\alias{MassMedical.no_control}
\title{Methods for mass medical}
\usage{
\method{MassMedical}{no_control}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{MassMedical} for the no_control model (do nothing)
}
