% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{MBionomics.basicM_xde}
\alias{MBionomics.basicM_xde}
\title{Set bionomic parameters to baseline}
\usage{
\method{MBionomics}{basicM_xde}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{MBionomics} for the basicM xde model
}
