% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{MBionomics.GeRM}
\alias{MBionomics.GeRM}
\title{Set mosquito bionomics to baseline}
\usage{
\method{MBionomics}{GeRM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the model as a \link{list}
}
\description{
Implements \link{MBionomics} for models with no forcing on the baseline
}
