% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{MBionomics.macdonald}
\alias{MBionomics.macdonald}
\title{macdonald-style adult mosquito bionomics}
\usage{
\method{MBionomics}{macdonald}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Reset the effect sizes for static models

When modules are added to compute effect sizes
from baseline parameters, those functions store
an effect size. The total effect size is the
product of the effect sizes for each intervention.
Since coverage could be changing dynamically, these
must be reset each time the derivatives are computed.
}
