% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{MBionomics}
\alias{MBionomics}
\title{Adult mosquito bionomics}
\usage{
MBionomics(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the adult mosquito bionomic parameter values
}
\keyword{internal}
