% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{MBionomics.SEI}
\alias{MBionomics.SEI}
\title{Set mosquito bionomics to baseline}
\usage{
\method{MBionomics}{SEI}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
the model as a \link{list}
}
\description{
Implements \link{MBionomics} for models with no forcing on the baseline
}
