% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{MBionomics.static}
\alias{MBionomics.static}
\title{Set mosquito bionomics to baseline}
\usage{
\method{MBionomics}{static}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the model as a \link{list}
}
\description{
Implements \link{MBionomics} for models with no forcing on the baseline
}
