% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{MBionomics.trace}
\alias{MBionomics.trace}
\title{Compute bloodfeeding and mortality rates}
\usage{
\method{MBionomics}{trace}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{MBionomics} for the trace model.
}
