% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{MBionomics.trivial}
\alias{MBionomics.trivial}
\title{Bionomics for \code{trivial} (\strong{MY})}
\usage{
\method{MBionomics}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{MBionomics} for the \code{trivial} module
}
\keyword{internal}
