% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{MEffectSizes.basicM}
\alias{MEffectSizes.basicM}
\title{Apply vector control effect sizes}
\usage{
\method{MEffectSizes}{basicM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{MEffectSizes} for models with no exogenous forcing
}
\keyword{internal}
