% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{ModifiedBionomics}
\alias{ModifiedBionomics}
\title{Set bionomic parameter rates relative to baseline}
\usage{
ModifiedBionomics(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list}
}
\description{
This calls Mbionomics and Lbionmics for each species. This function
resets bionomic parameters to their pre-control baseline value, which can later be
modified by vector control. In some models, the pre-control baseline is computed in
here as a function of resource availability.
}
