% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_vector_X.R
\name{modify_vector_X.list}
\alias{modify_vector_X.list}
\title{Replace Values in a List}
\usage{
\method{modify_vector_X}{list}(X, ix, vector)
}
\arguments{
\item{X}{the new values}

\item{ix}{indices to be replaced}

\item{vector}{the vector to modify}
}
\value{
the modified vector
}
\description{
If \code{ix} is a list,
replace the values in the indices in the
\eqn{i^{th}} element of the list
with the \eqn{i^{th}} element of \code{X}
}
\examples{
a = list()
a[[1]] = c(1,7)
a[[2]] = c(3,9)
modify_vector_X(c(6,4), a, 1:10)

}
