% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_vector_X.R
\name{modify_vector_X.NULL}
\alias{modify_vector_X.NULL}
\title{Replace Values in a List}
\usage{
\method{modify_vector_X}{`NULL`}(X, ix, vector)
}
\arguments{
\item{X}{the new values}

\item{ix}{indices to be replaced}

\item{vector}{the vector to modify}
}
\value{
the modified vector
}
\description{
If \code{ix} is NULL, return \code{X}
}
\examples{

modify_vector_X(6:15, c(), 1:10)

}
