% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_vector_X.R
\name{modify_vector_X.numeric}
\alias{modify_vector_X.numeric}
\title{Replace Values in a List}
\usage{
\method{modify_vector_X}{numeric}(X, ix, vector)
}
\arguments{
\item{X}{the new values}

\item{ix}{indices to be replaced}

\item{vector}{the vector to modify}
}
\value{
the modified vector
}
\description{
If \code{ix} is numeric
}
\examples{

modify_vector_X(6, c(1,7), 1:10)

}
