% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_vector_X.R
\name{modify_vector_X}
\alias{modify_vector_X}
\title{Modify some values in a vector}
\usage{
modify_vector_X(X, ix, vector)
}
\arguments{
\item{X}{the new values}

\item{ix}{indices to be replaced}

\item{vector}{the vector to modify}
}
\value{
the modified vector
}
\description{
Replace some values in a vector
and return the modified vector
}
