% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{mosquito_dispersal}
\alias{mosquito_dispersal}
\title{Mosquito Dispersal}
\description{
Mosquito Dispersal
}
\section{Mosquito Dispersal Matrix, \eqn{K}}{


In adult mosquito modules, mosquito dispersal is described by
a square matrix \eqn{K} with \code{nPatches} \eqn{(=n)} rows and columns. By convention,
all \eqn{K} matrices have the form:
\deqn{
 K = \left[ \begin{array}{ccccc}
 -1 & k_{1,2} & k_{1,3} & \cdots & k_{1, n} \\
 k_{2,1} & -1 & k_{2,3} & \cdots & k_{2, n} \\
 k_{3,1} & k_{3,2} & -1 & \cdots & k_{3, n} \\
 \vdots & \vdots & \vdots & \ddots  & \vdots \\
 k_{n,1} & k_{n,2} & k_{n,3} & \cdots & -1 \\
 \end{array} \right]
}
We interpret \eqn{K} as the destinations of
emigrating mosquitoes that survive and stay in the system.
The values of the elements \eqn{k_{i,j}} are thus
constrained such that there is no net dispersal loss from the system: \eqn{\forall i,}
\deqn{\sum_j k_{i,j} = 1.}
The columns of \eqn{K} sum up to zero.

Emigration-related loss (including mortality and emigration from the spatial domain)
is modeled with another parameter (see \link{mosquito_demography}).
}

\section{Basic Setup}{

\describe{
\item{\code{K_matrix}}{a matrix or a named list: options that set up \eqn{K}}
}

There are several ways to configure  \eqn{K} during basic setup.

By \strong{default:} \code{K_matrix} is an empty list. When the \code{MY_obj} is set up, the \eqn{K}
matrix is set to \code{diag(g)}, a model with no dispersal. That is not changed unless one of the
following holds:
\itemize{
\item \code{is.matrix(K_matrix)}: if the user to passes a proper \eqn{K} matrix, then it is used as-is
\item \code{with(Kmatrix, exists("Kname"))}: If a string in the named list passed as \code{K_matrix} is called \code{Kname}, then basic setup calls \link{setup_K_matrix} with \code{options=K_matrix}
}
}

\seealso{
\link{setup_K_matrix}
}
