% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{MosquitoBionomics}
\alias{MosquitoBionomics}
\title{Compute mosquito bionomic parameters}
\usage{
MosquitoBionomics(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list}
}
\description{
This calls MBionomics and LBionomics for each species.

This function computes the bionomic parameters for each vector species,
before any vector control effect sizes are applied. In some models, these
parameters are computed as a function of resource availability.
}
\keyword{internal}
