% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-help.R
\name{MY_change}
\alias{MY_change}
\title{Change Functions for the MY Component}
\description{
write me.

\itemize{
\item \code{\link{change_MY_inits}} -- change initial values
\item \code{\link{change_MY_pars}} -- change parameters
}
}
