% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-help.R
\name{MY_get}
\alias{MY_get}
\title{Get Functions for the MY Component}
\description{
write me.

\itemize{
\item \code{\link{get_MY_vars}} -- get state variables from \eqn{y}
\item \code{\link{get_MY_ix}} -- get variable indices
\item \code{\link{get_MY_inits}} -- get initial values
\item \code{\link{get_MY_pars}} -- get parameters
\item \code{\link{get_MY_orbits}} -- get saved orbits
\item \code{\link{get_f}} -- get blood feeding rate
\item \code{\link{get_q}} -- get human fraction
\item \code{\link{get_g}} -- get mosquito mortality rate
\item \code{\link{get_sigma}} -- get emigration rate
\item \code{\link{get_M}} -- get mosquito population density
\item \code{\link{get_Z}} -- get infective mosquito density
}
}
