% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-modules.R
\name{MY_module_list}
\alias{MY_module_list}
\title{MY Modules}
\description{
A List of Available Modules for the \strong{MY}-Component from \strong{\code{ramp.xds}} and \strong{\code{ramp.library.}}
}
\section{In \strong{\code{ramp.xds}}}{

\itemize{
\item \code{trivial} - the trivial module: see \link{trivial_MY}
\item \code{SI} - an \emph{SI} compartmental model: see \link{SI}
\item \code{macdonald} - an implementation Macdonald's model: see \link{macdonald}
\item \code{GeRM} - A Generalized Ross-Macdonald model: see \link{GeRM}
\item \code{basicM} - A Basic Model for mosquito ecology: see \link{basicM}
\item \code{RMdts} - A Discrete Time Ross-Macdonald model: see \link{RMdts}
}
}

\section{In \strong{\code{ramp.library}}}{


\code{ramp.library} \href{https://github.com/dd-harp/ramp.library}{source code} & \href{https://github.com/dd-harp/ramp.library}{github.io}
\itemize{
\item \code{SEI} - A compartmental model
\item \code{BQ} - A behavioral state model with blood feeding, egg laying, and infection dynamics
\item \code{BQm} - A behavioral state model with blood feeding & egg laying (\strong{M} only)
\item \code{BQRS} - A behavioral state model with blood feeding, egg laying, resting, sugar feeding and infection dynamics
}
}

