% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-help.R
\name{MY_setup}
\alias{MY_setup}
\title{Setup Functions for the MY Component}
\description{
write me.

\itemize{
\item \code{\link{setup_MY_obj}} -- set up the MY module object
\item \code{\link{setup_MY_ix}} -- set up variable indices
\item \code{\link{setup_MY_inits}} -- set up initial values
}
}
