% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{MYZ_rates2probs}
\alias{MYZ_rates2probs}
\title{Compute probabilities from rates}
\usage{
MYZ_rates2probs(MYZpar, runtime)
}
\arguments{
\item{MYZpar}{the \code{MYZ} model object}

\item{runtime}{the model \code{runtime} parameters}
}
\value{
a \link{list} with baseline values
}
\description{
This method dispatches on \code{MYZname}.
}
