% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-bionomics.R
\name{MYZ_rates2probs_RM.modified}
\alias{MYZ_rates2probs_RM.modified}
\title{Compute probabilities from rates}
\usage{
\method{MYZ_rates2probs_RM}{modified}(MYZpar, runtime)
}
\arguments{
\item{MYZpar}{the \code{MYZ} model object}

\item{runtime}{the model \code{runtime} parameters}
}
\value{
a \link{list} with baseline values
}
\description{
This method dispatches on \code{MYZname}.
}
