% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-bionomics.R
\name{MYZ_rates2probs_RM}
\alias{MYZ_rates2probs_RM}
\title{Compute probabilities from rates}
\usage{
MYZ_rates2probs_RM(MYZpar, runtime)
}
\arguments{
\item{MYZpar}{an adult mosquito \code{xds} model object}

\item{runtime}{the model \code{runtime} parameters}
}
\value{
a \link{list} with baseline values
}
\description{
This method dispatches on \code{MYZname}.
}
