% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_to_dde.R
\name{ode_to_dde.dts}
\alias{ode_to_dde.dts}
\title{Compute as DDE}
\usage{
\method{ode_to_dde}{dts}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} modelobject}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If \code{class(xds_obj$xde) == "dts"} don't change anything
}
