% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_to_dde.R
\name{ode_to_dde.ode}
\alias{ode_to_dde.ode}
\title{Compute as DDE}
\usage{
\method{ode_to_dde}{ode}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
If \code{class(xds_obj$xde) == "ode"} change it to \code{dde}
}
