% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_to_dde.R
\name{ode_to_dde}
\alias{ode_to_dde}
\title{Compute as DDE}
\usage{
ode_to_dde(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Setup a differential equation model
to be solved using \code{deSolve:dede}
}
\details{
For differential equations, the value of \code{xds_obj$xde}
is set to \code{ode} by default. This utility gets called by delay
differential equation modules to change
\code{xds_obj$xde} from \code{ode} to \code{dde}
}
