% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{One_tV}
\alias{One_tV}
\title{The trivial function}
\usage{
One_tV(t, V = list())
}
\arguments{
\item{t}{an arbitrary input}

\item{V}{an arbitrary input}
}
\value{
a vector of ones of length x
}
\description{
A function that returns 0
}
