% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{other_blood_hosts}
\alias{other_blood_hosts}
\title{Other blood hosts}
\usage{
other_blood_hosts(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{S3} definition for the
function that computes other_blood_hosts
}
\keyword{internal}
