% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{Other_State_Variables}
\alias{Other_State_Variables}
\title{Other State Variables}
\description{
For extensibility, \strong{\code{ramp.xds}} is capable of setting up and computing \emph{other state variables,}
or \emph{dependent variables} that are not part of one of the core dynamical components.
}
\seealso{
\link{variables}
}
