% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{ovitraps.static}
\alias{ovitraps.static}
\title{Compute availability of ovitraps}
\usage{
\method{ovitraps}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update nothing
}
\keyword{internal}
