% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-own-func.R
\name{OwnBedNets.func}
\alias{OwnBedNets.func}
\title{Set no own_bednets}
\usage{
\method{OwnBedNets}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for own_bednets
}
