% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-own-interface.R
\name{OwnBedNets}
\alias{OwnBedNets}
\title{Set the own_bednets}
\usage{
OwnBedNets(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
own_bednets
}
