% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_bionomics}
\alias{parse_bionomics}
\title{Pull bionomic parameters}
\usage{
parse_bionomics(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Pull the stored values of the
baseline and modified values of parameters.
Note that the total effect size of control
is the ratio.
}
