% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{parse_deout_L.basicL_xde}
\alias{parse_deout_L.basicL_xde}
\title{Parse the variable names for the basicL_xde model}
\usage{
\method{parse_deout_L}{basicL_xde}(deout, pars, s)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_deout_L} for basicL_xde competition model.
}
