% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{parse_deout_MYZ.basicM_xde}
\alias{parse_deout_MYZ.basicM_xde}
\title{Parse the output of deSolve and return variables for the basicM_xde model}
\usage{
\method{parse_deout_MYZ}{basicM_xde}(deout, pars, s)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_deout_MYZ} for the basicM_xde model.
}
