% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI-xde.R
\name{parse_deout_X.hMoI_xde}
\alias{parse_deout_X.hMoI_xde}
\title{Parse the output of deSolve and return variables for the hMoI_xde model}
\usage{
\method{parse_deout_X}{hMoI_xde}(deout, pars, i)
}
\arguments{
\item{deout}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_deout_X} for the hMoI_xde model
}
