% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{parse_L_orbits.basicL}
\alias{parse_L_orbits.basicL}
\title{parse \strong{L} Component Variables for \code{basicL}}
\usage{
\method{parse_L_orbits}{basicL}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
The function returns the column representing
the variable \eqn{L} from a matrix where each row is a state variable.
The variable is returned as a named list.
}
