% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{parse_L_orbits}
\alias{parse_L_orbits}
\title{parse \strong{L} Component Outputs}
\usage{
parse_L_orbits(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
After solving a dynamical system, parse the outputs and
return the variables by name in a list.
}
\note{
This method dispatches on the class of \code{xds_obj$L_obj[[s]]}
}
