% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{parse_L_orbits.trivial}
\alias{parse_L_orbits.trivial}
\title{Parse for \code{trivial} (\strong{L})}
\usage{
\method{parse_L_orbits}{trivial}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an empty \link{list}
}
\description{
Returns an empty list
}
\keyword{internal}
