% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{parse_Lorbits.basicL}
\alias{parse_Lorbits.basicL}
\title{Parse \strong{L} Component Variables for \code{basicL}}
\usage{
\method{parse_Lorbits}{basicL}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
The function returns the column representing
the variable \eqn{L} from a matrix where each row is a state variable.
The variable is returned as a named list.
}
