% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{parse_Lorbits}
\alias{parse_Lorbits}
\title{Parse \strong{L} Component Outputs}
\usage{
parse_Lorbits(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
After solving a dynamical system, parse the outputs and
return the variables by name in a list.
}
\note{
This method dispatches on the class of \code{pars$Lpar[[s]]}
}
