% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{parse_Lorbits.trace}
\alias{parse_Lorbits.trace}
\title{Parse the variable names for the trace model}
\usage{
\method{parse_Lorbits}{trace}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_Lorbits} for the trace model.
}
