% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{parse_Lorbits.trivial}
\alias{parse_Lorbits.trivial}
\title{Parse \strong{L} Component Variables for \code{basicL}}
\usage{
\method{parse_Lorbits}{trivial}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
none
}
\description{
Return a numeric vector of length 0
}
