% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{parse_MY_orbits.GeRM}
\alias{parse_MY_orbits.GeRM}
\title{parse the output of deSolve and return variables for the GeRM model}
\usage{
\method{parse_MY_orbits}{GeRM}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_MY_orbits} for the GeRM model
}
