% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{parse_MY_orbits}
\alias{parse_MY_orbits}
\title{parse the outputs and return the variables by name in a list}
\usage{
parse_MY_orbits(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj}.
It computes the variables by name and returns a named list.
}
