% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{parse_MY_orbits.RMdts}
\alias{parse_MY_orbits.RMdts}
\title{Parse outputs for \code{RMdts} (\strong{MY})}
\usage{
\method{parse_MY_orbits}{RMdts}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_MY_orbits} for the RMdts model
}
\keyword{internal}
