% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{parse_MY_orbits.SI}
\alias{parse_MY_orbits.SI}
\title{parse the output of deSolve and return variables for the \code{SI} model}
\usage{
\method{parse_MY_orbits}{SI}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_MY_orbits} for the \code{SI} model
}
