% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{parse_MY_orbits.trivial}
\alias{parse_MY_orbits.trivial}
\title{Parse outputs for \code{trivial} (\strong{MY})}
\usage{
\method{parse_MY_orbits}{trivial}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an empty \link{list}
}
\description{
Return an empty list
}
\keyword{internal}
