% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_MY_terms}
\alias{parse_MY_terms}
\title{Pull MY Terms}
\usage{
parse_MY_terms(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Pull the stored values of the
Lambda, G, fqZ, and kappa
}
