% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{parse_MYZorbits.basicM}
\alias{parse_MYZorbits.basicM}
\title{Parse outputs for basicM}
\usage{
\method{parse_MYZorbits}{basicM}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_MYZorbits} for the basicM model.
}
