% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{parse_MYZorbits.MY}
\alias{parse_MYZorbits.MY}
\title{Parse the output of deSolve and return variables for the \code{MY} model}
\usage{
\method{parse_MYZorbits}{MY}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_MYZorbits} for the \code{MY} model
}
