% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{parse_MYZorbits.RM_dts}
\alias{parse_MYZorbits.RM_dts}
\title{Parse the output of deSolve and return variables for the RM_dts model}
\usage{
\method{parse_MYZorbits}{RM_dts}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_MYZorbits} for the RM_dts model
}
