% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{parse_MYZorbits.trace}
\alias{parse_MYZorbits.trace}
\title{Parse the output of deSolve and return variables for the trace model}
\usage{
\method{parse_MYZorbits}{trace}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_MYZorbits} for trace
}
