% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{parse_MYZorbits.trivial}
\alias{parse_MYZorbits.trivial}
\title{Parse the output of deSolve and return variables for the trivial model}
\usage{
\method{parse_MYZorbits}{trivial}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_MYZorbits} for trivial
}
