% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_orbits}
\alias{parse_orbits}
\title{Parse outputs after solving}
\usage{
parse_orbits(outputs, xds_obj)
}
\arguments{
\item{outputs}{a \link{matrix} of _orbits returned by deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a named \link{list}
}
\description{
Parse outputs after solving
}
\keyword{internal}
