% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_outputs.mosy}
\alias{parse_outputs.mosy}
\title{Make Outputs}
\usage{
\method{parse_outputs}{mosy}(xds_obj, de_vars, tm)
}
\arguments{
\item{xds_obj}{an \code{xds} object}

\item{de_vars}{the solutions \link{matrix}}

\item{tm}{the time}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make Outputs
}
