% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{parse_outputs_L.basicL_xde}
\alias{parse_outputs_L.basicL_xde}
\title{Parse the variable names for the basicL_xde model}
\usage{
\method{parse_outputs_L}{basicL_xde}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_outputs_L} for basicL_xde competition model.
}
