% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{parse_outputs_L.stages_dts}
\alias{parse_outputs_L.stages_dts}
\title{Parse the variable names for the stages_dts model}
\usage{
\method{parse_outputs_L}{stages_dts}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_outputs_L} for stages_dts competition model.
}
