% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{parse_outputs_L.trace}
\alias{parse_outputs_L.trace}
\title{Parse the variable names for the trace model}
\usage{
\method{parse_outputs_L}{trace}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_outputs_L} for the trace model.
}
