% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{parse_outputs_MYZ.basicM}
\alias{parse_outputs_MYZ.basicM}
\title{Parse the output of deSolve and return variables for the basicM model}
\usage{
\method{parse_outputs_MYZ}{basicM}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_outputs_MYZ} for the basicM model.
}
