% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM-xde.R
\name{parse_outputs_MYZ.GeRM_xde}
\alias{parse_outputs_MYZ.GeRM_xde}
\title{Parse the output of deSolve and return variables for the GeRM_xde model}
\usage{
\method{parse_outputs_MYZ}{GeRM_xde}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_outputs_MYZ} for the GeRM_xde model
}
