% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{parse_outputs_MYZ.mz}
\alias{parse_outputs_MYZ.mz}
\title{Parse the output of deSolve and return variables for the mz model}
\usage{
\method{parse_outputs_MYZ}{mz}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_outputs_MYZ} for the mz model
}
