% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{parse_outputs_MYZ}
\alias{parse_outputs_MYZ}
\title{Parse the outputs and return the variables by name in a list}
\usage{
parse_outputs_MYZ(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
It computes the variables by name and returns a named list.
}
