% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outputs.R
\name{parse_outputs_vec}
\alias{parse_outputs_vec}
\title{Parse the output of an object returned by deSolve}
\usage{
parse_outputs_vec(vec, pars)
}
\arguments{
\item{vec}{a \link{vector} with the variables, as returned by rootsolve}

\item{pars}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
Parse the output of an object returned by deSolve
}
