% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{parse_outputs_X}
\alias{parse_outputs_X}
\title{Parse the output of deSolve and return the variables by name in a list}
\usage{
parse_outputs_X(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{i}{the host species index}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}. Adds the variables
from the X model to a list and returns it
}
