% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{parse_outputs_X.SEIS}
\alias{parse_outputs_X.SEIS}
\title{Parse the output of deSolve and return variables for the SEIS model}
\usage{
\method{parse_outputs_X}{SEIS}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_outputs_X} for the SEIS model
}
