% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-xde.R
\name{parse_outputs_X.SIP_xde}
\alias{parse_outputs_X.SIP_xde}
\title{Parse the output of deSolve and return variables for the SIP_xde model}
\usage{
\method{parse_outputs_X}{SIP_xde}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{a \link{list} that defines a model}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_outputs_X} for the SIP_xde model
}
