% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{parse_XH_orbits.SIS}
\alias{parse_XH_orbits.SIS}
\title{Parse outputs for \code{SIS} (\strong{XH})}
\usage{
\method{parse_XH_orbits}{SIS}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Implements \link{parse_XH_orbits} for the SIS model
}
\keyword{internal}
