% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_XH_terms}
\alias{parse_XH_terms}
\title{Pull XH Terms}
\usage{
parse_XH_terms(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Pull the stored values of the
EIR and the FoI and compute NI
}
