% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{parse_Xorbits.SEIS}
\alias{parse_Xorbits.SEIS}
\title{Parse the output of deSolve and return variables for the SEIS model}
\usage{
\method{parse_Xorbits}{SEIS}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_Xorbits} for the SEIS model
}
