% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{parse_Xorbits.trace}
\alias{parse_Xorbits.trace}
\title{Parse the output of deSolve and return variables for the trace model}
\usage{
\method{parse_Xorbits}{trace}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_Xorbits} for the trace model
}
