% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{parse_Xorbits.trivial}
\alias{parse_Xorbits.trivial}
\title{Parse the output of deSolve and return variables for the trivial model}
\usage{
\method{parse_Xorbits}{trivial}(outputs, pars, i)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_Xorbits} for the trivial model
}
