% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outputs.R
\name{parse_y}
\alias{parse_y}
\title{Parse the output of an object returned by deSolve}
\usage{
parse_y(y, pars)
}
\arguments{
\item{y}{a \link{vector} with the variables}

\item{pars}{an \code{xds} object}
}
\value{
varslist a \link{list}
}
\description{
Parse the output of an object returned by deSolve
}
