% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_dynamics.R
\name{patch_dynamics}
\alias{patch_dynamics}
\title{Patch Dynamics}
\description{
Spatial dynamics in \strong{\code{ramp.xds}} are patch-based, also known as
\emph{metapopulation dynamics.}
A \emph{patch} is the primary spatial unit for modeling transmission, and
the number of patches is a core structural parameter:

\describe{
\item{\code{nPatches}}{the number of patches }
}

In patch-based models, a patch is a
spatial unit that structures
adult mosquito populations.
A patch could have any number of
aquatic habitats (see \link{aquatic_habitats}), and mosquitoes disperse among patches as
they search for blood hosts, aquatic habitats to lay eggs, and other resources
they need (see \link{mosquito_dispersal}).

Patches are used to model blood feeding (see \link{blood_feeding}), so there is a corresponding human model
for
time spent (see \link{time_spent}).
}
