% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{port}
\alias{port}
\title{Ports}
\description{
In \strong{\verb{ramp.xds,}} a \emph{port} is any parameter or term that has
a flexible implementation, usually by calling an \code{S3} function.
Ports are an important design feature that make the software
extensible through the addition of new \code{S3} methods.

By policy, the default static values assigned to ports at basic
setup are the "simplest" method. Other methods in \strong{\code{ramp.xds}}
are easily configured through a \verb{setup_*} function.

Examples:
\itemize{
\item \link{mosquito_bionomics} parameters in some \strong{MY} modules are handled as ports
}
}
