% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.xds_obj}
\alias{print.xds_obj}
\title{Print the \strong{\code{xds}} model object}
\usage{
\method{print}{xds_obj}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a description of the model
}
\description{
Print the \strong{\code{xds}} model object
}
