% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.xds_obj}
\alias{print.xds_obj}
\title{Print for xds_objects}
\usage{
\method{print}{xds_obj}(x)
}
\arguments{
\item{x}{an xds_object}
}
\value{
a description of the model
}
\description{
Print for xds_objects
}
