% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{put_Lvars}
\alias{put_Lvars}
\title{Put Lvars in place of the L variables in y}
\usage{
put_Lvars(Lvars, y, pars, s)
}
\arguments{
\item{Lvars}{the L variables to insert into y}

\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the class of \code{pars$Lpar[[s]]}.
}
