% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{put_Lvars.trivial}
\alias{put_Lvars.trivial}
\title{Return the variables as a list}
\usage{
\method{put_Lvars}{trivial}(Lvars, y, pars, s)
}
\arguments{
\item{Lvars}{the L variables to insert into y}

\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}
}
