% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{put_MYZvars}
\alias{put_MYZvars}
\title{Put MYZvars in place of the MYZ variables in y}
\usage{
put_MYZvars(MYZvars, y, pars, s)
}
\arguments{
\item{MYZvars}{the variables}

\item{y}{the variables}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar[[s]]}.
}
