% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{put_Xvars}
\alias{put_Xvars}
\title{Put Xvars in place of the X variables in y}
\usage{
put_Xvars(Xvars, y, pars, i)
}
\arguments{
\item{Xvars}{the X variables to insert into y}

\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[s]]}.
}
