% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{put_Xvars.SEIS}
\alias{put_Xvars.SEIS}
\title{Return the SEIS model variables as a list, returned from Update_Xt.SISd}
\usage{
\method{put_Xvars}{SEIS}(Xvars, y, pars, i)
}
\arguments{
\item{Xvars}{the X variables to insert into y}

\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}
}
