% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{put_Xvars.SIS}
\alias{put_Xvars.SIS}
\title{Return the SIS model variables as a list, returned from Update_Xt.SIS}
\usage{
\method{put_Xvars}{SIS}(Xvars, y, pars, i)
}
\arguments{
\item{Xvars}{the X variables to insert into y}

\item{y}{the variables}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Xpar}
}
