% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Rainfall.no_forcing}
\alias{Rainfall.no_forcing}
\title{Methods for exogenous variables describing rainfall}
\usage{
\method{Rainfall}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Rainfall} for the no_forcing model (no variables)
}
