% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_forcing_functions.R
\name{rebuild_forcing_functions.XH}
\alias{rebuild_forcing_functions.XH}
\title{Rebuild Forcing Functions}
\usage{
\method{rebuild_forcing_functions}{XH}(xds_obj, ix)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Rebuild forcing functions on the X object using \code{make_function}:
\itemize{
\item \code{F_season} is made from \code{season_par}
\item \code{F_trend} is made from \code{trend_par}
\item \code{F_shock} is made from \code{shock_par}
}
}
\keyword{internal}
