% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{Relative_Humidity.no_forcing}
\alias{Relative_Humidity.no_forcing}
\title{Methods for exogenous variables describing relative humidity}
\usage{
\method{Relative_Humidity}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Relative_Humidity} for the no_forcing model (no variables)
}
