% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_state.R
\name{reset_state_i}
\alias{reset_state_i}
\title{Compute other variables at time t}
\usage{
reset_state_i(i, tm, y_matrix, pars)
}
\arguments{
\item{i}{current simulation time}

\item{tm}{current simulation time}

\item{y_matrix}{state vector}

\item{pars}{a \link{list}}
}
\value{
\strong{pars} a \link{list}
}
\description{
Compute everything but the derivatives for the generalized
spatial differential equation model
}
