% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources.none}
\alias{Resources.none}
\title{Resources (none)}
\usage{
\method{Resources}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The default no-op for Resources. All resource parameters
retain their static defaults (zero-initialized by \link{setup_blood_feeding}
and \link{setup_egg_laying}).
}
\keyword{internal}
