% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Resources}
\alias{Resources}
\title{Resources}
\usage{
Resources(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Computes availability of resources potentially
affecting adult mosquito bionomics: blood hosts,
aquatic habitats, and sugar. There are ports to
compute variables that are part of the blood feeding,
egg laying, or exposure interfaces.
\itemize{
\item \link{BloodHosts} availability of alternative blood hosts
\item \link{HabitatDynamics} to modify habitat search weights
\item \link{Travel} to compute variables associated with exposure while traveling
\item \link{Visitors} to compute parasite/pathogen blood feeding on visitors
\item \link{Traps} to compute availability of oviposition traps
\item \link{Sugar} availability of sugar
}
}
