% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts-runtime.R
\name{runt}
\alias{runt}
\title{A runtime switch function for mismatched dynamical component runtimes}
\usage{
runt(t, Dday, xday)
}
\arguments{
\item{t}{current simulation time}

\item{Dday}{the runtime time step for the simulation}

\item{xday}{the runtime time step for a component}
}
\value{
\link{logical} TRUE or FALSE
}
\description{
Determine whether to update the variables at time t
}
