% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveXDS}
\alias{saveXDS}
\title{saveRDS for \code{xds} Objects}
\usage{
saveXDS(xds_obj, filename)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{filename}{the file name}
}
\value{
invisible()
}
\description{
Removes forcing functions (to reduce the file size)
and stores the \code{xds} object using \code{saveRDS}
}
\note{
Forcing functions (\emph{e.g.} F_trend) are stored
on the \code{xds} object in a form that
takes up enormous space. This removes the functions
before saving.

The function \code{readRDS} in \code{ramp.func}
provides one method for saving the parameter
sets for trace functions, and rebuilds the
functions
}
\seealso{
\code{ramp.func}
}
