% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveXDS.XH}
\alias{saveXDS.XH}
\title{saveRDS for \code{xds} Objects}
\usage{
\method{saveXDS}{XH}(xds_obj, filename)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{filename}{the file name}
}
\value{
invisible()
}
\description{
Removes forcing functions from the \code{EIR_obj} (to reduce the size)
and stores the \code{xds} object using \code{saveRDS}
}
\keyword{internal}
