% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{School.none}
\alias{School.none}
\title{Methods for distributing interventions during schoolal visits}
\usage{
\method{School}{none}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{School} for the none model (do nothing)
}
