% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{School}
\alias{School}
\title{Methods for distributing interventions during schoolal visits}
\usage{
School(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$school}.
}
