% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{search_weights}
\alias{search_weights}
\title{Search Weights}
\description{
Mosquitoes search for resources, including vertebrate animals for
blood feeding and aquatic habitats for egg laying.
In developing a modular design, a
rigorous interface was needed to guarantee mathematical
consistency for blood feeding
and parasite transmission, and for adult mosquito ecology
and behaviors around egg laying.
Models for \emph{heterogeneous biting} among host strata and
\emph{heterogeneous egg laying} among habitats are implemented through \emph{search weights.}

A search
weight is a scalar value assigned to each human population stratum
or habitat that is used to compute \emph{availability}
to mosquitoes and relative biting rates or relative laying rates.
The value of these search weights are configurable during basic setup.
}
\section{Basic Setup Options}{


\describe{
\item{\code{searchB}}{a vector of search weights for blood feeding: \code{length(searchB)=nStrata}}
\item{\code{searchQ}}{a vector of search weights for egg laying: \code{length(searchQ)=nHabitats}}
}

By default, these search values are all set to 1. The
}

\section{Blood Search Weights}{


Blood feeding is an interaction among humans and mosquitoes:
in this model, humans are spending time, and
mosquitoes are searching. To model exposure, each human (or host)
population stratum is assigned a \emph{search weight,} a
number used to weigh \emph{time spent} and get a measure of
\emph{availability.}
Search weights for blood feeding thus play a key role in blood feeding (see \link{blood_feeding})
}

\section{Habitat Search Weights}{

where
}

\seealso{
\link{blood_feeding}, \link{transmission}, and \link{egg_laying}
}
