% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{set_Dday}
\alias{set_Dday}
\title{Set up a model for dts_diffeqn}
\usage{
set_Dday(Xday, MYZday, Lday, Lname = "trace")
}
\arguments{
\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{Lname}{is the S3 class of the L model}
}
\value{
a \link{list}
}
\description{
Set up a model for dts_diffeqn
}
