% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{set_eir.cohort}
\alias{set_eir.cohort}
\title{Reset the mean daily EIR}
\usage{
\method{set_eir}{cohort}(eir, pars)
}
\arguments{
\item{eir}{the mean eir}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For cohort models, reset the EIR
}
