% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{set_eir}
\alias{set_eir}
\title{Reset the mean daily EIR}
\usage{
set_eir(eir, pars)
}
\arguments{
\item{eir}{the mean eir}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For cohort models, reset the EIR
}
