% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{set_habitat_wts_static}
\alias{set_habitat_wts_static}
\title{Set static habitat search weights}
\usage{
set_habitat_wts_static(pars, searchQ = 1, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \code{xds} object}

\item{searchQ}{the habitat search weights}

\item{s}{the vector species index}

\item{Lopts}{a \link{list} of options to override defaults}
}
\value{
the modified \code{xds} object
}
\description{
Set the search weights, \eqn{\omega}, for a set of aquatic habitats
}
