% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{set_Linits.basicL}
\alias{set_Linits.basicL}
\title{Set the Initial Values for \code{basicL} (\strong{L} Component)}
\usage{
\method{set_Linits}{basicL}(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Initial values of the variable \eqn{L} are reset if they are
passed as a named component of \code{Lopts}
}
