% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{set_Linits}
\alias{set_Linits}
\title{Return the parameters as a list}
\usage{
set_Linits(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}.
}
