% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{set_Linits.trivial}
\alias{set_Linits.trivial}
\title{Set the Initial Values for \code{trivial} (\strong{L} Component)}
\usage{
\method{set_Linits}{trivial}(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns the unmodified \strong{\code{xds}} object
}
