% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{set_Lpars.basicL}
\alias{set_Lpars.basicL}
\title{Set \strong{L} Component parameters for \code{basicL}}
\usage{
\method{set_Lpars}{basicL}(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of \strong{L} component parameters
\itemize{
\item \code{psi} or \eqn{\psi}
\item \code{xi}  or \eqn{\xi}
\item \code{phi} or \eqn{\phi}
\item \code{theta} or \eqn{\theta}
}
}
\seealso{
\link{dLdt.basicL} or \link{create_Lpar_basicL}
}
