% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{set_Lpars}
\alias{set_Lpars}
\title{Set \strong{L} Component Parameters}
\usage{
set_Lpars(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \code{xds} object
}
\description{
Set \strong{L} Component Parameters
}
