% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mean_forcing.R
\name{set_mean_forcing.Lambda}
\alias{set_mean_forcing.Lambda}
\title{Set mean forcing}
\usage{
\method{set_mean_forcing}{Lambda}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{ramp.xds}} model object
}
\description{
Set \code{Lambda} to \code{X} for the \eqn{s^{th}} species
and return the \strong{\code{ramp.xds}} model object
}
