% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_mean_forcing.R
\name{set_mean_forcing}
\alias{set_mean_forcing}
\title{Set mean forcing}
\usage{
set_mean_forcing(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the mean forcing parameter to \code{X}
}
