% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season}
\alias{set_season}
\title{Get seasonal pattern}
\usage{
set_season(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new bottom parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Get the parameters set the seasonal pattern
}
