% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_bottom.R
\name{set_season_bottom.cohort}
\alias{set_season_bottom.cohort}
\title{set bottom}
\usage{
\method{set_season_bottom}{cohort}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new bottom parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\description{
Set the bottom for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
