% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_bottom.R
\name{set_season_bottom.none}
\alias{set_season_bottom.none}
\title{Set bottom}
\usage{
\method{set_season_bottom}{none}(bottom, xds_obj, s = 1)
}
\arguments{
\item{bottom}{the new bottom parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{set_season_bottom} for a model
with no forcing
}
