% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_phase.R
\name{set_season_phase.cohort}
\alias{set_season_phase.cohort}
\title{set phase}
\usage{
\method{set_season_phase}{cohort}(phase, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{phase}{the new phase parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\description{
Set the phase for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
