% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season_phase.eir}
\alias{set_season_phase.eir}
\title{Set phase}
\usage{
\method{set_season_phase}{eir}(phase, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{phase}{the new phase parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the phase for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
