% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_phase.R
\name{set_season_phase.Lambda}
\alias{set_season_phase.Lambda}
\title{Set phase}
\usage{
\method{set_season_phase}{Lambda}(phase, xds_obj, s = 1)
}
\arguments{
\item{phase}{the new phase parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{ramp.xds}} model object with an updated phase parameter
}
\description{
Set the phase for the seasonal pattern for the
\eqn{s^{th}} species
and return the \strong{\code{ramp.xds}} model object
}
