% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_pw.R
\name{set_season_pw.cohort}
\alias{set_season_pw.cohort}
\title{set pw}
\usage{
\method{set_season_pw}{cohort}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{a named list with the new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the pw for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
