% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_pw.R
\name{set_season_pw.none}
\alias{set_season_pw.none}
\title{Set pw}
\usage{
\method{set_season_pw}{none}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{a named list with the new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{set_season_pw} for a model
with no forcing
}
