% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trend.R
\name{set_spline.cohort}
\alias{set_spline.cohort}
\title{set yy}
\usage{
\method{set_spline}{cohort}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new interpolating points}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\description{
Set the yy for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
