% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trend.R
\name{set_spline.eir}
\alias{set_spline.eir}
\title{Set yy}
\usage{
\method{set_spline}{eir}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new interpolating points}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the yy for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
