% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spline.R
\name{set_spline_y.none}
\alias{set_spline_y.none}
\title{Set yy}
\usage{
\method{set_spline_y}{none}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new yy parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{set_spline_y} for a model
with no forcing
}
