% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_spline.R
\name{set_spline_y}
\alias{set_spline_y}
\title{Set yy}
\usage{
set_spline_y(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new yy parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the yy parameter to \code{X}
}
