% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{set_Xinits}
\alias{set_Xinits}
\title{Set new X parameter values}
\usage{
set_Xinits(pars, i = 1, Xopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{Xopts}{a named list}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{pars$Xpar[[s]]}.
}
