% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-coverage-func.R
\name{setup_area_spray_coverage.func}
\alias{setup_area_spray_coverage.func}
\title{Set up dynamic forcing}
\usage{
\method{setup_area_spray_coverage}{func}(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
