% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-effectsizes-none.R
\name{setup_area_spray_effectsizes.none}
\alias{setup_area_spray_effectsizes.none}
\title{Set up no area spray effectsizes}
\usage{
\method{setup_area_spray_effectsizes}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up no area spray effectsizes
}
