% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{setup_bad_habitats.as_is}
\alias{setup_bad_habitats.as_is}
\title{Set up bad habitats}
\usage{
\method{setup_bad_habitats}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured bad_habitats. If it passes
the checks, it replaces the current bad_habitats.

If called with \code{name = "as_is"}, the bad_habitats
must be at \code{options$bad_habitats}
}
\keyword{internal}
