% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{setup_bad_habitats.list}
\alias{setup_bad_habitats.list}
\title{Set up bad habitats}
\usage{
\method{setup_bad_habitats}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{BHname = name$name}
\item \code{options = name}
and call \code{setup_bad_habitats(BHname, xds_obj, options, s)}
}
}
\keyword{internal}
