% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{setup_bad_habitats.random}
\alias{setup_bad_habitats.random}
\title{Set up bad habitats}
\usage{
\method{setup_bad_habitats}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "method1" case for \link{setup_bad_habitats}. See \link{make_bad_habitats_random}
}
\keyword{internal}
