% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{setup_bad_habitats}
\alias{setup_bad_habitats}
\title{Set up bad habitats}
\usage{
setup_bad_habitats(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up bad habitats
}
