% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-coverage-func.R
\name{setup_bednet_coverage_func}
\alias{setup_bednet_coverage_func}
\title{Set up dynamic forcing}
\usage{
setup_bednet_coverage_func(
  pars,
  opts = list(),
  mean = 0,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}

\item{mean}{the mean bednet_coverage}

\item{F_season}{the seasonal signal in bednet coverage}

\item{F_trend}{a temporal trend in bednet coverage}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
