% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effects-interface.R
\name{setup_bednet_effects}
\alias{setup_bednet_effects}
\title{Set up dynamic forcing}
\usage{
setup_bednet_effects(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
