% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-effectsizes-lemenach.R
\name{setup_bednet_effectsizes_lemenach}
\alias{setup_bednet_effectsizes_lemenach}
\title{Set up dynamic forcing}
\usage{
setup_bednet_effectsizes_lemenach(
  pars,
  opts = list(),
  tau0_frac = c(0.68/3, 2.32/3),
  rr = 0.56,
  ss = 0.03
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}

\item{tau0_frac}{a \link{numeric} vector giving the proportion of a feeding cycle spent
host seeking/bloodfeeding \emph{vs.} resting/oviposition}

\item{rr}{probability of mosquito being repelled upon contact with ITN}

\item{ss}{probability of mosquito successfully feeding upon contact with ITN}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
