% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behavior.R
\name{setup_behavior_no_behavior}
\alias{setup_behavior_no_behavior}
\title{Make parameters for the no_behavior model for resource availability (do nothing)}
\usage{
setup_behavior_no_behavior(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the no_behavior model for resource availability (do nothing)
}
