% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography.R
\name{setup_births}
\alias{setup_births}
\title{Setup births}
\usage{
setup_births(name, xds_obj, i = 1)
}
\arguments{
\item{name}{the name of the births function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Configure the births function
}
