% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_hosts.R
\name{setup_blood_hosts.ts_func}
\alias{setup_blood_hosts.ts_func}
\title{Set up no blood_hosts}
\usage{
\method{setup_blood_hosts}{ts_func}(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list to configure \code{F_blood_hosts}}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup a model for no time spent blood_hostsing
and no exposure while blood_hostsing
}
