% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{setup_blood_search_weights.as_is}
\alias{setup_blood_search_weights.as_is}
\title{Set up a blood_search_weights}
\usage{
\method{setup_blood_search_weights}{as_is}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured blood_search_weights. If it passes
the checks, it replaces the current blood_search_weights.

If called with \code{name = "as_is"}, the blood_search_weights
must be at \code{options$blood_search_weights}
}
\keyword{internal}
