% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{setup_blood_search_weights.default}
\alias{setup_blood_search_weights.default}
\title{Set up no blood_search_weights}
\usage{
\method{setup_blood_search_weights}{default}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Don't change anything
}
\keyword{internal}
