% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{setup_blood_search_weights.list}
\alias{setup_blood_search_weights.list}
\title{Set up a blood_search_weights}
\usage{
\method{setup_blood_search_weights}{list}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
a numeric vector
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{BSWname = name$name}
\item \code{options = name}
and call \code{setup_blood_search_weights(BSWname, xds_obj, options, s)}
}
}
\keyword{internal}
