% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{setup_blood_search_weights.random}
\alias{setup_blood_search_weights.random}
\title{Set up a blood_search_weights}
\usage{
\method{setup_blood_search_weights}{random}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "method1" case for \link{setup_blood_search_weights}. See \link{make_blood_search_weights_random}
}
\keyword{internal}
