% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{setup_blood_search_weights}
\alias{setup_blood_search_weights}
\title{Set up a blood_search_weights}
\usage{
setup_blood_search_weights(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a blood_search_weights
}
