% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{setup_blood_traps.as_is}
\alias{setup_blood_traps.as_is}
\title{Set up a blood_traps}
\usage{
\method{setup_blood_traps}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured blood_traps. If it passes
the checks, it replaces the current blood_traps.

If called with \code{name = "as_is"}, the blood_traps
must be at \code{options$blood_traps}
}
\keyword{internal}
