% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{setup_blood_traps.list}
\alias{setup_blood_traps.list}
\title{Set up a blood_traps}
\usage{
\method{setup_blood_traps}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{xds} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{OBHname = name$name}
\item \code{options = name}
and call \code{setup_blood_traps(OBHname, xds_obj, options, s)}
}
}
\keyword{internal}
