% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{setup_blood_traps.random}
\alias{setup_blood_traps.random}
\title{Set up a blood_traps}
\usage{
\method{setup_blood_traps}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "random" case for \link{setup_blood_traps}. See \link{make_blood_traps_random}
}
\keyword{internal}
