% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{setup_blood_traps}
\alias{setup_blood_traps}
\title{Set up a blood_traps}
\usage{
setup_blood_traps(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a blood_traps
}
