% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/care_seeking.R
\name{setup_care_seeking_no_behavior}
\alias{setup_care_seeking_no_behavior}
\title{Make parameters for the no_behavior model for care seeking (do nothing)}
\usage{
setup_care_seeking_no_behavior(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the no_behavior model for care seeking (do nothing)
}
