% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/civil.R
\name{setup_civil_forced}
\alias{setup_civil_forced}
\title{Make parameters for the no_forcing model of civil forcing (do nothing)}
\usage{
setup_civil_forced(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the no_forcing model of civil forcing (do nothing)
}
