% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{setup_clinic_no_control}
\alias{setup_clinic_no_control}
\title{Set up the no_control model for clinical distribution (do nothing)}
\usage{
setup_clinic_no_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_control model for clinical distribution (do nothing)
}
