% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{setup_control_no_control}
\alias{setup_control_no_control}
\title{Set up the no_control model for control forcing (do nothing)}
\usage{
setup_control_no_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_control model for control forcing (do nothing)
}
