% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-distribute-func.R
\name{setup_distribute_bednets_func}
\alias{setup_distribute_bednets_func}
\title{Set up dynamic bednets}
\usage{
setup_distribute_bednets_func(
  pars,
  opts = list(),
  eventT = 365,
  F_distribute = NULL
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}

\item{eventT}{the time when a distribute occurs}

\item{F_distribute}{the effects of the distribute}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic bednets has not
already been set up, then turn on dynamic
bed net  and set all the
}
