% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-none.R
\name{setup_distribute_sugar_baits.none}
\alias{setup_distribute_sugar_baits.none}
\title{Set up distribution for the no sugar baits model}
\usage{
\method{setup_distribute_sugar_baits}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up distribution for the no sugar baits model
}
