% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-distribute-interface.R
\name{setup_distribute_sugar_baits}
\alias{setup_distribute_sugar_baits}
\title{Set up dynamic sugar_baits}
\usage{
setup_distribute_sugar_baits(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic sugar_baits has not
already been set up, then turn on dynamic
sugar_baits and set all the
}
