% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{setup_dynamic_travel}
\alias{setup_dynamic_travel}
\title{Set up a dynamic model for travel}
\usage{
setup_dynamic_travel(
  pars,
  traveling = 0,
  F_season = F_flat,
  F_trend = F_flat,
  i = 1
)
}
\arguments{
\item{pars}{a \link{list}}

\item{traveling}{the time spent traveling}

\item{F_season}{a function describing a seasonal pattern}

\item{F_trend}{a function describing a trend}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Setup a dynamical model for time spent traveling
}
