% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{setup_egg_laying_static}
\alias{setup_egg_laying_static}
\title{Setup egg laying for an static (autonomous) model of egg laying}
\usage{
setup_egg_laying_static(pars, searchQ = 1, s = 1, Lopts = list())
}
\arguments{
\item{pars}{a \code{xds} object}

\item{searchQ}{the membership matrix}

\item{s}{the vector species index}

\item{Lopts}{a \link{list} of options to override defaults}
}
\value{
a \link{list}  -- the modified \code{xds} object
}
\description{
Sets the search weights, \eqn{\omega} for the habitats
}
