% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-fixedlag-dts.R
\name{setup_EIP.fixedlag_dts}
\alias{setup_EIP.fixedlag_dts}
\title{Set up a fixedlag_dts model for the EIP}
\usage{
\method{setup_EIP}{fixedlag_dts}(EIPopts, MYZpar)
}
\arguments{
\item{EIPopts}{a list with options to setup EIPmod; must include EIPmod$EIPname}

\item{MYZpar}{the MYZ parameters}
}
\value{
\link{list}
}
\description{
Set up a fixedlag_dts model for the EIP
}
