% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static-xde.R
\name{setup_EIP.static_xde}
\alias{setup_EIP.static_xde}
\title{Set up a static_xde model for the EIP}
\usage{
\method{setup_EIP}{static_xde}(EIPopts, MYZpar)
}
\arguments{
\item{EIPopts}{a list with options to setup EIPmod; must include EIPmod$EIPname}

\item{MYZpar}{the MYZ parameters}
}
\value{
\link{list}
}
\description{
Set up a static_xde model for the EIP
}
