% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{setup_eip_obj}
\alias{setup_eip_obj}
\title{Set up the extrinsic incubation period object}
\usage{
setup_eip_obj(eip, MY_obj)
}
\arguments{
\item{eip}{the extrinsic incubation period in days}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the EIP
}
\keyword{internal}
