% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static-xde.R
\name{setup_eip_static_xde}
\alias{setup_eip_static_xde}
\title{Set up a static_xde model for the EIP}
\usage{
setup_eip_static_xde(EIPopts, MYZpar, eip = 12)
}
\arguments{
\item{EIPopts}{a \link{list}}

\item{MYZpar}{a \link{list}}

\item{eip}{the extrinsic incubation period (in days)}
}
\value{
\link{list}
}
\description{
Set up a static_xde model for the EIP
}
