% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure.pois}
\alias{setup_exposure.pois}
\title{Set up a Poisson model for Exposure and Infection}
\usage{
\method{setup_exposure}{pois}(EHname, xds_obj, i = 1, options = list())
}
\arguments{
\item{EHname}{environmental heterogeneity model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{set up options list}
}
\value{
an  \strong{\code{xds}} object
}
\description{
Set up a Poisson model for Exposure and Infection
}
\seealso{
Also, see \link{F_foi.pois} and \link{F_ar.pois}
}
