% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{setup_exposure}
\alias{setup_exposure}
\title{Set Up Exposure}
\usage{
setup_exposure(EHname, xds_obj, i = 1, options = list())
}
\arguments{
\item{EHname}{environmental heterogeneity model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{set up options list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up the model for
exposure. Current options for \code{EHname}
\itemize{
\item \code{pois} - a Poisson model
\item \code{nb} - a Negative Binomial model family
}
}
\note{
The daily EIR is an expected value, but that expectation
can have a distribution in a population. For
example, if the expectation is gamma distributed,
then we would get a negative binomial distribution
of bites per person.

Note that \link{Exposure} handles local exposure and
exposure while traveling separately.
}
