% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-multiday.R
\name{setup_exposure_multiday}
\alias{setup_exposure_multiday}
\title{Make parameters for the null model of exposure}
\usage{
setup_exposure_multiday(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
