% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb.dts}
\alias{setup_exposure_nb.dts}
\title{Set up the negative binomial model for exposure for discrete time models}
\usage{
\method{setup_exposure_nb}{dts}(pars, sz)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{sz}{the negative binomial size parameter (see \link{pnbinom})}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up the negative binomial model for exposure for discrete time models
}
\seealso{
Also, see \link{F_ar.nb}
}
