% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois.dts}
\alias{setup_exposure_pois.dts}
\title{Set up the Poisson model of exposure for discrete time models}
\usage{
\method{setup_exposure_pois}{dts}(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Set up the Poisson model of exposure for discrete time models
}
