% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois.xde}
\alias{setup_exposure_pois.xde}
\title{Make parameters for the null model of exposure}
\usage{
\method{setup_exposure_pois}{xde}(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
