% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_f}
\alias{setup_f}
\title{Set up constant blood feeding rate}
\usage{
setup_f(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a blood feeding rate or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Configure and validate models with
constant blood feeding rates. This calls \link{setup_F_f}
to configure a function \eqn{F_f}, and call
\link{change_f}.
}
\note{
Use \link{setup_F_f} and not \link{setup_f} if
the baseline values of \eqn{f} will vary over time.
}
\seealso{
\link{setup_F_f}
}
