% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-circadian.R
\name{setup_F_circadian.setup}
\alias{setup_F_circadian.setup}
\title{Show Circadian}
\usage{
\method{setup_F_circadian}{setup}(F_circadian, xds_obj, s = 1)
}
\arguments{
\item{F_circadian}{a circadian function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the circadian for the \eqn{s^{th}} host species
}
\keyword{internal}
