% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{setup_F_eip.as_is}
\alias{setup_F_eip.as_is}
\title{Set up static EIP values}
\usage{
\method{setup_F_eip}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{an EIP value or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set up supplied static extrinsic incubation periods.
}
\keyword{internal}
