% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-eip.R
\name{setup_F_eip.list}
\alias{setup_F_eip.list}
\title{Set up F_eip from options}
\usage{
\method{setup_F_eip}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{an EIP value or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
If an options list is passed as the first argument, use its
\code{name} field to select the setup method.
}
\keyword{internal}
