% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_F_f.as_is}
\alias{setup_F_f.as_is}
\title{Set up blood feeding rates}
\usage{
\method{setup_F_f}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the blood feeding rates to a new
set, either passed as the first argument or \code{name = "as_is"} and
the new values as \code{options$f}
}
\keyword{internal}
