% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_F_f.list}
\alias{setup_F_f.list}
\title{Set up blood feeding rates}
\usage{
\method{setup_F_f}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed as the first argument, use its
\code{name} field to select the setup method.
\itemize{
\item \code{Ff_name = name$name}
\item \code{options = name}
and call \code{setup_F_f(Ff_name, xds_obj, options, s)}.
}
}
\keyword{internal}
