% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_F_f.no_setup}
\alias{setup_F_f.no_setup}
\title{Set up blood feeding rates}
\usage{
\method{setup_F_f}{no_setup}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Leave the existing blood feeding configuration unchanged.
}
\keyword{internal}
