% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_F_f.type2}
\alias{setup_F_f.type2}
\title{Set up type 2 blood feeding rates}
\usage{
\method{setup_F_f}{type2}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Configure a type 2 functional response for blood feeding.
}
\keyword{internal}
