% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_F_habitat_weights}
\alias{setup_F_habitat_weights}
\title{Set up no habitats}
\usage{
setup_F_habitat_weights(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list to configure \code{F_habitat_weights}}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup a model for no time spent searching
and no exposure while searching
}
