% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{setup_F_mu.random}
\alias{setup_F_mu.random}
\title{Set up random emigration-loss fraction}
\usage{
\method{setup_F_mu}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "random" case for \link{setup_F_mu}. See \link{make_F_mu_random}.
}
\keyword{internal}
