% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{setup_F_nu.list}
\alias{setup_F_nu.list}
\title{Set up F_nu}
\usage{
\method{setup_F_nu}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{Fnu_name = name$name}
\item \code{options = name}
and call \code{setup_F_nu(Fnu_name, xds_obj, options, s)}
}
}
\keyword{internal}
