% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-egg_laying.R
\name{setup_F_nu}
\alias{setup_F_nu}
\title{Set up F_nu}
\usage{
setup_F_nu(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up F_nu
}
