% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_f_obj_B2}
\alias{setup_f_obj_B2}
\title{Setup Blood Feeding Bionomic Object}
\usage{
setup_f_obj_B2(MY_obj, options = list(), fx = 0.35, sf = 1)

setup_f_obj_B2(MY_obj, options = list(), fx = 0.35, sf = 1)
}
\arguments{
\item{MY_obj}{an \strong{\code{MY}} model object}

\item{options}{a list of options}

\item{fx}{the maximum blood feeding rate}

\item{sf}{a shape parameter}
}
\value{
a \strong{\code{MY}} model object

a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute dynamic blood feeding rates
as a type 2 functional response to
availability of blood hosts, \deqn{f=F_feeding_rate(B) = \frac{f_x s_f B}{1+s_f B}}

Set up an object
to compute dynamic blood feeding rates
as a type 2 functional response to
availability of blood hosts, \eqn{B}
}
