% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{setup_F_q.random}
\alias{setup_F_q.random}
\title{Set up F_q}
\usage{
\method{setup_F_q}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "random" case for \link{setup_F_q}. See \link{make_F_q_random}.
}
\keyword{internal}
