% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{setup_F_q.WVB}
\alias{setup_F_q.WVB}
\title{Set up WVB human feeding fraction}
\usage{
\method{setup_F_q}{WVB}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Configure the WVB human feeding fraction method.
}
\keyword{internal}
