% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{setup_F_sigma.as_is}
\alias{setup_F_sigma.as_is}
\title{Set up sigma}
\usage{
\method{setup_F_sigma}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the static values.
}
\keyword{internal}
