% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{setup_F_sigma.type2}
\alias{setup_F_sigma.type2}
\title{Set up F_sigma}
\usage{
\method{setup_F_sigma}{type2}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a type 2 functional response
to model the patch emigration rate
}
\keyword{internal}
