% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_F_travel_eir.ts_func}
\alias{setup_F_travel_eir.ts_func}
\title{Setup the Travel EIR}
\usage{
\method{setup_F_travel_eir}{ts_func}(setup_name, xds_obj, i, options)
}
\arguments{
\item{setup_name}{a}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{options list to set up \code{F_travel_eir}}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup a model for no time spent traveling
and no exposure while traveling
}
