% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_F_vis_kappa.ts_func}
\alias{setup_F_vis_kappa.ts_func}
\title{Set}
\usage{
\method{setup_F_vis_kappa}{ts_func}(mod_name, xds_obj, s, options)
}
\arguments{
\item{mod_name}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a list to configure \code{F_vis_kappa}}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup a model for the net infectiousness
of visitors
}
