% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{setup_forcing_object}
\alias{setup_forcing_object}
\title{Setup the Junction for Exogenous Forcing}
\usage{
setup_forcing_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{none} option
for exogenous forcing: no forcing.
}
