% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{setup_g}
\alias{setup_g}
\title{Set up mosquito mortality rate}
\usage{
setup_g(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a mosquito mortality rate or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set up the mosquito mortality strategy and validate its rates.
}
