% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{setup_habitat_object}
\alias{setup_habitat_object}
\title{Setup the Habitats Object}
\usage{
setup_habitat_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
During basic setup, the habitats membership
matrix is set up.

The \code{habitat_obj} handles:
\itemize{
\item dynamically changing habitat weights
\item availability of bad habitats
\item with multiple vector species, weights can be assigned to bad habitats
}
}
