% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{setup_habitat_search_weights.list}
\alias{setup_habitat_search_weights.list}
\title{Set up a habitat_search_weights}
\usage{
\method{setup_habitat_search_weights}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the mosquito species index}
}
\value{
a numeric vector
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{HSWname = name$name}
\item \code{options = name}
and call \code{setup_habitat_search_weights(HSWname, xds_obj, options, s)}
}
}
\keyword{internal}
