% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{setup_habitat_search_weights.setup}
\alias{setup_habitat_search_weights.setup}
\title{Set habitat search weights}
\usage{
\method{setup_habitat_search_weights}{setup}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the mosquito species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The default version called by setup
}
\keyword{internal}
