% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography-dHdt.R
\name{setup_Hmatrix}
\alias{setup_Hmatrix}
\title{Setup a matrix for \code{dHdt}}
\usage{
setup_Hmatrix(pars, i, Hmatrix)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Hmatrix}{a demographic matrix}
}
\value{
an \strong{\code{xds}} object
}
\description{
A demographic matrix should be a
square with dimensions \eqn{n_h \times n_h},
where \eqn{n_n=} \code{nStrata.} The column sums give
the background death rate for individuals
in the stratum.
}
