% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography-interface.R
\name{setup_Hpar_static}
\alias{setup_Hpar_static}
\title{A utility to set up Hpar}
\usage{
setup_Hpar_static(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
A utility to set up Hpar
}
