% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_human_frac}
\alias{setup_human_frac}
\title{Setup a Human Fraction Bionomic Object}
\usage{
setup_human_frac(q, MY_obj)
}
\arguments{
\item{q}{the human fraction}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{q}
}
