% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrology.R
\name{setup_hydrology_no_forcing}
\alias{setup_hydrology_no_forcing}
\title{Make parameters for the no_forcing model for hydrology (do nothing)}
\usage{
setup_hydrology_no_forcing(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Make parameters for the no_forcing model for hydrology (do nothing)
}
