% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{setup_importation}
\alias{setup_importation}
\title{Setup importation}
\usage{
setup_importation(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function, called by \link{make_xds_object_template}, sets up
terms and objects associated with malaria importation
}
\details{
This implements a framework to model blood feeding
described by Wu SL, \emph{et al.}, (2023).

Modular computation in \strong{\code{ramp.xds}} requires a rigid interface to
guarantee mathematical consistency in computing quantites related to blood feeding and transmission.

Several terms were developed to model importation by traveling humans and visitors. The terms
associated with time away and the travel EIR are human activities, so they are set up on the \strong{XH} object.
This sets up visitors and infectiousness.

\strong{Mulit-Host Models}

In models with multiple host species, the availability of visitors might vary by host.
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\seealso{
\link{setup_transmission}
}
\keyword{internal}
