% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{setup_importation_object}
\alias{setup_importation_object}
\title{Setup the Importation Object}
\usage{
setup_importation_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup an object to handle malaria importation
through travel and visitors. Initializes the junction and
zero-initializes the importation parameters on the \code{XY_interface}.
}
\keyword{internal}
