% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{setup_indices_L.basicL}
\alias{setup_indices_L.basicL}
\title{Setup Variable Indices for \code{basicL} (\strong{L} Component)}
\usage{
\method{setup_indices_L}{basicL}(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of the indices for the \strong{L} component variables
for the \code{basicL} module
}
