% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effects-interface.R
\name{setup_irs_effects}
\alias{setup_irs_effects}
\title{Set up dynamic irs}
\usage{
setup_irs_effects(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic irs has not
already been set up, then turn on dynamic
irs and set all the
}
