% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-effectsizes-none.R
\name{setup_irs_effectsizes.none}
\alias{setup_irs_effectsizes.none}
\title{Set up "no irs_effectsizes"}
\usage{
\method{setup_irs_effectsizes}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no irs_effectsizes"
}
