% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix.as_matrix}
\alias{setup_K_matrix.as_matrix}
\title{Setup a Here-There Dispersal Matrix}
\usage{
\method{setup_K_matrix}{as_matrix}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a method name: or a matrix, or a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Implements \link{setup_K_matrix} for the here and there model:
dispersal to every other patch, with equal probability
}
\keyword{internal}
