% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix.list}
\alias{setup_K_matrix.list}
\title{Setup K_matrix}
\usage{
\method{setup_K_matrix}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a method name: or a matrix, or a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
If the options list is passed
as the first argument, the set
\itemize{
\item \code{Kname = name$name}
\item \code{options = name}
and call \code{setup_K_matrix(Kname, xds_obj, options, s)}
}
}
\keyword{internal}
