% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix.N}
\alias{setup_K_matrix.N}
\title{Setup K_matrix}
\usage{
\method{setup_K_matrix}{N}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a method name: or a matrix, or a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Call \link{setup_K_matrix} \eqn{N} times.
The options for the \eqn{i^{th}} call are \code{options[[i]]}

of options
}
\keyword{internal}
