% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix}
\alias{setup_K_matrix}
\title{Setup Mosquito Dispersal Matrix}
\usage{
setup_K_matrix(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a method name: or a matrix, or a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A flexible function to set up or change the mosquito dispersal matrix (see \link{xds_info_mosquito_dispersal}).

The function was designed to dispatch on the first argument, \code{name}:
\itemize{
\item \code{name} is a method name
\item \code{options} is a named list that sets the parameters in a function \code{make_K_matrix_name}
\item Before dispatching, the function sets \code{class(option) = "name"}
}

Pre-dispatch cases were developed to make the function call more flexible: any
matrix can be passed as the first argument: or the user could set up an options list
and pass it
(\emph{e.g.} \code{Koptions} is passed to \code{setup_K_matrix} in \code{xds_setup}). The pre-dispatch
parsing:
\itemize{
\item if \code{name} is a method name, set \code{class(options) = "name"}
\item if \code{name} is a matrix, set \code{class(options) = "as_matrix"}
\item if \code{name} is a list of options,
}

Available methods are:
\itemize{
\item "as_matrix" --- sets up the matrix
\item "herethere" --- calls \link{make_K_matrix_herethere}
\item "xy" -- calls \link{make_K_matrix_xy}
\item "list" --- for options lists
\item "zero" --- sets up the zero matrix
\item "no_setup" --- returns the \strong{\code{xds}} object without modification
}
}
