% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix.zero}
\alias{setup_K_matrix.zero}
\title{Setup no dispersal matrix}
\usage{
\method{setup_K_matrix}{zero}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a method name: or a matrix, or a list}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Implements \link{setup_K_matrix} for the "no_setup" case
}
\keyword{internal}
