% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_obj}
\alias{setup_K_obj}
\title{Setup K_obj}
\usage{
setup_K_obj(MY_obj)
}
\arguments{
\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
an \strong{\code{MY}} model object
}
\description{
Set up a port object
for mosquito dispersal
}
\keyword{internal}
