% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{setup_L_inits}
\alias{setup_L_inits}
\title{Setup initial values (\strong{L})}
\usage{
setup_L_inits(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Sets the initial values of the \strong{L} Component
variables from \code{options}, a named list.
}
\keyword{internal}
