% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{setup_L_inits.trivial}
\alias{setup_L_inits.trivial}
\title{Setup Initial Values for the \strong{L} Component \code{trivial} Module}
\usage{
\method{setup_L_inits}{trivial}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
The \code{trivial} module initial values are an empty list
}
