% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{setup_L_ix.basicL}
\alias{setup_L_ix.basicL}
\title{Setup Variable Indices for \code{basicL} (\strong{L} Component)}
\usage{
\method{setup_L_ix}{basicL}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of the indices for the \strong{L} component variables
for the \code{basicL} module
}
