% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{setup_L_ix}
\alias{setup_L_ix}
\title{Set up Indices for \strong{L} Component Variables}
\usage{
setup_L_ix(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up Indices for \strong{L} Component Variables
}
\keyword{internal}
