% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{setup_L_ix.trivial}
\alias{setup_L_ix.trivial}
\title{Set up indices for \code{trivial} (\strong{L})}
\usage{
\method{setup_L_ix}{trivial}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the \strong{\code{xds}} object unmodified
}
\seealso{
\link{trivial_L}
}
\keyword{internal}
