% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{setup_L_obj}
\alias{setup_L_obj}
\title{Set up \strong{L} model object}
\usage{
setup_L_obj(Lname, membership, xds_obj, s, options = list())
}
\arguments{
\item{Lname}{the class name of the \strong{L} module}

\item{membership}{the habitats' patch membership vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named list to configure \strong{\code{L_obj}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Each instance of \verb{setup_L_obj.*} calls a function \verb{create_L_obj_*} that
creates an object \strong{\code{L_obj}.}
It is attached the \strong{\code{xds}} object as \verb{xds_obj$L_obj[[s]].} Each instance of \verb{create_L_obj_*}
should assign default parameter values that will be over-written by \code{options}
}
\note{
This method assigns \code{Lname} to class(\code{Lname}) and dispatches on \code{Lname}.
}
\keyword{internal}
