% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-search.R
\name{setup_L_ports}
\alias{setup_L_ports}
\title{Set up L ports}
\usage{
setup_L_ports(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Adds port objects and default values for the habitat search weights.
}
\keyword{internal}
