% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-dts.R
\name{setup_Linits.basicL_dts}
\alias{setup_Linits.basicL_dts}
\title{Setup the basicL_dts model}
\usage{
\method{setup_Linits}{basicL_dts}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_Linits} for the basicL_dts model
}
