% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{setup_Linits}
\alias{setup_Linits}
\title{Setup Initial Values for the \strong{L} Component}
\usage{
setup_Linits(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{Lname}.
}
