% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{setup_Linits.trivial}
\alias{setup_Linits.trivial}
\title{Setup Initial Values for the \strong{L} Component \code{trivial} Module}
\usage{
\method{setup_Linits}{trivial}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
The \code{trivial} module initial values are an empty list
}
