% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{setup_Lix}
\alias{setup_Lix}
\title{Set the Values of the Indices for \strong{L} Component Modules}
\usage{
setup_Lix(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the Values of the Indices for \strong{L} Component Modules
}
