% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{setup_Lix.trivial}
\alias{setup_Lix.trivial}
\title{Setup Variable Indices for \code{trivial} (\strong{L} Component)}
\usage{
\method{setup_Lix}{trivial}(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_Lix} for trivial (forced emergence) model.
}
