% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{setup_Lpar.basicL}
\alias{setup_Lpar.basicL}
\title{Setup \code{Lpar} for \code{basicL} (\strong{L} Component)}
\usage{
\method{setup_Lpar}{basicL}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a named \link{list} to configure \strong{\code{Lpar}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
The function sets up \code{Lpar} for the \eqn{s^{th}} species
by calling \link{make_Lpar_basicL}
}
\seealso{
\link{make_Lpar_basicL}
}
