% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-stages-dts.R
\name{setup_Lpar.stages_dts}
\alias{setup_Lpar.stages_dts}
\title{Setup Lpar for the stages_dts model}
\usage{
\method{setup_Lpar}{stages_dts}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Lpar} for the stages_dts model
}
