% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-none.R
\name{setup_lsm_coverage.none}
\alias{setup_lsm_coverage.none}
\title{Set up "no lsm_coverage"}
\usage{
\method{setup_lsm_coverage}{none}(name = "none", pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no lsm_coverage"
}
