% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-coverage-func.R
\name{setup_lsm_coverage_func}
\alias{setup_lsm_coverage_func}
\title{Set up dynamic forcing}
\usage{
setup_lsm_coverage_func(
  pars,
  opts = list(),
  mean = 0,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}

\item{mean}{the mean lsm_coverage}

\item{F_season}{the seasonal signal in lsm coverage}

\item{F_trend}{a temporal trend in lsm coverage}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
