% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_medical.R
\name{setup_mass_medical_no_control}
\alias{setup_mass_medical_no_control}
\title{Set up the no_control model for mass medical (do nothing)}
\usage{
setup_mass_medical_no_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_control model for mass medical (do nothing)
}
