% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_treatment.R
\name{setup_mass_treatment}
\alias{setup_mass_treatment}
\title{Setup mass treatment}
\usage{
setup_mass_treatment(
  xds_obj,
  mda = F_zero,
  msat = F_zero,
  options = list(),
  i = 1
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{mda}{a function of the form \eqn{F(t,V)}}

\item{msat}{a function of the form \eqn{F(t,V)}}

\item{options}{options list (overrides command line arguments)}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Add mass treatment ports to modules.
}
