% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography.R
\name{setup_mortality_matrix}
\alias{setup_mortality_matrix}
\title{Setup demographic matrix}
\usage{
setup_mortality_matrix(name, xds_obj, options, i = 1)
}
\arguments{
\item{name}{a setup method}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{options list (overrides command line arguments)}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Setup the demographic matirx
}
