% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mozy_mort.R
\name{setup_mozy_mort}
\alias{setup_mozy_mort}
\title{Setup a Mosquito Mortality Bionomic Object}
\usage{
setup_mozy_mort(g, MY_obj)
}
\arguments{
\item{g}{the mosquito mortality rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object to return a
constant baseline mosquito mortality rate, \eqn{g}
}
