% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{setup_mu}
\alias{setup_mu}
\title{Set up emigration-loss fraction}
\usage{
setup_mu(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{an emigration-loss fraction or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Configure and validate the emigration-loss fraction.
}
