% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigration_loss.R
\name{setup_mu_obj}
\alias{setup_mu_obj}
\title{Set up a dispersal loss bionomic object}
\usage{
setup_mu_obj(mu, MY_obj)
}
\arguments{
\item{mu}{the emigration loss fraction}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the dispersal loss fraction, \eqn{\mu}
}
\keyword{internal}
