% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{setup_MY_inits.basicM}
\alias{setup_MY_inits.basicM}
\title{Setup the basicM model}
\usage{
\method{setup_MY_inits}{basicM}(xds_obj, s, options)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_inits} for the basicM model
}
\keyword{internal}
