% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{setup_MY_inits.GeM}
\alias{setup_MY_inits.GeM}
\title{Setup initial values for \code{GeM} (\strong{MY})}
\usage{
\method{setup_MY_inits}{GeM}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_inits} for the GeM model
}
\keyword{internal}
