% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{setup_MY_inits.macdonald}
\alias{setup_MY_inits.macdonald}
\title{Setup initial values for the macdonald model}
\usage{
\method{setup_MY_inits}{macdonald}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_inits} for the macdonald model
}
\keyword{internal}
