% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{setup_MY_inits}
\alias{setup_MY_inits}
\title{A function to set up adult mosquito models}
\usage{
setup_MY_inits(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYname}.
}
