% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{setup_MY_inits.RMdts}
\alias{setup_MY_inits.RMdts}
\title{Setup initial values for the RMdts model}
\usage{
\method{setup_MY_inits}{RMdts}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_inits} for the RMdts model
}
