% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{setup_MY_inits.SI}
\alias{setup_MY_inits.SI}
\title{Setup initial values for the \code{SI} model}
\usage{
\method{setup_MY_inits}{SI}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_inits} for the \code{SI} model
}
