% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{setup_MY_inits.trivial}
\alias{setup_MY_inits.trivial}
\title{Setup inits for \code{trivial} (\strong{MY})}
\usage{
\method{setup_MY_inits}{trivial}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the \strong{\code{xds}} object unmodified
}
\keyword{internal}
