% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{setup_MY_ix.GeM}
\alias{setup_MY_ix.GeM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{GeM}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_ix} for the GeM model.
}
\keyword{internal}
