% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{setup_MY_ix.GeRM}
\alias{setup_MY_ix.GeRM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{GeRM}(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_ix} for the GeRM model.
}
