% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{setup_MY_ix.macdonald}
\alias{setup_MY_ix.macdonald}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{macdonald}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_ix} for the macdonald model.
}
\keyword{internal}
