% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{setup_MY_ix.SI}
\alias{setup_MY_ix.SI}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{SI}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MY_ix} for the \code{SI} model.
}
