% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{setup_MY_ix.trivial}
\alias{setup_MY_ix.trivial}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{trivial}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MY_ix} for trivial (forced emergence) model.
}
