% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{setup_MY_obj.basicM}
\alias{setup_MY_obj.basicM}
\title{Setup MY_obj for the basicM xde model}
\usage{
\method{setup_MY_obj}{basicM}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_obj} for the basicM xde model
}
