% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{setup_MY_obj.GeM}
\alias{setup_MY_obj.GeM}
\title{Setup MY_obj for the GeM model}
\usage{
\method{setup_MY_obj}{GeM}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_obj} for the GeM model
}
\keyword{internal}
