% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{setup_MY_obj.macdonald}
\alias{setup_MY_obj.macdonald}
\title{Set up \code{macdonald} (\strong{MY})}
\usage{
\method{setup_MY_obj}{macdonald}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_obj} for the macdonald model
}
\keyword{internal}
