% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{setup_MY_obj}
\alias{setup_MY_obj}
\title{Setup an \strong{MY} Model Object}
\usage{
setup_MY_obj(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
\link{list}
}
\description{
This function is a structured interface, a way
of building a model objects describing adult mosquito
ecology & infection dynamics. The string \emph{MYname}
is assigned its own class (\code{class(MYname) <- MYname}).

Since each model will require different arguments,
the dispatched function \code{setup_MY_obj.MYname}
calls \verb{make_MY_obj_MYname,} passing
the \code{options} to set the values of parameters or
terms.
}
\keyword{internal}
