% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{setup_MYZinits.basicM_xde}
\alias{setup_MYZinits.basicM_xde}
\title{Setup the basicM_xde model}
\usage{
\method{setup_MYZinits}{basicM_xde}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the basicM_xde model
}
