% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ-dts.R
\name{setup_MYZinits.ENBRQ_dts}
\alias{setup_MYZinits.ENBRQ_dts}
\title{Setup initial values for the ENBRQ_dts model}
\usage{
\method{setup_MYZinits}{ENBRQ_dts}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the ENBRQ_dts model
}
