% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-macdonald.R
\name{setup_MYZinits.macdonald}
\alias{setup_MYZinits.macdonald}
\title{Setup initial values for the macdonald model}
\usage{
\method{setup_MYZinits}{macdonald}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the macdonald model
}
