% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{setup_MYZinits.MY}
\alias{setup_MYZinits.MY}
\title{Setup initial values for the \code{MY} model}
\usage{
\method{setup_MYZinits}{MY}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the \code{MY} model
}
