% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{setup_MYZinits.mz}
\alias{setup_MYZinits.mz}
\title{Setup initial values for the mz model}
\usage{
\method{setup_MYZinits}{mz}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the mz model
}
