% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{setup_MYZinits.si_old}
\alias{setup_MYZinits.si_old}
\title{Setup initial values for the \code{si_old} model}
\usage{
\method{setup_MYZinits}{si_old}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZinits} for the \code{si_old} model
}
