% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{setup_MYZinits.trace}
\alias{setup_MYZinits.trace}
\title{Setup the trace model}
\usage{
\method{setup_MYZinits}{trace}(pars, s, MYZopts = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the trace model
}
