% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{setup_MYZinits.trivial}
\alias{setup_MYZinits.trivial}
\title{Setup the trivial model}
\usage{
\method{setup_MYZinits}{trivial}(pars, s, MYZopts = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZinits} for the trivial model
}
