% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM.R
\name{setup_MYZix.basicM}
\alias{setup_MYZix.basicM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{basicM}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MYZix} for the basic M model.
}
