% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-macdonald.R
\name{setup_MYZix.macdonald}
\alias{setup_MYZix.macdonald}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{macdonald}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZix} for the macdonald model.
}
