% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{setup_MYZix}
\alias{setup_MYZix}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
setup_MYZix(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.
}
