% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{setup_MYZix.RM_dts}
\alias{setup_MYZix.RM_dts}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{RM_dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZix} for the RM_dts model.
}
