% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{setup_MYZix.SI}
\alias{setup_MYZix.SI}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{SI}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_MYZix} for the \code{SI} model.
}
