% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{setup_MYZix.trivial}
\alias{setup_MYZix.trivial}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
\method{setup_MYZix}{trivial}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MYZix} for trivial (forced emergence) model.
}
