% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-ENBRQ-dts.R
\name{setup_MYZpar.ENBRQ_dts}
\alias{setup_MYZpar.ENBRQ_dts}
\title{Setup MYZpar for the ENBRQ_dts model}
\usage{
\method{setup_MYZpar}{ENBRQ_dts}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the ENBRQ_dts model
}
