% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{setup_MYZpar.MY}
\alias{setup_MYZpar.MY}
\title{Setup MYZpar for the MY model}
\usage{
\method{setup_MYZpar}{MY}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the MY model
}
