% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{setup_MYZpar.si_old}
\alias{setup_MYZpar.si_old}
\title{Setup MYZpar for the si_old model}
\usage{
\method{setup_MYZpar}{si_old}(MYZname, pars, s, MYZopts = list())
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the si_old model
}
