% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{setup_MYZpar.trivial}
\alias{setup_MYZpar.trivial}
\title{Setup the trivial}
\usage{
\method{setup_MYZpar}{trivial}(MYZname, pars, s, MYZopts = NULL)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MYZpar} for the trivial model
}
