% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_case_detection.R
\name{setup_no_active_case_detection}
\alias{setup_no_active_case_detection}
\title{Set up the none model for active_case_detectional distribution (do nothing)}
\usage{
setup_no_active_case_detection(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the none model for active_case_detectional distribution (do nothing)
}
