% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray.R
\name{setup_no_area_spray}
\alias{setup_no_area_spray}
\title{Set up "no AreaSpray"}
\usage{
setup_no_area_spray(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no AreaSpray"
}
