% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{setup_no_bednets}
\alias{setup_no_bednets}
\title{Set up "no bednet"}
\usage{
setup_no_bednets(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no bednet"
}
