% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinic.R
\name{setup_no_clinic}
\alias{setup_no_clinic}
\title{Set up the none model for clinical distribution (do nothing)}
\usage{
setup_no_clinic(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the none model for clinical distribution (do nothing)
}
