% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/development.R
\name{setup_no_development}
\alias{setup_no_development}
\title{Set up the no_dev model for developments (do nothing)}
\usage{
setup_no_development(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_dev model for developments (do nothing)
}
