% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{setup_no_forcing}
\alias{setup_no_forcing}
\title{Basic set up for exogenous forcing}
\usage{
setup_no_forcing(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{basic} option
for exogenous forcing: no forcing.
}
