% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{setup_no_health}
\alias{setup_no_health}
\title{none set up for exogenous health}
\usage{
setup_no_health(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{none} option
for exogenous health: no health.
}
