% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs.R
\name{setup_no_irs}
\alias{setup_no_irs}
\title{Set up "no irs"}
\usage{
setup_no_irs(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no irs"
}
