% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{setup_no_lsm}
\alias{setup_no_lsm}
\title{Set up "no LSM"}
\usage{
setup_no_lsm(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no LSM"
}
