% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass_health.R
\name{setup_no_mass_health}
\alias{setup_no_mass_health}
\title{Set up the none model for mass_healthal distribution (do nothing)}
\usage{
setup_no_mass_health(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the none model for mass_healthal distribution (do nothing)
}
