% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-null.R
\name{setup_no_nets}
\alias{setup_no_nets}
\title{Make parameters for the no_nets model of bed nets (do nothing)}
\usage{
setup_no_nets(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Make parameters for the no_nets model of bed nets (do nothing)
}
