% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school.R
\name{setup_no_school}
\alias{setup_no_school}
\title{Set up the none model for schoolal distribution (do nothing)}
\usage{
setup_no_school(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the none model for schoolal distribution (do nothing)
}
