% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{setup_no_sugar_baits}
\alias{setup_no_sugar_baits}
\title{Set up no sugar baits}
\usage{
setup_no_sugar_baits(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up no sugar baits
}
