% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{setup_no_temperature}
\alias{setup_no_temperature}
\title{Set up "no temperature"}
\usage{
setup_no_temperature(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no temperature"
}
