% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{setup_no_travel}
\alias{setup_no_travel}
\title{A function to set up malaria importation}
\usage{
setup_no_travel(pars, i = 1)
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Setup a static model for travel malaria
}
