% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_no_vector_control}
\alias{setup_no_vector_control}
\title{none set up for exogenous vector_control}
\usage{
setup_no_vector_control(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{none} option
for exogenous vector_control: no vector_control.
}
