% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_nu_obj_Q2}
\alias{setup_nu_obj_Q2}
\title{Setup Blood Feeding Bionomic Object}
\usage{
setup_nu_obj_Q2(MY_obj, options = list(), nux = 0.35, snu = 1)
}
\arguments{
\item{MY_obj}{an \strong{\code{MY}} model object}

\item{options}{a list of options}

\item{nux}{the maximum egg laying rate}

\item{snu}{a shape parameter}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute dynamic egg laying rates
as a type 2 functional response to
available habitat, \eqn{Q}
}
