% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-demography.R
\name{setup_Omega.dts}
\alias{setup_Omega.dts}
\title{Make the mosquito demography matrix}
\usage{
\method{setup_Omega}{dts}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
Make the mosquito demography matrix
}
\note{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}
}
