% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-demography.R
\name{setup_Omega.xde}
\alias{setup_Omega.xde}
\title{Make the mosquito demography matrix}
\usage{
\method{setup_Omega}{xde}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
This method dispatches on the type of \code{MY_obj}.
}
