% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{setup_other_blood_hosts.as_is}
\alias{setup_other_blood_hosts.as_is}
\title{Set up a other_blood_hosts}
\usage{
\method{setup_other_blood_hosts}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured other_blood_hosts. If it passes
the checks, it replaces the current other_blood_hosts.

If called with \code{name = "as_is"}, the other_blood_hosts
must be at \code{options$other_blood_hosts}
}
\keyword{internal}
