% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{setup_other_blood_hosts.list}
\alias{setup_other_blood_hosts.list}
\title{Set up a other_blood_hosts}
\usage{
\method{setup_other_blood_hosts}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{OBHname = name$name}
\item \code{options = name}
and call \code{setup_other_blood_hosts(OBHname, xds_obj, options, s)}
}
}
\keyword{internal}
