% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{setup_other_blood_hosts.no_setup}
\alias{setup_other_blood_hosts.no_setup}
\title{Set up no other_blood_hosts}
\usage{
\method{setup_other_blood_hosts}{no_setup}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Don't change anything
}
\keyword{internal}
