% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{setup_other_blood_hosts.random}
\alias{setup_other_blood_hosts.random}
\title{Set up a other_blood_hosts}
\usage{
\method{setup_other_blood_hosts}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "method1" case for \link{setup_other_blood_hosts}. See \link{make_other_blood_hosts_random}
}
\keyword{internal}
