% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-setup.R
\name{setup_other_variables}
\alias{setup_other_variables}
\title{Set Up the first (null) other variable}
\usage{
setup_other_variables(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \code{xds} object
}
\description{
The description for each method should include the equations.
}
