% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{setup_ovitraps.as_is}
\alias{setup_ovitraps.as_is}
\title{Set up ovitraps}
\usage{
\method{setup_ovitraps}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured ovitraps. If it passes
the checks, it replaces the current ovitraps.

If called with \code{name = "as_is"}, the ovitraps
must be at \code{options$ovitraps}
}
\keyword{internal}
