% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{setup_ovitraps.list}
\alias{setup_ovitraps.list}
\title{Set up ovitraps}
\usage{
\method{setup_ovitraps}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{OTname = name$name}
\item \code{options = name}
and call \code{setup_ovitraps(OTname, xds_obj, options, s)}
}
}
\keyword{internal}
