% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{setup_ovitraps}
\alias{setup_ovitraps}
\title{Set up ovitraps}
\usage{
setup_ovitraps(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up ovitraps
}
