% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-own-none.R
\name{setup_own_bednets.none}
\alias{setup_own_bednets.none}
\title{Set up "no own_bednets"}
\usage{
\method{setup_own_bednets}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no own_bednets"
}
