% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{setup_q}
\alias{setup_q}
\title{Set up human feeding fraction}
\usage{
setup_q(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a human feeding fraction or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Configure and validate the human feeding fraction.
}
