% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{setup_q_obj}
\alias{setup_q_obj}
\title{Set up a human fraction bionomic object}
\usage{
setup_q_obj(q, MY_obj)
}
\arguments{
\item{q}{the human fraction}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{q}
}
\keyword{internal}
