% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{setup_rainfall_func}
\alias{setup_rainfall_func}
\title{Set up dynamic forcing}
\usage{
setup_rainfall_func(
  pars,
  Topts = list(),
  mean = 30,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{Topts}{a list of options to override defaults}

\item{mean}{the mean rainfall}

\item{F_season}{the seasonal signal in rainfall}

\item{F_trend}{a temporal trend in rainfall}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
