% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall.R
\name{setup_rainfall_no_forcing}
\alias{setup_rainfall_no_forcing}
\title{Set up the no_forcing model for RAINFALL}
\usage{
setup_rainfall_no_forcing(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_forcing model for RAINFALL
}
