% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{setup_relative_humidity_no_forcing}
\alias{setup_relative_humidity_no_forcing}
\title{Set up the no_forcing model for HUMIDITY}
\usage{
setup_relative_humidity_no_forcing(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_forcing model for HUMIDITY
}
