% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_setup.dynamic}
\alias{setup_setup.dynamic}
\title{Setup setup}
\usage{
\method{setup_setup}{dynamic}(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
the modified object
}
\description{
Don't change the class of an object to
\code{setup} if it is \code{dynamic}
}
