% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_setup.setup}
\alias{setup_setup.setup}
\title{Setup setup}
\usage{
\method{setup_setup}{setup}(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
the modified object
}
\description{
Don't change the class of an object if it is
already \code{setup}
}
