% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_setup.static}
\alias{setup_setup.static}
\title{Setup setup}
\usage{
\method{setup_setup}{static}(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
the modified object
}
\description{
Change the class of an object from
\code{static} to \code{setup}
}
