% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{setup_sigma}
\alias{setup_sigma}
\title{Set up patch emigration rate}
\usage{
setup_sigma(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a patch emigration rate or setup method name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Configure and validate the patch emigration rate.
}
