% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_sigma_obj_BQS}
\alias{setup_sigma_obj_BQS}
\title{Setup Blood Feeding Bionomic Object}
\usage{
setup_sigma_obj_BQS(
  MY_obj,
  options = list(),
  sigma_x = 1,
  sigma_B = 3,
  sigma_Q = 3,
  sigma_S = 3,
  sB = 1,
  sQ = 1,
  sS = 1
)
}
\arguments{
\item{MY_obj}{an \strong{\code{MY}} model object}

\item{options}{an \strong{\code{MY}} model object}

\item{sigma_x}{a scaling parameter}

\item{sigma_B}{maximum rate, blood searching}

\item{sigma_Q}{maximum rate, habitat searching}

\item{sigma_S}{maximum rate, sugar searching}

\item{sB}{shape parameter, blood searching}

\item{sQ}{shape parameter, habitat searching}

\item{sS}{shape parameter, sugar searching}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute dynamic blood feeding rates
as a type 2 functional response to
availability of blood hosts, \eqn{B}
}
