% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{setup_skillset_MY.macdonald}
\alias{setup_skillset_MY.macdonald}
\title{The skill set}
\usage{
\method{setup_skillset_MY}{macdonald}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the \strong{\code{xds}} object
}
\description{
The skill set
}
\keyword{internal}
