% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{setup_skillset_MY}
\alias{setup_skillset_MY}
\title{Setup the skill set}
\usage{
setup_skillset_MY(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
The skill set for an \strong{MY} module
is a list that summarizes capabilities and
compatibilities
}
\keyword{internal}
