% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{setup_skillset_XH.hMoI}
\alias{setup_skillset_XH.hMoI}
\title{The \strong{XH} module skill set}
\usage{
\method{setup_skillset_XH}{hMoI}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
the \strong{\code{xds}} object
}
\description{
The \code{hMoI} module is an example
of a model that is not extensible. There is no
way to add either dynamic human population density
or mass treatment without violating the model
assumptions.

The model does output observed population density
and here it is
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
\keyword{internal}
