% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{setup_skillset_XH.trivial}
\alias{setup_skillset_XH.trivial}
\title{The \strong{XH} module skill set}
\usage{
\method{setup_skillset_XH}{trivial}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
the \strong{\code{xds}} object
}
\description{
The \strong{XH} skill set is a list of
a module's capabilities
}
\keyword{internal}
