% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-spray-interface.R
\name{setup_spray_area}
\alias{setup_spray_area}
\title{Set up dynamic lsm}
\usage{
setup_spray_area(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic lsm has not
already been set up, then turn on dynamic
lsm and set all the
}
