% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses-none.R
\name{setup_spray_houses.none}
\alias{setup_spray_houses.none}
\title{Set up "no spray_houses"}
\usage{
\method{setup_spray_houses}{none}(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no spray_houses"
}
