% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{setup_static_travel}
\alias{setup_static_travel}
\title{Set up static travel}
\usage{
setup_static_travel(pars, traveling_fraction = 0, i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{traveling_fraction}{the fraction of time spent traveling}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Setup a static model for time spent traveling

Setup a static model for travel malaria
}
