% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-func.R
\name{setup_sugar_bait_coverage.func}
\alias{setup_sugar_bait_coverage.func}
\title{Set up dynamic sugar bait coverage function}
\usage{
\method{setup_sugar_bait_coverage}{func}(name, pars, opts = list())
}
\arguments{
\item{name}{the module name}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override the defaults}
}
\description{
Set up a dynamic sugar bait
coverage function
}
