% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-none.R
\name{setup_sugar_bait_coverage.none}
\alias{setup_sugar_bait_coverage.none}
\title{Set up no sugar bait coverage}
\usage{
\method{setup_sugar_bait_coverage}{none}(name = "none", pars, opts = list())
}
\arguments{
\item{name}{the module name}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override the defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up no sugar bait coverage
}
