% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-interface.R
\name{setup_sugar_bait_coverage}
\alias{setup_sugar_bait_coverage}
\title{Set up dynamic forcing}
\usage{
setup_sugar_bait_coverage(name, pars, opts = list())
}
\arguments{
\item{name}{the module name}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override the defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
