% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-func.R
\name{setup_sugar_bait_coverage_func}
\alias{setup_sugar_bait_coverage_func}
\title{Set up dynamic sugar bait coverage function}
\usage{
setup_sugar_bait_coverage_func(
  pars,
  opts = list(),
  mean = 0,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}

\item{mean}{the mean sugar_bait_coverage}

\item{F_season}{the seasonal signal in sugar_bait coverage}

\item{F_trend}{a temporal trend in sugar_bait coverage}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up dynamic sugar bait coverage function
}
