% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effectsizes-interface.R
\name{setup_sugar_bait_effectsizes}
\alias{setup_sugar_bait_effectsizes}
\title{Set up sugar baits effect sizes}
\usage{
setup_sugar_bait_effectsizes(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up sugar baits effect sizes
}
