% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{setup_sugar_object}
\alias{setup_sugar_object}
\title{Setup the Junction for Exogenous sugar}
\usage{
setup_sugar_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{none} option
for exogenous sugar: no sugar.
}
