% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{setup_temperature_func}
\alias{setup_temperature_func}
\title{Set up dynamic forcing}
\usage{
setup_temperature_func(
  pars,
  Topts = list(),
  meanT = 30,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{Topts}{a list of options to override defaults}

\item{meanT}{the mean temperature}

\item{F_season}{the seasonal signal in temperature}

\item{F_trend}{a temporal trend in temperature}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
