% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{setup_temperature_no_forcing}
\alias{setup_temperature_no_forcing}
\title{Set up the no_forcing model for temperature}
\usage{
setup_temperature_no_forcing(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_forcing model for temperature
}
