% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{setup_time_away.list}
\alias{setup_time_away.list}
\title{Set up a time spent}
\usage{
\method{setup_time_away}{list}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{TAname = name$name}
\item \code{options = name}
and call \code{setup_time_away(TAname, xds_obj, options, s)}
}
}
\keyword{internal}
