% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{setup_time_away.random}
\alias{setup_time_away.random}
\title{Set up time away}
\usage{
\method{setup_time_away}{random}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up random values for time away
}
\keyword{internal}
