% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{setup_time_away.setup}
\alias{setup_time_away.setup}
\title{Set up a time spent}
\usage{
\method{setup_time_away}{setup}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured time spent . If it passes
the checks, it replaces the current time_away .

If called with \code{name = "as_is"}, the time spent
must be at \code{options$time_away}
}
\keyword{internal}
