% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespent.R
\name{setup_timespent.as_matrix}
\alias{setup_timespent.as_matrix}
\title{Set up a time spent matrix}
\usage{
\method{setup_timespent}{as_matrix}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured time spent matrix. If it passes
the checks, it replaces the current timespent matrix.

If called with \code{name = "as_matrix"}, the time spent
matrix must be at \code{options$timespent_matrix}
}
\keyword{internal}
