% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_TimeSpent.athome}
\alias{setup_TimeSpent.athome}
\title{Make a mosquito dispersal matrix, called TimeSpent with a here / away}
\usage{
\method{setup_TimeSpent}{athome}(TimeSpent, xds_obj, i, options = list())
}
\arguments{
\item{TimeSpent}{a matrix or a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_TimeSpent} for as_matrix
}
