% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{setup_transmission}
\alias{setup_transmission}
\title{Setup the interface for parasite / pathogen transmission}
\usage{
setup_transmission(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Sets up the object that defines the transmission interface for an \code{xds} model object.
}
\details{
This implements a model for egg laying described by Wu SL, \emph{et al.}, (2023).
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\seealso{
\link{make_xds_object_template}

\link{setup_blood_feeding}
}
\keyword{internal}
