% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps.R
\name{setup_traps_object}
\alias{setup_traps_object}
\title{Setup the Junction for Exogenous traps}
\usage{
setup_traps_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up the \code{none} option
for exogenous traps: no traps.
}
