% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-ar.R
\name{setup_travel_ar}
\alias{setup_travel_ar}
\title{Set up parameters and functions to compute the Attack-Rate for a travel_malaria model}
\usage{
setup_travel_ar(pars, travel_Opts = list(), delta_scale = 0, delta_t = NULL)
}
\arguments{
\item{pars}{a \link{list}}

\item{travel_Opts}{a list, overwrites default values}

\item{delta_scale}{a non-negative numeric value to scale the mean travel related malaria Attack Rate}

\item{delta_t}{the temporal pattern for travel related malaria Attack-Rate}
}
\value{
none
}
\description{
Set up parameters and functions to compute the Attack-Rate for a travel_malaria model
}
