% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{setup_travel_eir.list}
\alias{setup_travel_eir.list}
\title{Set up a travel eir}
\usage{
\method{setup_travel_eir}{list}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{Teirname = name$name}
\item \code{options = name}
and call \code{setup_travel_eir(Teirname, xds_obj, options, s)}
}
}
\keyword{internal}
