% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{setup_travel_eir.no_travel}
\alias{setup_travel_eir.no_travel}
\title{Set up a travel eir}
\usage{
\method{setup_travel_eir}{no_travel}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured travel eir . If it passes
the checks, it replaces the current travel_eir .

If called with \code{name = "as_is"}, the travel eir
must be at \code{options$travel_eir}
}
\keyword{internal}
