% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{setup_travel_eir.random}
\alias{setup_travel_eir.random}
\title{Set up time away}
\usage{
\method{setup_travel_eir}{random}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a  or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up random values for time away
}
\keyword{internal}
