% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{setup_travel_object}
\alias{setup_travel_object}
\title{Setup the Travel Object}
\usage{
setup_travel_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup the object to handle
travel
\itemize{
\item \code{F_travel} should have the form \eqn{F(t,V)} where \eqn{V}
is returned by \link{get_variables} dispatching on \code{class(F_travel)}
\item \code{F_travel_eir} should have the form \eqn{F(t,V)} where \eqn{V}
is returned by \link{get_variables} dispatching on \code{class(F_travel)}
}
}
