% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria.R
\name{setup_travel_static.xde}
\alias{setup_travel_static.xde}
\title{A function to set up malaria importation}
\usage{
\method{setup_travel_static}{xde}(pars, delta = 0)
}
\arguments{
\item{pars}{a \link{list}}

\item{delta}{the travel FoI}
}
\value{
a \link{list}
}
\description{
Setup a static model for travel malaria
}
