% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{setup_Upsilon_obj}
\alias{setup_Upsilon_obj}
\title{Set up the Upsilon object}
\usage{
setup_Upsilon_obj(xds_obj, s = 1, dede = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}

\item{dede}{is it a delay differential equation that needs \eqn{\Upsilon}}
}
\value{
an \strong{xds} model object
}
\description{
Set to \code{dde} if \strong{MY} is a
delay differential equation
}
\keyword{internal}
