% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use-none.R
\name{setup_use_bednets.none}
\alias{setup_use_bednets.none}
\title{Set up "no use_bednets"}
\usage{
\method{setup_use_bednets}{none}(name, pars, opts)
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up "no use_bednets"
}
