% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use-interface.R
\name{setup_use_bednets}
\alias{setup_use_bednets}
\title{Set up dynamic bednets}
\usage{
setup_use_bednets(name, pars, opts = list())
}
\arguments{
\item{name}{the name of a model to set up}

\item{pars}{an \strong{\code{xds}} object}

\item{opts}{a list of options to override defaults}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic bednets has not
already been set up, then turn on dynamic
bednets and set all the
}
