% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{setup_V_inits}
\alias{setup_V_inits}
\title{Setup Initial Values for \strong{V} Modules}
\usage{
setup_V_inits(xds_obj, H, i, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{xds_obj$V_obj[[i]]}.
}
