% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-macdonald.R
\name{setup_V_ix.macdonald}
\alias{setup_V_ix.macdonald}
\title{Initialize VC}
\usage{
\method{setup_V_ix}{macdonald}(xds_obj, i)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{i}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
To compute vectorial capacity, we
need to
}
\keyword{internal}
