% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-setup.R
\name{setup_V_ix.setup}
\alias{setup_V_ix.setup}
\title{Compute Other Variables}
\usage{
\method{setup_V_ix}{setup}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the variable index}
}
\value{
an \code{xds} object
}
\description{
The description for each method should include the equations.
}
\note{
Index the variables
}
