% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity-GeM.R
\name{setup_VC.GeM}
\alias{setup_VC.GeM}
\title{Initialize VC}
\usage{
\method{setup_VC}{GeM}(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
To compute vectorial capacity, we
need to
}
\keyword{internal}
