% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorial_capacity.R
\name{setup_VC}
\alias{setup_VC}
\title{Setup VC initial values}
\usage{
setup_VC(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a numeric \link{matrix}
}
\description{
To compute vectorial capacity, we
need to
}
\keyword{internal}
