% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-no_control.R
\name{setup_vc_no_control}
\alias{setup_vc_no_control}
\title{Distribute vector control, the no_control model}
\usage{
setup_vc_no_control(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
no_control
}
\description{
Distribute vector control, the no_control model
}
