% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_vector_control_object}
\alias{setup_vector_control_object}
\title{Setup a Junction for Vector Control}
\usage{
setup_vector_control_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This creates \code{xds_obj$vector_control}
and sets \verb{class(vector_control) = 'none'.}

Functions to implement vector control are
in a companion package called
\href{https://github.com/dd-harp/ramp.control}{\strong{\code{ramp.control}}}.
}
\seealso{
\link{VectorControl.none}
}
