% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{setup_vis_kappa.as_is}
\alias{setup_vis_kappa.as_is}
\title{Set up a vis_kappa}
\usage{
\method{setup_vis_kappa}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured vis_kappa. If it passes
the checks, it replaces the current vis_kappa.

If called with \code{name = "as_is"}, the vis_kappa
must be at \code{options$vis_kappa}
}
\keyword{internal}
