% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{setup_vis_kappa.random}
\alias{setup_vis_kappa.random}
\title{Set up a vis_kappa}
\usage{
\method{setup_vis_kappa}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "random" case for \link{setup_vis_kappa}. See \link{make_vis_kappa_random}
}
\keyword{internal}
