% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{setup_vis_kappa}
\alias{setup_vis_kappa}
\title{Set up a vis_kappa}
\usage{
setup_vis_kappa(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a vis_kappa
}
