% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitor_object}
\alias{setup_visitor_object}
\title{Setup the Visitors Object}
\usage{
setup_visitor_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup a an object
to model blood feeding and importation
from visitors
}
