% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors.as_is}
\alias{setup_visitors.as_is}
\title{Set up a visitors}
\usage{
\method{setup_visitors}{as_is}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Pass a pre-configured visitors. If it passes
the checks, it replaces the current visitors.

If called with \code{name = "as_is"}, the visitors
must be at \code{options$visitors}
}
\keyword{internal}
