% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors.list}
\alias{setup_visitors.list}
\title{Set up a visitors}
\usage{
\method{setup_visitors}{list}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
a \link{matrix}
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{Vname = name$name}
\item \code{options = name}
and call \code{setup_visitors(Vname, xds_obj, options, s)}
}
}
\keyword{internal}
