% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors.no_setup}
\alias{setup_visitors.no_setup}
\title{Set up no visitors}
\usage{
\method{setup_visitors}{no_setup}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Don't change anything
}
\keyword{internal}
