% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors.random}
\alias{setup_visitors.random}
\title{Set up a visitors}
\usage{
\method{setup_visitors}{random}(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements the "random" case for \link{setup_visitors}. See \link{make_visitors_random}
}
\keyword{internal}
