% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{setup_visitors}
\alias{setup_visitors}
\title{Set up a visitors}
\usage{
setup_visitors(name, xds_obj, options = list(), s = 1)
}
\arguments{
\item{name}{a or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{s}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a visitors
}
