% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{setup_Vix}
\alias{setup_Vix}
\title{Compute Other Variables}
\usage{
setup_Vix(pars, i)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{i}{the variable index}
}
\value{
an \code{xds} object
}
\description{
The description for each method should include the equations.
}
\note{
Index the variables
}
