% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{setup_XH_inits.hMoI}
\alias{setup_XH_inits.hMoI}
\title{Setup Xinits.hMoI}
\usage{
\method{setup_XH_inits}{hMoI}(xds_obj, H, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_XH_inits} for the hMoI model
}
