% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{setup_XH_inits}
\alias{setup_XH_inits}
\title{Setup Initial Values for \strong{XH} Modules}
\usage{
setup_XH_inits(xds_obj, H, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{xds_obj$XH_obj[[i]]}.
}
