% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{setup_XH_inits.SIS}
\alias{setup_XH_inits.SIS}
\title{Setup initial values for \emph{SIS}}
\usage{
\method{setup_XH_inits}{SIS}(xds_obj, H, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{H}{initial host population density}

\item{i}{the host species index}

\item{options}{a \link{list}}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Setup initial values for \emph{SIS}
}
