% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{setup_XH_ix.hMoI}
\alias{setup_XH_ix.hMoI}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_XH_ix}{hMoI}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_XH_ix} for the hybrid MoI model.
}
\keyword{internal}
