% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{setup_XH_ix}
\alias{setup_XH_ix}
\title{Add indices for human population to parameter list}
\usage{
setup_XH_ix(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}
}
