% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{setup_XH_ix.SIS}
\alias{setup_XH_ix.SIS}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_XH_ix}{SIS}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_XH_ix} for the SIS model.
}
