% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{setup_XH_obj.SIS}
\alias{setup_XH_obj.SIS}
\title{Set up \code{SIS} (\strong{XH})}
\usage{
\method{setup_XH_obj}{SIS}(Xname, residence, HPop, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{residence}{the residence vector}

\item{HPop}{the initial human population size}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up an \code{SIS} model object
for the \strong{XH} component
}
\keyword{internal}
