% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{setup_XH_obj.trivial}
\alias{setup_XH_obj.trivial}
\title{Set up \code{trivial} (\strong{XH})}
\usage{
\method{setup_XH_obj}{trivial}(Xname, residence, HPop, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{residence}{the residence vector}

\item{HPop}{the initial human population size}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_XH_obj} for the trivial model
}
\keyword{internal}
