% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-ports.R
\name{setup_XH_ports}
\alias{setup_XH_ports}
\title{Set up XH ports}
\usage{
setup_XH_ports(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Adds port objects and default values for the time spent matrix,
time away, and blood feeding search weights.
}
\keyword{internal}
