% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{setup_Xinits.SEIS}
\alias{setup_Xinits.SEIS}
\title{Setup Xinits.SEIS}
\usage{
\method{setup_Xinits}{SEIS}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xinits} for the SEIS model
}
