% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{setup_Xinits.SIS_xde}
\alias{setup_Xinits.SIS_xde}
\title{Setup Xinits.SIS_xde}
\usage{
\method{setup_Xinits}{SIS_xde}(pars, i, Xopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xinits} for the SIS_xde model
}
