% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{setup_Xix}
\alias{setup_Xix}
\title{Add indices for human population to parameter list}
\usage{
setup_Xix(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
