% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{setup_Xix.trivial}
\alias{setup_Xix.trivial}
\title{Add indices for human population to parameter list}
\usage{
\method{setup_Xix}{trivial}(pars, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{setup_Xix} for the trivial model.
}
