% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{setup_Xpar}
\alias{setup_Xpar}
\title{Setup an \strong{X} Module}
\usage{
setup_Xpar(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the parameter values and configure a model
for the \strong{X} Component
}
