% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{setup_Xpar.SIS_dts}
\alias{setup_Xpar.SIS_dts}
\title{Setup Xpar.SIS_dts}
\usage{
\method{setup_Xpar}{SIS_dts}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the SIS_dts model
}
