% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{setup_Xpar.SIS_xde}
\alias{setup_Xpar.SIS_xde}
\title{Setup Xpar.SIS_xde}
\usage{
\method{setup_Xpar}{SIS_xde}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_Xpar} for the SIS_xde model
}
