% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{setup_XY_interface}
\alias{setup_XY_interface}
\title{Setup the Blood Feeding Interface}
\usage{
setup_XY_interface(xds_obj, residency)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object template}

\item{residency}{the residency vector}
}
\value{
a \strong{\code{xds}} model object template with a blood feeding object
}
\description{
This function, called by \link{make_xds_object_template}, sets up
the blood feeding interface, \code{xds_obj$XY_interface.}
}
\details{
This implements a framework to model blood feeding
described by Wu SL, \emph{et al.}, (2023).

Modular computation in \strong{\code{ramp.xds}} requires a rigid interface to
guarantee mathematical consistency in computing quantites related to blood feeding and transmission.

The interface for blood feeding is defined by an object called
\code{XY_interface},
attached to the \strong{\code{xds}} model object as \code{xds_obj$XY_interface.}

The blood feeding interface sets up several objects:
\itemize{
\item the residency matrix \eqn{J} (see \link{make_residency_matrix})
\item a time spent (TiSp) matrix \eqn{\Theta} (see \link{setup_TimeSpent})
\item a circadian function \code{F_circadian} for each vector species
\item a time at risk (TaR) matrix \eqn{\Psi,} the product of the TiSp matrix and the circadian function
\item blood feeding search weights \eqn{\omega}
\item a vector describing the available of humans (or hosts), \eqn{W} (see \link{F_W_available})
\item a vector describing the available of visitors, \eqn{V}
\item a vector describing the available of other blood hosts, \eqn{O}
\item a vector describing the total available of blood hosts, \eqn{B} (see \link{F_B_available})
}

These quantities are used to compute the transmission matrix, \eqn{\beta}, to models transmission (see \code{\link[=setup_transmission]{setup_transmission()}}).

Mosquito bionomic parameters \emph{ought} to be constrained. If bionomic parameters are
assigned, there's no guarantee they are internally mathematically consistent or sensible.
To guarantee internal consistency, the the concept of resource available should be
used to compute the blood feeding rates (\eqn{f})
using \emph{functional responses.} The human fraction ought to be \eqn{q=W/B}.
available can also be used to model mosquito movement.

\strong{Mulit-Host Models}

In models with multiple host species, let \eqn{W_i} denote the available of the \eqn{i^{th}} host species.
Total availablity of blood hosts is \deqn{B = \sum_i W_i + O + V,}
so the the fraction of bites on each host species is \eqn{W_i/B}.

In models with multiple vector species and one host species, each vector species could have different search habits and preferences.
so blood feeding available is computed for each species, denoted \eqn{B_s} and \eqn{W_{s}}.

In models with multiple vector and multiple host species, \eqn{W_{i,s}}
is the available of the \eqn{i^{th}} host species to the \eqn{s^{th}} vector species.

For hosts, available is based on \emph{time spent} in each patch, and \emph{time at risk,} or
time spent by time of day weighted by mosquito species-specific \emph{search weights} reflecting different preferences
and a circadian function describing relative mosquito blood feeding rates by time of day.
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\seealso{
\link{setup_transmission}
}
