% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{Shock.no_shock}
\alias{Shock.no_shock}
\title{Set up shocks}
\usage{
\method{Shock}{no_shock}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Shock} for the no_shock model (do nothing)
}
