% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian.R
\name{show_F_circadian}
\alias{show_F_circadian}
\title{Get availability of circadian}
\usage{
show_F_circadian(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Get availability of circadian
}
