% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_pattern.R
\name{show_forcing}
\alias{show_forcing}
\title{Plot the forcing pattern}
\usage{
show_forcing(xds_obj, tm = seq(0, 730, by = 5), add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
the seasonal pattern, invisibly
}
\description{
Plot the composed time
series function forcing a model
}
