% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{show_season.eir}
\alias{show_season.eir}
\title{Plot the seasonal pattern}
\usage{
\method{show_season}{eir}(xds_obj, tm = seq(0, 730, by = 5), add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{`ramp.xds} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
the seasonal pattern, invisibly
}
\description{
For a model forced by the EIR
plot the seasonal pattern
}
