% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{show_season.none}
\alias{show_season.none}
\title{Plot the seasonal pattern}
\usage{
\method{show_season}{none}(xds_obj, tm = seq(0, 730, by = 5), add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
a null result
}
\description{
The \code{show_season} case for models
with no forcing.
}
