% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{show_season}
\alias{show_season}
\title{Plot the seasonal pattern}
\usage{
show_season(xds_obj, tm, add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{`ramp.xds} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
the seasonal pattern, invisibly
}
\description{
Plot the seasonal pattern
}
