% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend.eir}
\alias{show_trend.eir}
\title{Plot the temporal trend}
\usage{
\method{show_trend}{eir}(xds_obj, tm = seq(0, 3650, by = 10), add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{`ramp.xds} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
the temporal trend, invisibly
}
\description{
For a model forced by the EIR
plot the temporal trend
}
