% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend.Lambda}
\alias{show_trend.Lambda}
\title{Plot the temporal trend}
\usage{
\method{show_trend}{Lambda}(xds_obj, tm, add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{`ramp.xds} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
the temporal trend, invisibly
}
\description{
For a model forced by emergence,
show the temporal trend
}
