% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend.none}
\alias{show_trend.none}
\title{Plot the temporal trend}
\usage{
\method{show_trend}{none}(xds_obj, tm = seq(0, 3650, by = 10), add = FALSE)
}
\arguments{
\item{xds_obj}{a \strong{`ramp.xds} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}
}
\value{
a null result
}
\description{
The \code{show_trend} case for models
with no forcing.
}
