% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend}
\alias{show_trend}
\title{Plot the Fitted Trend}
\usage{
show_trend(model, dt = 10, add = FALSE)
}
\arguments{
\item{model}{a \strong{`ramp.xds} model object}

\item{dt}{the outputs interval}

\item{add}{add to existing plot}
}
\value{
a \strong{`ramp.xds} model object
}
\description{
Plot the Fitted Trend
}
