% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{SI}
\alias{SI}
\title{The \code{SI} Module for the MY Component}
\description{
Implements the \strong{MY} component using a simple SI (Susceptible-Infected)
model of adult mosquito infection dynamics.
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{Y}}{density of infected adult mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{eip}}{extrinsic incubation period (\eqn{\tau})}
\item{\code{g}}{mosquito mortality rate}
\item{\code{sigma}}{emigration rate}
\item{\code{mu}}{emigration loss rate}
\item{\code{K}}{mosquito dispersal matrix}
\item{\code{Omega}}{adult mosquito demographic matrix (mortality + migration)}
}
}

\section{Dynamics}{

The density of infectious mosquitoes is given by:
\deqn{Z = e^{-\Omega \tau} \cdot Y}
\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\
\end{array}}
}

