% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{skill_set_L.basicL}
\alias{skill_set_L.basicL}
\title{The \strong{L} Module Skill Set}
\usage{
\method{skill_set_L}{basicL}(Lname = "basicL")
}
\arguments{
\item{Lname}{the name of the \strong{L} module}
}
\value{
\emph{L} module skill set, as a list
}
\description{
The \strong{L} skill set is a list of
an module's capabilities
}
