% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{skill_set_L}
\alias{skill_set_L}
\title{The skill set (\strong{L} module)}
\usage{
skill_set_L(Lname)
}
\arguments{
\item{Lname}{the \strong{L} module name}
}
\value{
\emph{L} module skill set, as a list
}
\description{
The skill set for an \strong{L} module
is a list of that summarizes capabilities and
compatabilities
}
