% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{skill_set_MY.GeRM}
\alias{skill_set_MY.GeRM}
\title{The \strong{GeRM} Module Skill Set}
\usage{
\method{skill_set_MY}{GeRM}(MYname)
}
\arguments{
\item{MYname}{the \strong{MY} module name}
}
\value{
\emph{MY} module skill set, as a list
}
\description{
The \strong{MY} skill set is a list of
an module's capabilities:
\itemize{
\item \code{demography} is
}
}
