% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{skill_set_XH.hMoI}
\alias{skill_set_XH.hMoI}
\title{The \strong{XH} Module Skill Set}
\usage{
\method{skill_set_XH}{hMoI}(Xname = "hMoI")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the skill set, as a list
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
