% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{skill_set_XH}
\alias{skill_set_XH}
\title{The \strong{XH} Module Skill Set}
\usage{
skill_set_XH(Xname)
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
\emph{XH} module skill set, as a list
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities:
\itemize{
\item \code{demography} is
}
}
