% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{skill_set_XH.SIS}
\alias{skill_set_XH.SIS}
\title{The \strong{XH} Module Skill Set}
\usage{
\method{skill_set_XH}{SIS}(Xname = "SIS")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the \code{SIS} \emph{XH} module skill set, a list
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities.
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
\keyword{internal}
