% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{skill_set_XH.trivial}
\alias{skill_set_XH.trivial}
\title{The \strong{XH} Module Skill Set}
\usage{
\method{skill_set_XH}{trivial}(Xname)
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
a list describing the skill set
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities
}
