% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-spray-interface.R
\name{SprayArea}
\alias{SprayArea}
\title{Set the spray_area_lsm}
\usage{
SprayArea(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
spray_area_lsm
}
