% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses-none.R
\name{SprayHouses.none}
\alias{SprayHouses.none}
\title{Set no spray_houses}
\usage{
\method{SprayHouses}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for spray_houses
}
