% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irs-spray_houses-interface.R
\name{SprayHouses}
\alias{SprayHouses}
\title{Set the spray_houses_irs}
\usage{
SprayHouses(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
spray_houses_irs
}
