% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{steady_state_L}
\alias{steady_state_L}
\title{Compute steady states for \strong{L} Component Modules}
\usage{
steady_state_L(eta, L_obj)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{L_obj}{a list that defines an xde model}
}
\value{
a named \link{list}: values of the state variables at the steady state
}
\description{
For differential equation models, compute
steady states as a function of daily eggs laid, \eqn{\eta}
}
\note{
This method dispatches on the class of \code{L_obj}.
}
