% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{steady_state_L.trivial}
\alias{steady_state_L.trivial}
\title{Compute the Steady State of \code{dLdt.trivial} (\strong{L} Component)}
\usage{
\method{steady_state_L}{trivial}(eta, L_obj)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{L_obj}{a list that defines an xde model}
}
\value{
none
}
\description{
Return a numeric vector of length 0
}
