% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{steady_state_M.basicM_ode}
\alias{steady_state_M.basicM_ode}
\title{Compute the steady states as a function of the daily EIR}
\usage{
\method{steady_state_M}{basicM_ode}(Lambda, xds_obj, s = 1)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MY_obj}
}
