% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{steady_state_M}
\alias{steady_state_M}
\title{Compute steady states for \strong{M}}
\usage{
steady_state_M(Lambda, MY_obj)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{MY_obj}{a list that defines an adult model}
}
\value{
the steady states
}
\description{
Given the emergence rate of adult mosquitoes (\eqn{\Lambda} or \code{Lambda}) compute
the population density of adult mosquitoes at the steady state
}
