% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{steady_state_MY.GeM}
\alias{steady_state_MY.GeM}
\title{Compute the steady states as a function of the daily EIR}
\usage{
\method{steady_state_MY}{GeM}(Lambda, kappa, xds_obj, s = 1)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MY_obj}.
}
\keyword{internal}
