% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{steady_state_MY}
\alias{steady_state_MY}
\title{Compute steady states for \strong{MY}}
\usage{
steady_state_MY(Lambda, kappa, MY_obj)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{MY_obj}{a list that defines an adult model}
}
\value{
the steady states
}
\description{
Given the emergence rate of adult mosquitoes (\eqn{\Lambda} or \code{Lambda}) and the
net infectiousness of humans (\eqn{\kappa} or \code{kappa}), compute the steady states
for the system
}
