% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{steady_state_X.hMoI}
\alias{steady_state_X.hMoI}
\title{Compute Steady States for \code{hMoI} (\strong{X}-Model)}
\usage{
\method{steady_state_X}{hMoI}(foi, H, XH_obj)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{XH_obj}{an \strong{XH} model object}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the hMoI model as a function of the daily foi
}
