% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{steady_state_X}
\alias{steady_state_X}
\title{Steady States for \strong{X}}
\usage{
steady_state_X(foi, H, XH_obj)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{XH_obj}{an \strong{XH} model object}
}
\value{
steady states
}
\description{
Compute the steady states as a function of the daily FoI for a
static value of human population density
}
