% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{steady_state_XH.trivial}
\alias{steady_state_XH.trivial}
\title{Compute the steady states for the trivial model as a function of the daily EIR}
\usage{
\method{steady_state_XH}{trivial}(foi, XH_obj)
}
\arguments{
\item{foi}{the daily FoI}

\item{XH_obj}{an \strong{XH} model object}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the trivial model as a function of the daily eir.
}
