% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{steady_state_Y}
\alias{steady_state_Y}
\title{Compute steady states for \strong{Y}}
\usage{
steady_state_Y(M, kappa, MY_obj)
}
\arguments{
\item{M}{adult mosquito population density}

\item{kappa}{net infectiousness}

\item{MY_obj}{a list that defines an adult model}
}
\value{
none
}
\description{
Given adult mosquito population density (\eqn{M} or \code{M}) and the
net infectiousness of humans (\eqn{\kappa} or \code{kappa}), compute the steady states
for the other states
}
