% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Sugar.none}
\alias{Sugar.none}
\title{Set the values of exogenous variables}
\usage{
\method{Sugar}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{dependent variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After none setup, no exogenous
variables are configured so sugar returns
the \strong{\code{xds}} object without modification
}
