% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar.R
\name{Sugar}
\alias{Sugar}
\title{Set up Mosquito sugar}
\usage{
Sugar(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{dependent variables vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Create a junction
for mosquito sugar
}
\note{
Mosquito trap models are in
\code{ramp.control}
}
