% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-func.R
\name{SugarBaitCoverage.func}
\alias{SugarBaitCoverage.func}
\title{Set up sugar bait coverage function}
\usage{
\method{SugarBaitCoverage}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
Set up sugar bait coverage function
}
