% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-none.R
\name{SugarBaitCoverage.none}
\alias{SugarBaitCoverage.none}
\title{Set no sugar bait coverage}
\usage{
\method{SugarBaitCoverage}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The model for no sugar bait coverage
}
