% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-coverage-interface.R
\name{SugarBaitCoverage}
\alias{SugarBaitCoverage}
\title{Set the sugar bait coverage}
\usage{
SugarBaitCoverage(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the coverage levels of
sugar baits
}
