% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effects-none.R
\name{SugarBaitEffects.none}
\alias{SugarBaitEffects.none}
\title{Set up no sugar bait effects}
\usage{
\method{SugarBaitEffects}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up no sugar bait effects
}
