% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effects-interface.R
\name{SugarBaitEffects}
\alias{SugarBaitEffects}
\title{Compute sugar bait effects}
\usage{
SugarBaitEffects(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the effects of a sugar bait
}
