% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effectsizes-none.R
\name{SugarBaitEffectSizes.none}
\alias{SugarBaitEffectSizes.none}
\title{No sugar bait effect sizes}
\usage{
\method{SugarBaitEffectSizes}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The no sugar bait model for sugar bait effect sizes
}
