% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits.R
\name{SugarBaitEffectSizes}
\alias{SugarBaitEffectSizes}
\title{Set the SugarBaitEffectSizes}
\usage{
SugarBaitEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
Sugar Baits
}
