% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait.R
\name{SugarBaits.dynamic}
\alias{SugarBaits.dynamic}
\title{Set no sugar baits}
\usage{
\method{SugarBaits}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
