% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits.R
\name{SugarBaits.none}
\alias{SugarBaits.none}
\title{No sugar baits}
\usage{
\method{SugarBaits}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
