% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{Temperature.func}
\alias{Temperature.func}
\title{Set no temperature}
\usage{
\method{Temperature}{func}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for temperature
}
