% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{time_away}
\alias{time_away}
\title{time spent}
\usage{
time_away(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute and store a
time spent
}
\keyword{internal}
