% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{time_away.static}
\alias{time_away.static}
\title{Compute time spent}
\usage{
\method{time_away}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the time spent objecTA unmodified
}
\keyword{internal}
