% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{time_spent}
\alias{time_spent}
\title{Time Spent}
\description{
The risk of exposure to mosquito-borne pathogens
is related to time spent in places where vectors are
blood feeding. In \strong{\code{ramp.xds},} this is implemented
through the concepts of \strong{time spent} and \strong{time at risk}
(see \link{time_at_risk})
}
