% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespent.R
\name{timespent}
\alias{timespent}
\title{time spent}
\usage{
timespent(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute and store a
time spent matrix
}
\keyword{internal}
