% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{TimeSpent.setup}
\alias{TimeSpent.setup}
\title{Compute time spent objects: setup for static models}
\usage{
\method{TimeSpent}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This sets up host available, \eqn{W},
total blood host available, \eqn{B},
the time spent matrix \eqn{\Theta}, and the time-at-risk matrix \eqn{\Psi}
for static models.
}
\details{
The mixing matrix, \eqn{\beta}, depends on
time spent terms, so the class of \code{xds_obj$beta} must also
be updated, if they are not dynamic, so \link{trigger_setup} is called.
}
