% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission.dynamic}
\alias{Transmission.dynamic}
\title{Compute transmission, the dynamic case}
\usage{
\method{Transmission}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Compute transmission terms with a dynamic mixing matrix
}
