% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission}
\alias{Transmission}
\title{Transmission}
\usage{
Transmission(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Transmission computes:
\itemize{
\item \code{beta} -- the bite distribution matrix
\item \code{eir} -- the daily EIR
\item \code{kappa} -- the net infectiousness
\item \link{local_frac}
}

This method dispatches on the type of \code{xds_obj$beta}
}
\keyword{internal}
