% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission.static}
\alias{Transmission.static}
\title{Compute transmission terms with a static mixing matrix}
\usage{
\method{Transmission}{static}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
In the static case, \eqn{\beta} is not updated
}
