% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps.R
\name{Traps.none}
\alias{Traps.none}
\title{Set the values of exogenous variables}
\usage{
\method{Traps}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After none setup, no exogenous
variables are configured so traps returns
the \strong{\code{xds}} object without modification
}
