% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traps.R
\name{Traps}
\alias{Traps}
\title{Set up Mosquito traps}
\usage{
Traps(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Create a junction
for mosquito traps
}
\note{
Mosquito trap models are in
\code{ramp.control}
}
