% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel}
\alias{Travel}
\title{Time Spent F_travel}
\usage{
Travel(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a parameter
describing time spent F_travel. It is computed in \link{compute_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
