% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{Travel.setup}
\alias{Travel.setup}
\title{Time Spent F_travel}
\usage{
\method{Travel}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets a static value for the parameter
describing time spent F_travel, resets the class to \code{static} and
then triggers an update to the blood feeding model.
It is computed in \link{compute_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
