% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{travel_dynamics}
\alias{travel_dynamics}
\title{Time Spent Here}
\usage{
travel_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a term that
is the time spent \emph{here,} stored as \code{time_at_home}
on the \code{XY_interface.} Here is defined as time
spent in one of the patches in this model, so
travle is time spent elsewhere.

A function \code{F_travel} computes the fraction of time spent traveling
stores its complement as \code{time_at_home.}
}
\note{
The quantity \code{time_at_home} is part of the \code{XY_interface}
It is used to compute the Time at Risk matrix, and it is also
used to weight local exposure \emph{vs.}  traveling exposure.
}
