% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{travel_eir.static}
\alias{travel_eir.static}
\title{Compute travel eir}
\usage{
\method{travel_eir}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the travel eir objecTeir unmodified
}
\keyword{internal}
