% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria-multiday.R
\name{travel_malaria.multiday}
\alias{travel_malaria.multiday}
\title{A model for the travel Attack Rate}
\usage{
\method{travel_malaria}{multiday}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{numeric}
}
\description{
Implements \link{travel_malaria} through a model for the travel Attack Rate
}
