% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{traveling.dynamic}
\alias{traveling.dynamic}
\title{A model for the travel FoI}
\usage{
\method{traveling}{dynamic}(t, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}#}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{traveling} through a model for the travel FoI
}
