% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{traveling}
\alias{traveling}
\title{A model for the travel FoI}
\usage{
traveling(t, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}#}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{traveling}. The function is
called by \link{make_TaR} to set the fraction of time spent
traveling
}
