% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{traveling.static}
\alias{traveling.static}
\title{Time Spent Traveling}
\usage{
\method{traveling}{static}(t, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}#}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
For a static model for time spent traveling, the function
does not update anything.
}
