% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-treat_habitats-none.R
\name{TreatHabitats.none}
\alias{TreatHabitats.none}
\title{Set no treat_habitats}
\usage{
\method{TreatHabitats}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for treat_habitats
}
