% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-treat_habitats-interface.R
\name{TreatHabitats}
\alias{TreatHabitats}
\title{Set the treat_habitats_lsm}
\usage{
TreatHabitats(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
treat_habitats_lsm
}
