% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_setup.R
\name{trigger_setup}
\alias{trigger_setup}
\title{Trigger \code{setup}}
\usage{
trigger_setup(obj)
}
\arguments{
\item{obj}{an object}
}
\value{
the \code{obj}
}
\description{
For various \code{S3} functions, change
the class of the dispatching object from \code{static}
to \code{setup}
}
