% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_setup.R
\name{trigger_setup.setup}
\alias{trigger_setup.setup}
\title{Trigger \code{setup}}
\usage{
\method{trigger_setup}{setup}(obj)
}
\arguments{
\item{obj}{part of an \strong{\code{xds}} object}
}
\value{
the \code{obj}
}
\description{
For various \code{S3} functions, change
leave the class of the dispatching object as \code{setup}
}
