% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trigger_setup.R
\name{trigger_setup.static}
\alias{trigger_setup.static}
\title{Trigger \code{setup}}
\usage{
\method{trigger_setup}{static}(obj)
}
\arguments{
\item{obj}{part of an \strong{\code{xds}} object}
}
\value{
the \code{obj}
}
\description{
For various \code{S3} functions, change
the class of the dispatching object from \code{static}
to \code{setup}
}
