% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{trivial_MY}
\alias{trivial_MY}
\title{\code{trivial} --- \strong{MY} module}
\description{
The trivial \strong{MY} module configures two trace functions:
\itemize{
\item infectious biting:
\deqn{F_{fqZ}(t) = fqZ S(t) T(t) K(t)}
\item egg laying:
\deqn{F_G(t) = G S(t) T(t) K(t)}
}

where
\itemize{
\item \eqn{G} or \code{eggs} is the mean egg laying rate
\item \eqn{fqZ} or \code{fqZ} is the mean number of infectious bites on humans, per patch
\item \eqn{S(t)} or \code{F_season} is a seasonal pattern
\item \eqn{T(t)} or \code{F_trend} is a trend pattern
\item \eqn{K(t)} or \code{F_shock} is a perturbation
}
}
\section{Parameters}{

\describe{
\item{\code{eggs}}{the mean number of eggs laid, per day}
\item{\code{Z}}{the mean density of infectious mosquitoes}
\item{\code{f}}{the blood feeding rate}
\item{\code{q}}{the human fraction}
\item{\code{F_season}}{a seasonal pattern function, \eqn{{S(t)}}}
\item{\code{F_trend}}{a trend function, \eqn{T(t)}}
\item{\code{F_shock}}{a shock function, \eqn{K(t)}}
}

The default values are \code{F_season=F_trend=F_shock=F_one}

Setup also adds the objects \code{season_par} and \code{trend_par} and \code{shock_par} for
use by \code{ramp.trace}

For the bionomic parameters, \code{f=q=Z=eggs=1}.
}

\section{Get}{

\itemize{
\item \code{get_MY_pars} --- the \code{trivial} method returns all the parameters
\item \code{get_mean_forcing} --- returns \code{Z}
\item \code{get_f} --- returns \code{f}
\item \code{get_q} --- returns \code{q}
\item \code{get_season} --- returns \code{season_par}
\item \code{get_trend}  --- returns \code{trend_par}
\item \code{get_shock}  --- returns \code{shock_par}
}

Note: use \code{get_MY_pars} to inspect \code{eggs}
}

\section{Change}{

\itemize{
\item \code{change_MY_pars} --- change bionomic parameters by name
\item \code{change_mean_forcing} --- changes \code{Z}
\item \code{change_F_season} --- changes \code{F_season}
\item \code{change_F_trend}  --- changes \code{F_trend}
\item \code{change_F_shock}  --- changes \code{F_shock}
}
}

\section{Notes}{

\enumerate{
\item The module has no state variables.
\item The size of an object saved by \code{saveRDS} balloons if it saves a function,
so \code{saveXDS} function strips the functions and \code{readRDS} remakes the function
from the stored parameters.
}

\code{F_season}, \code{F_trend}, and \code{F_shock} can be set up manually by passing any
user defined function. If so, the user should use \code{saveRDS} and \code{readRDS}
rather than \code{saveXDS} and \code{readXDS}
}

