% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-bad.R
\name{update_bad_habitats}
\alias{update_bad_habitats}
\title{Update other blood hosts}
\usage{
update_bad_habitats(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the bad_habitats, \eqn{\Theta}.
Dispatches on \code{class(xds_obj$patches$bad_habitats_obj[[s]])}.
}
\keyword{internal}
