% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{update_blood_search_weights}
\alias{update_blood_search_weights}
\title{Update the blood_search_weights Matrix}
\usage{
update_blood_search_weights(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the blood_search_weights, \eqn{w}.
Dispatches on \code{xds_obj$XH_obj[[i]]$blood_search_weights_obj}.
}
\keyword{internal}
