% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-search_weights.R
\name{update_blood_search_weights.static}
\alias{update_blood_search_weights.static}
\title{Update the blood_search_weights}
\usage{
\method{update_blood_search_weights}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{static} method returns \code{xds_obj} unmodified
}
\keyword{internal}
