% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{update_F_season.Lambda}
\alias{update_F_season.Lambda}
\title{Update the seasonality function}
\usage{
\method{update_F_season}{Lambda}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Update \code{F_season}
}
