% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_F_season.R
\name{update_F_season}
\alias{update_F_season}
\title{Update the seasonality function}
\usage{
update_F_season(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Update \code{F_season}
}
