% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_trend.R
\name{update_F_trend}
\alias{update_F_trend}
\title{Update the trend function}
\usage{
update_F_trend(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Update \code{F_trend}
}
