% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_function_X.R
\name{update_function_X.mean_forcing}
\alias{update_function_X.mean_forcing}
\title{Update Mean Forcing}
\usage{
\method{update_function_X}{mean_forcing}(X, ix, xds_obj, update = "mean_forcing", s = 1)
}
\arguments{
\item{X}{new parameter values}

\item{ix}{indices or a list}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{update}{a string to dispatch}

\item{s}{the species index}
}
\value{
sum of squared differences
}
\description{
Update the mean value
that forces a xds_obj. This dispatches
on class(xds_obj$forced_by)
}
