% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_function_X.R
\name{update_function_X.season}
\alias{update_function_X.season}
\title{Update \code{F_season}}
\usage{
\method{update_function_X}{season}(X, ix, xds_obj, update = "season", s = 1)
}
\arguments{
\item{X}{new parameter values}

\item{ix}{indices or a list}

\item{xds_obj}{a \strong{\code{ramp.xds}} xds_obj object}

\item{update}{a string to dispatch}

\item{s}{the species index}
}
\value{
sum of squared differences
}
\description{
This updates three shape parameters
for a function \code{F_season} using \link{modify_vector_X} and \link{update_F_season}
}
\note{
This assumes the vector, \code{X}, has got
three sets of parameters describing three
parameters in a seasonal function with
parameter names phase, bottom, and pw.
}
