% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{update_habitat_search_weights}
\alias{update_habitat_search_weights}
\title{Update the habitat_search_weights Matrix}
\usage{
update_habitat_search_weights(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the habitat_search_weights, \eqn{w}.
Dispatches on \code{xds_obj$L_obj[[s]]$habitat_search_obj}.
}
\keyword{internal}
