% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-search_weights.R
\name{update_habitat_search_weights.static}
\alias{update_habitat_search_weights.static}
\title{Update the habitat_search_weights}
\usage{
\method{update_habitat_search_weights}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{static} method returns \code{xds_obj} unmodified
}
\keyword{internal}
