% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-basicM-xde.R
\name{update_inits_MYZ.basicM_xde}
\alias{update_inits_MYZ.basicM_xde}
\title{Make inits for RM adult mosquito model}
\usage{
\method{update_inits_MYZ}{basicM_xde}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Make inits for RM adult mosquito model
}
