% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-dts.R
\name{update_inits_MYZ.RM_dts}
\alias{update_inits_MYZ.RM_dts}
\title{Make inits for RM_dts adult mosquito model}
\usage{
\method{update_inits_MYZ}{RM_dts}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for RM_dts adult mosquito model
}
