% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{update_inits_MYZ.trace}
\alias{update_inits_MYZ.trace}
\title{Update inits for trace}
\usage{
\method{update_inits_MYZ}{trace}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for trace
}
