% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{update_inits_X.SEIS}
\alias{update_inits_X.SEIS}
\title{Update inits for the SEIS xde human model from a vector of states}
\usage{
\method{update_inits_X}{SEIS}(pars, y0, i)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SEIS xde human model from a vector of states
}
