% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{update_interfaces}
\alias{update_interfaces}
\title{Update interfaces}
\usage{
update_interfaces(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update interfaces
}
