% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{update_Linits.basicL}
\alias{update_Linits.basicL}
\title{Update Initial Values for \code{basicL} from a state vector \code{y}}
\usage{
\method{update_Linits}{basicL}(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Extract the values of the variable \eqn{L} from
a state vector \code{y} and use them to set the initial value for \eqn{L}
}
