% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{update_Linits}
\alias{update_Linits}
\title{Set the initial values from a vector of model states}
\usage{
update_Linits(pars, y0, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the class of \code{pars$Lpar}.
}
