% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{update_Linits.trace}
\alias{update_Linits.trace}
\title{Update inits for the basic aquatic mosquito competition model}
\usage{
\method{update_Linits}{trace}(pars, y0, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for the basic aquatic mosquito competition model
}
