% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{update_Linits.trivial}
\alias{update_Linits.trivial}
\title{Update inits for the basic aquatic mosquito competition model}
\usage{
\method{update_Linits}{trivial}(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for the basic aquatic mosquito competition model
}
