% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{Update_Lt}
\alias{Update_Lt}
\title{Update state variables (\strong{L} module)}
\usage{
Update_Lt(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the states, a \link{numeric} vector of length \eqn{n_q=}\code{nHabitats}
}
\description{
This method updates the state variables
for \strong{L} Component modules
}
\note{
Dispatches on the class of \code{xds_obj$L_obj[[s]]}
}
