% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{Update_MYt.basicM}
\alias{Update_MYt.basicM}
\title{Update state variables for \code{basicM} (\strong{MY})}
\usage{
\method{Update_MYt}{basicM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_MYt} for the basicM model.
}
\keyword{internal}
