% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{Update_MYt}
\alias{Update_MYt}
\title{Update States for a Adult Mosquito Model}
\usage{
Update_MYt(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
Updated states for an adult mosquito model, a \link{vector}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj[[s]]}.
}
