% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-RMdts.R
\name{Update_MYt.RMdts}
\alias{Update_MYt.RMdts}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYt}{RMdts}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_MYt} for the RMdts model.
}
