% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{Update_MYt.trivial}
\alias{Update_MYt.trivial}
\title{Update state variables for \code{trivial} (\strong{MY})}
\usage{
\method{Update_MYt}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length 0
}
\description{
Implements \link{Update_MYt} for the trivial (forced emergence) model.
}
\keyword{internal}
