% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{update_MYZinits.GeRM}
\alias{update_MYZinits.GeRM}
\title{Update inits for GeRM adult mosquito model}
\usage{
\method{update_MYZinits}{GeRM}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Update inits for GeRM adult mosquito model
}
