% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-macdonald.R
\name{update_MYZinits.macdonald}
\alias{update_MYZinits.macdonald}
\title{Update inits for macdonald adult mosquito model}
\usage{
\method{update_MYZinits}{macdonald}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Update inits for macdonald adult mosquito model
}
