% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SII.R
\name{update_MYZinits.MY}
\alias{update_MYZinits.MY}
\title{Make inits for \code{MY} adult mosquito model}
\usage{
\method{update_MYZinits}{MY}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for \code{MY} adult mosquito model
}
