% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{update_MYZinits.RM}
\alias{update_MYZinits.RM}
\title{Update inits for RM adult mosquito model}
\usage{
\method{update_MYZinits}{RM}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Update inits for RM adult mosquito model
}
