% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{update_MYZinits.si}
\alias{update_MYZinits.si}
\title{Make inits for si adult mosquito model}
\usage{
\method{update_MYZinits}{si}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for si adult mosquito model
}
