% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{update_MYZinits.trivial}
\alias{update_MYZinits.trivial}
\title{Update inits for trivial}
\usage{
\method{update_MYZinits}{trivial}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for trivial
}
