% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-bionomics.R
\name{update_MYZpar_RM.modified}
\alias{update_MYZpar_RM.modified}
\title{Update Ross-Macdonald Bionomic Parameters}
\usage{
\method{update_MYZpar_RM}{modified}(MYZpar)
}
\arguments{
\item{MYZpar}{an \code{xds} adult mosquito  model object}
}
\value{
the model as a \link{list}
}
\description{
If there is no modification, return
the parameters without modification
}
