% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM-bionomics.R
\name{update_MYZpar_RM}
\alias{update_MYZpar_RM}
\title{Update Ross-Macdonald Bionomic Parameters}
\usage{
update_MYZpar_RM(MYZpar)
}
\arguments{
\item{MYZpar}{an \code{xds} adult mosquito  model object}
}
\value{
the model as a \link{list}
}
\description{
Update bionomic parameter values as the product of the baseline and
the effect sizes
}
