% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{Update_MYZt.GeRM}
\alias{Update_MYZt.GeRM}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYZt}{GeRM}(t, y, pars, s)

\method{Update_MYZt}{GeRM}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector

a \link{numeric} vector
}
\description{
Implements \link{Update_MYZt} for the GeRM_dts model.

Implements \link{Update_MYZt} for the GeRM_dts model.
}
