% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R, R/runtime.R
\name{Update_MYZt}
\alias{Update_MYZt}
\title{Derivatives for adult mosquitoes}
\usage{
Update_MYZt(t, y, pars, s)

Update_MYZt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
the derivatives a \link{vector}

a \link{vector} containing the vector of all state derivatives
}
\description{
This method dispatches on the type of \code{pars$MYZpar}.

Compute and update the state variables for
a model with only humans
}
