% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{Update_MYZt.RM}
\alias{Update_MYZt.RM}
\title{Derivatives for adult mosquitoes}
\usage{
Update_MYZt.RM(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_MYZt} for the RM_dts model.
}
