% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{Update_MYZt.si_old}
\alias{Update_MYZt.si_old}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYZt}{si_old}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_MYZt} for the si_old model.
}
