% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{Update_MYZt.trivial}
\alias{Update_MYZt.trivial}
\title{Handle state updates for the \code{trivial} \strong{MYZ} module}
\usage{
\method{Update_MYZt}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a null vector
}
\description{
Implements \link{Update_MYZt} for the trivial (forced emergence) model.
}
