% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-demography.R
\name{update_Omega.dts}
\alias{update_Omega.dts}
\title{Make the mosquito demography matrix}
\usage{
\method{update_Omega}{dts}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
Make the mosquito demography matrix
}
\note{
This method dispatches on the type of \code{pars$MYZpar[[s]]}
}
