% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{update_Omega.dynamic}
\alias{update_Omega.dynamic}
\title{Change Omega}
\usage{
\method{update_Omega}{dynamic}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Update a static demographic matrix
}
\keyword{internal}
