% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{update_Omega_dts.dynamic}
\alias{update_Omega_dts.dynamic}
\title{Update Omega for dts}
\usage{
\method{update_Omega_dts}{dynamic}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Compute \eqn{\Omega}
}
\keyword{internal}
