% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{update_Omega_dts}
\alias{update_Omega_dts}
\title{Update Omega for dts}
\usage{
update_Omega_dts(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Update the demographic matrix
for differential equations, if needed
}
\keyword{internal}
