% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{update_Omega_dts.static}
\alias{update_Omega_dts.static}
\title{Update Omega for dts}
\usage{
\method{update_Omega_dts}{static}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
The static case for \link{update_Omega_dts}
}
\keyword{internal}
