% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{update_Omega_xde.setup}
\alias{update_Omega_xde.setup}
\title{Update Omega for xde}
\usage{
\method{update_Omega_xde}{setup}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Compute \eqn{\Omega} and change the class to "setup"
}
\keyword{internal}
