% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{update_other_blood_hosts}
\alias{update_other_blood_hosts}
\title{Update other blood hosts}
\usage{
update_other_blood_hosts(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the other_blood_hosts, \eqn{\Theta}.
Dispatches on \code{class(xds_obj$XY_interface$other_blood_hosts_obj[[s]])}.
}
\keyword{internal}
