% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-other_hosts.R
\name{update_other_blood_hosts.static}
\alias{update_other_blood_hosts.static}
\title{Update other blood hosts}
\usage{
\method{update_other_blood_hosts}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{static} method returns \code{xds_obj} unmodified
}
\keyword{internal}
