% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats-traps.R
\name{update_ovitraps.static}
\alias{update_ovitraps.static}
\title{Update ovitraps}
\usage{
\method{update_ovitraps}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{static} method returns \code{xds_obj} unmodified
}
\keyword{internal}
