% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{update_time_away.dynamic}
\alias{update_time_away.dynamic}
\title{Update time_away}
\usage{
\method{update_time_away}{dynamic}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update time_away when it changes dynamically
}
\keyword{internal}
