% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{update_time_away}
\alias{update_time_away}
\title{Update the time spent}
\usage{
update_time_away(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the time spent , \eqn{\Theta}.
Dispatches on \code{class(xds_obj$XY_interface$time_away_obj[[i]])}.
}
\keyword{internal}
