% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{update_time_away.setup}
\alias{update_time_away.setup}
\title{Update Time Away (setup)}
\usage{
\method{update_time_away}{setup}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Acknowledges a one-time update to \code{time_away}
and sets the port back to \code{"static"}.
}
\keyword{internal}
