% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-time_away.R
\name{update_time_away.static}
\alias{update_time_away.static}
\title{Update time_away}
\usage{
\method{update_time_away}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A utility to ensure that the time_away
is updated properly after a change
}
\keyword{internal}
