% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{update_timespent.setup}
\alias{update_timespent.setup}
\title{Update the time spent Matrix (setup)}
\usage{
\method{update_timespent}{setup}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Acknowledges a one-time update to the time spent matrix
and sets the port back to \code{"static"}.
}
\keyword{internal}
