% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timespent.R
\name{update_timespent.static}
\alias{update_timespent.static}
\title{Update timespent}
\usage{
\method{update_timespent}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A utility to ensure that the timespent
is updated properly after a change
}
\keyword{internal}
