% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{update_travel_eir}
\alias{update_travel_eir}
\title{Update the travel eir}
\usage{
update_travel_eir(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the travel eir , \eqn{\Theta}.
Dispatches on \code{class(xds_obj$XY_interface$travel_eir_obj[[i]])}.
}
\keyword{internal}
