% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-travel_eir.R
\name{update_travel_eir.static}
\alias{update_travel_eir.static}
\title{Update travel_eir}
\usage{
\method{update_travel_eir}{static}(t, y, xds_obj, i)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A utility to ensure that the travel_eir
is updated properly after a change
}
\keyword{internal}
