% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{update_vis_kappa}
\alias{update_vis_kappa}
\title{Update NI of visitors}
\usage{
update_vis_kappa(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function to compute the availability of vis_kappa.
Dispatches on \code{class(xds_obj$patches$vis_kappa_obj[[s]])}.
}
\keyword{internal}
