% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{update_visitors}
\alias{update_visitors}
\title{Update visitors}
\usage{
update_visitors(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function to compute the availability of visitors.
}
\keyword{internal}
