% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{Update_XHt}
\alias{Update_XHt}
\title{Update X states for a discrete time system}
\usage{
Update_XHt(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
