% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{update_Xinits.hMoI}
\alias{update_Xinits.hMoI}
\title{Update inits for hybrid MoI human model from a vector of states}
\usage{
\method{update_Xinits}{hMoI}(pars, y0, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for hybrid MoI human model from a vector of states
}
