% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{update_Xinits}
\alias{update_Xinits}
\title{Set the initial values from a vector of states}
\usage{
update_Xinits(pars, y0, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
