% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{update_Xinits.SEISd}
\alias{update_Xinits.SEISd}
\title{Update inits for the SEISd xde human model from a vector of states}
\usage{
\method{update_Xinits}{SEISd}(pars, y0, i)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of initial values}

\item{i}{the host species index}
}
\value{
none
}
\description{
Update inits for the SEISd xde human model from a vector of states
}
