% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{Update_Xt}
\alias{Update_Xt}
\title{Update X states for a discrete time system}
\usage{
Update_Xt(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
