% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{Update_Xt.SEISd}
\alias{Update_Xt.SEISd}
\title{DTS updating for the SEISd model for human / vertebrate host infections}
\usage{
\method{Update_Xt}{SEISd}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Xt} for the SEISd model
}
