% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP.R
\name{Update_Xt.SIP}
\alias{Update_Xt.SIP}
\title{Derivatives for human population}
\usage{
\method{Update_Xt}{SIP}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Xt} for the SIP model.
}
