% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{Update_Xt.SIS}
\alias{Update_Xt.SIS}
\title{DTS updating for the SIS model for human / vertebrate host infections}
\usage{
Update_Xt.SIS(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Xt} for the SIS model
}
