% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{Update_Xt.trivial}
\alias{Update_Xt.trivial}
\title{Handle State Updating for the \code{trivial} \strong{X}-Module}
\usage{
\method{Update_Xt}{trivial}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
The trivial model has no state variables so it returns
a numeric vector of length 0
}
