% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-no_net.R
\name{UseBedNet.no_nets}
\alias{UseBedNet.no_nets}
\title{Bed net ownership}
\usage{
\method{UseBedNet}{no_nets}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{vector of state variables}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{UseBedNet} for the no_nets model of bed nets (do nothing)
}
