% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-use-none.R
\name{UseBedNets.none}
\alias{UseBedNets.none}
\title{Set no use_bednets}
\usage{
\method{UseBedNets}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for use_bednets
}
