% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control-no_control.R
\name{VectorControl.no_control}
\alias{VectorControl.no_control}
\title{Distribute vector control}
\usage{
\method{VectorControl}{no_control}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControl} for the no_control model of vector control (do nothing)
}
