% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl1.none}
\alias{VectorControl1.none}
\title{Compute Vector Control Effect Sizes}
\usage{
\method{VectorControl1}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Do no vector control:
the \strong{\code{xds}} object is returned unmodified
}
\note{
No control or \code{none} is the default setting
}
