% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl1}
\alias{VectorControl1}
\title{Vector Control, Stage 1}
\usage{
VectorControl1(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
A junction to implement
stage one vector control functions.
}
\note{
Non-trivial vector control modules are in
\href{https://github.com/dd-harp/ramp.control}{\strong{\code{ramp.control}}}.
}
