% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControlEffectSizes.none}
\alias{VectorControlEffectSizes.none}
\title{Set the values of exogenous variables}
\usage{
\method{VectorControlEffectSizes}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
After none setup, no exogenous
variables are configured so the function returns
the unmodified \strong{\code{xds}} object
}
