% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{VectorControlEffectSizes}
\alias{VectorControlEffectSizes}
\title{Apply vector control effect sizes}
\usage{
VectorControlEffectSizes(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list}
}
\description{
This calls \link{MEffectSizes} and \link{LEffectSizes}
for each species.
This function applies the effect sizes of vector control to the bionomic
parameters for each vector species.
}
\keyword{internal}
