% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{view_residence_matrix}
\alias{view_residence_matrix}
\title{View residence membership}
\usage{
view_residence_matrix(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
a \code{nStrata} \eqn{\times 2} \link{matrix}: col 1 is the residence index, and col 2 is the patch index
}
\description{
Shows the residence membership information from \eqn{\cal J}
}
\seealso{
\link{create_residence_matrix}
}
