% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{view_residency_matrix}
\alias{view_residency_matrix}
\title{View residence membership}
\usage{
view_residency_matrix(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \code{xds} object}

\item{i}{the host species index}
}
\value{
a named \link{list}
}
\description{
Shows the residence membership information (from the residence matrix)
}
\seealso{
\link{make_residency_matrix}
}
