% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{vis_kappa}
\alias{vis_kappa}
\title{Compute NI of visitors}
\usage{
vis_kappa(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
The \code{S3} definition for the
function that computes vis_kappa
}
\keyword{internal}
