% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors-kappa.R
\name{vis_kappa.static}
\alias{vis_kappa.static}
\title{Compute NI of visitors}
\usage{
\method{vis_kappa}{static}(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets up...
}
\keyword{internal}
