% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visiting.basic}
\alias{Visiting.basic}
\title{Visiting, the basic model}
\usage{
\method{Visiting}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{Visiting} for the basic model for Visitors
}
