% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visiting.static}
\alias{Visiting.static}
\title{Visiting, a static model}
\usage{
\method{Visiting}{static}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{Visiting} for the static model (do nothing)
}
