% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{visitor_dynamics}
\alias{visitor_dynamics}
\title{Availabilit of Visitors}
\usage{
visitor_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a term that
is the time spent \emph{here,} stored as \code{TiSp_frac_visitorsing}
on the \code{XY_interface.} Here is defined as time
spent in one of the patches in this model, so
travle is time spent elsewhere.

A function \code{F_visitors} computes \code{TiSp_frac_visitorsing} and
this function sets \code{TiSp_frac_here.}
}
\note{
The quantity \code{TiSp_frac_here} is part of the \code{XY_interface}
It is used to compute the Time at Risk matrix, and it is also
used to weight local exposure \emph{vs.}  visitorsing exposure.
}
