% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{Visitors.setup}
\alias{Visitors.setup}
\title{Visitors}
\usage{
\method{Visitors}{setup}(t, y, xds_obj)
}
\arguments{
\item{t}{current time}

\item{y}{the state variable vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets a static value for the parameter
describing time spent F_visitors, resets the class to \code{static} and
then triggers an update to the blood feeding model.
It is computed in \link{compute_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
