% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.basic}
\alias{Weather.basic}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{basic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements a null weather model
}
