% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{Weather.no_forcing}
\alias{Weather.no_forcing}
\title{Methods for exogenous variables describing weather}
\usage{
\method{Weather}{no_forcing}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Implements \link{Weather} for the no_forcing model (no variables)
}
