% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{xde_cohort_derivatives}
\alias{xde_cohort_derivatives}
\title{Differential equation models for human cohorts}
\usage{
xde_cohort_derivatives(age, y, xds_obj, birthday)
}
\arguments{
\item{age}{host age}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{birthday}{the cohort birthday}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
