% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{xde_cohort_desolve.dde}
\alias{xde_cohort_desolve.dde}
\title{Solve a system of equations as a dde}
\usage{
\method{xde_cohort_desolve}{dde}(birthday, inits, times, xds_obj)
}
\arguments{
\item{birthday}{a cohort birthday}

\item{inits}{initial values}

\item{times}{= the times}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
a \link{list}
}
\description{
Implements for delay differential equations
}
