% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{xde_cohort_desolve}
\alias{xde_cohort_desolve}
\title{Solve a system of equations as an ode}
\usage{
xde_cohort_desolve(birthday, inits, times, xds_obj)
}
\arguments{
\item{birthday}{a cohort birthday}

\item{inits}{initial values}

\item{times}{= the times}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \link{list}
}
\description{
Implements for ordinary differential equations
}
