% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_derivatives.R
\name{xde_derivatives.eir}
\alias{xde_derivatives.eir}
\title{Differential equation models for aquatic mosquito populations}
\usage{
\method{xde_derivatives}{eir}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
