% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_derivatives.R
\name{xde_derivatives.mosy}
\alias{xde_derivatives.mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
\method{xde_derivatives}{mosy}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Mirrors \link{xde_derivatives} but only includes the adult and aquatic
mosquito components.
}
