% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_derivatives.R
\name{xde_derivatives}
\alias{xde_derivatives}
\title{Compute Derivatives}
\usage{
xde_derivatives(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
The function to compute the derivatives dispatches on \code{class(frame)}
}
\seealso{
\link{xds_solve}
}
