% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-xde.R
\name{xde_lines_X_SIS_xde}
\alias{xde_lines_X_SIS_xde}
\title{Add lines for the density of infected individuals for the SIS_xde model}
\usage{
xde_lines_X_SIS_xde(XH, nStrata, clrs = c("darkblue", "darkred"), llty = 1)
}
\arguments{
\item{XH}{a list with the outputs of parse_deout_X_SIS_xde}

\item{nStrata}{the number of population strata}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected individuals for the SIS_xde model
}
