% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-mz.R
\name{xde_make_MYZpar_mz}
\alias{xde_make_MYZpar_mz}
\title{Make parameters for mz ODE adult mosquito model}
\usage{
xde_make_MYZpar_mz(
  nPatches,
  MYZopts = list(),
  calK,
  g = 1/12,
  sigma = 1/8,
  mu = 1,
  eip = 12,
  f = 0.3,
  q = 0.95,
  nu = 1,
  eggsPerBatch = 60
)
}
\arguments{
\item{nPatches}{is the number of patches, an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{calK}{a mosquito dispersal matrix of dimensions \code{nPatches} by \code{nPatches}}

\item{g}{mosquito mortality rate}

\item{sigma}{emigration rate}

\item{mu}{loss to emigration emigration rate}

\item{eip}{the extrinsic incubation periodw}

\item{f}{feeding rate}

\item{q}{human blood fraction}

\item{nu}{oviposition rate, per mosquito}

\item{eggsPerBatch}{eggs laid per oviposition}
}
\value{
a \link{list}
}
\description{
Make parameters for mz ODE adult mosquito model
}
