% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEIS.R
\name{xde_make_Xpar_SEIS}
\alias{xde_make_Xpar_SEIS}
\title{Make parameters for SEIS xde human model, with defaults}
\usage{
xde_make_Xpar_SEIS(
  nStrata,
  Xopts = list(),
  b = 0.55,
  r = 1/180,
  nu = 1/14,
  c = 0.15
)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{Xopts}{a \link{list} that could overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{nu}{1/latent period}

\item{c}{transmission probability (efficiency) from human to mosquito}
}
\value{
a \link{list}
}
\description{
Make parameters for SEIS xde human model, with defaults
}
