% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{xde_setup_exposure_pois}
\alias{xde_setup_exposure_pois}
\title{Make parameters for the null model of exposure}
\usage{
xde_setup_exposure_pois(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
