% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{xde_setup_Lpar.basicL_xde}
\alias{xde_setup_Lpar.basicL_xde}
\title{xde_setup Lpar for the basicL_xde model}
\usage{
\method{xde_setup_Lpar}{basicL_xde}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_Lpar} for the basicL_xde model
}
