% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM.R
\name{xde_setup_MYZpar.GeRM}
\alias{xde_setup_MYZpar.GeRM}
\title{Setup MYZpar for the GeRM model}
\usage{
\method{xde_setup_MYZpar}{GeRM}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_MYZpar} for the GeRM model
}
