% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-GeRM-xde.R
\name{xde_setup_MYZpar.GeRM_xde}
\alias{xde_setup_MYZpar.GeRM_xde}
\title{Setup MYZpar for the GeRM_xde model}
\usage{
\method{xde_setup_MYZpar}{GeRM_xde}(MYZname, pars, s, EIPopts, MYZopts = list(), calK)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_MYZpar} for the GeRM_xde model
}
