% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-Gtrace.R
\name{xde_setup_MYZpar.Gtrace}
\alias{xde_setup_MYZpar.Gtrace}
\title{Setup the Gtrace}
\usage{
\method{xde_setup_MYZpar}{Gtrace}(
  MYZname,
  pars,
  s,
  EIPopts = list(),
  MYZopts = NULL,
  calK = NULL
)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{EIPopts}{is a \link{list}}

\item{MYZopts}{a \link{list}}

\item{calK}{is a \link{matrix}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_MYZpar} for the Gtrace model
}
