% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIP-xde.R
\name{xde_setup_Xpar.SIP_xde}
\alias{xde_setup_Xpar.SIP_xde}
\title{Setup Xpar.SIP_xde}
\usage{
\method{xde_setup_Xpar}{SIP_xde}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_Xpar} for the SIP_xde model
}
