% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{xde_setup_Xpar.SIS}
\alias{xde_setup_Xpar.SIS}
\title{Setup Xpar.SIS}
\usage{
\method{xde_setup_Xpar}{SIS}(Xname, pars, i, Xopts = list())
}
\arguments{
\item{Xname}{a \link{character} string}

\item{pars}{a \link{list}}

\item{i}{the host species index}

\item{Xopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{xde_setup_Xpar} for the SIS model
}
