% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_cohort.R
\name{xde_solve_cohort}
\alias{xde_solve_cohort}
\title{Cohort dynamics for a human / host model}
\usage{
xde_solve_cohort(
  eir,
  bday,
  pars,
  F_season = NULL,
  F_trend = NULL,
  F_age = NULL,
  A = 10,
  da = 10
)
}
\arguments{
\item{eir}{the average daily EIR for a population}

\item{bday}{a cohort's birthday}

\item{pars}{an \strong{\code{xds}} object}

\item{F_season}{a function describing a seasonal pattern over time}

\item{F_trend}{a function describing a temporal trend over time}

\item{F_age}{a function assigning a biting weight by age}

\item{A}{the maximum age to compute (in years)}

\item{da}{the output interval (age, in days)}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the states for a model \eqn{\cal X} in a cohort of humans / hosts
as it ages, up to age \eqn{A} years of age
}
\details{
This method substitutes age for time: a model
\deqn{\cal X(t)}
is solved with respect to age \eqn{a}:
\deqn{\cal X(a),}
where the daily EIR is computed by a \emph{trace} function with four elements:
\itemize{
\item \eqn{\bar E} or \code{eir}, the mean daily EIR,
\item \eqn{\omega(a)} or \verb{F_age,} a function of age
\item \eqn{S(t)} or \verb{F_season,} a function of time of year
\item \eqn{T(t)} or \verb{F_trend,} a function describing a trend
}

For a cohort born on day \eqn{B},
the function creates a mesh on age / time, where time and age
are related by the formula:
\deqn{t = B + a}
and the trace function is:
\deqn{E(a, t) = \hat E \; \omega(a) \; S(t)\; T(t) }
The output is returned as \code{pars$outputs$cohort}
}
