% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_solving.R
\name{xde_stable_orbit}
\alias{xde_stable_orbit}
\title{Compute the stable orbit for a system of differential equations}
\usage{
xde_stable_orbit(pars, Ymax = 10)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{Ymax}{the number of years to burn-in}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$xde}.
}
