% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_solving.R
\name{xde_steady.dde}
\alias{xde_steady.dde}
\title{Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}}
\usage{
\method{xde_steady}{dde}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$xde}
}
