% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{xde_steady_state_L.basicL}
\alias{xde_steady_state_L.basicL}
\title{Compute the Steady State for \link{dLdt.basicL}}
\usage{
\method{xde_steady_state_L}{basicL}(eta, Lpar)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{Lpar}{a list that defines an xde model}
}
\value{
the values of \eqn{L} at the steady state
}
\description{
Given an egg deposition rate \verb{eta,}
return a steady state value for the equations in \link{dLdt.basicL}
}
\note{
This function does not use deSolve
}
