% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{xde_steady_state_L}
\alias{xde_steady_state_L}
\title{Compute the steady states as a function of the daily EIR}
\usage{
xde_steady_state_L(eta, Lpar)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{Lpar}{a list that defines an xde model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{Lpar}.
}
