% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{xde_steady_state_L.trivial}
\alias{xde_steady_state_L.trivial}
\title{Compute the steady state as a function of the egg deposition rate eta}
\usage{
\method{xde_steady_state_L}{trivial}(eta, Lpar)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{Lpar}{a list that defines an xde model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{Lpar}.
}
