% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{xde_steady_state_M}
\alias{xde_steady_state_M}
\title{Compute the steady states as a function of the daily EIR}
\usage{
xde_steady_state_M(Lambda, MYZpar)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{MYZpar}{a list that defines an adult model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
