% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{xde_steady_state_MYZ}
\alias{xde_steady_state_MYZ}
\title{Compute steady states for \eqn{\cal MYZ} models}
\usage{
xde_steady_state_MYZ(Lambda, kappa, MYZpar)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{MYZpar}{a list that defines an adult model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
