% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{xde_steady_state_MYZ.SI}
\alias{xde_steady_state_MYZ.SI}
\title{Steady States: MYZ-SI}
\usage{
\method{xde_steady_state_MYZ}{SI}(Lambda, kappa, MYZpar)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{MYZpar}{a list that defines an adult model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
