% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trace.R
\name{xde_steady_state_MYZ.trace}
\alias{xde_steady_state_MYZ.trace}
\title{Steady States: MYZ-trace}
\usage{
\method{xde_steady_state_MYZ}{trace}(Lambda, kappa, MYZpar)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{MYZpar}{a list that defines an adult model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
