% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-hMoI.R
\name{xde_steady_state_X.hMoI}
\alias{xde_steady_state_X.hMoI}
\title{Compute Steady States for \code{hMoI} (\strong{X}-Model)}
\usage{
\method{xde_steady_state_X}{hMoI}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \strong{X} model object}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the hMoI model as a function of the daily foi
}
