% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{xde_steady_state_X}
\alias{xde_steady_state_X}
\title{Compute the steady states as a function of the daily EIR}
\usage{
xde_steady_state_X(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{host density}

\item{Xpar}{a list that defines a X-class model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{Xpar}.
}
