% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIPd.R
\name{xde_steady_state_X.SIPd}
\alias{xde_steady_state_X.SIPd}
\title{Compute the steady states for the SIPd model as a function of the daily foi}
\usage{
\method{xde_steady_state_X}{SIPd}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \eqn{\cal X} model}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIPd model as a function of the daily eir.
}
