% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{xde_steady_state_X.trace}
\alias{xde_steady_state_X.trace}
\title{Compute the steady states for the trace model as a function of the daily EIR}
\usage{
\method{xde_steady_state_X}{trace}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \eqn{\cal X} model}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the trace model as a function of the daily eir.
}
