% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{xde_steady_state_X.trivial}
\alias{xde_steady_state_X.trivial}
\title{Compute the steady states for the trivial model as a function of the daily EIR}
\usage{
\method{xde_steady_state_X}{trivial}(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \strong{X} model object}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the trivial model as a function of the daily eir.
}
