% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{xde_steady_state_XH}
\alias{xde_steady_state_XH}
\title{Steady States for \eqn{\cal XH}-Models}
\usage{
xde_steady_state_XH(foi, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{Xpar}{a list that defines an \eqn{\cal X} model}
}
\value{
none
}
\description{
Compute the steady states as a function of the daily FoI for
a model that has a steady state human / host population density

This method dispatches on the type of \code{Xpar}.
}
