% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_compute_terms.R
\name{xds_compute_terms.aquatic}
\alias{xds_compute_terms.aquatic}
\title{Differential equation models for aquatic mosquito populations}
\usage{
\method{xds_compute_terms}{aquatic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Compute derivatives for \link[deSolve:ode]{deSolve::ode} or \link[deSolve:dede]{deSolve::dede} using
generic methods for each model component.
}
