% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_compute_terms.R
\name{xds_compute_terms.mosy}
\alias{xds_compute_terms.mosy}
\title{Generalized spatial differential equation model (mosquito only)}
\usage{
\method{xds_compute_terms}{mosy}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
Mirrors \link{xds_compute_terms} but only includes the adult and aquatic
mosquito components.
}
