% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-template.R
\name{xds_dde.dts}
\alias{xds_dde.dts}
\title{Set \code{xds} for \code{dts}}
\usage{
\method{xds_dde}{dts}(pars)
}
\arguments{
\item{pars}{a \strong{\code{ramp.xds}} object}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Do not change \code{xds}
}
