% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-template.R
\name{xds_dde.ode}
\alias{xds_dde.ode}
\title{Set \code{xds} to \code{dde}}
\usage{
\method{xds_dde}{ode}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Change \code{xds} from \code{ode} to \code{dde}
}
