% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-template.R
\name{xds_dde}
\alias{xds_dde}
\title{Set \code{xds} to \code{dde}}
\usage{
xds_dde(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Creates and returns structured template for an
}
