% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xds_flatten}
\alias{xds_flatten}
\title{Set the initial values to the last values of the last simulation}
\usage{
xds_flatten(vars)
}
\arguments{
\item{vars}{a \link{list}}
}
\value{
\code{vars} as an unnamed \link{vector}
}
\description{
Set the initial values to the last values of the last simulation
}
