% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_help_basic_setup.R
\name{xds_help_basic_setup}
\alias{xds_help_basic_setup}
\title{\strong{\code{ramp.xds}}: Guide to Basic Setup}
\description{
\strong{Basic Setup} uses one of the following functions to build an
\strong{\code{xds}} model:
\itemize{
\item \link{xds_setup} sets up a full model with:
\itemize{
\item an \strong{XH} Component
\item an \strong{MY} Component
\item an \strong{L} Component
}
\item \link{xds_setup_human} sets up a models to study human epidemiology, given exposure:
\itemize{
\item an \strong{XH} component
\item a trivial \strong{MY} component is configured for infectious biting in patches
\item the time spent matrix is used
\item the \strong{L} component is not used
}
\item \link{xds_setup_eir} sets up models to study human epidemiology, given exposure:
\itemize{
\item an \strong{XH} component with one population stratum
\item a \emph{trace function} to output the daily EIR
\item a function configures relative biting rates by age for cohort dynamics
\item the \strong{MY} component is not used
\item the \strong{L} component is not used
}
\item \link{xds_setup_mosy} sets up models for mosquito ecology:
\itemize{
\item the \strong{L} component
\item the \strong{M} component (\emph{eg} \link{basicM})
\item a trivial \strong{H} module (for blood feeding)
\item the \strong{X} and \strong{Y} components are not used
}
\item \link{xds_setup_aquatic} sets up models for aquatic mosquito ecology:
\itemize{
\item the \strong{L} component
\item a trivial \strong{MY} component is configured for egg laying
\item the \strong{XH} component is not used
}
}
}
\section{The Frame}{

Each one of the basic setup functions assigns a different value to \verb{class(xds_obj$frame),}
which dispatches the \code{S3} methods that solve and parse differential equations:
\itemize{
\item \link{xds_setup} sets \code{class(frame) = "full"}
\item \link{xds_setup_human} sets \code{class(frame) = "human"}
\item \link{xds_setup_eir} sets \code{class(frame) = "eir"}
\item \link{xds_setup_mosy} sets \code{class(frame) = "mosy"}
\item \link{xds_setup_aquatic} sets \code{class(frame) = "aquatic"}
}
}

\seealso{
\link{dynamical_components}, \link{trivial_forcing} and \link{xds_object}
}
