% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_info.R
\name{xds_help_info}
\alias{xds_help_info}
\title{\strong{\code{ramp.xds}}: Info Pages}
\description{
A list of all \verb{xds_info_*} documentation pages in \strong{\code{ramp.xds}}.

\strong{Setup & Structure}
\itemize{
\item \link{xds_info_basic_setup} -- basic setup overview
\item \link{xds_info_setup_options} -- setup options
\item \link{xds_info_patch_dynamics} -- patch dynamics
\item \link{xds_info_strata} -- human/host population strata
\item \link{xds_info_aquatic_habitats} -- aquatic habitats
\item \link{xds_info_dynamical_components} -- dynamical components
\item \link{xds_info_trivial_forcing} -- trivial forcing
}

\strong{Human / Host}
\itemize{
\item \link{xds_info_interfaces} -- dynamical interfaces
\item \link{xds_info_search_weights} -- search weights
\item \link{xds_info_availability} -- resource availability
\item \link{xds_info_transmission} -- transmission
}

\strong{Human / Host}
\itemize{
\item \link{xds_info_human_populations} -- human populations
\item \link{xds_info_exposure} -- exposure
\item \link{xds_info_time_spent} -- time spent
\item \link{xds_info_time_at_risk} -- time at risk
\item \link{xds_info_local_fraction} -- the local fraction
\item \link{xds_info_travel_malaria} -- travel malaria
\item \link{xds_info_heterogeneity} -- heterogeneity
\item \link{xds_info_environmental_heterogeneity} -- environmental heterogeneity
\item \link{xds_info_heterogeneous_biting_rates} -- heterogeneous biting rates
\item \link{xds_info_relative_biting_rates} -- relative biting rates
}

\strong{Mosquitoes}
\itemize{
\item \link{xds_info_mosquito_bionomics} -- mosquito bionomics
\item \link{xds_info_mosquito_demography} -- mosquito demography
\item \link{xds_info_mosquito_dispersal} -- mosquito dispersal
\item \link{xds_info_blood_feeding} -- blood feeding
\item \link{xds_info_search_weights_blood} -- blood search weights
\item \link{xds_info_available_blood_hosts} -- available blood hosts
}

\strong{Habitats & Egg Laying}
\itemize{
\item \link{xds_info_egg_laying} -- egg laying
\item \link{xds_info_aquatic_habitats} -- aquatic habitats
\item \link{xds_info_search_weights_habitat} -- habitat search weights
\item \link{xds_info_available_habitats} -- habitat availability
}

\strong{Resources}
\itemize{
\item \link{xds_info_resources} -- resources
\item \link{xds_info_available_sugar} -- sugar availability
\item \link{xds_info_mosquito_traps} -- mosquito traps
}

\strong{Malaria Importation}
\itemize{
\item \link{xds_info_malaria_importation} -- malaria importation
\item \link{xds_info_travel_malaria} -- travel malaria
\item \link{xds_info_visitors} -- visitors
}

\strong{Malaria Control}
\itemize{
\item \link{xds_info_vector_control} -- vector control
\item \link{xds_info_health} -- health-based Interventions
}

\strong{Internals}
\itemize{
\item \link{xds_info_port} -- ports
\item \link{xds_info_junction} -- junctions
\item \link{xds_info_parsing} -- parsing outputs
}
}
