% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_help_setup_options.R
\name{xds_help_setup_options}
\alias{xds_help_setup_options}
\title{\strong{\code{ramp.xds}}: All Setup Options}
\description{
\strong{Basic Setup} uses \link{xds_setup} or one of its variants (see \link{xds_help_basic_setup})
to create an xds object.

The following is a list of configurable options
and where to look for more help.
}
\section{Structural Parameters}{

\itemize{
\item \code{nPatches} -- see \link{patch_dynamics}
\item \code{nStrata} -- see \link{human_populations}
\item \code{nHabitats} -- see \link{aquatic_habitats}
}
}

\section{Modules}{

\itemize{
\item \code{Xname} -- see \link{XH_module_list}
\item \code{MYname}-- see \link{MY_module_list}
\item \code{Lname} -- see \link{L_module_list}
}

Parameter values and initial values for the variables are
passed as named lists. The module help pages have information
about how to configure:
\verb{Xoptions,}
\verb{MYoptions,} &
\code{Loptions.}
}

\section{Basic options}{

\itemize{
\item \code{Koptions} see \link{mosquito_dispersal} and \link{setup_K_matrix}
\item \link{search_weights}
\item \link{time_spent}
}
}

\section{Blood Feeding and Transmission}{

\itemize{
\item \link{available_blood_hosts}
\itemize{
\item \link{blood_search_weights}
\item \link{other_blood_hosts}
\item \link{blood_traps}
}
\item \link{local_frac}
}
}

\section{Exposure}{

\itemize{
\item \link{environmental_heterogeneity}
\item \link{travel_malaria}
}
}

\section{Malaria Importation}{

\itemize{
\item \link{time_spent_here}
\item \link{travel_malaria}
}
}

\section{Mosquito Bionomics}{


\link{xds_info_mosquito_bionomics}
}

\section{Forcing}{


\link{xds_forcing}
\link{trivial_forcing}
\link{xds_junction_forcing}
}

\section{Resources}{


\link{xds_junction_resources}
}

\section{Habitats and Egg Laying}{

\itemize{
\item \link{bad_habitats}
\item \link{ovitraps}
}
}

