% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_basic_setup.R, R/xds_setup-info.R
\name{xds_info_basic_setup}
\alias{xds_info_basic_setup}
\title{\strong{\code{ramp.xds}}: Basic Setup}
\description{
\strong{Basic Setup} for \strong{\code{ramp.xds}} uses
\link{xds_setup} or one of its variants.

\strong{Basic Setup} uses one of the following functions to build an
\strong{\code{xds}} model:
\itemize{
\item \link{xds_setup} sets up a full model with:
\itemize{
\item an \strong{XH} Component
\item an \strong{MY} Component
\item an \strong{L} Component
}
\item \link{xds_setup_human} sets up models to study human epidemiology, given exposure:
\itemize{
\item an \strong{XH} component
\item a trivial \strong{MY} component is configured for infectious biting in patches
\item the time spent matrix is used
\item the \strong{L} component is not used
}
\item \link{xds_setup_eir} sets up models to study human epidemiology, given exposure:
\itemize{
\item an \strong{XH} component with one population stratum
\item a \emph{trace function} to output the daily EIR
\item a function configures relative biting rates by age for cohort dynamics
\item the \strong{MY} component is not used
\item the \strong{L} component is not used
}
\item \link{xds_setup_mosy} sets up models for mosquito ecology:
\itemize{
\item the \strong{L} component
\item the \strong{M} component (\emph{eg} \link{basicM})
\item a trivial \strong{H} module (for blood feeding)
\item the \strong{X} and \strong{Y} components are not used
}
\item \link{xds_setup_aquatic} sets up models for aquatic mosquito ecology:
\itemize{
\item the \strong{L} component
\item a trivial \strong{MY} component is configured for egg laying
\item the \strong{XH} component is not used
}
}
}
\section{Modularity}{


\strong{\code{ramp.xds}} is designed for modular computation. The
mathematical framework has
five core biological processes organized
into three chunks called \strong{dynamical components} (see \link{xds_info_dynamical_components}).
The modular design
makes it possible to build models that fully define all
components (using \link{xds_setup}), or to
develop systems of equations to isolate and study some part
of a system.

To implement \emph{plug-and-play} modularity, a trivial module was developed
for each component:
a \emph{trace function} in
the trivial \emph{upstream} component is configured to pass known inputs (see \link{xds_info_trivial_forcing}).
}

\section{The Frame}{


\link{xds_setup} sets up a model
that has all three components. Functions were developed to setup and solve systems that
making another component completely unnecessary.

All \verb{xds_setup_*} functions define \code{xds_obj$frame}
and the functions used to solve and parse differential equations
use the \code{S3} system, with methods
that dispatch on \code{class(xds_obj$frame)}.
\itemize{
\item \link{xds_setup} (\code{frame="full"}) sets up all three components.
\item \link{xds_setup_mosy} (\code{frame="mosy"}) sets up models for mosquito ecology:
\itemize{
\item the model includes an \strong{L} component
\item the model has an \strong{M} component but no \strong{Y} component (\emph{eg} \link{basicM})
\item a trivial \strong{H} module can be configured
\item the \strong{X} component is not used
}
\item \link{xds_setup_aquatic} (\code{frame="aquatic"}) sets up models aquatic mosquito ecology:
\itemize{
\item a trivial \strong{MY} component is configured for egg laying
\item the \strong{XH} component is not used
}
\item \link{xds_setup_human} (\code{frame="human"}) sets up models to study epidemiology
\itemize{
\item a trivial \strong{MY} component is configured for infectious biting in patches
\item the \strong{L} component is not used
}
\item \link{xds_setup_eir} (\code{frame="eir"}) sets up models to study epidemiology
\itemize{
\item a trace function computes the EIR, denoted \eqn{F_E(t)}
\item the \strong{MY} component is not used
\item the \strong{XH} component is not used
}
}


Each one of the basic setup functions assigns a different value to \verb{class(xds_obj$frame),}
which dispatches the \code{S3} methods that solve and parse differential equations:
\itemize{
\item \link{xds_setup} sets \code{class(frame) = "full"}
\item \link{xds_setup_human} sets \code{class(frame) = "human"}
\item \link{xds_setup_eir} sets \code{class(frame) = "eir"}
\item \link{xds_setup_mosy} sets \code{class(frame) = "mosy"}
\item \link{xds_setup_aquatic} sets \code{class(frame) = "aquatic"}
}
}

\seealso{
\link{xds_info_dynamical_components}, \link{xds_info_trivial_forcing} and \link{xds_object}
}
