% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-info.R
\name{xds_info_blood_feeding}
\alias{xds_info_blood_feeding}
\title{Blood Feeding}
\description{
Blood feeding is an interaction among humans and mosquitoes:
in this framework, humans spend time and
mosquitoes search. To model exposure, each human (or host)
population stratum is assigned a \emph{search weight,} a
number used to weigh \emph{time spent} and get a measure of
\emph{availability.}
}
