% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying-concepts.R
\name{xds_info_egg_laying}
\alias{xds_info_egg_laying}
\title{Egg Laying}
\description{
Egg laying
}
\section{Habitat Availability}{


Habitat availability is a concept developed
to model egg laying behavior.
\itemize{
\item Habitats
\item Bad Habitats
\item Ovitraps
}
}

