% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{xds_info_environmental_heterogeneity}
\alias{xds_info_environmental_heterogeneity}
\title{Environmental Heterogeneity}
\description{
Humans can be exposed to malaria at home or as they move around.
The number of infectious bites can vary by chance, depending on
where and when humans encounter mosquitoes.
Environmental heterogeneity models
exposure for homogeneous population strata under a model for the distribution
of the expectation. In simple terrms, it models super-spreading without super-spreaders.
\itemize{
\item Poisson
\item Negative binomial
}
}
