% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{xds_info_exposure}
\alias{xds_info_exposure}
\title{Exposure}
\description{
The EIR computed by \link{xds_info_transmission} is a measure of \emph{local} risk: it is
the expected number of infected
bites, per person, and it is computed for every population stratum. In
\strong{exposure,} the daily local EIR is computed into a measure
of the population FoI in three steps:
\itemize{
\item the total EIR is the weighted sum of local exposure and travel exposure, using
the concept of time here \emph{vs.} time away;
\item the distribution of the number of bites per person is computed under a
model of environmental heterogeneity ;
\item the number of infections is computed under a model of partial immunity.
}
}
\section{Environmental Heterogeneity}{

While EIR is an expected value, the expected value can have a distribution
in a human population stratum. For example, if the expectation has a Gamma
distribution, then the expected number of bites per person would be negatively
binomially distributed. For more, see \link{xds_info_environmental_heterogeneity}.
}

\section{Travel Malaria}{

Travel malaria is defined by exposure to malaria while spending
time outside of the spatial domain represented by the patches.
See \link{xds_info_travel_malaria}
}

\section{Pre-Erythrocytic Immunity}{

Depending on the X module, a parameter called \eqn{b} describes the probability
of infection per infectious bite.
}

\seealso{
\link{xds_info_environmental_heterogeneity} | \link{xds_info_travel_malaria}
}
