% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{xds_info_forcing}
\alias{xds_info_forcing}
\title{\strong{\code{ramp.xds}}: Forcing}
\description{
In \strong{\code{ramp.xds}}, forcing is implemented in several ways:
\itemize{
\item through \code{trivial} modules: see \link{xds_info_trivial_forcing}
\item a library of modulesas functional responses to other variables: models are
developed by activiting the \link{Forcing} junction
\item through malaria control:
\itemize{
\item see \link{xds_info_vector_control}
\item see \link{xds_info_health}
}
}
}
\seealso{
\link{xds_info_trivial_forcing}
}
