% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{xds_info_heterogeneity}
\alias{xds_info_heterogeneity}
\title{Heterogeneity}
\description{
\strong{\code{ramp.xds}} was designed to explore heterogeneous exposure & transmission:
\itemize{
\item \emph{heterogeneous biting rates} - differences in the expected number of expected
bites, per person, among heterogeneous biting rates:
see \link{xds_info_heterogeneous_biting_rates}.
\item \emph{environmental heterogeneity} - describes variability in the distribution
of the number of bites, per person in a homogenous population stratum:
see \link{xds_info_environmental_heterogeneity}
\item heterogeneous mixing describes
}
}
