% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{xds_info_heterogeneous_biting_rates}
\alias{xds_info_heterogeneous_biting_rates}
\title{Heterogeneous Biting Rates}
\description{
Heterogeneous biting rates are closely
related to \emph{relative biting
rates:} for a stratum, the relative biting rate is
a value \eqn{\omega} such that if the
population average daily EIR is \eqn{E}, then the
daily EIR for the stratum is \eqn{\omega E}.

Models for heterogeneous biting are constructed
by passing \emph{blood search weights} for human population
strata, \eqn{w}. If the time spent in a patch
is \eqn{\theta}, then
the availability of humans in that patch is is:
\deqn{W_j = \sum_i w_i \theta_{i,j} H_i.}
The fraction of bites received by the \eqn{i^{th}}
stratum in patch \eqn{j} is:
\deqn{\frac{w_i \theta_{i,j} H_i}{W_j}}
The daily EIR is:
\deqn{E_i = w_i \sum_j \frac{\theta_{i,j} }{W_j}f_j q_j Z_j}
The population average EIR is
\deqn{E = \sum_i E_i H_i/ \sum_i H_i}
The relative biting rate is
\deqn{\omega_i = E_i / E}
}
