% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-info.R
\name{xds_info_human_populations}
\alias{xds_info_human_populations}
\title{The human population}
\description{
A human (or host) population is sub-divided into \emph{strata.}
The human population (or other host population) and its structure, size and behavior
are critical aspects of transmission.

\describe{
\item{\code{HPop}}{initial human population size: \code{length(HPop) = nStrata}}
}
}
\section{Human Population Size, \eqn{H}}{


The human (or host) population size plays a key role
in the blood feeding and transmission interface. In
most \strong{XH} modules, it is handled as a variable. Setup
requires having some assigning a value to H, so it is passed
at setup as \code{HPop.} Since \eqn{H} is part of the
blood feeding and transmission interface, changing its
initial value triggers functions to refresh the dynamical
terms for models where
\eqn{dH/dt = 0}.
}

