% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_info-interfaces.R
\name{xds_info_interfaces}
\alias{xds_info_interfaces}
\title{Dynamical Interfaces}
\description{
In the modular design, the \link{xds_info_dynamical_components} interact through
two well-designed interfaces:
\itemize{
\item \strong{Infection Dynamics:} the \strong{X} and \strong{Y} components communicate through an interface with three parts:
\itemize{
\item \strong{Blood Feeding} -- a coherent framework for modeling blood feeding based on the concept of blood host availability. Models can set
blood feeding rates without using it, but the parameter values should satisfy some constraints:
\itemize{
\item the blood feeding rate is 0 (\eqn{f=0}) if there are no hosts available;
\item the human fraction is 0 (\eqn{q=0}) if there are no humans available;
\item the human fraction is 1 (\eqn{q=1}) if only humans are available;
}
\item \strong{Transmission} -- under the blood feeding model, compute the biting
distribution matrix (\eqn{\beta}) that guarantees every bite and blood meal
is accounted for: it updates the local fraction (if there are visitors),
and then outputs the local
daily entomological inoculation rate (EIR or \eqn{E}) for
all the strata; and the net infectiousness (NI or \eqn{\kappa}) in each patch.
\item \strong{Exposure} -- compute the force of infection (FoI) from:
\itemize{
\item the local daily EIR (\eqn{E})
\item the travel EIR (\eqn{\delta})
\item a model of partial immunity (\eqn{b}) from the \strong{X} component
}
}
\item \strong{Mosquito Ecology:} the \strong{L} and \strong{M} components communicate through an interface with three parts:
\itemize{
\item \strong{Patch Dynamics} -- adult mosquito populations are distributed in patches (see \link{xds_info_patch_dynamics})
\item \strong{Aquatic Habitats} -- aquatic habitats are located in patches (see \link{xds_info_aquatic_habitats})
\item \strong{Egg Laying} -- The egg laying interface uses search weights to distribute eggs laid by adults in a patch to the aquatic habitats in that patch
}
}
}
