% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/junctions.R
\name{xds_info_junction}
\alias{xds_info_junction}
\title{Junctions}
\description{
In \strong{\verb{ramp.xds,}} clusters of ports are sometimes bundled
into a \strong{junction} (see \link{xds_info_port}). A junction is a structural element,
a function that is designed to handle one or more ports.
\itemize{
\item \link{Other_State_Variables} -- a junction to add new state variables, effectively extending the dynamical system
\item \link{xds_info_forcing} -- a junction to handle exogenous forcing by weather, \emph{etc.}
\item \link{xds_info_vector_control} -- a junction to handle vector control
\item \link{xds_info_health} -- a junction to handle vector control
\item \link{xds_info_resources} -- a junction to handle resource dynamics
\item \link{xds_info_importation} -- a junction to handle resource dynamics
}
}
\seealso{
\link{xds_info_port}
}
