% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{xds_info_local_fraction}
\alias{xds_info_local_fraction}
\title{The Local Fraction}
\description{
The blood feeding and transmission interface is set up to ensure
mathematical consistency: all mosquito blood feeding should
be accounted for. In models with visitors, the visitor population
contributes to local host availability and absorbs a fraction of
blood feeding. The fraction on local hosts is computed as \code{local_frac}
}
