% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel-info.R
\name{xds_info_malaria_importation}
\alias{xds_info_malaria_importation}
\title{Malaria Importation}
\description{
Importation of malaria in \strong{\code{ramp.xds}} is defined as any
malaria acquired outside the spatial domain represented by
the patches. Imported malaria can be modeled in two ways:
\itemize{
\item residents of the spatial domain acquire malaria while
traveling;
\item infected visitors
}
}
