% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics.R
\name{xds_info_mosquito_bionomics}
\alias{xds_info_mosquito_bionomics}
\title{Mosquito Bionomics}
\description{
The value of a mosquito bionomic parameters gets computed in two stages:
\itemize{
\item \link{MosquitoBionomics} computes the bionomic parameter values;
\item \link{VectorControlEffectSizes} applies the effect sizes of vector control
}
}
\section{Baseline}{

The concept of a \emph{baseline} here is not exactly corre
}

\section{Modified by Control}{

The concept of a \emph{baseline} is
The value of a parameter
}

