% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{xds_info_mosquito_demography}
\alias{xds_info_mosquito_demography}
\title{Mosquito Demography}
\description{
Mosquito survival and dispersal is described by a \emph{demographic matrix,} denoted \eqn{\Omega.} It is
computed using several parameters:
\describe{
\item{\code{g}}{mortality rate}
\item{\code{sigma}}{patch emigration}
\item{\code{mu}}{emigration-related loss}
\item{\code{K}}{a dispersal matrix}
}

The matrix is computed as:

\deqn{
 \Omega = \mbox{diag} \left( g + \sigma \mu \right) - K \cdot \mbox{diag} \left( \sigma \left(1-\mu\right) \right)
}

In delay differential equations with a constant EIP (\eqn{\tau}),
survival and dispersal through the EIP is given by:
\deqn{
 \Upsilon = e^{-\Omega \tau} 
}
}
\seealso{
\link{xds_info_mosquito_dispersal}
}
