% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{xds_info_mosquito_dispersal}
\alias{xds_info_mosquito_dispersal}
\title{Mosquito Dispersal}
\description{
Mosquito dispersal is handled
in a standard way across all \strong{MY} modules.
\itemize{
\item A mosquito dispersal matrix, called \eqn{K} or \code{K_matrix}, is set up using \link{setup_K_matrix};
\item \eqn{K} is used to make a mosquito demographic matrix, called \eqn{\Omega} or \code{Omega} (see \link{xds_info_mosquito_demography}).
}
}
\section{Setup}{

\describe{
\item{\code{K_matrix}}{either a \eqn{K} matrix: or options for \link{setup_K_matrix}}
}

There are several ways to configure  \eqn{K} during basic setup.

By \strong{default:} \code{K_matrix} is an empty list. When the \code{MY_obj} is set up, the \eqn{K}
matrix is set to \code{diag(g)}, a model with no dispersal. That is not changed unless one of the
following holds:
}

\section{Mosquito Dispersal Matrix, \eqn{K}}{


In adult mosquito modules, mosquito dispersal is described by
a square matrix \eqn{K} with \code{nPatches} \eqn{(=n)} rows and columns. By convention,
all \eqn{K} matrices have the form:
\deqn{
 K = \left[ \begin{array}{ccccc}
 -1 & k_{1,2} & k_{1,3} & \cdots & k_{1, n} \\
 k_{2,1} & -1 & k_{2,3} & \cdots & k_{2, n} \\
 k_{3,1} & k_{3,2} & -1 & \cdots & k_{3, n} \\
 \vdots & \vdots & \vdots & \ddots  & \vdots \\
 k_{n,1} & k_{n,2} & k_{n,3} & \cdots & -1 \\
 \end{array} \right]
}
The values of the elements \eqn{k_{i,j}} are thus
constrained such that there is no net dispersal loss from the system: \eqn{\forall i,}
\deqn{\sum_j k_{i,j} = 1.}
The columns of \eqn{K} sum up to zero.
\eqn{K} thus describes the destinations of
emigrating mosquitoes that survive and stay in the system.

Emigration rates and emigration-related loss -- emigration from the spatial domain and mortality that is conditioned on emigration --
are handled separately (see \link{xds_info_mosquito_demography}).
}

\seealso{
\link{setup_K_matrix}, \link{change_K_matrix} & \link{xds_info_mosquito_demography}
}
