% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogeneity.R
\name{xds_info_relative_biting_rates}
\alias{xds_info_relative_biting_rates}
\title{Relative Biting Rate}
\description{
In \strong{\code{ramp.xds}}, heterogeneous
biting rates are handled by assigning \emph{search weights}
to various population strata. The \emph{relative biting
rate} for a stratum is a value \eqn{\omega} such that if the
population average daily EIR is \eqn{E}, then the
daily EIR for the stratum is \eqn{\omega E}.

Relative biting rates are handled in two ways, depending on the frame:
\itemize{
\item In model that define multiple human population strata, the RBRs are computed from the
stratum specific EIR: \deqn{\omega_i = E_i / E}
\item In models that are forced by the EIR, the user configures a function
to compute the population average EIR over time, and another function to
compute the relative biting
rate by age.
}
}
\seealso{
\link{xds_info_search_weights_blood} | \link{xds_info_heterogeneous_biting_rates} | \link{xds_info_heterogeneity}
}
