% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_weights.R
\name{xds_info_search_weights}
\alias{xds_info_search_weights}
\title{Search Weights}
\description{
\describe{
\item{\code{searchB}}{a vector of search weights for blood feeding: \code{length(searchB)=nStrata}}
\item{\code{searchQ}}{a vector of search weights for egg laying: \code{length(searchQ)=nHabitats}}
}

By default, the values are all set to 1.

Mosquitoes search for resources, including vertebrate animals for
blood feeding and aquatic habitats for egg laying.
In developing a modular design, a
rigorous interface was needed to guarantee mathematical
consistency for blood feeding
and parasite transmission, and for adult mosquito ecology
and behaviors around egg laying.
Models for \emph{heterogeneous biting} among host strata and
\emph{heterogeneous egg laying} among habitats are implemented through \emph{search weights.}

A search
weight is a scalar value assigned to each human population stratum
or habitat that is used to compute \emph{availability}
to mosquitoes and relative biting rates or relative laying rates.
The value of these search weights are configurable during basic setup.
}
