% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_weights.R
\name{xds_info_search_weights_habitat}
\alias{xds_info_search_weights_habitat}
\title{Habitat Search Weights}
\description{
where
}
\seealso{
\link{xds_info_blood_feeding}, \link{xds_info_transmission}, and \link{xds_info_egg_laying}
}
