% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_help_setup_options.R,
%   R/xds_info-setup_options.R
\name{xds_info_setup_options}
\alias{xds_info_setup_options}
\title{\strong{\code{ramp.xds}}: All Setup Options}
\description{
\strong{Basic Setup} uses \link{xds_setup} or one of its variants (see \link{xds_info_basic_setup})
to create an xds object.

The following is a list of configurable options
and where to look for more help.

\strong{Basic Setup} uses \link{xds_setup} or one of its variants (see \link{xds_info_basic_setup})
to create an xds object.

The following is a list of configurable options
and where to look for more help.
}
\section{Structural Parameters}{

\itemize{
\item \code{nPatches} -- see \link{xds_info_patch_dynamics}
\item \code{nStrata} -- see \link{xds_info_human_populations}
\item \code{nHabitats} -- see \link{xds_info_aquatic_habitats}
}


\itemize{
\item \code{nPatches} -- see \link{xds_info_patch_dynamics}
\item \code{nStrata} -- see \link{xds_info_human_populations}
\item \code{nHabitats} -- see \link{xds_info_aquatic_habitats}
}
}

\section{Modules}{

\itemize{
\item \code{Xname} -- see \link{XH_module_list}
\item \code{MYname}-- see \link{MY_module_list}
\item \code{Lname} -- see \link{L_module_list}
}

Parameter values and initial values for the variables are
passed as named lists. The module help pages have information
about how to configure:
\verb{Xoptions,}
\verb{MYoptions,} &
\code{Loptions.}


\itemize{
\item \code{Xname} -- see \link{XH_module_list}
\item \code{MYname}-- see \link{MY_module_list}
\item \code{Lname} -- see \link{L_module_list}
}

Parameter values and initial values for the variables are
passed as named lists. The module help pages have information
about how to configure:
\verb{Xoptions,}
\verb{MYoptions,} &
\code{Loptions.}
}

\section{Basic options}{

\itemize{
\item \code{Koptions} see \link{xds_info_mosquito_dispersal} and \link{setup_K_matrix}
\item \link{xds_info_search_weights}
\item \link{xds_info_time_spent}
}


\itemize{
\item \code{Koptions} see \link{xds_info_mosquito_dispersal} and \link{setup_K_matrix}
\item \link{xds_info_search_weights}
\item \link{xds_info_time_spent}
}
}

\section{Blood Feeding and Transmission}{

\itemize{
\item \link{xds_info_available_blood_hosts}
\itemize{
\item \link{xds_info_search_weights_blood}
\item \link{xds_port_other_blood_hosts}
\item \link{xds_port_blood_traps}
}
\item \link{xds_info_local_fraction}
}


\itemize{
\item \link{xds_info_available_blood_hosts}
\itemize{
\item \link{xds_info_search_weights_blood}
\item \link{xds_port_other_blood_hosts}
\item \link{xds_port_blood_traps}
}
\item \link{xds_info_local_fraction}
}
}

\section{Exposure}{

\itemize{
\item \link{xds_info_environmental_heterogeneity}
\item \link{xds_info_travel_malaria}
}


\itemize{
\item \link{xds_info_environmental_heterogeneity}
\item \link{xds_info_travel_malaria}
}
}

\section{Malaria Importation}{

\itemize{
\item \link{xds_info_malaria_importation}
\item \link{xds_info_travel_malaria}
\item \link{xds_info_visitors}
}


\itemize{
\item \link{xds_info_malaria_importation}
\item \link{xds_info_travel_malaria}
\item \link{xds_info_visitors}
}
}

\section{Mosquito Bionomics}{


\link{xds_info_mosquito_bionomics}



\link{xds_info_mosquito_bionomics}
}

\section{Forcing}{


\link{xds_info_forcing}
\link{xds_info_trivial_forcing}



\link{xds_info_forcing}
\link{xds_info_trivial_forcing}
}

\section{Resources}{


\link{xds_info_resources}



\link{xds_info_resources}
}

\section{Habitats and Egg Laying}{

\itemize{
\item \link{xds_info_aquatic_habitats}
\item \link{xds_port_ovitraps}
}


\itemize{
\item \link{xds_info_aquatic_habitats}
\item \link{xds_port_ovitraps}
}
}

