% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-info.R
\name{xds_info_strata}
\alias{xds_info_strata}
\title{Population Strata}
\description{
A human (or host) population is sub-divided into \emph{strata.}
The human population (or other host population) and its structure, size and behavior
are critical aspects of transmission.

\describe{
\item{\code{residence}}{a vector: the patch index where each human stratum lives}
\item{\code{nStrata}}{\code{nStrata = length(residence)}}
}

Human / host populations are heterogeneous, and some of that heterogeneity is important
for accurately describing the epidemiology of malaria and other mosquito-borne pathogens.
\strong{\code{ramp.xds}} was thus designed with the capability of segmenting a population
into as many strata
as needed to achieve a degree of accuracy. In designing a study, decisions about
how to construct a model are translated into a set of population strata. Information
about the number of strata is configured through the \code{residence} vector.
}
\section{Residency and Time Spent}{

In designing \strong{\code{ramp.xds}}, the first concern was how to handle spatial dynamics.
Since mosquitoes and humans are very different, a metapopulation that is set up for
adult mosquito ecology might not work as well for the humans, whose behaviors follow
very different rules.  Humans are enumerated by where
they live, and most people spend most of their time at risk in or around home.

The human population is thus stratified around the concept of patch residency -- the
patch where home is found. Exposure is based on the notions of \emph{time spent} (see \link{xds_info_time_spent})
and \emph{time at risk} (see \link{xds_info_time_at_risk}). The model is configured by passing
a vector, called \code{residence.}
The residence vector is the index of the patch where each
human population resides, and multiple strata can reside in the same patch.
Conversely, there might be some patches where no one resides, such as the mosquito
habitats in a buffer around the places where humans live.
}

\section{Residency Matrix}{

The residence vector is used to construct a
residency matrix, \eqn{J}, that is the same shape as the time
spent and time at risk matrices. It can be used to sum quantities
by patch. For example, if \eqn{X} is the density of infected individuals,
then \eqn{J \cdot X} is the number of infected individuals by patch, so
true prevalence would be \deqn{\frac{J \cdot X}{J \cdot H}}.
}

\section{Demography}{


Utilities to handle host age and \emph{cohort dynamics} are handled in
\strong{\code{ramp.demog}}. Similarly, the functionality to handle
}

\section{Principled Stratification}{


Principled stratification is the algorithm:
\itemize{
\item interventions
}
}

