% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{xds_info_transmission}
\alias{xds_info_transmission}
\title{Transmission}
\description{
In transmission, a quantity called the \emph{local fraction} gets computed, and then
the model computes terms describing parasite transmission:
\itemize{
\item \code{beta} -- the bite distribution matrix
\item \code{eir} -- the daily EIR
\item \code{kappa} -- the net infectiousness
}
}
\seealso{
\link{xds_info_local_fraction}
}
