% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_interfaces.R
\name{xds_interfaces}
\alias{xds_interfaces}
\title{Dynamical Interfaces}
\description{
In the modular design, the \link{dynamical_components} interact through
two well-designed interfaces:
\itemize{
\item \strong{Infection Dynamics:} the \strong{X} and \strong{Y} components communicate through an interface with three parts:
\itemize{
\item \strong{Blood Feeding} --
\item \strong{Transmission} --
\item \strong{Exposure} --
}
\item \strong{Mosquito Ecology:} the \strong{L} and \strong{M} components communicate through an interface with three parts:
\itemize{
\item \strong{Patch Dynamics} -- adult mosquito populations are distributed in patches (see \link{patch_dynamics})
\item \strong{Aquatic Habitats} -- aquatic habitats are located in patches (see \link{aquatic_habitats})
\item \strong{Egg Laying} --
}
}
}
