% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-plotting.R
\name{xds_lines_L}
\alias{xds_lines_L}
\title{Add lines for aquatic mosquito population density}
\usage{
xds_lines_L(xds_obj, s = 1, clrs = "darkblue", llty = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\value{
invisible(NULL)
}
\description{
Add lines for aquatic mosquito population density
}
