% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_M}
\alias{xds_lines_M}
\title{Add lines for adult mosquito population density}
\usage{
xds_lines_M(MY, xds_obj, clrs = "darkblue", llty = 1)
}
\arguments{
\item{MY}{xds_objed ouptuts}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for adult mosquito population density
}
