% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-plotting.R
\name{xds_lines_PR}
\alias{xds_lines_PR}
\title{Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity}
\usage{
xds_lines_PR(xds_obj, i, clrs = "black", llty = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) that specifies \code{lty} for plotting}
}
\description{
Add lines for the prevalence / parasite rate (PR) from a model of human infection and immunity
}
