% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_Y}
\alias{xds_lines_Y}
\title{Add lines for the density of infected and infective mosquitoes}
\usage{
xds_lines_Y(xds_obj, s = 1, clrs = "purple", llty = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected and infective mosquitoes
}
