% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_Y_fracs}
\alias{xds_lines_Y_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
xds_lines_Y_fracs(MY, xds_obj, clrs = "purple", llty = 1)
}
\arguments{
\item{MY}{xds_objed ouptuts}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{clrs}{a vector of colors for infected mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
