% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_YZ_fracs}
\alias{xds_lines_YZ_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
xds_lines_YZ_fracs(
  time,
  MYZ,
  pars,
  Yclrs = "purple",
  Zclrs = "darkred",
  llty = 1
)
}
\arguments{
\item{time}{time points for the observations}

\item{MYZ}{a list the output of \code{ramp.xde::parse_deout()}}

\item{pars}{a list that defines an \code{ramp.xde} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{Yclrs}{a vector of colors for infected mosquitoes}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
