% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_Z}
\alias{xds_lines_Z}
\title{Add lines for the density of infected and infective mosquitoes}
\usage{
xds_lines_Z(times, MYZ, pars, clrs = "darkred", llty = 1)
}
\arguments{
\item{times}{a sequence of times when variables were output}

\item{MYZ}{parsed ouptuts}

\item{pars}{an \strong{\code{xds}} object}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected and infective mosquitoes
}
