% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_object.R
\name{xds_object}
\alias{xds_object}
\title{The \strong{\code{xds}} Model Object}
\description{
A dynamical system is encoded and stored
on an \strong{\code{xds}} model
object,
a compound \link{list} with
a specific set of features.

A properly constructed xds object is
returned by \link{xds_setup} or one of its variants (see \link{xds_help_basic_setup}).
The object is created in several steps:
\itemize{
\item a basic template for an xds object is returned by
\link{make_xds_object_template}, which
\itemize{
\item sets the value of structural parameters;
\item creates empty lists to hold the module objects: \code{XH_obj} \verb{MY_obj,} and \code{L_obj};
\item sets up empty \link{port}s and \link{junction}s;
}
\item the core dynamical components are setup; and
\item the values of a few modifiable parameters affecting transmission are set.
}

While xds object internals are useful for developers,
functions have been developed to
set up (\verb{setup_*}), inspect (\verb{get_*}), and modify (`change_*) various
aspects of xds objects.
}
\seealso{
Vignettes: \href{https://dd-harp.github.io/ramp.xds/articles/GettingStarted.html}{Getting Started}, \href{https://dd-harp.github.io/ramp.xds/articles/Working.html}{Working}, and \href{https://dd-harp.github.io/ramp.xds/articles/ex_534.html}{The 5-3-4 Model}
}
