% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_object.R
\name{xds_object_frame}
\alias{xds_object_frame}
\title{The xds object frame}
\description{
There are several variants of \link{xds_setup} that set the value of \code{xds_obj$frame.}

The modular design makes it possible to study a
dynamical component in isolation, or to focus on some
combination of those elements.

To facilitate those studies, each component has a \link{trivial} module
that computes the dynamical term that gets passed to
another component. The user configures the \emph{trace function}
of the upstream component.

With \link{xds_setup}, a user defines all three components, but other variants
were developed to make it easier to set up partial systems. The functions that solve
differential equations and parse outputs are written as \code{S3} functions with methods
that dispatch on \code{xds_obj$frame}
\itemize{
\item \link{xds_setup} (\code{frame="full"}) sets up all three components.
\item \link{xds_setup_mosy} (\code{frame="mosy"}) sets up models for mosquito ecology:
\itemize{
\item the model includes an \strong{L} component
\item the model has an \strong{M} component but no \strong{Y} component (\emph{eg} \link{basicM})
\item a trivial \strong{H} module can be configured, but the \strong{XH} derivatives are not computed
}
\item \link{xds_setup_aquatic} (\code{frame="aquatic"}) sets up models aquatic mosquito ecology:
\itemize{
\item a trivial \strong{MY} component is configured for egg laying
\item the \strong{XH} component is missing
}
\item \link{xds_setup_human} (\code{frame="human"}) sets up models to study epidemiology
\itemize{
\item a trivial \strong{MY} component is configured for infectious biting in patches
\item the \strong{L} component is missing
}
\item \link{xds_setup_eir} (\code{frame="eir"}) sets up models to study epidemiology
\itemize{
\item a function \eqn{F_E(t),} a trace function that passes the EIR.  These models lack the \strong{MY} and \strong{L} components.
\item a trivial \strong{MY} component is configured for infectious biting in patches
\item the \strong{XH} component is missing
}
}
}
\seealso{
\link{trivial_forcing} and \link{xds_object}
}
