% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-terms.R
\name{xds_plot_aEIR}
\alias{xds_plot_aEIR}
\title{Plot the annualized EIR \emph{vs.} time}
\usage{
xds_plot_aEIR(
  xds_obj,
  i = 1,
  clrs = "black",
  llty = 1,
  stable = FALSE,
  add = FALSE
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and FALSE for \code{stable_orbits}}

\item{add}{a logical: plot axes only if FALSE}
}
\description{
Plot the annualized EIR \emph{vs.} time
}
