% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-terms.R
\name{xds_plot_EIR}
\alias{xds_plot_EIR}
\title{Plot the EIR \emph{vs.} time}
\usage{
xds_plot_EIR(pars, i = 1, clrs = "black", llty = 1, add = FALSE)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.xds} model (\emph{e.g.},  generated by \code{ramp.xds::xds_setup()})}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{a logical: plot axes only if FALSE}
}
\description{
Plot the EIR \emph{vs.} time
}
