% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-plotting.R
\name{xds_plot_L}
\alias{xds_plot_L}
\title{Plot aquatic mosquito population density}
\usage{
xds_plot_L(xds_obj, s = 1, clrs = "darkblue", llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{do not plot axes if TRUE}
}
\value{
invisible(NULL)
}
\description{
Plot aquatic mosquito population density
}
