% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{xds_plot_X}
\alias{xds_plot_X}
\title{Default plotting for epidemiology (\strong{X})}
\usage{
xds_plot_X(xds_obj, i = 1, clrs = "black", llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Each module can define a plotting method
for the epidemiology
}
\keyword{internal}
