% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{xds_plot_X.SEISd}
\alias{xds_plot_X.SEISd}
\title{Plot the density of infected individuals for the SEISd model}
\usage{
\method{xds_plot_X}{SEISd}(pars, i = 1, clrs = c("darkblue", "darkred"), llty = 1, add = FALSE)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the SEISd model
}
