% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{xds_plot_X.SIS}
\alias{xds_plot_X.SIS}
\title{Plot the density of infected individuals for the SIS model}
\usage{
\method{xds_plot_X}{SIS}(
  pars,
  i = 1,
  clrs = c("darkblue", "darkred"),
  llty = 1,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the density of infected individuals for the SIS model
}
