% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS-dts.R
\name{xds_plot_X.SIS_dts}
\alias{xds_plot_X.SIS_dts}
\title{Plot the density of infected individuals for the SIS_dts model}
\usage{
\method{xds_plot_X}{SIS_dts}(
  pars,
  i = 1,
  clrs = c("darkblue", "darkred"),
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.xde} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the density of infected individuals for the SIS_dts model
}
