% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-adult.R
\name{xds_plot_Y}
\alias{xds_plot_Y}
\title{Plot the density of infected and infective mosquitoes}
\usage{
xds_plot_Y(xds_obj, s = 1, clrs = "purple", llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{clrs}{a vector of colors for infected mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{do not plot axes if TRUE}
}
\description{
Plot the density of infected and infective mosquitoes
}
