% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-adult.R
\name{xds_plot_Z_fracs}
\alias{xds_plot_Z_fracs}
\title{Plot the fraction infective}
\usage{
xds_plot_Z_fracs(xds_obj, s = 1, clrs = "darkred", llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{clrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{do not plot axes if TRUE}
}
\description{
Plot the fraction infective
}
