% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria.R
\name{xds_port_time_spent_here}
\alias{xds_port_time_spent_here}
\title{Time Spent Here}
\description{
A variable \code{time_spent_here} is used in computing
\link{Transmission} to weight
\code{time_spent} and in \link{Exposure} to weight the local \emph{vs.}
travel EIR
}
