% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet.R
\name{xds_setup_bednets}
\alias{xds_setup_bednets}
\title{Set up dynamic forcing}
\usage{
xds_setup_bednets(
  pars,
  distribute_name = "none",
  distribute_opts = list(),
  own_name = "none",
  own_opts = list(),
  use_name = "none",
  use_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  effectsizes_name = "none",
  effectsizes_opts = list()
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{distribute_name}{the name of a model for mass bed net distribution}

\item{distribute_opts}{options for the bed net distribution model}

\item{own_name}{the name of a model for bed net ownership}

\item{own_opts}{options for the bed net ownership model}

\item{use_name}{the name of a model for bed net usage}

\item{use_opts}{options for the bed net usage model}

\item{effects_name}{the name of a model for bed net effects}

\item{effects_opts}{options for the bed net effects model}

\item{coverage_name}{the name of a model for bed net coverage}

\item{coverage_opts}{options for the bed net coverage model}

\item{effectsizes_name}{the name of a model for bed net effect sizes}

\item{effectsizes_opts}{options for the bed net effect sizes model}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
