% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_cohort}
\alias{xds_setup_cohort}
\title{Set up an \strong{\code{xds}} object to model human cohort dynamics}
\usage{
xds_setup_cohort(
  eir,
  bday = 0,
  F_season = F_flat,
  F_trend = F_flat,
  F_age = F_flat,
  xds = "xde",
  dlay = "ode",
  Xname = "SIS",
  Xopts = list(),
  Xday = 1,
  HPop = 1000,
  searchB = 1,
  model_name = "unnamed"
)
}
\arguments{
\item{eir}{is the entomological inoculation rate}

\item{bday}{the birthday of a cohort}

\item{F_season}{a function describing a seasonal pattern over time}

\item{F_trend}{a function describing a temporal trend over time}

\item{F_age}{a assigning a biting weight by age}

\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{dlay}{is either "ode" or "dde"}

\item{Xname}{is a character string defining a X model}

\item{Xopts}{a list to configure the X model}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{HPop}{is the number of humans in each stratum}

\item{searchB}{is a vector of search weights for blood feeding}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
a \link{list}
}
\description{
The user supplies a function \code{F_eir(time, birthday, ...)} and specifies a model, and
the \code{xds} is set up to simulate dynamics for cohorts of different
ages experiencing exposure.
}
\details{
Please write me
}
