% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm.R
\name{xds_setup_lsm}
\alias{xds_setup_lsm}
\title{Set up dynamic forcing}
\usage{
xds_setup_lsm(
  pars,
  treat_habitats_name = "none",
  treat_habitats_opts = list(),
  effects_name = "none",
  effects_opts = list(),
  coverage_name = "none",
  coverage_opts = list(),
  effectsizes_name = "none",
  effectsizes_opts = list()
)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{treat_habitats_name}{the name of a model for mass LSM distribution}

\item{treat_habitats_opts}{options for the LSM distribution model}

\item{effects_name}{the name of a model for LSM effects}

\item{effects_opts}{options for the LSM effects model}

\item{coverage_name}{the name of a model for LSM coverage}

\item{coverage_opts}{options for the LSM coverage model}

\item{effectsizes_name}{the name of a model for LSM effect sizes}

\item{effectsizes_opts}{options for the LSM effect sizes model}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic forcing has not
already been set up, then turn on dynamic
forcing and set all the
}
