% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_mosy}
\alias{xds_setup_mosy}
\title{Build a Model of Mosquito Ecology}
\usage{
xds_setup_mosy(
  model_name = "unnamed",
  xds = "ode",
  MYname = "basicM",
  Lname = "basicL",
  nPatches = 1,
  membership = 1,
  HPop = 1000,
  searchQ = list(),
  kappa = list(),
  MYoptions = list(),
  Koptions = list(),
  Loptions = list()
)
}
\arguments{
\item{model_name}{is a name for the model}

\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{MYname}{is a character string defining a \strong{MY} Component module}

\item{Lname}{is a character string defining a \strong{L} Component module}

\item{nPatches}{the number of patches in the model}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{HPop}{is the human / host population density}

\item{searchQ}{is a vector of search weights for egg laying}

\item{kappa}{is a vector describing net infectiousness}

\item{MYoptions}{a list to configure the \strong{MY} Component module}

\item{Koptions}{a K matrix, or options for \link{setup_K_matrix} (also see \link{xds_info_mosquito_dispersal})}

\item{Loptions}{a list to configure the \strong{L} Component module}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax
A modified version of \link{xds_setup} that streamlines setup for models without parasite / pathogen infection
dynamics. These models lack the \strong{Y} component and an \strong{X} component, but they will often need an \strong{H}
component (host density).

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'mosy'}
to dispatch functions that compute derivatives (\code{xde}), update variables (\code{dts}),
and parse outputs

The \strong{X} Component module is \code{trivial}, but since humans / vertebrate hosts can be a
resource, \code{HPop} must be set.
}
\seealso{
\link{xds_setup} and \link{dMYdt.basicM}
}
