% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{xds_shrink.cohort}
\alias{xds_shrink.cohort}
\title{Shrink an xds model object}
\usage{
\method{xds_shrink}{cohort}(model)
}
\arguments{
\item{model}{an xds model object}
}
\value{
a smaller xds model object
}
\description{
Shrink an xds model object
}
