% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{xds_shrink}
\alias{xds_shrink}
\title{Shrink an xds xds_obj object}
\usage{
xds_shrink(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
invisible()
}
\description{
Shrink an xds xds_obj object
}
