% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve.dde}
\alias{xds_solve.dde}
\title{Solve a System of Delay Differential Equations}
\usage{
\method{xds_solve}{dde}(xds_obj, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time step}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{xds_solve} for delay differential equations
}
