% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve.dts}
\alias{xds_solve.dts}
\title{Solve a discrete-time system}
\usage{
\method{xds_solve}{dts}(pars, Tmax = 365, dt = NULL, times = NULL)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}

\item{times}{the times}
}
\value{
a \link{list}
}
\description{
Implements \link{xds_solve} for equations
}
