% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve}
\alias{xds_solve}
\title{Solve a system of differential equations}
\usage{
xds_solve(pars, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}

\item{times}{the times}
}
\value{
pars an \strong{\code{xds}} object
}
\description{
This method solves a system of equations and returns the values of
the dependent variables at \code{times} or at  \eqn{t=0, dt, 2\;dt, ... Tmax}.
}
\note{
Note that \link{xde_derivatives} dispatches on \code{class(frame)}
}
\seealso{
\code{\link[=xde_derivatives]{xde_derivatives()}}
}
