% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve_cohort.R
\name{xds_solve_cohort}
\alias{xds_solve_cohort}
\title{Solve Cohort Dynamics}
\usage{
xds_solve_cohort(xds_obj, birthday = 0, Amax = 365, da = 1, ages = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{birthday}{a cohort's birthday}

\item{Amax}{the oldest year, run from 0...Amax}

\item{da}{the age interval}

\item{ages}{a set of ages}
}
\description{
Given a trace function describing the average EIR in a population over time,
simulate malaria in a cohort as it ages: the independent variable is age
\eqn{(a)} and not time \eqn{(t)}.

Relative biting rates by
age are set by \eqn{F_\omega(a)} or \code{F_age} with associated parameters \code{age_par}.

To solve, the argument \code{birthday} sets the birthday for a cohort, \eqn{(d)}.
The user ether passes \code{ages} or \code{Amax} and \code{da} to configure a vector of
ages at which output is wanted \eqn{(a_i)}, at times \eqn{t_i=a_i+d}.

In effect, since \eqn{a = t-d}, the function solves the
solves the system for for an exposure function with the pattern:
\deqn{F_w(a) \times F_S(t-d) \times F_T(t-d) \times F_K(t-d)}
}
\note{
Use \code{xds_setup_eir} to set up a model for cohort dynamics.

During setup, the variable \code{xds_obj$EIR_obj$bday} is set to 0,
and \code{F_age} is set to \code{F_one}.
}
