% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_stable_orbit.dts}
\alias{xds_stable_orbit.dts}
\title{Compute the stable orbit for a discrete time system}
\usage{
\method{xds_stable_orbit}{dts}(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of configurable options}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run a discrete time system for 10 years, and save the last year's orbits.
}
\keyword{internal}
