% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_stable_orbit}
\alias{xds_stable_orbit}
\title{Compute stable orbits}
\usage{
xds_stable_orbit(xds_obj, Y = 10, tol = 1e-05)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{Y}{the number of years to run once}

\item{tol}{the desired accuracy}
}
\value{
an \strong{\code{xds}} object
}
\description{
Run the system for 10 years, and save the last year's orbits.
}
