% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_steady.dde}
\alias{xds_steady.dde}
\title{Compute the steady state for DDEs}
\usage{
\method{xds_steady}{dde}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Call \code{runsteady} to get the steady state of a
system of delay differential equations.
}
\keyword{internal}
