% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_steady.ode}
\alias{xds_steady.ode}
\title{Compute the steady state for ODEs}
\usage{
\method{xds_steady}{ode}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Call \code{rootSolve} to get the steady state of a
system of ordinary differential equations.
}
\keyword{internal}
