% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-help.R
\name{XH_change}
\alias{XH_change}
\title{A List of Functions to Change the XH Component}
\description{
write me.

\itemize{
\item \code{\link{change_H}} -- change host population density
\item \code{\link{change_XH_inits}} -- change initial values
\item \code{\link{change_XH_pars}} -- change parameters
}
}
