% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-help.R
\name{XH_get}
\alias{XH_get}
\title{Get Functions for XH}
\description{
A list of functions to inspect XH objects.
}
\section{Outputs}{

\itemize{
\item \code{\link{get_XH_orbits}} -- get the XH orbits
\item \code{\link{get_EIR}} -- get the EIR orbits
\item \code{\link{get_PR}} -- get the PR orbits
\item \code{\link{get_H}} -- get the PR orbits
}
}

\section{Variables}{

\itemize{
\item \code{\link{get_XH_inits}} -- get initial values
\item \code{\link{get_XH_vars}} -- get state variables from \eqn{y}
\item \code{\link{get_XH_ix}} -- get variable indices
}
}

\section{Parameters}{

\itemize{
\item \code{\link{get_XH_pars}} -- get parameters
\item \code{\link{get_TimeSpent_matrix}} -- get the Time Spent Matrix
}
}

