% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-modules.R
\name{XH_module_list}
\alias{XH_module_list}
\title{XH modules}
\description{
A List of Modules for the \strong{XH}-Component
}
\section{In \strong{\code{ramp.xds}}}{

\itemize{
\item \code{trivial} - the trivial module: see \link{trivial_XH}
\item \code{SIS} - an \strong{SIS} compartmental model: see \link{SIS}
\item \code{hMoI} - a hybrid model for exposure: see \link{hMoI}
}
}

\section{In \strong{\code{ramp.library}}}{

\itemize{
\item \strong{Malaria}
\itemize{
\item \code{garki} - an implementation of the Garki Model for malaria
\item \code{SIP}   - a compartmental model with treatment and chemoprotection
\item \code{SIPd}  - an implementation of the Garki Model for malaria
}
\item \strong{Compartmental Models}
\itemize{
\item \code{SIR} - an \strong{SIR} compartmental model
\item \code{SIRS} - an \strong{SIRS} compartmental model
\item \code{SEIR} - an \strong{SEIR} compartmental model
\item \code{SEIRV} - an \strong{SEIRV} compartmental model
\item \code{SEIS} - an \strong{SEIS} compartmental model
\item \code{SEISd} - an \strong{SEIS} compartmental model, as a delay differential equation
}
}

\strong{\code{ramp.library}} --- \href{https://github.com/dd-harp/ramp.library}{source code} & \href{https://dd-harp.github.io/ramp.library}{github.io}
}

