% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-setup.R
\name{add_variable}
\alias{add_variable}
\title{Add Variable}
\usage{
add_variable(var_object, xds_obj)
}
\arguments{
\item{var_object}{an object to dispatch and compute a variable}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the \strong{\code{xds}} model object with a new variable
}
\description{
This is a generic method to add new dependent state variables
that are computed by \code{dVdt.}

The model object for the new variable should fully define
an other variables module, including methods to set up
initial values.

The description for each method should include the equations.
}
