% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_spray-effectsizes-interface.R
\name{AreaSprayEffectSizes}
\alias{AreaSprayEffectSizes}
\title{Set the area spray effectsizes}
\usage{
AreaSprayEffectSizes(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of exogenous variables related to
area spray effectsizes
}
