% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{basicL}
\alias{basicL}
\title{The \code{basicL} Module for the L Component}
\description{
Implements the \strong{L} component using a basic model of aquatic mosquito
ecology with density-dependent mortality and delayed maturation, based
on Smith DL, \emph{et al.} (2013).
}
\section{State Variables}{

\describe{
\item{\code{L}}{density of mosquito larvae in each habitat}
}
}

\section{Parameters}{

\describe{
\item{\code{psi}}{maturation rate (\eqn{\psi})}
\item{\code{xi}}{delayed maturation parameter in response to mean crowding (\eqn{\xi})}
\item{\code{phi}}{density-independent death rate (\eqn{\phi})}
\item{\code{theta}}{slope of density-dependent mortality (\eqn{\theta})}
}
}

\section{Dynamics}{

\deqn{dL/dt = \eta - (\psi e^{-\xi L} + \phi + \theta L) L}

Emergence rate: \deqn{\alpha = \psi e^{-\xi L} L}

Setting \eqn{\xi = 0} recovers the model of Smith DL, \emph{et al.} (2013).
}

\references{
Smith DL, \emph{et al.} (2013)
}
