% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography-interface.R
\name{Births}
\alias{Births}
\title{A function that computes the birth rate for human populations}
\usage{
Births(t, y, Hpar)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{Hpar}{a \link{list}}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of pars$Hpar$Births
}
