% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography.R
\name{Births}
\alias{Births}
\title{Human (or Host) Population Birth Rate}
\usage{
Births(t, H, births)
}
\arguments{
\item{t}{current simulation time}

\item{H}{population density}

\item{births}{the birth function object}
}
\value{
see help pages for specific methods
}
\description{
This method dispatches on the type of xds_obj$Hpar$Births
}
