% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-demography.R
\name{Births.zero}
\alias{Births.zero}
\title{Human (or Host) Population Birth Rate}
\usage{
\method{Births}{zero}(t, H, births = "zero")
}
\arguments{
\item{t}{current simulation time}

\item{H}{population density}

\item{births}{the birth function object}
}
\value{
see help pages for specific methods
}
\description{
a function
}
