% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_weights.R
\name{blood_search_weights}
\alias{blood_search_weights}
\title{Blood Search Weights}
\description{
Blood feeding is an interaction among humans and mosquitoes:
in this model, humans are spending time, and
mosquitoes are searching. To model exposure, each human (or host)
population stratum is assigned a \emph{search weight,} a
number used to weigh \emph{time spent} and get a measure of
\emph{availability.}
Search weights for blood feeding thus play a key role in blood feeding (see \link{blood_feeding})
}
