% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\title{Burn-in}
\usage{
burnin(model, t_neg_inf = -3650)
}
\arguments{
\item{model}{a \strong{\code{ramp.xds}} model object}

\item{t_neg_inf}{run from -abs(t_neg_inf) days to \eqn{t=0}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Use the defined hindcast to
run a model from the past up to the present.
Reset the initial conditions.
}
