% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{change_bad_habitat}
\alias{change_bad_habitat}
\title{Change Habitat Search Weights}
\usage{
change_bad_habitat(Qbad, xds_obj, s = 1)
}
\arguments{
\item{Qbad}{availability of bad habitat}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \code{xds} object
}
\description{
Set the search weights, \eqn{\omega}, for a set of aquatic habitats
}
