% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{change_blood_search_weights}
\alias{change_blood_search_weights}
\title{Change Blood Search Weights}
\usage{
change_blood_search_weights(wts, xds_obj, s = 1, i = 1)
}
\arguments{
\item{wts}{the blood feeding search weights}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
an \code{xds} object
}
\description{
Change the blood feeding search weights, \eqn{\omega},
for a set of host strata
}
\keyword{internal}
