% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_eip}
\alias{change_eip}
\title{Change the EIP}
\usage{
change_eip(eip, xds_obj, s)
}
\arguments{
\item{eip}{the extrinsic incubation period}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the extrinsic incubation
period for a static model
}
