% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_f}
\alias{change_f}
\title{Change the blood feeding rate}
\usage{
change_f(f, xds_obj, s)
}
\arguments{
\item{f}{the blood feeding rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the blood feeding
rate for a static model
}
