% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_season.R
\name{change_F_season}
\alias{change_F_season}
\title{Change F_season}
\usage{
change_F_season(F, xds_obj, s = 1)
}
\arguments{
\item{F}{a seasonal pattern function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species indeF}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the seasonality function
}
