% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_shock.R
\name{change_F_shock}
\alias{change_F_shock}
\title{Change F_shock}
\usage{
change_F_shock(F, xds_obj, s = 1)
}
\arguments{
\item{F}{new shock function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species indeF}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the shock function
}
