% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_trend.R
\name{change_F_trend.eir}
\alias{change_F_trend.eir}
\title{Change F_trend}
\usage{
\method{change_F_trend}{eir}(F, xds_obj, s = 1)
}
\arguments{
\item{F}{new trend function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species indeF}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the trend function
when \code{forced_by = "eir"}
}
\keyword{internal}
