% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_trend.R
\name{change_F_trend.none}
\alias{change_F_trend.none}
\title{Change F_trend}
\usage{
\method{change_F_trend}{none}(F, xds_obj, ix = 1)
}
\arguments{
\item{F}{new trend function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the trend function
when \code{forced_by = "none"}
}
\keyword{internal}
