% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_F_trend.R
\name{change_F_trend}
\alias{change_F_trend}
\title{Change F_trend}
\usage{
change_F_trend(F, xds_obj, ix = 1)
}
\arguments{
\item{F}{new trend function}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change  the trend function. The function \code{F} should have the form
\eqn{F(t,V)}. If the variable \eqn{V} is not used, it's default value should be set to an empty list.
}
\seealso{
\link{F_one}
}
