% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_g}
\alias{change_g}
\title{Change the mosquito mortality rate}
\usage{
change_g(g, xds_obj, s)
}
\arguments{
\item{g}{the mosquito mortality rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the mosquito mortality
rate for a static model
}
