% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{change_H}
\alias{change_H}
\title{Change human population density}
\usage{
change_H(H, xds_obj, i = 1)
}
\arguments{
\item{H}{human population density}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Change human population density
}
