% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-change.R
\name{change_H}
\alias{change_H}
\title{Change human population density}
\usage{
change_H(H, pars, i = 1)
}
\arguments{
\item{H}{human population density}

\item{pars}{a \link{list}}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
Change human population density
}
