% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-search.R
\name{change_habitat_search_weights}
\alias{change_habitat_search_weights}
\title{Change habitat search weights}
\usage{
change_habitat_search_weights(wts, xds_obj, s = 1)
}
\arguments{
\item{wts}{the habitat feeding search weights}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the habitat feeding search weights, \eqn{\omega},
for a set of host strata
}
