% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_dynamics.R
\name{change_habitat_weights}
\alias{change_habitat_weights}
\title{Change Habitat Search Weights}
\usage{
change_habitat_weights(wts, xds_obj, s = 1)
}
\arguments{
\item{wts}{habitat search weights}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set the search weights, \eqn{\omega}, for a set of aquatic habitats
}
