% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{change_K_matrix}
\alias{change_K_matrix}
\title{Change Mosquito Dispersal Matrix}
\usage{
change_K_matrix(K_matrix, xds_obj, behave = "0", s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{behave}{the behavioral state}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Check that \code{K_matrix}
\itemize{
\item is an \code{nPatches} \eqn{\times} \code{nPatches} matrix;
\item and columns sum to 0.
}

After passing checks, \code{xds_obj} is updated.

In models with multiple species, use \code{s} to
specify the species to update.
}
\seealso{
\link{xds_info_mosquito_dispersal}; \link{setup_K_matrix}
}
