% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{change_K_matrix}
\alias{change_K_matrix}
\title{Change Mosquito Dispersal Matrix}
\usage{
change_K_matrix(K_matrix, xds_obj, s = 1)
}
\arguments{
\item{K_matrix}{a mosquito dispersal \link{matrix}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change Mosquito Dispersal Matrix
}
