% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{change_L_inits}
\alias{change_L_inits}
\title{Set \strong{L} Component Initial Values}
\usage{
change_L_inits(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \code{xds} object
}
\description{
This sets initial values for an \strong{L} Component
module from a named list
}
