% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{change_L_pars.trivial}
\alias{change_L_pars.trivial}
\title{Set \strong{L} Component parameters for \code{trivial}}
\usage{
\method{change_L_pars}{trivial}(xds_obj, s = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
If \code{Lambda} or \code{F_season} or \code{F_trend}
are named in a list \code{options}, the old value is replaced
}
