% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_mean_forcing.R
\name{change_mean_forcing.eir}
\alias{change_mean_forcing.eir}
\title{Change mean forcing}
\usage{
\method{change_mean_forcing}{eir}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Change the mean daily EIR for an \code{eir} model
}
\keyword{internal}
