% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_mean_forcing.R
\name{change_mean_forcing.none}
\alias{change_mean_forcing.none}
\title{Set mean forcing}
\usage{
\method{change_mean_forcing}{none}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new mean forcing parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{change_mean_forcing} for a model
with no forcing
}
