% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_mu}
\alias{change_mu}
\title{Change the emigration loss fraction}
\usage{
change_mu(mu, xds_obj, s)
}
\arguments{
\item{mu}{the emigration loss fraction}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the emigration loss
fraction for a static model
}
