% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_nu}
\alias{change_nu}
\title{Change the egg laying rate}
\usage{
change_nu(nu, xds_obj, s)
}
\arguments{
\item{nu}{the egg laying rate (# batches, per mosquito, per day)}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the egg laying
rate for a static model
}
