% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{change_other_blood_hosts}
\alias{change_other_blood_hosts}
\title{Set static blood feeding search weights}
\usage{
change_other_blood_hosts(other_blood_hosts, xds_obj, s)
}
\arguments{
\item{other_blood_hosts}{availability of other blood hosts}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the blood feeding search weights, \eqn{\omega}, for a set of host strata
}
