% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_q}
\alias{change_q}
\title{Change the human fraction}
\usage{
change_q(q, xds_obj, s)
}
\arguments{
\item{q}{the human fraction}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the human fraction
for a static model
}
