% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_season.R
\name{change_season.Lambda}
\alias{change_season.Lambda}
\title{Get seasonal pattern}
\usage{
\method{change_season}{Lambda}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{a list with new parameters for bottom, phase, and pw}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Get the parameters set the seasonal pattern
}
