% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_shock.R
\name{change_shock.eir}
\alias{change_shock.eir}
\title{Set yy}
\usage{
\method{change_shock}{eir}(shock_par, xds_obj, s = 1)
}
\arguments{
\item{shock_par}{parameters for \link{make_function}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the yy for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
