% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_shock.R
\name{change_shock.Lambda}
\alias{change_shock.Lambda}
\title{Set the \eqn{y} value for interpolating points}
\usage{
\method{change_shock}{Lambda}(shock_par, xds_obj, s = 1)
}
\arguments{
\item{shock_par}{parameters for \link{make_function}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{ramp.xds}} model object
}
\description{
and return the \strong{\code{ramp.xds}} model object
}
