% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-change.R
\name{change_sigma}
\alias{change_sigma}
\title{Change the mosquito patch emigration rate}
\usage{
change_sigma(sigma, xds_obj, s)
}
\arguments{
\item{sigma}{the mosquito patch emigration rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Change the mosquito patch emigration
rate for a static model
}
