% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_trend.R
\name{change_spline.Lambda}
\alias{change_spline.Lambda}
\title{Set the \eqn{y} value for interpolating points}
\usage{
\method{change_spline}{Lambda}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new interpolating points}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the \strong{\code{ramp.xds}} model object
}
\description{
Set the yy for the seasonal pattern for the
\eqn{s^{th}} species
and return the \strong{\code{ramp.xds}} model object
}
