% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_trend.R
\name{change_spline_y.eir}
\alias{change_spline_y.eir}
\title{Set yy}
\usage{
\method{change_spline_y}{eir}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new yy parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the yy for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
