% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_trend.R
\name{change_spline_y}
\alias{change_spline_y}
\title{Set yy}
\usage{
change_spline_y(X, xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{yy}{new y values for the interpolation points}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the yy parameter to \code{X}
}
