% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{change_timespent_matrix}
\alias{change_timespent_matrix}
\title{Change the time spent matrix}
\usage{
change_timespent_matrix(timespent, xds_obj, i = 1)
}
\arguments{
\item{timespent}{a time spent matrix}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the time spent matrix for the \eqn{i^{th}} host species, and
trigger updates for the \code{XY} interface.
}
