% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{change_TimeSpent_matrix}
\alias{change_TimeSpent_matrix}
\title{Set up (or change) a Time Spent matrix}
\usage{
change_TimeSpent_matrix(TimeSpent, xds_obj, i = 1)
}
\arguments{
\item{TimeSpent}{a time spent matrix}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a modified \strong{\code{xds}} model object
}
\description{
Set up (or change) a Time Spent matrix
}
