% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{change_travel_EIR}
\alias{change_travel_EIR}
\title{Change the travel EIR}
\usage{
change_travel_EIR(teir, xds_obj, i)
}
\arguments{
\item{teir}{the travel entomological inoculation rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set a new \emph{static} value for the travel EIR
}
