% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{change_travel_EIR}
\alias{change_travel_EIR}
\title{Change the travel EIR}
\usage{
change_travel_EIR(teir, xds_obj, i)
}
\arguments{
\item{teir}{availability of other blood hosts}

\item{xds_obj}{an \code{xds} object}

\item{i}{the host species index}
}
\value{
an \code{xds} object
}
\description{
Set a new \emph{static} value for travel EIR
}
