% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{change_visitors}
\alias{change_visitors}
\title{Change Availability of Visitors}
\usage{
change_visitors(visitors, xds_obj, s)
}
\arguments{
\item{visitors}{availability of visitors}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the value of the parameter describing
available visitors
}
