% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{change_XH_inits.SIS}
\alias{change_XH_inits.SIS}
\title{Return the parameters as a list}
\usage{
\method{change_XH_inits}{SIS}(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Return the parameters as a list
}
