% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{change_XH_pars.SIS}
\alias{change_XH_pars.SIS}
\title{Set new values for the SIS model}
\usage{
\method{change_XH_pars}{SIS}(xds_obj, i = 1, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the vector species index}

\item{options}{a named list}
}
\value{
an \strong{\code{xds}} model object
}
\description{
A utility to change parameters
values for the SIS model (\strong{XH} component)

\code{options} is a named list, and:
\itemize{
\item \eqn{b} is set to \code{options$b}
\item \eqn{c} is set to \code{options$c}
\item \eqn{r} is set to \code{options$r}
}
}
