% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{check_K_matrix}
\alias{check_K_matrix}
\title{Check K Matrix}
\usage{
check_K_matrix(K, Np, tol = 1e-12)
}
\arguments{
\item{K}{a mosquito dispersal matrix}

\item{Np}{the number of patches}

\item{tol}{tolerance}
}
\description{
Check that
\itemize{
\item \eqn{K} is a \eqn{N_p \times N_p} matrix
\item if not zero, the diagonal elements are all \eqn{-1}
\item the columns sum to 0: tolerance is set by \code{tol}
}
}
\seealso{
\link{xds_info_mosquito_dispersal}
}
