% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bednet-lemenach.R
\name{compute_bednet_effect_sizes_lemenach}
\alias{compute_bednet_effect_sizes_lemenach}
\title{Modify baseline values due to vector control}
\usage{
compute_bednet_effect_sizes_lemenach(ix, phi, ff, qq, gg, tau0_frac, r, ss)
}
\arguments{
\item{ix}{an index over nPatches}

\item{phi}{ITN coverage}

\item{ff}{baseline blood feeding rate}

\item{qq}{baseline human fraction}

\item{gg}{baseline mosquito mortality rate}

\item{tau0_frac}{a \link{numeric} vector giving the proportion of a feeding cycle spent
host seeking/bloodfeeding \emph{vs.} resting/oviposition}

\item{r}{probability of mosquito being repelled upon contact with ITN}

\item{ss}{probability of mosquito successfully feeding upon contact with ITN}
}
\value{
a named \link{list}
}
\description{
Implements \link{BedNetEffectSizes} for the Le Menach ITN model of vector control
}
\references{
{This implements the model for ITN effect sizes
from \insertRef{LeMenachA2007_ITN}{ramp.xds}}
}
