% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_eggs_laid}
\alias{compute_eggs_laid}
\title{Compute eggs laid}
\usage{
compute_eggs_laid(eggs_laid, calU)
}
\arguments{
\item{eggs_laid}{the number of eggs laid in each patch, a vector of length \code{nPatches}}

\item{calU}{the egg distribution matrix}
}
\value{
a \link{vector}, \eqn{\eta} where \eqn{\left|\eta\right|=}\code{nHabitats}
}
\description{
Computes egg distribution for the aquatic habitats
}
\seealso{
The egg distribution matrix \link{compute_calU}
}
