% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{compute_eggs_laid}
\alias{compute_eggs_laid}
\title{Compute eggs laid}
\usage{
compute_eggs_laid(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Computes eggs laid in aquatic habitats
}
\seealso{
\link{compute_eggs_laid}
}
\keyword{internal}
