% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{compute_local_frac}
\alias{compute_local_frac}
\title{Compute the local fraction}
\usage{
compute_local_frac(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Compute the availability for the pathogen's hosts for blood feeding
}
\keyword{internal}
