% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{compute_Q}
\alias{compute_Q}
\title{Compute the total availability of egg-laying habitats, \eqn{Q}}
\usage{
compute_Q(habitat_matrix, search_weights)
}
\arguments{
\item{habitat_matrix}{the membership matrix, \eqn{N}}

\item{search_weights}{the habitat search weights, \eqn{\omega_q}}
}
\value{
a \link{vector} of describing habitat availability, \eqn{Q}, of length \code{nPatches}
}
\description{
Compute the availability of aquatic habitats.
}
\details{
The availability of the habitats that we have defined in the model, denoted \eqn{Q}, sums search weights, \eqn{\omega_q}, by patch
using the habitat membership matrix, \eqn{N}, and we can compute
\deqn{Q = {N} \cdot \omega_q.}
}
\seealso{
This function is called by \link{make_Q}

\link{create_habitat_matrix} discusses \eqn{N}

The availability of ovitraps and bad habitats is setup in \link{setup_EGG_LAYING}
}
