% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_RBR}
\alias{compute_RBR}
\title{Compute relative biting rates}
\usage{
compute_RBR(search_weights, H)
}
\arguments{
\item{search_weights}{blood feeding search weights for the host strata}

\item{H}{host density}
}
\value{
host availability, a \link{vector}
}
\description{
Relative biting rates translate biting weights into a
frailty parameter -- a multiplicative term.
}
