% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_TaR}
\alias{compute_TaR}
\title{Compute blood feeding availability of all vertebrate hosts}
\usage{
compute_TaR(t, TiSp, F_circadian)
}
\arguments{
\item{t}{the time}

\item{TiSp}{a time spent matrix}

\item{F_circadian}{a function to compute relative activity rates by time of day}
}
\value{
a TaR \link{matrix}
}
\description{
Computes availability all vertebrate hosts to blood feeding
mosquitoes
}
\details{
The availability of other vertebrate hosts is a sum of available local
hosts \eqn{W}, and visitors \eqn{W_\delta}, and other available vertebrate hosts, \eqn{O}.
Total availability is a simple sum: \deqn{B = W + W_\delta + O}
}
\references{
{This function implements equation 3
from \insertRef{WuSL2023SpatialDynamics}{ramp.xds} }
}
