% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_TaR}
\alias{compute_TaR}
\title{Make TaR}
\usage{
compute_TaR(xds_obj, t = 0)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{t}{the time}
}
\value{
an \code{xds} object
}
\description{
Make a time at risk matrix (TaR) from a time spent matrix and a circadian function
}
\keyword{internal}
