% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{compute_W}
\alias{compute_W}
\title{Compute Host Availability for Blood Feeding}
\usage{
compute_W(search_weights, H, TaR)
}
\arguments{
\item{search_weights}{blood feeding search weights for the host strata}

\item{H}{host density}

\item{TaR}{the host species index}
}
\value{
host availability, a \link{vector}
}
\description{
Compute the availability of the population strata defined
in the model as hosts for blood feeding by mosquitoes
}
\details{
Host availability to blood searching mosquitoes in patches is
the sum of search weights of the human strata, a vector \eqn{\omega},
weighted by time at risk, defined by a matrix \eqn{\Psi}
that is \eqn{n_p \times n_h}. The search weight is a \emph{per-capita} measure
so we weight it by human population density, \eqn{H}. Availability, \eqn{W}
is computed as \deqn{\Psi \cdot (\omega H).}
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds} }
}
\seealso{
Availability of all vertebrate hosts for blood feeding is computed by \code{\link[=compute_B]{compute_B()}}
}
