% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{create_Lpar_trivial}
\alias{create_Lpar_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
create_Lpar_trivial(
  nHabitats,
  Lopts = list(),
  Lambda = 1000,
  F_season = F_no_season,
  phase = 0,
  season_opts = list(),
  F_trend = F_no_trend,
  trend_opts = list()
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{F_season}{a F_seasonality function}

\item{phase}{a parameter to set the phase of F_season}

\item{season_opts}{a \link{list} of options to pass to F_season}

\item{F_trend}{a F_trend function}

\item{trend_opts}{a \link{list} of options to pass to F_trend}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
