% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-sei.R
\name{create_MYZinits_sei}
\alias{create_MYZinits_sei}
\title{Make inits for sei adult mosquito model}
\usage{
create_MYZinits_sei(nPatches, MYZopts = list(), M = 5, Y = 1, Z = 0)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{Y}{total infected mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for sei adult mosquito model
}
