% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{create_MYZinits_si}
\alias{create_MYZinits_si}
\title{Make inits for si adult mosquito model}
\usage{
create_MYZinits_si(nPatches, MYZopts = list(), M = 5, Z = 1)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{M}{total mosquito density at each patch}

\item{Z}{infectious mosquito density at each patch}
}
\value{
a \link{list}
}
\description{
Make inits for si adult mosquito model
}
