% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{create_MYZpar_trivial}
\alias{create_MYZpar_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
create_MYZpar_trivial(
  nPatches,
  MYZopts,
  f = 1,
  q = 1,
  Z = 1,
  eggs = 1,
  F_season = F_no_season,
  phase = 0,
  season_opts = list(),
  F_trend = F_no_trend,
  trend_opts = list()
)
}
\arguments{
\item{nPatches}{an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Z}{the human fraction}

\item{eggs}{the human fraction}

\item{F_season}{a F_seasonality function}

\item{phase}{a parameter to set the phase of F_season}

\item{season_opts}{a \link{list} of options to pass to F_season}

\item{F_trend}{a F_trend function}

\item{trend_opts}{a \link{list} of options to pass to F_trend}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
