% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{create_residence_matrix}
\alias{create_residence_matrix}
\title{Create the residence matrix, \eqn{\cal J}}
\usage{
create_residence_matrix(nPatches, residence)
}
\arguments{
\item{nPatches}{the number of patches}

\item{residence}{a vector describing the patch index for each habitat}
}
\value{
the residence \link{matrix}, denoted \eqn{\cal J} where \eqn{\left|\cal J\right|=}\code{nPatches} \eqn{\times} \code{nStrata}
}
\description{
The residence matrix, \eqn{\cal J}, holds
information about the location of home for each human population stratum.
If \eqn{w} is any vector describing a quantity in strata (\emph{i.e.}, \eqn{\left|w\right|=} \code{nStrata}), then
\eqn{W={\cal J}\cdot w} is a vector that has summed \eqn{w} by residency for the strata, and \eqn{\left|W\right|=} \code{nPatches}.
}
\details{
Information about residence in a patch location for each stratum
is passed as the residence vector, an ordered list of patch locations. If
the \eqn{i^{th}} stratum lives in the \eqn{j^{th}} patch, then
\eqn{{\cal J}[j,i]=1.} other_bloodwise, \eqn{{\cal J}[j,i]=0.}
}
\seealso{
to extract habitat residence information, see \link{view_residence_matrix}
}
