% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{create_Xpar_trivial}
\alias{create_Xpar_trivial}
\title{Make parameters for trivial human model}
\usage{
create_Xpar_trivial(
  nPatches,
  Xopts,
  kappa = 0.1,
  HPop = 1,
  F_season = F_no_season,
  phase = 0,
  season_opts = list(),
  F_trend = F_no_trend,
  trend_opts = list()
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{Xopts}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}

\item{F_season}{a F_seasonality function}

\item{phase}{a parameter to set the phase of F_season}

\item{season_opts}{a \link{list} of options to pass to F_season}

\item{F_trend}{a F_trend function}

\item{trend_opts}{a \link{list} of options to pass to F_trend}
}
\value{
a \link{list}
}
\description{
Make parameters for trivial human model
}
