% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{dLdt.basicL}
\alias{dLdt.basicL}
\title{Derivatives for \code{basicL} (\strong{L} Component)}
\usage{
\method{dLdt}{basicL}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
\loadmathjax
This implements differential equation model for aquatic mosquito ecology.
The equations have been modified slightly from a version published by
Smith DL, \emph{et al.} (2013); this version includes delayed maturation i
in response to mean crowding.

\strong{Variables:}
\itemize{
\item \eqn{L}: the density of mosquito larvae in each habitat
}

\strong{Parameters and Terms:}
\itemize{
\item \eqn{\eta} or \code{eta}: egg deposition rate (from \link{F_eggs})
\item \eqn{\psi} or \code{psi}: maturation rate
\item \eqn{\xi} or \code{xi}: delayed maturation parameter in response to mean crowding
\item \eqn{\phi} or \code{phi}: density-independent death rate
\item \eqn{\theta} or \code{theta}: the slope of the mortality rate in response to mean crowding
}

\strong{Dynamics:}

\deqn{dL/dt = \eta - (\psi\;e^{-\xi L} + \phi + \theta L)L}

Per-capita mortality is thus \eqn{\phi + \theta L}, and the emergence rate
of adult mosquitoes is \eqn{\psi e^{-\xi L} L }
}
\references{
{\insertRef{SmithDL2013LarvalDynamics}{ramp.xds} }
}
\seealso{
\link{make_Lpar_basicL}
}
