% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{dLdt.basicL}
\alias{dLdt.basicL}
\title{\eqn{\cal L} Component Derivatives for a \code{basicL}}
\usage{
\method{dLdt}{basicL}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
This implements differential equation model for aquatic mosquito ecology.
It has been modified slightly from a version published by Smith DL, \emph{et al.} (2013):
\deqn{dL/dt = \eta - (\psi + \phi + \theta L)L}

\strong{Parameters:}
\itemize{
\item \eqn{\eta}: egg deposition rate
\item \eqn{\psi}: maturation rate
\item \eqn{\phi}: density-independent death rate
\item \eqn{\theta}: density dependence in mortality: the slope of the response to mean crowding
}

Per-capita mortality is thus \eqn{\phi + \theta L}, and the emergence rate
of adult mosquitoes is \eqn{\psi L}
}
\references{
{\insertRef{SmithDL2013LarvalDynamics}{ramp.xds} }
}
\seealso{
\link{create_Lpar_basicL}
}
