% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{dLdt.trivial}
\alias{dLdt.trivial}
\title{Derivatives for \code{trivial} (\strong{L} Component)}
\usage{
\method{dLdt}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an empty \link{list}
}
\description{
Return a numeric vector of length 0
}
