% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{dLdt.trivial}
\alias{dLdt.trivial}
\title{Derivatives for the \code{trivial} \strong{L} module}
\usage{
\method{dLdt}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an empty \link{list}
}
\description{
Returns a numeric vector of length 0
}
