% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{dMYdt.GeRM}
\alias{dMYdt.GeRM}
\title{Compute Derivatives for \strong{MY} module \code{GeRM}}
\usage{
\method{dMYdt}{GeRM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for the generalized, non-autonomous Ross-Macdonald model
for mosquito ecology and infection dynamics.
}
