% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{dMYdt.trivial}
\alias{dMYdt.trivial}
\title{Handle derivatives for the \code{trivial} \strong{MY} module}
\usage{
\method{dMYdt}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a null value
}
\description{
Implements \link{dMYdt} for the trivial (forced emergence) model.
}
