% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{dMYZdt}
\alias{dMYZdt}
\title{Compute Derivatives for an Adult Mosquito Model}
\usage{
dMYZdt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
Derivatives for an adult mosquito model, a \link{vector}
}
\description{
In the description, each method should include the equations and
a reference to relevant papers in the literature.
}
\note{
This is the \code{S3} generic. Methods dispatch on \code{MYZpar[[s]]}
}
