% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-RM.R
\name{dMYZdt.RM}
\alias{dMYZdt.RM}
\title{\eqn{\cal MYZ} Component Derivatives for a Ross-Macdonald Mosquito Model}
\usage{
\method{dMYZdt}{RM}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for a generalized, non-autonomous dynamic model
for mosquito ecology and infection dynamics.
}
