% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{dMYZdt.SEI}
\alias{dMYZdt.SEI}
\title{\eqn{\cal MYZ} Component Derivatives for the \code{SEI} Mosquito Model}
\usage{
\method{dMYZdt}{SEI}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dMYZdt} for the SEI ODE model.
}
\details{
The dynamics of adult mosquitoes:
\deqn{\frac{dM}{dt} = \Lambda - \Omega \cdot M}
The density of infected but not infectious mosquitoes:
\deqn{\frac{dY}{dt} = fq\kappa(M-Y) - \Omega \cdot Y}
The density of infectious mosquitoes:
\deqn{\frac{dZ}{dt} = (Y-Z)/\tau - \Omega \cdot Z}
}
