% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{dMYZdt.trivial}
\alias{dMYZdt.trivial}
\title{Handle derivatives for the \code{trivial} \strong{MYZ} module}
\usage{
\method{dMYZdt}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a null value
}
\description{
Implements \link{dMYZdt} for the trivial (forced emergence) model.
}
