% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dts_steady_state_X}
\alias{dts_steady_state_X}
\title{Compute the steady states as a function of the daily EIR}
\usage{
dts_steady_state_X(eir, H, pars, i)
}
\arguments{
\item{eir}{the daily EIR}

\item{H}{host density}

\item{pars}{a list that defines an \code{ramp.xde} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{i}{the host species index}
}
\value{
none
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
