% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts_equations.R
\name{dts_update.full}
\alias{dts_update.full}
\title{Generalized spatial differential equation model}
\usage{
\method{dts_update}{full}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
dts_update_ the state variables
}
