% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-setup.R
\name{dVdt.setup}
\alias{dVdt.setup}
\title{Compute Other Variables}
\usage{
\method{dVdt}{setup}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}

\item{i}{the i^th auxiliary variable}
}
\value{
Derivatives for auxiliary variables as a \link{vector}
}
\description{
The description for each method should include the equations.
}
