% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dXdt}
\alias{dXdt}
\title{Derivatives for \eqn{\cal X} Models}
\usage{
dXdt(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute and return the derivatives
for a model of class \eqn{\cal X} - dynamics of infection and
immunity in humans or other vertebrate host
}
