% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{dXdt}
\alias{dXdt}
\title{Compute Derivatives for the \strong{X} Component}
\usage{
dXdt(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
the derivatives, a \link{numeric} vector
}
\description{
Compute and return the derivatives
for an \strong{X}-Component module.
}
\note{
The function dispatches on \verb{class(Xpar[[i]]).}
}
