% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{dXdt.SEISd}
\alias{dXdt.SEISd}
\title{\eqn{\cal X} Component Derivatives for the \code{SEISd} Model}
\usage{
\method{dXdt}{SEISd}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SEISd model
}
