% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dXdt.SIS}
\alias{dXdt.SIS}
\title{\eqn{\cal X} Component Derivatives for an SIS Model}
\usage{
\method{dXdt}{SIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for SIS compartmental model, defined as:
\deqn{\frac{dH}{dt} = B(t) - h S - r I + d{\cal H}(S)}
\deqn{\frac{dI}{dt} = h S - r I + d{\cal H}(I)}
where \eqn{H = S+I}, \eqn{B(t)} is the time-dependent birth rate, and \eqn{d{\cal H}} is a linear demographic operator.
}
