% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dXdt.SIS}
\alias{dXdt.SIS}
\title{Derivatives for the SIS \eqn{\cal X} Compartment Model}
\usage{
\method{dXdt}{SIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{dXdt} for the SIS model for human / vertebrate host infections
}
\details{
The core dynamics of infection
\deqn{\frac{dS}{dt} = -h S + r I}
}
