% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{dXdt.SIS}
\alias{dXdt.SIS}
\title{\strong{X} Component Derivatives for the \code{SIS} Model}
\usage{
\method{dXdt}{SIS}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
Compute the derivatives for SIS compartmental model. Here, the model includes human demographic changes,
and it is computed in an equivalent form:
\deqn{
\begin{array}{rl}
dH/dt = & B(t,H)  + {\cal D} \cdot H \\
dI/dt = & h (H-I) - r I - \xi(t) + {\cal D} \cdot I
\end{array}
}
where \eqn{S=H-I} ;
\eqn{\xi(t)} is a function to simulate mass treatment;
\eqn{B(t, H)} is the time-dependent birth rate; and the \eqn{\cal D} is a matrix describing demographic changes,
including mortality, migration, and aging;
}
