% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trivial.R
\name{dXdt.trivial}
\alias{dXdt.trivial}
\title{Compute Derivatives for the \code{trivial} \eqn{\cal X} Model}
\usage{
\method{dXdt}{trivial}(t, y, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{numeric} vector
}
\description{
The trivial model has no state variables so it returns
a numeric vector of length 0
}
