% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{dXHdt.SIS}
\alias{dXHdt.SIS}
\title{Derivatives function for the \code{SIS} model (\emph{XH}* Model)}
\usage{
\method{dXHdt}{SIS}(t, y, xds_obj, i)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
the derivatives, as a vector
}
\description{
Compute the derivatives for SIS compartmental model:
\itemize{
\item \eqn{H}  is human (or host) population density
\item \eqn{I} is the density of infected humans (or hosts)
\item \eqn{S=H-I} is the density of susceptible humans (or hosts)
}

The clearance rate for infections is \eqn{r}, and by assumption,
individuals are assumed to be susceptible to infection after clearing
infections.

The model has a port to model mass treatment, in equations \eqn{\xi(t)}.

The \strong{\code{xds}} implementation computes \eqn{dH/dt}
rather than \eqn{dS/dt.} In the functions \link{get_XH_vars}
and \link{parse_XH_orbits},
\eqn{S} is computed as \eqn{S=H-I} and listed as a variable. The
derivatives computed are:

\deqn{
\begin{array}{rl}
dH/dt = & B(t,H)  + D \cdot H \\
dI/dt = & h (H-I) - r I - \xi(t) +  D \cdot I
\end{array}
}
where \eqn{S=H-I};

\eqn{\xi(t)} is a function to simulate mass treatment;
\eqn{B(t, H)} is the time-dependent birth rate; and \eqn{D} is a linear operator, a matrix describing demographic changes,
including mortality, migration, and aging;
}
