% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control.dynamic}
\alias{dynamic_vector_control.dynamic}
\title{Set up dynamic vector_control}
\usage{
\method{dynamic_vector_control}{dynamic}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
If dynamic vector_control has not
already been set up, then turn on dynamic
vector_control and set all the
}
