% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{dynamic_vector_control}
\alias{dynamic_vector_control}
\title{Turn On Vector Control}
\usage{
dynamic_vector_control(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Any function that sets up
non-trivial vector control must
calls this function.
If \code{class(pars$vector_control) == 'none'}
then it is set to \code{dynamic} and the
trivial module for every mode of
vector control is set up. Otherwise,
nothing happens.
}
