% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{EggLaying}
\alias{EggLaying}
\title{Compute eggs laid}
\usage{
EggLaying(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Compute eggs laid at time \eqn{t}
}
\seealso{
\code{\link[=setup_ML_interface]{setup_ML_interface()}}
}
\keyword{internal}
