% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{EggLaying}
\alias{EggLaying}
\title{Compute eggs laid}
\usage{
EggLaying(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Compute eggs laid at time \eqn{t}
}
\seealso{
\code{\link[=setup_EGG_LAYING]{setup_EGG_LAYING()}}
}
