% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.dts}
\alias{Exposure.dts}
\title{Compute Attack Rates}
\usage{
\method{Exposure}{dts}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
For \code{dts} models, compute
the attack rates, the expected probability of infection
over one time step, is a weighted sum of local and travel exposure.
}
\details{
Local attack rates (AR or \eqn{\alpha}) are
a function of local daily entomological inoculation
rates (dEIR or \eqn{E}), and
a model for the probability of infection
per infectious bite, \eqn{b} (from \eqn{F_infectivity},
as defined by an \eqn{\cal X} model).
The total attack rates also consider
exposure to malaria while traveling, (\eqn{T_\delta}):
\deqn{\alpha = 1-((1-\delta) \; F_\alpha(E, b))(1-\delta\; T_\alpha(b))}
}
\seealso{
Related: \link{Exposure} & \link{F_ar.pois} & \link{F_ar.nb} & \link{Travel}
}
