% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.dts}
\alias{Exposure.dts}
\title{Compute the Attack Rates}
\usage{
\method{Exposure}{dts}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
the modified \strong{\code{xds}} object
}
\description{
For dts models, compute
the attack rates (AR) as a function
of local daily entomological inoculation
rates (dEIR), and a model for exposure
to malaria while traveling.
}
\details{
Local attack rates are computed
under a model for immunity that
computes the probability of infection
per infectious bite, \eqn{b}, returned
from \eqn{F_b}, as defiend by the
\eqn{\cal X} model.
}
