% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{Exposure and Infection}
\usage{
Exposure(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object as a \link{list}}
}
\value{
the function modifies \strong{pars} and returns it: the computed FoI are stored as \code{pars$FoI}
}
\description{
Function that model the FoI or AR as a function of
the local daily entomological inoculation rate (dEIR).
exposure while traveling.
}
