% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{Compute local exposure and travel}
\usage{
Exposure(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
the modified \strong{\code{xds}} object
}
\description{
Compute the force of infection (FoI) or
attack rates (AR) as a function of the local
daily entomological inoculation rate (dEIR),
and exposure to malaria while traveling.
}
