% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{Compute local exposure and travel}
\usage{
Exposure(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object; FoI are stored as \code{pars$FoI}
}
\description{
Function that model the FoI or AR as a function of
the local daily entomological inoculation rate (dEIR),
and a model for exposure to malaria while traveling.
}
