% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure}
\alias{Exposure}
\title{Compute Infection Rates}
\usage{
Exposure(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the force of infection (FoI) or
attack rates (AR) as a function of the local
daily entomological inoculation rate (dEIR),
immunity, and exposure to malaria while traveling.
}
\seealso{
Cases: \link{Exposure.xde} & \link{Exposure.dts}. Related: \link{F_ar} & \link{F_foi}
}
