% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.xde}
\alias{Exposure.xde}
\title{Compute the FoI}
\usage{
\method{Exposure}{xde}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
the modified \strong{\code{xds}} object
}
\description{
For xde models, compute
the force of infection (FoI) for
as a function of the daily entomological
inoculation rates (dEIR), and
exposure to malaria while traveling.
}
