% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.xde}
\alias{Exposure.xde}
\title{Compute the force of infection (FoI) given the local EIR and a travel model}
\usage{
\method{Exposure}{xde}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object; FoIs are stored as \code{pars$FoI}
}
\description{
For xde models, compute the force of infection (FoI) for
all the population strata. The FoI is a function of the daily entomological
inoculation rates (dEIR), and a model for exposure to malaria while traveling.
}
