% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{Exposure.xde}
\alias{Exposure.xde}
\title{Exposure and Infection}
\usage{
\method{Exposure}{xde}(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{the model object as a \link{list}}
}
\value{
the function modifies \strong{pars} and returns it: the computed FoI are stored as \code{pars$FoI}
}
\description{
For xde models, compute the daily force of infection (dFoI) for
a population. The dFoI is a function of the local daily entomological
inoculation rate (dEIR), and a function \link{travel_malaria} that accounts for
exposure while traveling.
}
