% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{F_ar.nb}
\alias{F_ar.nb}
\title{A negative binomial model for the attack rate as a function of the daily EIR.}
\usage{
\method{F_ar}{nb}(eir, b, pars)
}
\arguments{
\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ar} for a negative binomial model
}
\details{
Compute the daily AR, \eqn{h},
given the daily EIR, \eqn{E},
under a negative binomial model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{\alpha = 1-(1 + b E/\phi)^{-\phi} }
The model is parameterized using the \strong{mu} \eqn{= E} and \strong{sz} \eqn{=\phi}
parameterization. The expression is the complement of the zero term.
}
