% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{F_ar.nb}
\alias{F_ar.nb}
\title{Negative Binomial Attack Rates}
\usage{
\method{F_ar}{nb}(eir, b, env_het_obj)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{env_het_obj}{an environmental heterogeneity model object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A negative binomial model for the attack rate as a function of the daily EIR.
}
\details{
This function computes the daily AR, \eqn{\alpha}, as a function
of the daily EIR, \eqn{E} or \code{eir},
under a negative binomial model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{\alpha = 1-(1 + b E/\phi)^{-\phi} }
This negative binomial model uses the \strong{mu} \eqn{= bE} and \strong{size} \eqn{=\phi}
parameterization (see \link{pnbinom}).
The formula is equivalent to:

\code{pnbinom(0, mu=b*eir, size=phi, lower.tail=FALSE)}
}
\seealso{
Related topics: \link{Exposure.dts}, \link{make_exposure_nb} & \link{F_foi.nb}
}
