% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{F_ar.pois}
\alias{F_ar.pois}
\title{A Poisson model for Exposure}
\usage{
\method{F_ar}{pois}(eir, b, pars)
}
\arguments{
\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
dEIR as a \link{numeric} vector of length \eqn{n_h=}\code{nStrata}
}
\description{
Compute the daily AR, \eqn{\alpha},
given the daily EIR, \eqn{E},
under a Poisson modelfor the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{\alpha = 1 - e^{- b E}}
}
