% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{F_ar.pois}
\alias{F_ar.pois}
\title{Poisson Attack Rates}
\usage{
\method{F_ar}{pois}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector: attack rates for the population strata
}
\description{
A Poisson model for attack rates as a function of the daily EIR.
}
\details{
This function computes the daily attack rates (\eqn{\alpha}), the fraction
of a cohort that gets infected in a day,
as a function of the daily EIR (\eqn{E})
under a Poisson model for the distribution
of bites per person, and the probability of
infection per infective bite (\eqn{b}):
\deqn{\alpha = 1 - e^{- b E}}
The expression is equivalent to

\code{ppois(0, b*eir, lower.tail=FALSE)}
}
\seealso{
Related: \link{F_foi} & \link{F_foi.pois} & \link{ar2eir.pois}. Called from \link{Exposure.dts}
}
