% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{F_available_habitat}
\alias{F_available_habitat}
\title{Compute Habitat Availability}
\usage{
F_available_habitat(habitat_matrix, search_weights)
}
\arguments{
\item{habitat_matrix}{the membership matrix, \eqn{N}}

\item{search_weights}{the habitat search weights, \eqn{\omega_q}}
}
\value{
a \link{vector} of describing habitat availability, \eqn{Q}, of length \code{nPatches}
}
\description{
Compute the availability of aquatic habitats.
}
\details{
The availability of the habitats that we have defined in the model, denoted \eqn{Q}, sums search weights, \eqn{\omega_q}, by patch
using the habitat membership matrix, \eqn{N}, and we can compute
\deqn{Q = {N} \cdot \omega_q.}
}
\seealso{
This function is called by \link{compute_Qall}

\link{make_habitat_matrix} discusses \eqn{N}

The availability of ovitraps and bad habitats is setup in \link{setup_egg_laying}
}
\keyword{internal}
