% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{F_eggs.SI}
\alias{F_eggs.SI}
\title{Compute Component Egg Laying Rates for \code{SI}}
\usage{
\method{F_eggs}{SI}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
The density of adult mosquitoes is \eqn{M}.
The \strong{parameters} describing egg laying by adult mosquitoes are:
\itemize{
\item \eqn{\nu} or \code{nu} is the egg laying rate
\item \eqn{\xi} or \code{eggsPerBatch} is the number of eggs per batch
}

The egg laying rate, per patch, is \deqn{\nu \xi M}
}
