% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{F_eip}
\alias{F_eip}
\title{Compute the EIP}
\usage{
F_eip(t, vars, eip_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{eip_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method ...
}
