% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-static.R
\name{F_eip.static}
\alias{F_eip.static}
\title{Modify parameters due to exogenous forcing by all kinds of control}
\usage{
\method{F_eip}{static}(t, vars, eip_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{eip_par}{a \link{list}}
}
\value{
\link{numeric}
}
\description{
Implements \link{F_eip} for the static_xde model (the EIP is constant)
}
