% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{F_eip.static}
\alias{F_eip.static}
\title{Static model patch emigration}
\usage{
\method{F_eip}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{eip}, the patch emigration rate
}
\description{
Implements \link{F_eip} for a static model
}
