% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{F_emerge.basicL}
\alias{F_emerge.basicL}
\title{Compute emergent adults}
\usage{
\method{F_emerge}{basicL}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
The function computes the number of
emergent adults from aquatic habitats for \code{basicL}
}
