% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{F_emerge.basicL}
\alias{F_emerge.basicL}
\title{Compute Emergent Adults for \code{basicL} (\strong{L} Component)}
\usage{
\method{F_emerge}{basicL}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
The number of adults emerging from the habitats,
per day, is:
\deqn{\psi e^{-\xi L} L.}
}
\seealso{
\link{dLdt.basicL}
}
