% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{F_emerge}
\alias{F_emerge}
\title{Number of newly emerging adults from each larval habitat}
\usage{
F_emerge(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
