% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{F_emerge}
\alias{F_emerge}
\title{Compute Emergent Adults}
\usage{
F_emerge(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
This function computes the rate or number of emerging adults: a
rate for differential equations, or a number for difference equations.
}
\note{
This method dispatches on the class of \code{pars$Lpar[[s]]}
}
