% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{F_emerge.trivial}
\alias{F_emerge.trivial}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_emerge}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Implements \link{F_emerge} for the trivial (forced emergence) model.
}
