% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{F_emerge.trivial}
\alias{F_emerge.trivial}
\title{Number of newly emerging adults from each larval habitat}
\usage{
\method{F_emerge}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
The number of emerging adults is a function \deqn{\Lambda S(t) T(t)} where
\itemize{
\item \eqn{\Lambda} or \code{Lambda} is the mean number of adult female mosquitoes emerging per day
\item \eqn{S(t)} or \code{F_season} is a seasonal signal (ideally, with an average annual mean of 1)
\item \eqn{T(t)} or \code{F_trend} is a function returning a trend (ideally, with an average value of 1)
}
}
