% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{F_f.static}
\alias{F_f.static}
\title{Constant baseline blood feeding rate}
\usage{
\method{F_f}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{f}, the baseline blood feeding rate
}
\description{
Implements \link{F_f} for a static model
}
\note{
This method dispatches on the type of \code{f_obj} attached to the \code{MY_obj}.
}
\keyword{internal}
