% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{F_foi.nb}
\alias{F_foi.nb}
\title{Negative Binomial Force of Infection}
\usage{
\method{F_foi}{nb}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Compute the daily FoI from the daily EIR
under a negative binomial model for exposure
}
\details{
This function computes the local daily FoI, \eqn{h} as a function of
the local daily EIR, \eqn{E},
under a negative binomial model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{h = \phi \log{\frac{1 + b E}\phi}}
The expression is derived by solving the expression in \link{F_ar.nb} for \eqn{E}
This negative binomial model uses the \strong{mu} \eqn{= bE} and \strong{size} \eqn{=\phi}
parameterization (see \link{pnbinom}).
}
\seealso{
Related topics: \link{Exposure.xde} and \link{F_ar.nb} and \link{setup_exposure_nb}
}
