% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{F_foi.pois}
\alias{F_foi.pois}
\title{Poisson Force of Infection}
\usage{
\method{F_foi}{pois}(eir, b, pars)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A Poisson model for the force of infection as a function of the daily EIR.
}
\details{
Compute the daily FoI, \eqn{h},
as a given the daily EIR, \eqn{E},
under a Poisson model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{h = b E}
}
