% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure.R
\name{F_foi}
\alias{F_foi}
\title{Compute the Local FoI}
\usage{
F_foi(eir, b, env_het_obj)
}
\arguments{
\item{eir}{the daily eir for each stratum}

\item{b}{the probability of infection, per bite}

\item{env_het_obj}{an environmental heterogeneity model object}
}
\value{
the local FoI as a \link{numeric} vector of length \eqn{n_h =} \code{nStrata}
}
\description{
Compute the daily local FoI as a function
of the daily EIR and effects of partial immunity.
}
\details{
This function computes the daily
force of infection (FoI)
as a function of the daily EIR under
a probabilistic model for environmentally
heterogeneous exposure --
the distribution of bites per person.

The model for human / host infection (\eqn{\cal X}) provide
a term \eqn{b}, from \link{F_infectivity} that describe the probability of infection
per infectious bite, possibly affected by
pre-erythrocytic immunity.
}
\seealso{
Cases: \link{F_foi.pois} & \link{F_foi.nb}. Related: \link{Exposure.xde} & \link{foi2eir}
}
