% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{F_fqZ.trivial}
\alias{F_fqZ.trivial}
\title{Net Infectious Biting Rate}
\usage{
\method{F_fqZ}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Returns
\deqn{F_{fqZ}(t) = fqZ S(t) T(t) K(t)}
where
\itemize{
\item \eqn{f} is the feeding rate
\item \eqn{q} the human fraction
\item \eqn{Z} is the density of infectious mosquitoes, per patch
\item \eqn{S(t)} or \code{F_season} is a seasonal pattern
\item \eqn{T(t)} or \code{F_trend} is a trend pattern
\item \eqn{K(t)} or \code{F_shock} is a perturbation
}
}
\keyword{internal}
