% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-mosquito_mortality.R
\name{F_g}
\alias{F_g}
\title{Compute the blood geeding rate, g}
\usage{
F_g(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector og length \code{nPatches}
}
\description{
This method dispatches on the type of \code{g_obj}. It should
set the values og the bionomic parameters to baseline values
}
\keyword{internal}
