% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{F_K_matrix.static}
\alias{F_K_matrix.static}
\title{Constant baseline blood feeding rate}
\usage{
\method{F_K_matrix}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Implements \link{F_K_matrix} for a static model
}
\note{
This method dispatches on the type of \code{f_obj} attached to the \code{MY_obj}.
}
