% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-RM-functions.R
\name{F_mu}
\alias{F_mu}
\title{Compute the emigration loss}
\usage{
F_mu(t, vars, mu_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{mu_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Compute the fraction lost to emigration
}
