% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-GeRM.R
\name{F_mu.static}
\alias{F_mu.static}
\title{Static model for mosquito survival}
\usage{
\method{F_mu}{static}(t, vars, mu_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{mu_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_mu} for a static model
}
