% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_ni}
\alias{F_ni}
\title{Compute Net Infectiousness (NI)}
\usage{
F_ni(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A function to compute NI as an output
}
