% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{F_ni.SIS}
\alias{F_ni.SIS}
\title{Compute the net infectiousness}
\usage{
\method{F_ni}{SIS}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_ni} for the SIS model.
}
\keyword{internal}
