% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_nu}
\alias{F_nu}
\title{Compute the egg laying rate}
\usage{
F_nu(t, vars, nu_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{nu_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{q_par}. It should
set the values of nu to (possibly changing) baseline value(s).
}
