% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-GeRM.R
\name{F_nu.static}
\alias{F_nu.static}
\title{Static model for the egg laying rate}
\usage{
\method{F_nu}{static}(t, vars, nu_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{nu_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_nu} for a static model
}
