% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{F_Omega_xde}
\alias{F_Omega_xde}
\title{Update Upsilon for xde}
\usage{
F_Omega_xde(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
an \strong{xds} model object
}
\description{
Compute \eqn{\Upsilon}
}
\keyword{internal}
