% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{F_pfpr_by_pcr}
\alias{F_pfpr_by_pcr}
\title{Compute infection prevalence by PCR}
\usage{
F_pfpr_by_pcr(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A function that translates the state variables into
the predicted \emph{Pf}PR by PCR
}
\note{
This method dispatches on the type of \code{xds_obj$XH_obj[[i]]}.
}
