% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_pfpr_by_pcr.SIS}
\alias{F_pfpr_by_pcr.SIS}
\title{Compute \emph{Pf}PR by PCR}
\usage{
\method{F_pfpr_by_pcr}{SIS}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the SIS model.
}
