% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{F_pfpr_by_rdt}
\alias{F_pfpr_by_rdt}
\title{Compute the prevalence of infection by RDT}
\usage{
F_pfpr_by_rdt(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
A function that translates the state variables into
the predicted \emph{Pf}PR by rapid diagnostic test (RDT)
}
