% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SEISd.R
\name{F_pr_by_lm.SEISd}
\alias{F_pr_by_lm.SEISd}
\title{Compute the prevalence of infection by light microscopy}
\usage{
\method{F_pr_by_lm}{SEISd}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pr} for the SEISd model.
}
