% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{F_pr_by_rdt}
\alias{F_pr_by_rdt}
\title{Compute the prevalence of infection by RDT}
\usage{
F_pr_by_rdt(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
This method dispatches on the type of \code{pars$Xpar[[i]]}.
}
