% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-trace.R
\name{F_pr_by_rdt.trace}
\alias{F_pr_by_rdt.trace}
\title{Compute the prevalence of infection by RDT}
\usage{
\method{F_pr_by_rdt}{trace}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector numeric(0)
}
\description{
Implements \link{F_pr} for the trace model.
}
