% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{F_prevalence.hMoI}
\alias{F_prevalence.hMoI}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_prevalence}{hMoI}(vars, XH_obj)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{XH_obj}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_prevalence} for the hMoI model.
}
