% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-SIS.R
\name{F_prevalence.SIS}
\alias{F_prevalence.SIS}
\title{Compute the "true" prevalence of infection / parasite rate}
\usage{
\method{F_prevalence}{SIS}(vars, Xpar)
}
\arguments{
\item{vars}{a list with the variables attached by name}

\item{Xpar}{a list defining a model for human}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{F_pfpr} for the SIS model.
}
