% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{F_psi}
\alias{F_psi}
\title{Setup for}
\usage{
F_psi(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector of length \code{nHabitats}
}
\description{
Set up an object
to compute
}
\keyword{internal}
