% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-human_fraction.R
\name{F_q}
\alias{F_q}
\title{Compute the blood qeeding rate, q}
\usage{
F_q(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
a \link{numeric} vector oq length \code{nPatches}
}
\description{
This method dispatches on the type of \code{q_obj}. It should
set the values oq the bionomic parameters to baseline values
}
\keyword{internal}
