% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-RM-functions.R
\name{F_q}
\alias{F_q}
\title{Compute the human blood fraction}
\usage{
F_q(t, vars, q_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{q_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
This method dispatches on the type of \code{q_par$q_par}. It should
set the values of the bionomic parameters to baseline values.
}
