% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics-functions-static.R
\name{F_q.static}
\alias{F_q.static}
\title{Static model for human blood fraction}
\usage{
\method{F_q}{static}(t, vars, q_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{q_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_q} for a static model
}
