% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{F_season.eir}
\alias{F_season.eir}
\title{Plot the seasonal pattern}
\usage{
\method{F_season}{eir}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the seasonal pattern, invisibly
}
\description{
For a model forced by the EIR
plot the seasonal pattern
}
