% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{F_season.Lambda}
\alias{F_season.Lambda}
\title{Plot the seasonal pattern}
\usage{
\method{F_season}{Lambda}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the seasonal pattern, invisibly
}
\description{
For a model forced by emergence,
show the seasonal pattern
}
