% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_season.R
\name{F_season.none}
\alias{F_season.none}
\title{Plot the seasonal pattern}
\usage{
\method{F_season}{none}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a null result
}
\description{
The \code{F_season} case for models
with no forcing.
}
