% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_shock.eir}
\alias{F_shock.eir}
\title{Compute the temporal shock}
\usage{
\method{F_shock}{eir}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal shock, invisibly
}
\description{
Evaluate the function \code{F_shock} for a forced model when \code{forced_by} is "eir"
}
\keyword{internal}
