% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_shock.eir}
\alias{F_shock.eir}
\title{Plot the temporal trend}
\usage{
\method{F_shock}{eir}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend, invisibly
}
\description{
For a model forced by the EIR
plot the temporal trend
}
