% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_shock}
\alias{F_shock}
\title{Plot the Temporal Trend}
\usage{
F_shock(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend, invisibly
}
\description{
For a model with temporal forcing,
show the temporal trend
}
