% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-emigrate.R
\name{F_sigma.static}
\alias{F_sigma.static}
\title{Static model patch emigration}
\usage{
\method{F_sigma}{static}(t, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{vector species index}
}
\value{
\eqn{sigma}, the patch emigration rate
}
\description{
Implements \link{F_sigma} for a static model
}
\keyword{internal}
