% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{F_sigma.static}
\alias{F_sigma.static}
\title{Static model for mosquito emigration}
\usage{
\method{F_sigma}{static}(t, vars, sigma_par)
}
\arguments{
\item{t}{current simulation time}

\item{vars}{exogenous variables}

\item{sigma_par}{a \link{list}}
}
\value{
a \link{numeric} vector of length \code{nPatches}
}
\description{
Implements \link{F_sigma} for a static model
}
