% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend.Lambda}
\alias{F_trend.Lambda}
\title{Plot the temporal trend}
\usage{
\method{F_trend}{Lambda}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the temporal trend
}
\description{
For a model forced by emergence,
show the temporal trend
}
