% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{F_trend.none}
\alias{F_trend.none}
\title{Plot the temporal trend}
\usage{
\method{F_trend}{none}(tm, xds_obj)
}
\arguments{
\item{tm}{the time points}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a null result
}
\description{
The \code{F_trend} case for models
with no forcing.
}
