% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{F_Upsilon_xde}
\alias{F_Upsilon_xde}
\title{F_Upsilon for dts}
\usage{
F_Upsilon_xde(eip, Omega)
}
\arguments{
\item{eip}{the extrinsic incubation period}

\item{Omega}{the demographic matrix}
}
\value{
a \link{matrix} of dimensions \code{nPatches} by \code{nPatches}
}
\description{
F_Upsilon for dts
}
\keyword{internal}
