% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{foi2eir.nb}
\alias{foi2eir.nb}
\title{A negative binomial model for the daily FoI as a function of the daily EIR.}
\usage{
\method{foi2eir}{nb}(foi, b, pars)
}
\arguments{
\item{foi}{the daily FoI for each stratum}

\item{b}{the probability of infection, per bite}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a \link{numeric} vector of length \code{nStrata}
}
\description{
Implements \link{foi2eir} for a negative binomial model
}
\details{
Compute the daily AR, \eqn{h},
given the daily EIR, \eqn{E},
under a negative binomial model for the distribution
of bites per person, and \eqn{b}, the probability of
infection per infective bite:
\deqn{\alpha = \left(e^{h/\phi} -1\right) \frac \phi b }
The model is parameterized using the \strong{mu} \eqn{= E} and \strong{sz} \eqn{= 1/\phi}
parameterization. The expression is the complement of the zero term.
}
