% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{foi2eir.pois}
\alias{foi2eir.pois}
\title{Convert FoI to EIR under a Poisson model for Exposure}
\usage{
\method{foi2eir}{pois}(foi, b, env_het_obj)
}
\arguments{
\item{foi}{the daily FoI for each stratum}

\item{b}{the probability of infection, per bite}

\item{env_het_obj}{an environmental heterogeneity model object}
}
\value{
dEIR as a \link{numeric} vector of length \eqn{n_h=}\code{nStrata}
}
\description{
The inverse of \link{F_foi.pois} is \eqn{E=h/b}
}
\seealso{
Also, see \link{F_foi.pois}
}
\keyword{internal}
