% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.dynamic}
\alias{Forcing.dynamic}
\title{Set the values of exogenous variables}
\usage{
\method{Forcing}{dynamic}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
With dynamic forcing, exogenous variables
can be set in one of four function calls:
\itemize{
\item Weather
\item Hydrology
\item Shock
\item Development
}
}
\seealso{
\link{dynamic_forcing}
}
