% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing.none}
\alias{Forcing.none}
\title{Set the values of exogenous variables}
\usage{
\method{Forcing}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After none setup, no exogenous
variables are configured so forcing returns
the \strong{\code{xds}} object without modification
}
