% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing}
\alias{Forcing}
\title{Set the values of exogenous variables}
\usage{
Forcing(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of
exogenous variables.
}
