% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{Forcing}
\alias{Forcing}
\title{Exogenous Forcing}
\usage{
Forcing(t, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a junction for exogenous forcing
by weather and hydrology. Non-trivial
examples are in the satellite package
\code{ramp.forcing}
}
