% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmittion-get.R
\name{get_beta}
\alias{get_beta}
\title{Get xi, delayed maturation}
\usage{
get_beta(xds_obj, s = 1, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
the mixing matrix, beta
}
\description{
Get xi, delayed maturation
}
