% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-terms.R
\name{get_EIR}
\alias{get_EIR}
\title{Get the}
\usage{
get_EIR(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\description{
Get the
}
