% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{get_habitat_matrix}
\alias{get_habitat_matrix}
\title{Get habitat matrix, \eqn{N}}
\usage{
get_habitat_matrix(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{matrix}
}
\description{
Output the habitat membership information as a list
}
\seealso{
\code{\link[=make_habitat_matrix]{make_habitat_matrix()}}
}
