% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{get_habitats}
\alias{get_habitats}
\title{Get the habitat membership vector}
\usage{
get_habitats(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Output the habitat membership information as a list
}
\seealso{
\code{\link[=make_habitat_matrix]{make_habitat_matrix()}}
}
