% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inits}
\alias{get_inits}
\title{Get the stored initial values, \eqn{y_0}}
\usage{
get_inits(xds_obj, flatten = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{flatten}{a \link{logical}: if true, results are returned as an unnamed vector}
}
\value{
a named \link{list} or \verb{if(flatten==TRUE)} a \link{vector}
}
\description{
Get the stored initial values, \eqn{y_0}
}
