% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{get_K_matrix}
\alias{get_K_matrix}
\title{Get the Mosquito Dispersal Matrix}
\usage{
get_K_matrix(xds_obj, behave = "0", s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{behave}{the behavioral state}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get the Mosquito Dispersal Matrix
}
\note{
The \code{behave} argument is used for
behavioral state models in \code{ramp.library}
}
