% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{get_K_matrix}
\alias{get_K_matrix}
\title{Get the Mosquito Dispersal Matrix}
\usage{
get_K_matrix(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Get the Mosquito Dispersal Matrix
}
