% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_inits}
\alias{get_L_inits}
\title{Get Initial Values for the \strong{L} Component}
\usage{
get_L_inits(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Get Initial Values for the \strong{L} Component
}
\note{
This method dispatches on the class of \code{xds_obj$L_obj}.
}
