% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_ix}
\alias{get_L_ix}
\title{Get L indices}
\usage{
get_L_ix(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Get and display the
values of the indices
for the \strong{L} variables
}
\keyword{internal}
