% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{get_L_pars}
\alias{get_L_pars}
\title{Get parameters for the \strong{L} Component module}
\usage{
get_L_pars(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get parameters for the \strong{L} Component module
}
