% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{get_Linits}
\alias{get_Linits}
\title{Get Initial Values for the \strong{L} Component}
\usage{
get_Linits(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Get Initial Values for the \strong{L} Component
}
\note{
This method dispatches on the class of \code{pars$Lpar}.
}
