% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{get_Linits}
\alias{get_Linits}
\title{Return initial values as a vector}
\usage{
get_Linits(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
