% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{get_Lpars.basicL}
\alias{get_Lpars.basicL}
\title{Return the parameters as a list}
\usage{
\method{get_Lpars}{basicL}(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}.
}
