% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{get_Lpars.basicL}
\alias{get_Lpars.basicL}
\title{Get \strong{L} component parameters}
\usage{
\method{get_Lpars}{basicL}(pars, s = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
Get the \strong{L} component parameters
}
\seealso{
\link{dLdt.basicL}
}
