% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{get_MY_inits.trivial}
\alias{get_MY_inits.trivial}
\title{Return initial values as a vector}
\usage{
get_MY_inits.trivial(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{get_MY_inits} for the GeRM model.
}
