% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_MY_ix}
\alias{get_MY_ix}
\title{Get \strong{MY} Variable Indices}
\usage{
get_MY_ix(xds_obj, s)
}
\arguments{
\item{xds_obj}{a \link{list}}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$MY_obj}.
}
