% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.lm}
\alias{get_PR.lm}
\title{Get the \emph{Pf}PR from a Malaria Model}
\usage{
\method{get_PR}{lm}(pars, method = "lm", i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{method}{the method used for computing \emph{Pf}PR}

\item{i}{the host species index}
}
\value{
none
}
\description{
Return the PR by light microscopy
}
