% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.lm}
\alias{get_PR.lm}
\title{Get the \emph{Pf}PR from a Malaria Model}
\usage{
\method{get_PR}{lm}(xds_obj, i = 1, method = "lm")
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{method}{the method used for computing \emph{Pf}PR}
}
\value{
none
}
\description{
Return the PR by light microscopy
}
