% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-get.R
\name{get_PR.pcr}
\alias{get_PR.pcr}
\title{Get the \emph{Pf}PR from a MalariasModel}
\usage{
\method{get_PR}{pcr}(xds_obj, i = 1, method = "pcr")
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{method}{the method used for computing \emph{Pf}PR}
}
\value{
none
}
\description{
Return the \emph{Pf}PR by PCR
}
