% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_PR.pcr}
\alias{get_PR.pcr}
\title{Get the \emph{Pf}PR from a Malaria Model}
\usage{
\method{get_PR}{pcr}(pars, method = "pcr", i = 1)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{method}{the method used for computing \emph{Pf}PR}

\item{i}{the host species index}
}
\value{
none
}
\description{
Return the \emph{Pf}PR by PCR
}
