% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season.MY}
\alias{get_season.MY}
\title{Get seasonal pattern}
\usage{
\method{get_season}{MY}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
seasonal pattern parameters
}
\description{
Get the parameters that set the seasonal pattern
for an \strong{MY} trivial model
}
\keyword{internal}
