% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season.none}
\alias{get_season.none}
\title{Get seasonal pattern}
\usage{
\method{get_season}{none}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
seasonal pattern parameters
}
\description{
Get the parameters set the seasonal pattern
}
