% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season}
\alias{get_season}
\title{Get seasonal pattern}
\usage{
get_season(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
seasonal pattern parameters
}
\description{
Get the parameters set the seasonal pattern
}
