% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_bottom.R
\name{get_season_bottom}
\alias{get_season_bottom}
\title{Get the seasonal bottom}
\usage{
get_season_bottom(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
the bottom parameter(s)
}
\description{
Get the parameter that sets the bottom
for a seasonally forced model
}
