% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season.R
\name{get_season_bottom}
\alias{get_season_bottom}
\title{Get phase}
\usage{
get_season_bottom(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the phase parameter(s)
}
\description{
Get the parameter that sets the phase
for a seasonally forced model
}
