% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_season_pw.R
\name{get_season_pw}
\alias{get_season_pw}
\title{Get the seasonal pw}
\usage{
get_season_pw(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
the pw parameter(s)
}
\description{
Get the pw parameter
for a seasonally forced model
}
