% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{get_sigma}
\alias{get_sigma}
\title{Get the patch emigration rates}
\usage{
get_sigma(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector assigned the class "dynamic"
}
\description{
Get the patch emigration rates
}
