% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-SI.R
\name{get_sigma.SI}
\alias{get_sigma.SI}
\title{Get the feeding rate}
\usage{
\method{get_sigma}{SI}(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
y a \link{numeric} vector asSIgned the class "dynamic"
}
\description{
Get the feeding rate
}
