% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_spline.R
\name{get_spline_s.none}
\alias{get_spline_s.none}
\title{Get spline interpolation points}
\usage{
\method{get_spline_s}{none}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\description{
Return the value that sets mean forcing
for forced models.
}
