% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{get_TimeSpent_matrix}
\alias{get_TimeSpent_matrix}
\title{Get the Time Spent Matrix}
\usage{
get_TimeSpent_matrix(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a modified \strong{\code{xds}} model object
}
\description{
Get the Time Spent Matrix
}
