% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-demography.R
\name{get_Upsilon}
\alias{get_Upsilon}
\title{Make the mosquito demography matrix}
\usage{
get_Upsilon(xds_obj, s = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
This method dispatches on the type of \code{MY_obj}.
}
