% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mosquito.R
\name{get_Upsilon}
\alias{get_Upsilon}
\title{Make the mosquito demography matrix}
\usage{
get_Upsilon(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
