% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_variables.list}
\alias{get_variables.list}
\title{Get Variables}
\usage{
\method{get_variables}{list}(opts, t, y, xds_obj, ix)
}
\arguments{
\item{opts}{an options list and dispatching}

\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{ix}{the species index}
}
\value{
an empty list
}
\description{
The NULL case for \link{get_variables}
}
\keyword{internal}
