% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_variables}
\alias{get_variables}
\title{Get Variables}
\usage{
get_variables(t, y, func, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{variables}

\item{func}{a function that dispatches \link{get_variables} and pulls variables it needs}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a vector of variables
}
\description{
A utility to write functions
that respond to state variables or exogenous
variables
}
\note{
The method dispatches on \code{class(func)}
}
