% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{get_XH_pars.trivial}
\alias{get_XH_pars.trivial}
\title{Return the parameters as a list}
\usage{
\method{get_XH_pars}{trivial}(xds_obj, i = 1)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{xds_obj$XH_obj[[s]]}.
}
