% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health.none}
\alias{Health.none}
\title{Set no exogenous health variables}
\usage{
\method{Health}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
After none setup, no exogenous
variables are configured so \code{Health} returns
the \strong{\code{ramp.xds}} model object without modification
}
