% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/health.R
\name{Health}
\alias{Health}
\title{Set the values variables for health-based malaria control}
\usage{
Health(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the values of
variables for health interventions.
}
