% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-concepts.R
\name{human_populations}
\alias{human_populations}
\title{The Human Population}
\description{
A human (or host) population is sub-divided into \emph{strata.}
The human population (or other host population) and its structure, size and behavior
are critical aspects of transmission.

\describe{
\item{\code{residence}}{a vector: the patch index where each human stratum lives}
\item{\code{nStrata}}{\code{nStrata = length(residence)}}
\item{\code{HPop}}{initial human population size: \code{length(HPop) = nStrata}}
}
}
\section{Human Population Size, \eqn{H}}{

write me.
}

\section{The Residency Matrix}{

write me.
}

