% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importation.R
\name{Importation}
\alias{Importation}
\title{Importation Junction}
\usage{
Importation(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Dispatches on \code{class(xds_obj$importation_obj)} to update
importation-related parameters: time at home, travel EIR,
visitor availability, and visitor infectiousness.
}
\keyword{internal}
