% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-help.R
\name{L_change}
\alias{L_change}
\title{Change functions (\strong{L})}
\description{
A list of change functions for aquatic mosquito modules.

\itemize{
\item \code{\link{change_L_inits}} -- change initial values
\item \code{\link{change_L_pars}} -- change parameters
}
}
