% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{LBaseline.basicL}
\alias{LBaseline.basicL}
\title{Baseline Bionomics for \code{basicL} (\strong{L} Component)}
\usage{
\method{LBaseline}{basicL}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Set \strong{L} component parameters
to baseline values for \code{basicL}
}
\keyword{internal}
