% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{LBaseline}
\alias{LBaseline}
\title{Baseline Bionomics for an \strong{L} Component Module}
\usage{
LBaseline(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
Handle immature mosquito bionomic parameters as a baseline modified by control
}
