% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{LBaseline.trivial}
\alias{LBaseline.trivial}
\title{Reset aquatic parameters to baseline}
\usage{
\method{LBaseline}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
a named \link{list}
}
\description{
Implements \link{LBaseline} for the RM model
}
