% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{LBionomics.basicL}
\alias{LBionomics.basicL}
\title{Bionomics for \code{basicL} (\strong{L} Component)}
\usage{
\method{LBionomics}{basicL}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \strong{\code{ramp.xds}} object
}
\description{
Implements \link{LBionomics} for the \code{basicL}
}
\keyword{internal}
