% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{LBionomics}
\alias{LBionomics}
\title{Immature mosquito bionomics}
\usage{
LBionomics(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the values of bionomic parameter
values for an \strong{L} Component module
}
\keyword{internal}
