% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R, R/aquatic-interface.R
\name{LBionomics}
\alias{LBionomics}
\title{Larval and Aquatic Stage Bionomics}
\usage{
LBionomics(t, y, pars, s)

LBionomics(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the unmodified \code{xds} object

a \link{list}
}
\description{
Handle immature mosquito bionomic parameters as a baseline modified by control

This method dispatches on the type of \code{pars$Lpar}. It should
compute the values of parameters as a function of exogenous variables
or reset the values of the bionomic parameters to baseline values.
}
