% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R, R/bionomics.R
\name{LBionomics}
\alias{LBionomics}
\title{Set aquatic bionomic parameter rates relative to baseline}
\usage{
LBionomics(t, y, pars, s)

LBionomics(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object

an \code{xds} object
}
\description{
It should compute the values of parameters
as a function of exogenous variables
or reset the values of the bionomic parameters to baseline values.

Handle immature mosquito bionomic parameters as a baseline modified by control
}
\note{
This method dispatches on the class of \code{pars$Lpar[[s]]}
}
