% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{LBionomics}
\alias{LBionomics}
\title{Bionomics for an \strong{L} Component Module}
\usage{
LBionomics(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
Handle immature mosquito bionomic parameters as a baseline modified by control
}
