% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{LBionomics.trivial}
\alias{LBionomics.trivial}
\title{Reset aquatic parameters to baseline}
\usage{
\method{LBionomics}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{LBionomics} for the RM model
}
