% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{list_Lvars.basicL}
\alias{list_Lvars.basicL}
\title{Return the variables as a list}
\usage{
\method{list_Lvars}{basicL}(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}
}
