% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{list_Lvars}
\alias{list_Lvars}
\title{Return the variables as a list}
\usage{
list_Lvars(y, pars, s)
}
\arguments{
\item{y}{the variables}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}.
}
