% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsm-effectsizes-none.R
\name{LSMEffectSizes.none}
\alias{LSMEffectSizes.none}
\title{Set no lsm_effectsizes}
\usage{
\method{LSMEffectSizes}{none}(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
\link{list}
}
\description{
The null model for lsm_effectsizes
}
