% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-macdonald.R
\name{macdonald}
\alias{macdonald}
\title{The \code{macdonald} module (MY component)}
\description{
Implements a Macdonald-style delay differential
equation model of adult mosquito ecology and infection dynamics, consistent
with the model published by George Macdonald in 1952. This formulation is
actually closer to one published by Aron & May (1982).
}
\section{State Variables}{

\describe{
\item{\code{M}}{density of adult mosquitoes}
\item{\code{Y}}{density of infected adult mosquitoes}
\item{\code{Z}}{density of infectious adult mosquitoes}
}
}

\section{Parameters}{

\describe{
\item{\code{f}}{blood feeding rate}
\item{\code{q}}{human blood fraction}
\item{\code{g}}{mortality rate}
\item{\code{sigma}}{patch emigration}
\item{\code{mu}}{emigration-related loss}
\item{\code{K}}{the dispersal matrix}
\item{\code{tau}}{extrinsic incubation period (\eqn{\tau})}
}

The \emph{demographic matrix} \eqn{\Omega} is formulated as
\deqn{
 \Omega = \mbox{diag} \left( g + \sigma \mu \right) + K \cdot \mbox{diag} \left( \sigma \left(1-\mu\right) \right)
}
Survival and dispersal through the EIP is
\deqn{
 \Upsilon = e^{-\Omega \tau} 
}
}

\section{Dynamics}{

The dynamical system is described
by a coupled set of delay differential equations.
In these equations, a sub-scripted variable or term denotes its
lagged value: \eqn{M_\tau = M(t-\tau)}.

\deqn{
\begin{array}{rl}
dM/dt &= \Lambda - \Omega \cdot M \\\\
dY/dt &= fq\kappa(M-Y) - \Omega \cdot Y \\\\
dZ/dt &= \Upsilon \cdot (fq\kappa)_\tau(M_\tau-Y_\tau) - \Omega \cdot Z \\\\
\end{array}
}
}

\section{Note}{

This model is not capable of handling exogenous forcing by weather or
vector control. Use the \code{GeM} module instead.
}

\references{
\itemize{
\item Macdonald G (1952) The analysis of the sporozoite rate. Tropical Diseases Bulletin 49:569-586.
\item Aron JL, May RM (1982) The population dynamics of malaria. Chapter 5 in \emph{The Population Dynamics of Infectious Diseases: Theory and Applications,} Springer, Boston, MA.
}
}
