% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{make_eggs_laid}
\alias{make_eggs_laid}
\title{Compute eggs laid}
\usage{
make_eggs_laid(t, y, pars)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{pars}{an \code{xds} object}
}
\value{
a \link{list}  -- the modified \code{xds} object
}
\description{
Computes eggs laid in aquatic habitats
}
\seealso{
\link{compute_eggs_laid}
}
