% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_indices}
\alias{make_indices}
\title{Make indices for all the model variables}
\usage{
make_indices(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Make indices for all the model variables
}
