% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{make_indices_L.basicL}
\alias{make_indices_L.basicL}
\title{Add indices for aquatic mosquitoes to parameter list}
\usage{
\method{make_indices_L}{basicL}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{make_indices_L} for the basic M model.
}
