% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_indices_L}
\alias{make_indices_L}
\title{Add indices for aquatic stage mosquitoes to parameter list}
\usage{
make_indices_L(pars, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the class of \code{pars$Lpar}. Adds field \code{L_ix}
to parameter list.
}
