% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{make_indices_MYZ.SI}
\alias{make_indices_MYZ.SI}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{make_indices_MYZ}{SI}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Implements \link{make_indices_MYZ} for the \code{SI} model.
}
