% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{make_K_matrix_xy}
\alias{make_K_matrix_xy}
\title{make a Kernel-Based Mosquito Dispersal Matrix}
\usage{
make_K_matrix_xy(
  xy,
  ker = function(x) {
     exp(-x)
 }
)
}
\arguments{
\item{xy}{is a vector of the xy-coordinates of patch locations}

\item{ker}{is a function that weights putative locations by distance}
}
\description{
make a Kernel-Based Mosquito Dispersal Matrix
}
