% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{make_L_obj_trivial}
\alias{make_L_obj_trivial}
\title{Make \code{L_obj} for \code{trivial} (\strong{L} component)}
\usage{
make_L_obj_trivial(
  nHabitats,
  options = list(),
  Lambda = 1000,
  F_season = F_one,
  F_trend = F_one,
  F_shock = F_one,
  season_par = list(name = "F_one"),
  trend_par = list(name = "F_one"),
  shock_par = list(name = "F_one")
)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{options}{a \link{list} that overwrites default values}

\item{Lambda}{vector of mean emergence rates from each aquatic habitat}

\item{F_season}{the seasonal pattern function}

\item{F_trend}{the trend function}

\item{F_shock}{the shock function}

\item{season_par}{a list of options for F_season}

\item{trend_par}{a list of options for F_trend}

\item{shock_par}{a list of options for F_shock}
}
\value{
a \link{list}: an L module object
}
\description{
The number of emerging adults is a function \deqn{\Lambda S(t) T(t) K(t)} where
\itemize{
\item \eqn{\Lambda} or \code{Lambda} is the mean number of adult female mosquitoes emerging per day
\item \eqn{S(t)} or \code{F_season} is a seasonal pattern function (ideally, with an average annual mean of 1)
\item \eqn{T(t)} or \code{F_trend} is a trend pattern function (ideally, with an average value of 1)
\item \eqn{K(t)} or \code{F_shock} is a perturbation function (by default, it is set to \code{F_one})
}
}
\keyword{internal}
