% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{make_Linits.basicL}
\alias{make_Linits.basicL}
\title{Setup the basicL model}
\usage{
\method{make_Linits}{basicL}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_Linits} for the basicL model
}
