% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{make_Linits_basicL}
\alias{make_Linits_basicL}
\title{Make Initial Values for \code{basicL} (\strong{L} Component)}
\usage{
make_Linits_basicL(nHabitats, Lopts = list(), L = 1)
}
\arguments{
\item{nHabitats}{the number of habitats in the model}

\item{Lopts}{a \link{list} that overwrites default values}

\item{L}{initial conditions}
}
\value{
a \link{list} with Linits added
}
\description{
Initial values of the variable \eqn{L} can be set
}
