% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{make_Lpar.basicL}
\alias{make_Lpar.basicL}
\title{xde_setup Lpar for the basicL model}
\usage{
\method{make_Lpar}{basicL}(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{make_Lpar} for the basicL model
}
