% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{make_Lpar}
\alias{make_Lpar}
\title{A function to set up adult mosquito models}
\usage{
make_Lpar(Lname, pars, s, Lopts = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{Lname}.
}
