% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{make_MY_obj_trivial}
\alias{make_MY_obj_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
make_MY_obj_trivial(
  nPatches,
  options,
  f = 1,
  q = 1,
  Z = 1,
  eggs = 1,
  F_season = F_flat,
  season_par = list(),
  F_trend = F_flat,
  trend_par = list(),
  F_shock = F_flat,
  shock_par = list()
)
}
\arguments{
\item{nPatches}{an integer}

\item{options}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Z}{the human fraction}

\item{eggs}{the human fraction}

\item{F_season}{a function describing a seasonal pattern}

\item{season_par}{parameters to configure a \code{F_season} using \link{make_function}}

\item{F_trend}{a function describing a temporal trend}

\item{trend_par}{parameters to configure \code{F_trend} using \link{make_function}}

\item{F_shock}{a function describing a temporal shock}

\item{shock_par}{parameters to configure \code{F_shock} using \link{make_function}}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
