% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-sei.R
\name{make_MYZinits.sei}
\alias{make_MYZinits.sei}
\title{Setup initial values for the sei model}
\usage{
\method{make_MYZinits}{sei}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_MYZinits} for the sei model
}
