% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-si.R
\name{make_MYZinits.si}
\alias{make_MYZinits.si}
\title{Setup initial values for the si model}
\usage{
\method{make_MYZinits}{si}(pars, s, MYZopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{make_MYZinits} for the si model
}
