% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{make_MYZpar_trivial}
\alias{make_MYZpar_trivial}
\title{Make parameters for trivial aquatic mosquito model}
\usage{
make_MYZpar_trivial(
  nPatches,
  MYZopts,
  f = 1,
  q = 1,
  Z = 1,
  eggs = 1,
  F_season = F_flat,
  F_trend = F_flat
)
}
\arguments{
\item{nPatches}{an integer}

\item{MYZopts}{a \link{list} of values that overwrites the defaults}

\item{f}{the blood feeding rate}

\item{q}{the human fraction}

\item{Z}{the human fraction}

\item{eggs}{the human fraction}

\item{F_season}{a F_seasonality function}

\item{F_trend}{a F_trend function}
}
\value{
none
}
\description{
Make parameters for trivial aquatic mosquito model
}
