% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg-laying.R
\name{make_O_matrix}
\alias{make_O_matrix}
\title{Compute the Laying Matrix}
\usage{
make_O_matrix(search_weights, habitat_matrix, Q)
}
\arguments{
\item{search_weights}{the habitat search weights}

\item{habitat_matrix}{the habitat membership matrix}

\item{Q}{the availability of egg-laying habitats}
}
\value{
a \code{nHabitats} \eqn{\times} \code{nPatches} \link{matrix} describing egg distribution, \eqn{O}
}
\description{
The egg laying matrix, \eqn{O}, allocates
a portion of eggs laid
by adult mosquitoes in a patch to each
one of the aquatic habitats in the patch
}
\details{
The algorithm is motivated by
the notion of mosquito \emph{searching} for aquatic habitat
and a quantity we call \emph{habitat availability}, \eqn{Q}.
Each habitat is assigned a \emph{search weight,} \eqn{\omega}.
In the simplest case, habitat availability is the sum of \eqn{\omega} by patch, which uses
the habitat membership matrix, \eqn{N}. If the habitats in the model got
all the eggs, then \deqn{Q = N \cdot \omega.}

The fraction of eggs in the \eqn{j^{th}} patch laid
in the \eqn{i^{th}} habitat is \eqn{\omega_i / Q_j.}

To make the model extensible, habitat availability sums all the places where
mosquitoes might lay eggs, including ovitraps and unsuitable habitats.
If every element in \eqn{Q} were positive, the egg
distribution matrix  would be:
\deqn{{O} = \mbox{diag}\left(\omega \right) \cdot {N}^T \cdot \mbox{diag} \left(\frac{1}{Q}\right).}
To avoid dividing by zero, the zero elements in \eqn{Q} are set to an arbitrary positive value.
}
\seealso{
The membership matrix \eqn{N} is computed by \link{make_habitat_matrix}

Total habitat availability, \eqn{\cal Q}, is computed by \link{F_Q}

The availability of ovitraps and bad habitats is setup in \link{setup_ML_interface}
}
