% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult_mosquito-demography.R
\name{make_Omega.xde}
\alias{make_Omega.xde}
\title{Make the mosquito demography matrix}
\usage{
\method{make_Omega}{xde}(pars, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
the derivatives a \link{vector}
}
\description{
This method dispatches on the type of \code{MYZpar}.
}
