% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{make_residence_matrix}
\alias{make_residence_matrix}
\title{Make a residence Matrix}
\usage{
make_residence_matrix(nPatches, residence)
}
\arguments{
\item{nPatches}{the number of patches}

\item{residence}{the residence matrix}
}
\value{
a \code{nPatches} \eqn{\times} \code{nStrata} matrix
}
\description{
This function creates the residence matrix (see \link{get_residence_matrix}).
}
\keyword{internal}
