% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_timespent_at_home}
\alias{make_timespent_at_home}
\title{Make a time spent matrix, called timespent}
\usage{
make_timespent_at_home(
  nPatches,
  residence,
  options = list(),
  at_home = 1,
  not_at_risk = 0
)
}
\arguments{
\item{nPatches}{is the number of patches}

\item{residence}{is the home patch for each stratum}

\item{options}{is a set of options that overwrites the defaults}

\item{at_home}{is the fraction of time spent at home}

\item{not_at_risk}{is the fraction of time not at risk}
}
\value{
a \link{matrix}
}
\description{
Make a time spent matrix, called timespent
}
