% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{make_timespent_xy}
\alias{make_timespent_xy}
\title{Make a time spent matrix using location data}
\usage{
make_timespent_xy(xds_obj, xy, residence, stay, kern, kopts = list())
}
\arguments{
\item{xds_obj}{the \strong{\code{xds}} model object}

\item{xy}{is the xy-locations of the patches}

\item{residence}{is the home patch for each stratum}

\item{stay}{is the fraction of time spent at home}

\item{kern}{is a function to compute time spent away from home}

\item{kopts}{options to pass to the kernel}
}
\value{
a \link{matrix}
}
\description{
Make a time spent matrix using location data
}
