% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_ts_function}
\alias{make_ts_function}
\title{Make a Time Series Function}
\usage{
make_ts_function(
  options = list(),
  N = 1,
  scale = 1,
  season_par = list(),
  trend_par = list()
)
}
\arguments{
\item{options}{configurable options}

\item{N}{the length of the return value}

\item{scale}{scale parameter, usually the average}

\item{season_par}{seasonality function parameters}

\item{trend_par}{trend function parameters}
}
\value{
a function
}
\description{
Make a Time Series Function
}
