% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-object.R
\name{make_xds_object}
\alias{make_xds_object}
\title{Make an \code{xds} object}
\usage{
make_xds_object(
  xds = "xde",
  frame = "full",
  dlay = "ode",
  nPatches = 1,
  membership = 1,
  residence = 1,
  HPop = 1000
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{frame}{model component subset}

\item{dlay}{is \code{ode} for ordinary and \code{dde} for delay}

\item{nPatches}{is the number of patches}

\item{membership}{is the habitat membership vector}

\item{residence}{is the strata residence vector}

\item{HPop}{is the initial human population density}
}
\value{
an unspecified \code{xds} object as a compound \link{list}
}
\description{
Make an \code{xds} object
}
