% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-setup.R
\name{make_xds_object}
\alias{make_xds_object}
\title{Make an \code{xds} object}
\usage{
make_xds_object(xds = "xde", frame = "full")
}
\arguments{
\item{xds}{is \code{xde} for differential equations; or \code{dts} for discrete-time systems}

\item{frame}{the equation frame / subset of dynamical components in the model}
}
\value{
an unspecified \code{xds} object as a compound \link{list}
}
\description{
Make an \code{xds} object
}
