% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{make_XH_obj_SIS}
\alias{make_XH_obj_SIS}
\title{Make an SIS \strong{XH} model object}
\usage{
make_XH_obj_SIS(nStrata, options = list(), b = 0.55, r = 1/180, c = 0.15)
}
\arguments{
\item{nStrata}{is the number of population strata}

\item{options}{a named list with parameter values to overwrite defaults}

\item{b}{transmission probability (efficiency) from mosquito to human}

\item{r}{recovery rate}

\item{c}{transmission probability (efficiency) from human to mosquito}
}
\value{
an \strong{XH} model object
}
\description{
Make an SIS \strong{XH} model object
}
