% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{make_XH_obj_trivial}
\alias{make_XH_obj_trivial}
\title{Make parameters for trivial human model}
\usage{
make_XH_obj_trivial(
  nPatches,
  options,
  kappa = 0.1,
  HPop = 1,
  F_season = F_one,
  F_trend = F_one,
  F_shock = F_one,
  season_par = list(name = "F_one"),
  trend_par = list(name = "F_one"),
  shock_par = list(name = "F_one")
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{options}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}

\item{F_season}{the seasonal pattern function}

\item{F_trend}{the trend function}

\item{F_shock}{the shock function}

\item{season_par}{a list of options for F_season}

\item{trend_par}{a list of options for F_trend}

\item{shock_par}{a list of options for F_shock
#' @return a \link{list}}
}
\description{
Make parameters for trivial human model
}
\keyword{internal}
