% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{make_XH_obj_trivial}
\alias{make_XH_obj_trivial}
\title{Make parameters for trivial human model}
\usage{
make_XH_obj_trivial(
  nPatches,
  options,
  kappa = 0.1,
  HPop = 1,
  season_par = makepar_F_one(),
  trend_par = makepar_F_one()
)
}
\arguments{
\item{nPatches}{the number of patches}

\item{options}{a \link{list}}

\item{kappa}{net infectiousness}

\item{HPop}{initial human population density}

\item{season_par}{parameters to configure a \code{F_season} using \link{make_function}}

\item{trend_par}{parameters to configure \code{F_trend} using \link{make_function}}
}
\value{
a \link{list}
}
\description{
Make parameters for trivial human model
}
