% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-interface.R
\name{MBaseline}
\alias{MBaseline}
\title{Adult Mosquito Bionomics - Baseline}
\usage{
MBaseline(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
Compute the baseline adult mosquito bionomic parameter values
}
