% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{MBionomics.GeM}
\alias{MBionomics.GeM}
\title{Compute mosquito bionomic parameters}
\usage{
\method{MBionomics}{GeM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{MBionomics} for models with no exogenous forcing
}
\keyword{internal}
