% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{MBionomics}
\alias{MBionomics}
\title{Adult Mosquito Bionomics - Modified by Control}
\usage{
MBionomics(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
Modify the baseline adult mosquito bionomic parameters
}
