% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-bionomics.R
\name{MBionomics}
\alias{MBionomics}
\title{Adult Mosquito Bionomics}
\usage{
MBionomics(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}

\item{s}{the species index}
}
\value{
the modified \code{xds} object
}
\description{
Handle adult mosquito bionomics as a baseline modified by control
}
