% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{MEffectSizes}
\alias{MEffectSizes}
\title{Apply effect sizes (\strong{MY})}
\usage{
MEffectSizes(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Apply vector control effect sizes to adult mosquito bionomic parameters
}
\keyword{internal}
