% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{parse_L_orbits.trivial}
\alias{parse_L_orbits.trivial}
\title{parse \strong{L} Component Variables for \code{basicL}}
\usage{
\method{parse_L_orbits}{trivial}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
none
}
\description{
Return a numeric vector of length 0
}
