% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{parse_Lorbits.basicL}
\alias{parse_Lorbits.basicL}
\title{Parse the variable names for the basicL model}
\usage{
\method{parse_Lorbits}{basicL}(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_Lorbits} for basicL competition model.
}
