% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{parse_MY_orbits.basicM}
\alias{parse_MY_orbits.basicM}
\title{parse outputs for basicM}
\usage{
\method{parse_MY_orbits}{basicM}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{a \link{list} that defines a model}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_MY_orbits} for the basicM model.
}
