% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{parse_MY_orbits.trivial}
\alias{parse_MY_orbits.trivial}
\title{parse the output of deSolve and return variables for the trivial model}
\usage{
\method{parse_MY_orbits}{trivial}(outputs, xds_obj, s)
}
\arguments{
\item{outputs}{a \link{matrix} of outputs from deSolve}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
\link{list}
}
\description{
Implements \link{parse_MY_orbits} for trivial
}
