% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_orbits}
\alias{parse_orbits}
\title{parse the outputs of an object created by xde_solve or dts_solve}
\usage{
parse_orbits(outputs, xds_obj)
}
\arguments{
\item{outputs}{a \link{matrix} of _orbits returned by deSolve}

\item{xds_obj}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
parse the outputs of an object created by xde_solve or dts_solve
}
