% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_outputs.R
\name{parse_orbits}
\alias{parse_orbits}
\title{Parse the outputs of an object created by xde_solve or dts_solve}
\usage{
parse_orbits(outputs, pars)
}
\arguments{
\item{outputs}{a \link{matrix} of orbits returned by deSolve}

\item{pars}{a \link{list}}
}
\value{
varslist a \link{list}
}
\description{
Parse the outputs of an object created by xde_solve or dts_solve
}
