% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{parse_outputs_L}
\alias{parse_outputs_L}
\title{Parse the outputs of solving a dynamical system and return the variables by name in a list}
\usage{
parse_outputs_L(outputs, pars, s)
}
\arguments{
\item{outputs}{a \link{matrix} with the solutions to a dynamical system}

\item{pars}{a \link{list} that defines the model}

\item{s}{the species index}
}
\description{
This method dispatches on the type of \code{pars$Lpar}. Attaches the
state variables for the aquatic ecology model to a list and returns it
}
