% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables-V-interface.R
\name{parse_V_orbits}
\alias{parse_V_orbits}
\title{Parse \strong{V} Outputs}
\usage{
parse_V_orbits(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link{deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\description{
After solving, this function extracts
the values of the dependent variables for the \strong{V}
module from the matrix of solutions,
often called the "orbits."
The orbits are parsed
and attached by name to the \strong{\code{xds}} model object.

For differential equations
the solution matrix is returned by \link{deSolve}.
Discrete time systems return a matrix
that has the same shape.
}
