% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{parse_XH_orbits.hMoI}
\alias{parse_XH_orbits.hMoI}
\title{parse the output of deSolve and return variables for the hMoI model}
\usage{
\method{parse_XH_orbits}{hMoI}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link{deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_XH_orbits} for the hMoI model
}
