% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{parse_XH_orbits}
\alias{parse_XH_orbits}
\title{Parse outputs (\strong{XH})}
\usage{
parse_XH_orbits(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
a named \link{list} of parsed output variables
}
\description{
After solving, this function extracts
the values of the dependent variables for the \strong{XH}
module from the matrix of solutions,
often called the "orbits."
The orbits are parsed
and attached by name to the \strong{\code{xds}} model object.

For differential equations
the solution matrix is returned by \link[deSolve:deSolve]{deSolve::deSolve}.
Discrete time systems return a matrix
that has the same shape.
}
\keyword{internal}
