% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{parse_XH_orbits.SIS}
\alias{parse_XH_orbits.SIS}
\title{parse the output of deSolve and return variables for the SIS model}
\usage{
\method{parse_XH_orbits}{SIS}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link[deSolve:deSolve]{deSolve::deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_XH_orbits} for the SIS model
}
