% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{parse_XH_orbits.trivial}
\alias{parse_XH_orbits.trivial}
\title{parse the output of deSolve and return variables for the trivial model}
\usage{
\method{parse_XH_orbits}{trivial}(outputs, xds_obj, i)
}
\arguments{
\item{outputs}{an output matrix returned by \link{deSolve}}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
none
}
\description{
Implements \link{parse_XH_orbits} for the trivial model
}
