% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_y}
\alias{parse_y}
\title{parse the output of an object returned by deSolve}
\usage{
parse_y(y, xds_obj)
}
\arguments{
\item{y}{a \link{vector} with the variables}

\item{xds_obj}{a \strong{\code{xds}} model object}
}
\value{
varslist a \link{list}
}
\description{
parse the output of an object returned by deSolve
}
