% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch_dynamics.R
\name{patch_dynamics}
\alias{patch_dynamics}
\title{Patch Dynamics}
\description{
\describe{
\item{\code{nPatches}}{the number of patches}
}

A spatial domain is sub-divided into a set of patches.
The patches are used to model
adult mosquito ecology, \link{blood_feeding}, and \link{Transmission}.

Each patch could some number of
aquatic habitats (see \link{aquatic_habitats}). Mosquitoes disperse among patches as
they search for or resources and blood feed (if there are any blood hosts), lay eggs (if there are any aquatic habitats), nd other resources
they need (see \link{mosquito_dispersal}).

Patches are used to model blood feeding (see \link{blood_feeding}), so there is a corresponding human model
for
time spent (see \link{time_spent}).
}
