% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{port}
\alias{port}
\title{Ports}
\description{
\strong{port}
: a parameter or variable whose value is set by a function call with a standardized, flexible implementation

The \link{port} is a design template that facilitates nimble model
building by establishing a protocol for adding model \emph{features,}
including aspects of
heterogeneous transmission, exogenous forcing (\emph{eg} weather),
malaria control, malaria importation, and habitat dynamics.
Ports add functionality to an \link{xds_object}: default
values assigned during \link{xds_help_basic_setup} have no effect;
features can be
configured after basic setup (see \link{xds_help_setup_options}).

In some cases, ports were added to \link{dynamical_components} or the
the blood feeding or egg laying interfaces (see \link{xds_interfaces}).
In some cases, a \link{junction} is added to handle bundles of ports
(\emph{eg} Vector Control).
}
\section{Value}{

Each port is a flexible method for assigning a value parameter or variable.
The current value of the parameter
or variable has a location on some object or module object, and it is
always retrieved from that location:

\code{xds_obj$...$value}
}

\section{Functional Response}{

The value of the parameter is set by a function call with
a standard form:

\code{xds_obj <- this_case_port(t, xds_obj)}

where \code{t} is time.
The \code{xds_obj} is passed, modified, and returned.

Inside \code{this_case_port} functions can use
any variables that have been defined. The
value is set by a function call:

\code{xds_obj$...$value <- F_this_case(t, vars)}

where \code{F_this_case} is an \code{S3} object
that dispatches on \code{class(value)}.
}

\section{Default Case}{

The default case for \code{F_this_case} is a pair of methods that set
a constant value for the parameter:
\itemize{
\item \code{this_case_port.setup}
\itemize{
\item assigns a static value to \code{xds_obj$...$value}
\item sets \code{class(xds_obj$...$value) =  "static"}
}
\item \code{this_case_port.static} returns \code{xds_obj} unmodified
}
}

\section{Trace Case}{

For most ports, an alternative case sets the
value by a call to a trace function, configured
using the package standard decomposable time
series trace function (see \link{make_ts_function}).
}

\section{Setup}{

All methods should have well-defined setup functions:

\code{xds_obj <- setup_this_case(case_name, xds_obj, options, s)}

should work.
}

\section{Extensibility}{


Any port can be modified by adding a new \code{S3} method. Since most
of these examples involve some form of mechanistic forcing, new
cases are generally added to \strong{\code{ramp.forcing}.} All new methods
should be fully documented.

For every port, the mathematical consequences of \emph{exogenous forcing}
should be carefully examined.
For any
parameter or term that is implemented as a port, the supporting
mathematics should be fully worked out, so the user can
\emph{set it and forget it.}
}

\section{Example}{


In \link{macdonald}, the bionomic parameters are assigned constant values:
those parameters are not ports. That module does not have the \emph{skill set}
to handle exogenous forcing.

In the generalized Ross-Macdonald \link{GeRM}, several bionomic parameters
are set by calling an \code{S3} function. In this case, the equations have
been set up to handle time-varying survival through a time-varying EIP
(see \link{mosquito_bionomics}).
}

\seealso{
\link{xds_help_setup_options}, \link{junction}
}
\concept{port}
