% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runtime.R
\name{runt}
\alias{runt}
\title{A run-time switch function for mismatched dynamical component run-times}
\usage{
runt(t, Dday, xday)
}
\arguments{
\item{t}{current simulation time}

\item{Dday}{the run-time time step for the simulation}

\item{xday}{the run-time time step for a component}
}
\value{
\link{logical} TRUE or FALSE
}
\description{
Determine whether to update the variables at time t
}
