% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{saveXDS}
\alias{saveXDS}
\title{saveRDS for \code{xds} Objects}
\usage{
saveXDS(xds_obj, filename)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{filename}{the file name}
}
\value{
invisible()
}
\description{
Removes forcing functions (to reduce the file size)
and stores the \code{xds} object using \code{saveRDS}
}
\note{
Forcing functions (\emph{e.g.} F_trend) are stored
on the \code{xds} object in a form that
takes up enormous space.

Since they are created
by calling \code{\link[=make_function]{make_function()}} from the
stored \emph{function objects}
(\emph{e.g.} \code{trend_par}),
they are redundant.
}
\seealso{
\code{\link[=readXDS]{readXDS()}}
}
