% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R, R/aquatic-interface.R
\name{set_Linits}
\alias{set_Linits}
\title{Set \strong{L} Component Initial Values}
\usage{
set_Linits(pars, s = 1, Lopts = list())

set_Linits(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \code{xds} object

an \code{xds} object
}
\description{
This sets initial values for an \strong{L} Component
module from a named list
}
