% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL.R
\name{set_Lpars.basicL}
\alias{set_Lpars.basicL}
\title{Return the parameters as a list}
\usage{
\method{set_Lpars}{basicL}(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{pars$Lpar[[s]]}.
}
