% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{set_Lpars.trivial}
\alias{set_Lpars.trivial}
\title{Set \strong{L} Component parameters for \code{trivial}}
\usage{
\method{set_Lpars}{trivial}(pars, s = 1, Lopts = list())
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{s}{the vector species index}

\item{Lopts}{a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
If \code{Lambda} or \code{F_season} or \code{F_trend}
are named in a list \code{Lopts}, the old value is replaced
}
