% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season_bottom.cohort}
\alias{set_season_bottom.cohort}
\title{set bottom}
\usage{
\method{set_season_bottom}{cohort}(bottom, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{bottom}{the new bottom parameter}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\description{
Set the bottom for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
