% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season_bottom.Lambda}
\alias{set_season_bottom.Lambda}
\title{Set bottom}
\usage{
\method{set_season_bottom}{Lambda}(bottom, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{bottom}{the new bottom parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
the \strong{\code{ramp.xds}} model object
}
\description{
Set the bottom parameter(s) for the seasonal pattern for the
\eqn{s^{th}} species
and return the \strong{\code{ramp.xds}} model object
}
