% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_bottom.R
\name{set_season_bottom}
\alias{set_season_bottom}
\title{Set bottom}
\usage{
set_season_bottom(bottom, xds_obj, s = 1)
}
\arguments{
\item{bottom}{the new bottom parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the bottom parameter to \code{bottom}
}
