% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_phase.R
\name{set_season_phase.eir}
\alias{set_season_phase.eir}
\title{Set phase}
\usage{
\method{set_season_phase}{eir}(X, xds_obj, s = 1)
}
\arguments{
\item{X}{the new phase parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the phase for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
