% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_phase.R
\name{set_season_phase.none}
\alias{set_season_phase.none}
\title{Set phase}
\usage{
\method{set_season_phase}{none}(phase, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{phase}{the new phase parameter}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
the unmodified \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{set_season_phase} for a model
with no forcing
}
