% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season_pw.cohort}
\alias{set_season_pw.cohort}
\title{Set pw, a seasonality shape parameter}
\usage{
\method{set_season_pw}{cohort}(pw, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species inde}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the pw for the eir seasonal pattern for
a \code{cohort} model
and return the \strong{\code{ramp.xds}} model object
}
