% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_pw.R
\name{set_season_pw.eir}
\alias{set_season_pw.eir}
\title{Set pw, a seasonality shape parameter}
\usage{
\method{set_season_pw}{eir}(pw, xds_obj, s = 1)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the pw for the seasonal pattern for
an \code{eir} model
and return the \strong{\code{ramp.xds}} model object
}
