% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_pw.R
\name{set_season_pw.none}
\alias{set_season_pw.none}
\title{Set pw, a seasonality shape parameter}
\usage{
\method{set_season_pw}{none}(pw, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species inde}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Implement \code{set_season_pw} for a model
with no forcing
}
