% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season.R
\name{set_season_pw}
\alias{set_season_pw}
\title{Set pw, a seasonality shape parameter}
\usage{
set_season_pw(pw, xds_obj, s = 1, compile_F = TRUE)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species inde}

\item{compile_F}{if true, call \code{update_F_season}}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the pw parameter to \code{pw}
}
