% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_season_pw.R
\name{set_season_pw}
\alias{set_season_pw}
\title{Set pw, a seasonality shape parameter}
\usage{
set_season_pw(pw, xds_obj, s = 1)
}
\arguments{
\item{pw}{a named list with the new parameter values}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}

\item{s}{the vector species index}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Set the pw parameter to \code{pw}
}
