% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{setup_BLOOD_FEEDING}
\alias{setup_BLOOD_FEEDING}
\title{Set up the interface for blood feeding}
\usage{
setup_BLOOD_FEEDING(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
a \link{list}
}
\description{
Sets up the object that defines the blood feeding interface
for an \code{xds} model object.
}
\details{
Blood feeding is modeled as an interaction among humans and mosquitoes.
\code{ramp.xds} uses the concept of host availability for blood feeding that is based
on a notion of \emph{searching} for hosts. Each hosts is assigned search weights, \eqn{\omega}
and availability is the sum of those search weights. The total availability of all
vertebrate hosts for blood feeding is \eqn{B} and the availability of humans
(or a pathogen's host species) is \eqn{W}.
These can be used to compute mosquito bionomic parameters. For mosquitoes, blood feeding rates (\eqn{f})
can be computed using \emph{functional responses.} The human fraction is \eqn{W/B}. Availability can
also be used to model mosquito movement.
In models with multiple host species, with availability \eqn{W_i}, the fraction on
each host species would be \eqn{W_i/B}.
In models with multiple vector species, each species could have different search habits and preferences,
so blood feeding availability is indexed for each species: \eqn{B_s} and \eqn{W_{s}}.
In models with multiple host species, \eqn{W_{i,s}} is the availability of the
\eqn{i^{th}} host species to the \eqn{s^{th}} vector species.
For hosts, availability is based on \emph{time spent} in each patch, and \emph{time at risk,} or
time spent by time of day weighted by mosquito species-specific \emph{search weights} reflecting different preferences
and a circadian function describing relative mosquito blood feeding rates by time of day.
}
\seealso{
\link{setup_TRANSMISSION}
}
