% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{setup_EGG_LAYING}
\alias{setup_EGG_LAYING}
\title{Setup the interface for egg laying}
\usage{
setup_EGG_LAYING(pars, membership)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{membership}{is the habitat membership vector}
}
\value{
the modified \code{xds} object
}
\description{
Sets up the object that defines the egg laying interface for an \code{xds} object.
}
\details{
Modular computation \code{ramp.xds} defines a rigid interface
to compute egg laying and emergence for models of mosquito ecology, as described by
Implementation is defined by an object called \code{EGGpar} found
attached to the \code{xds} object \code{pars} as \code{pars$EGGpar}.
The interface was developed around a habitat
membership matrix, \eqn{\cal N}, a quantity called
habitat availability \eqn{Q}, and the egg distribution matrix \eqn{\cal U}.
This function is called by \code{make_xds_object} to set up \code{EGGpar} and the variables and parameters with all
the variables it might depend on.
}
\references{
{ This implements an enhanced version of the egg laying model in Equations 14-15 from \insertRef{WuSL2023SpatialDynamics}{ramp.xds} }
}
