% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{setup_egg_laying_static}
\alias{setup_egg_laying_static}
\title{Setup egg laying for most models}
\usage{
setup_egg_laying_static(pars, searchQ = 1, s = 1, Lopts = list())
}
\arguments{
\item{pars}{the model object}

\item{searchQ}{the membership matrix}

\item{s}{the vector species index}

\item{Lopts}{a \link{list} of options to override defaults}
}
\value{
a \link{list} vector
}
\description{
Sets up the egg-deposition matrix calU for the s^th species
}
