% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_EIP}
\alias{setup_EIP}
\title{Set up the fixed model for control forcing (do nothing)}
\usage{
setup_EIP(EIPopts, MYZpar)
}
\arguments{
\item{EIPopts}{a list with options to setup EIPmod; must include EIPmod$EIPname}

\item{MYZpar}{the MYZ parameters}
}
\value{
\link{list} MYZpar with the EIPmod attached
}
\description{
Set up the fixed model for control forcing (do nothing)
}
