% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip-fixedlag-dts.R
\name{setup_eip_fixedlag_dts}
\alias{setup_eip_fixedlag_dts}
\title{Set up a fixedlag_dts model for the EIP}
\usage{
setup_eip_fixedlag_dts(EIPopts, MYZpar, eip = 12)
}
\arguments{
\item{EIPopts}{a \link{list}}

\item{MYZpar}{the MYZ parameters}

\item{eip}{a default value for the eip}
}
\value{
\link{list}
}
\description{
Set up a fixedlag_dts model for the EIP
}
