% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eip.R
\name{setup_eip_obj}
\alias{setup_eip_obj}
\title{Setup an EIP Bionomic Object}
\usage{
setup_eip_obj(eip, MY_obj)
}
\arguments{
\item{eip}{the mosquito patch emigration rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{eip}
}
