% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb.dts}
\alias{setup_exposure_nb.dts}
\title{Set up the negative binomial model of exposure for discrete time models}
\usage{
\method{setup_exposure_nb}{dts}(pars, sz)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{sz}{the size parameter, as in dnbinom(mu=mu, size=size)}
}
\value{
the modified \code{xds} object
}
\description{
Set up the negative binomial model of exposure for discrete time models
}
