% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb}
\alias{setup_exposure_nb}
\title{Set up the negative binomial model of exposure}
\usage{
setup_exposure_nb(pars, sz)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{sz}{the size parameter, as in dnbinom(mu=mu, size=size)}
}
\value{
the modified \code{xds} object
}
\description{
Set up the negative binomial model of exposure
}
