% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb}
\alias{setup_exposure_nb}
\title{Set up a negative binomial model for environmental heterogeneity}
\usage{
setup_exposure_nb(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
none
}
\description{
Set up a negative binomial model for environmental heterogeneity
}
