% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb}
\alias{setup_exposure_nb}
\title{Set up the negative binomial model of exposure}
\usage{
setup_exposure_nb(pars, sz)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{sz}{the negative binomial size parameter (see \link{pnbinom})}
}
\value{
an \code{xds} object
}
\description{
Set up the negative binomial model of exposure
}
\seealso{
\link{F_foi.nb} and \link{F_ar.nb}
}
