% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb.xde}
\alias{setup_exposure_nb.xde}
\title{Make parameters for the null model of exposure}
\usage{
\method{setup_exposure_nb}{xde}(pars, sz)
}
\arguments{
\item{pars}{a \link{list}}

\item{sz}{the size parameter, as in dnbinom(mu=mu, size=size)}
}
\value{
none
}
\description{
Make parameters for the null model of exposure
}
