% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-nb.R
\name{setup_exposure_nb.xde}
\alias{setup_exposure_nb.xde}
\title{Set up the negative binomial model of exposure for continuous time models}
\usage{
\method{setup_exposure_nb}{xde}(pars, sz)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{sz}{the size parameter, as in \link{dnbinom}(mu=mu, size=sz)}
}
\value{
the modified \strong{\code{xds}} object
}
\description{
Set up the negative binomial model of exposure for continuous time models
}
