% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois.dts}
\alias{setup_exposure_pois.dts}
\title{Set up Poisson Exposure and Infection for \code{dts}}
\usage{
\method{setup_exposure_pois}{dts}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an  \strong{\code{xds}} object
}
\description{
Set up the Poisson model for
exposure for discrete time systems
}
\seealso{
Also, see \link{F_ar.pois}
}
