% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois}
\alias{setup_exposure_pois}
\title{Set up the Poisson model of exposure}
\usage{
setup_exposure_pois(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Set up the Poisson model of exposure
}
