% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois}
\alias{setup_exposure_pois}
\title{Set up a Poisson model for Exposure and Infection}
\usage{
setup_exposure_pois(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an  \strong{\code{xds}} object
}
\description{
Set up a Poisson model for Exposure and Infection
}
\seealso{
Also, see \link{F_foi.pois} and \link{F_ar.pois}
}
