% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois.xde}
\alias{setup_exposure_pois.xde}
\title{Set up the Poisson model of exposure for continuous time models}
\usage{
\method{setup_exposure_pois}{xde}(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Set up the Poisson model of exposure for continuous time models
}
