% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure-pois.R
\name{setup_exposure_pois.xde}
\alias{setup_exposure_pois.xde}
\title{Set up Poisson Exposure and Infection for \code{xde}}
\usage{
\method{setup_exposure_pois}{xde}(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an  \strong{\code{xds}} object
}
\description{
Set up the Poisson model for
exposure for continuous time models
}
\seealso{
Also, see \link{F_foi.pois}
}
