% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-circadian.R
\name{setup_F_circadian}
\alias{setup_F_circadian}
\title{Setup a circadian pattern function}
\usage{
setup_F_circadian(F_circadian, xds_obj, s = 1)
}
\arguments{
\item{F_circadian}{a circadian function or a method name (\emph{e.g.} "setup")}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Update the circadian for the \eqn{s^{th}} host species
}
