% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-feeding_rate.R
\name{setup_f_obj}
\alias{setup_f_obj}
\title{Setup Blood Feeding Rate}
\usage{
setup_f_obj(f, MY_obj)
}
\arguments{
\item{f}{the blood feeding rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute dynamic blood feeding rates
}
\keyword{internal}
