% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forcing.R
\name{setup_forcing}
\alias{setup_forcing}
\title{Set up exogenous forcing}
\usage{
setup_forcing(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
This method...
}
