% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_g_obj}
\alias{setup_g_obj}
\title{Setup a Mosquito Mortality Bionomic Object}
\usage{
setup_g_obj(g, MY_obj)
}
\arguments{
\item{g}{the mosquito mortality rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object to return a
constant baseline mosquito mortality rate, \eqn{g}
}
