% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{setup_K_matrix.as_matrix}
\alias{setup_K_matrix.as_matrix}
\title{Setup a Here-There Dispersal Matrix}
\usage{
\method{setup_K_matrix}{as_matrix}(Kname, s, xds_obj, options = list())
}
\arguments{
\item{Kname}{a matrix or a string}

\item{s}{the vector species index}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}
}
\value{
a \link{matrix}
}
\description{
Implements \link{setup_K_matrix} for the herethere model:
dispersal to every other patch, with equal probability
}
