% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-dispersal.R
\name{setup_K_matrix}
\alias{setup_K_matrix}
\title{Setup Mosquito Dispersal Matrix}
\usage{
setup_K_matrix(Kname, xds_obj, options = list(), s = 1)
}
\arguments{
\item{Kname}{a matrix or a string}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
A flexible setup function for mosquito dispersal. The first
argument \code{Kname} dispatches a method: \code{Kname} is used to set the \code{class} of \verb{options,} and
the method  dispatches on \code{class(options)}

Options for \code{Kname} are:
\itemize{
\item \code{is.matrix(Kname)}: if the user passes a matrix, then \code{class(Kname) <- "as_matrix"}
\item \code{Kname = "no_setup"} -- the \strong{\code{xds}} object is returned unmodified
\item \code{Kname = "as_matrix"} -- calls \link{change_K_matrix} and passes \code{K_matrix}
\item \code{Kname = "herethere"} -- calls \link{make_K_matrix_herethere}
\item \code{Kname = "xy"} -- calls \link{make_K_matrix_xy}
}
}
