% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{setup_K_matrix}
\alias{setup_K_matrix}
\title{Setup Mosquito Dispersal Matrix}
\usage{
setup_K_matrix(Kname, s, xds_obj, options = list())
}
\arguments{
\item{Kname}{a matrix or a string}

\item{s}{the vector species index}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Setup Mosquito Dispersal Matrix
}
