% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{setup_K_matrix.xy}
\alias{setup_K_matrix.xy}
\title{Setup a Kernel-Based Mosquito Dispersal Matrix}
\usage{
\method{setup_K_matrix}{xy}(Kname = "xy", s, xds_obj, options = list())
}
\arguments{
\item{Kname}{a matrix or a string}

\item{s}{the vector species index}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of options to configure K_matrix}
}
\value{
a \link{matrix}
}
\description{
Set up a mosquito dispersal matrix from
a set of \eqn{x,y} coordinates and a \emph{kernel,} a function
that assigns weights by distance. The fraction leaving from
each patch that arrive at other patch is the vector of normalized
weights.
}
