% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-dispersal.R
\name{setup_K_obj}
\alias{setup_K_obj}
\title{Setup a Human Fraction Bionomic Object}
\usage{
setup_K_obj(nPatches, MY_obj)
}
\arguments{
\item{nPatches}{the number of patches in the model}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{K}
}
