% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{setup_L_inits.basicL}
\alias{setup_L_inits.basicL}
\title{Setup Initial Values for \code{basicL} (\strong{L} Component)}
\usage{
\method{setup_L_inits}{basicL}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list}
}
\description{
This sets initial values of the variable \eqn{L} by
calling \link{make_L_inits_basicL}. Default values are used unless other values
are passed in \code{options} by name (\emph{i.e.} \code{options$L})
}
\seealso{
\link{make_L_inits_basicL}
}
