% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{setup_L_inits}
\alias{setup_L_inits}
\title{Setup Initial Values for the \strong{L} Component}
\usage{
setup_L_inits(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on \code{Lname}.
}
