% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{setup_L_ix.trivial}
\alias{setup_L_ix.trivial}
\title{Setup Variable Indices for \code{trivial} (\strong{L} Component)}
\usage{
\method{setup_L_ix}{trivial}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_L_ix} for trivial (forced emergence) model.
}
\keyword{internal}
