% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{setup_L_obj.basicL}
\alias{setup_L_obj.basicL}
\title{Setup \code{L_obj} for \code{basicL} (\strong{L} Component)}
\usage{
\method{setup_L_obj}{basicL}(Lname, xds_obj, s, options = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named \link{list} to configure \strong{\code{L_obj}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
The function sets up \code{L_obj} for the \eqn{s^{th}} species
by calling \link{make_L_obj_basicL}
}
\seealso{
\link{make_L_obj_basicL}
}
