% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{setup_L_obj.trivial}
\alias{setup_L_obj.trivial}
\title{Set up \code{trivial} (\strong{L})}
\usage{
\method{setup_L_obj}{trivial}(Lname, xds_obj, s, options = list())
}
\arguments{
\item{Lname}{the class name of the \strong{L} module}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a named list to configure \strong{\code{L_obj}}}
}
\value{
an \strong{\code{xds}} object
}
\description{
Call \link{make_L_obj_trivial} and set
\code{class(xds_obj$forced_by) = "Lambda"}
}
\keyword{internal}
