% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{setup_L_obj.trivial}
\alias{setup_L_obj.trivial}
\title{Setup \code{L_obj} for the \code{trivial} module}
\usage{
\method{setup_L_obj}{trivial}(Lname, xds_obj, s, options = list())
}
\arguments{
\item{Lname}{the class name of the aquatic model}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{s}{the species index}

\item{options}{a named \link{list} to configure \strong{\code{L_obj}}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_L_obj} for the trivial model
}
