% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{setup_Linits.basicL}
\alias{setup_Linits.basicL}
\title{Setup the basicL model}
\usage{
\method{setup_Linits}{basicL}(pars, s, Lopts = list())
}
\arguments{
\item{pars}{a \link{list}}

\item{s}{the species index}

\item{Lopts}{a \link{list}}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_Linits} for the basicL model
}
