% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{setup_MY_inits.trivial}
\alias{setup_MY_inits.trivial}
\title{Setup the trivial model}
\usage{
\method{setup_MY_inits}{trivial}(xds_obj, s, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_MY_inits} for the trivial model
}
\keyword{internal}
