% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-M-basicM.R
\name{setup_MY_ix.basicM}
\alias{setup_MY_ix.basicM}
\title{Add indices for adult mosquitoes to parameter list}
\usage{
\method{setup_MY_ix}{basicM}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
none
}
\description{
Implements \link{setup_MY_ix} for the basic M model.
}
