% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{setup_MY_obj}
\alias{setup_MY_obj}
\title{Setup an \strong{MY} Module (Adult Mosquito Ecology & Infection Dynamics)}
\usage{
setup_MY_obj(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{a \link{list}}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
\link{list}
}
\description{
This method dispatches on \code{MYname}.
}
