% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{setup_MY_obj.trivial}
\alias{setup_MY_obj.trivial}
\title{Setup the trivial model for an adult mosquito model}
\usage{
\method{setup_MY_obj}{trivial}(MYname, xds_obj, s, options = list())
}
\arguments{
\item{MYname}{the name of the model}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}

\item{options}{a \link{list}}
}
\value{
a modified \strong{\code{xds}} model object
}
\description{
Set up the trivial adult mosquito model.
In general, this should be used for aquatic mosquito
ecology or human / host epidemiology.  In the former case,
the user configures \code{F_eggs}; in the latter,
\code{F_fqZ}
}
