% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-trivial.R
\name{setup_MYZpar.trivial}
\alias{setup_MYZpar.trivial}
\title{Setup the trivial model for an adult mosquito model}
\usage{
\method{setup_MYZpar}{trivial}(MYZname, pars, s, MYZopts = NULL)
}
\arguments{
\item{MYZname}{the name of the model}

\item{pars}{a \link{list}}

\item{s}{the species index}

\item{MYZopts}{a \link{list}}
}
\value{
a modified \strong{\code{xds}} model object
}
\description{
Set up the trivial adult mosquito model.
In general, this should be used for aquatic mosquito
ecology or human / host epidemiology.  In the former case,
the user configures \code{F_eggs}; in the latter,
\code{F_fqZ}
}
