% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shock.R
\name{setup_no_shock}
\alias{setup_no_shock}
\title{Set up the no_shock model for shocks (do nothing)}
\usage{
setup_no_shock(pars)
}
\arguments{
\item{pars}{a \link{list}}
}
\value{
\link{list}
}
\description{
Set up the no_shock model for shocks (do nothing)
}
