% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{setup_no_vector_control}
\alias{setup_no_vector_control}
\title{Setup Function for No Vector Control (default)}
\usage{
setup_no_vector_control(pars)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
This sets \code{class(vector_control) <- 'none'}
to dispatch the no vector control option.
}
\seealso{
\link{VectorControl.none}
}
