% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{setup_Omega_obj}
\alias{setup_Omega_obj}
\title{Setup the Omega object}
\usage{
setup_Omega_obj(MY_obj)
}
\arguments{
\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to dispatch updating for the mosquito demographic
object, \eqn{\Omega}
}
\keyword{internal}
