% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{setup_other_variables}
\alias{setup_other_variables}
\title{Set Up the first (null) other variable}
\usage{
setup_other_variables(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
the \code{xds} object
}
\description{
The description for each method should include the equations.
}
