% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{setup_phi_obj}
\alias{setup_phi_obj}
\title{Setup phi object (density independent mortality)}
\usage{
setup_phi_obj(phi, L_obj)
}
\arguments{
\item{phi}{the human fraction}

\item{L_obj}{an \strong{\code{L}} model object}
}
\value{
a \strong{\code{L}} model object
}
\description{
Set up an object
to compute the density independent death rate, \eqn{phi}
}
\keyword{internal}
