% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{setup_psi_obj}
\alias{setup_psi_obj}
\title{Setup psi object (maturation rate)}
\usage{
setup_psi_obj(psi, L_obj)
}
\arguments{
\item{psi}{the maturation rate}

\item{L_obj}{an \strong{\code{L}} model object}
}
\value{
a \strong{\code{L}} model object
}
\description{
Set up an object
to compute the maturation
rate \eqn{psi}
}
\keyword{internal}
