% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{setup_psi_obj}
\alias{setup_psi_obj}
\title{Setup a Human Fraction Bionomic Object}
\usage{
setup_psi_obj(psi, L_obj)
}
\arguments{
\item{psi}{the human fraction}

\item{L_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute \link{F_maturation}, the maximum
maturation rate \eqn{psi}
}
