% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{setup_resources_object}
\alias{setup_resources_object}
\title{Setup Resources Object}
\usage{
setup_resources_object(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Resources is a junction with
ports for several objects that are part of the
blood feeding or habitat interface.
}
