% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-adult.R
\name{setup_sigma_obj}
\alias{setup_sigma_obj}
\title{Setup a Patch Emigration Bionomic Object}
\usage{
setup_sigma_obj(sigma, MY_obj)
}
\arguments{
\item{sigma}{the mosquito patch emigration rate}

\item{MY_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{sigma}
}
