% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.at_home}
\alias{setup_timespent.at_home}
\title{Make a time spent matrix, called timespent with a here / away}
\usage{
\method{setup_timespent}{at_home}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_timespent} for as_matrix
}
\keyword{internal}
