% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.list}
\alias{setup_timespent.list}
\title{Set up a time spent matrix}
\usage{
\method{setup_timespent}{list}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
a \link{matrix}
}
\description{
If an options list is passed
as the first argument, then set
\itemize{
\item \code{TSname = name$name}
\item \code{options = name}
and call \code{setup_timespent(TSname, xds_obj, options, s)}
}
}
\keyword{internal}
