% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.no_setup}
\alias{setup_timespent.no_setup}
\title{Setup no time spent matrix}
\usage{
\method{setup_timespent}{no_setup}(name, xds_obj, i, options = list())
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
an \strong{\code{xds}} object
}
\description{
Don't change anything
}
\keyword{internal}
