% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent}
\alias{setup_timespent}
\title{Make a time spent matrix, called timespent}
\usage{
setup_timespent(name, xds_obj, i, options = list())
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make a time spent matrix, called timespent
}
