% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_TimeSpent}
\alias{setup_TimeSpent}
\title{Make a time spent matrix, called TimeSpent}
\usage{
setup_TimeSpent(TimeSpent, xds_obj, i, options = list())
}
\arguments{
\item{TimeSpent}{a matrix or a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Make a time spent matrix, called TimeSpent
}
