% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.xy}
\alias{setup_timespent.xy}
\title{Set up a time spent matrix}
\usage{
\method{setup_timespent}{xy}(name, xds_obj, options = list(), i = 1)
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{configuration options}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_timespent} from a set of xy coordinates
and a spatial kernel. See \link{make_timespent_xy}
}
\keyword{internal}
