% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_timespent.xy}
\alias{setup_timespent.xy}
\title{Develop a time spent matrix from a kernel and xy-coordinates}
\usage{
\method{setup_timespent}{xy}(name, xds_obj, i, options = list())
}
\arguments{
\item{name}{a matrix or setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{setup_timespent} for kernels
}
\keyword{internal}
