% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{setup_TimeSpent.xy}
\alias{setup_TimeSpent.xy}
\title{Develop a mosquito dispersal matrix from a kernel and xy-coordinates}
\usage{
\method{setup_TimeSpent}{xy}(TimeSpent, xds_obj, i, options = list())
}
\arguments{
\item{TimeSpent}{a matrix or a setup function name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{configuration options}
}
\value{
a \link{list}
}
\description{
Implements \link{setup_TimeSpent} for kernels
}
