% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{setup_TRANSMISSION}
\alias{setup_TRANSMISSION}
\title{Setup the interface for parasite / pathogen transmission}
\usage{
setup_TRANSMISSION(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Sets up the object that defines the transmission interface for an \code{xds} model object.
}
\details{
This implements a model for egg laying described by Wu SL, \emph{et al.}, (2023).
}
\references{
{\insertRef{WuSL2023SpatialDynamics}{ramp.xds}}
}
\seealso{
\link{make_xds_object}

\link{setup_BLOOD_FEEDING}
}
