% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_eir.R
\name{setup_travel_eir}
\alias{setup_travel_eir}
\title{Setup Travel EIR}
\usage{
setup_travel_eir(
  pars,
  travelEIR = 0,
  F_season = F_flat,
  F_trend = F_flat,
  i = 1
)
}
\arguments{
\item{pars}{a \link{list}}

\item{travelEIR}{the time spent traveling}

\item{F_season}{a function describing a seasonal pattern}

\item{F_trend}{a function describing a trend}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set up a function to compute the travel EIR
}
