% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosquito-demography.R
\name{setup_Upsilon_obj}
\alias{setup_Upsilon_obj}
\title{Set up an Upsilon object}
\usage{
setup_Upsilon_obj(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Set up an object
to dispatch updating for the
object, \eqn{\Upsilon}
}
\keyword{internal}
