% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{setup_XH_obj.hMoI}
\alias{setup_XH_obj.hMoI}
\title{Setup XH_obj.hMoI}
\usage{
\method{setup_XH_obj}{hMoI}(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
a \link{list} vector
}
\description{
Implements \link{setup_XH_obj} for the hMoI model
}
