% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{setup_XH_obj}
\alias{setup_XH_obj}
\title{Setup an \strong{XH} Module (Human / Host Epidemiology & Demography)}
\usage{
setup_XH_obj(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the parameter values and configure a model
for the \strong{X} Component
}
\keyword{internal}
