% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{setup_XH_obj.SIS}
\alias{setup_XH_obj.SIS}
\title{Setup \code{SIS} (\strong{XH} component)}
\usage{
\method{setup_XH_obj}{SIS}(Xname, xds_obj, i, options = list())
}
\arguments{
\item{Xname}{the model name}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{options}{model options as a named list}
}
\value{
the \strong{\code{xds}} model object
}
\description{
Set up an \code{SIS} model object
for the \strong{XH} component
}
