% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bionomics-aquatic.R
\name{setup_xi_obj}
\alias{setup_xi_obj}
\title{Setup a Human Fraction Bionomic Object}
\usage{
setup_xi_obj(xi, L_obj)
}
\arguments{
\item{xi}{the human fraction}

\item{L_obj}{an \strong{\code{MY}} model object}
}
\value{
a \strong{\code{MY}} model object
}
\description{
Set up an object
to compute the human fraction, \eqn{xi}
}
