% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_seasonality.R
\name{show_season.eir}
\alias{show_season.eir}
\title{Plot the Fitted season}
\usage{
\method{show_season}{eir}(model, dt = 10, add = FALSE)
}
\arguments{
\item{model}{a \strong{`ramp.xds} model object}

\item{dt}{the outputs interval}

\item{add}{add to existing plot}
}
\value{
a \strong{`ramp.xds} model object
}
\description{
for a model forced by emergence,
output the fitted season
}
