% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_seasonality.R
\name{show_season}
\alias{show_season}
\title{Plot the Fitted season}
\usage{
show_season(model, dt = 10, add = FALSE)
}
\arguments{
\item{model}{a \strong{`ramp.xds} model object}

\item{dt}{the outputs interval}

\item{add}{add to existing plot}
}
\value{
a \strong{`ramp.xds} model object
}
\description{
Plot the Fitted season
}
