% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_shock}
\alias{show_shock}
\title{Plot the Temporal shock}
\usage{
show_shock(xds_obj, tm = 10 * c(0:365), add = FALSE, clr = "black", rng = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}

\item{clr}{the line color}

\item{rng}{if not NULL, range limits for plotting}
}
\value{
the temporal shock, invisibly
}
\description{
For a model with temporal forcing,
show the temporal shock
}
