% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend.eir}
\alias{show_trend.eir}
\title{Plot the Fitted Trend}
\usage{
\method{show_trend}{eir}(model, dt = 10, add = FALSE)
}
\arguments{
\item{model}{a \strong{`ramp.xds} model object}

\item{dt}{the outputs interval}

\item{add}{add to existing plot}
}
\value{
a \strong{`ramp.xds} model object
}
\description{
for a model forced by emergence,
output the fitted trend
}
