% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_trend.R
\name{show_trend}
\alias{show_trend}
\title{Plot the Temporal Trend}
\usage{
show_trend(
  xds_obj,
  tm = 10 * c(0:365),
  add = FALSE,
  clr = "black",
  rng = NULL,
  shock = FALSE
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{tm}{the time points}

\item{add}{add to existing plot}

\item{clr}{the line color}

\item{rng}{if not NULL, range limits for plotting}

\item{shock}{if 0,}
}
\value{
the temporal trend, invisibly
}
\description{
For a model with temporal forcing,
show the temporal trend
}
