% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L.R
\name{skill_set_L}
\alias{skill_set_L}
\title{The \strong{L} Module Skill Set}
\usage{
skill_set_L(Lname)
}
\arguments{
\item{Lname}{the \strong{L} module name}
}
\value{
\emph{L} module skill set, as a list
}
\description{
The \strong{L} skill set is a list of
an module's capabilities:
\itemize{
\item \code{demography} is
}
}
