% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{skill_set_MY}
\alias{skill_set_MY}
\title{The \strong{MY} Module Skill Set}
\usage{
skill_set_MY(MYname)
}
\arguments{
\item{MYname}{the \strong{MY} module name}
}
\value{
\emph{MY} module skill set, as a list
}
\description{
The \strong{MY} skill set is a list of
an module's capabilities:
\itemize{
\item \code{demography} is
}
}
