% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-X-hMoI.R
\name{skill_set_XH.hMoI}
\alias{skill_set_XH.hMoI}
\title{The \strong{XH} Module Skill Set}
\usage{
\method{skill_set_XH}{hMoI}(Xname = "hMoI")
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
the \code{hMoI} \emph{XH} module skill set, a list
}
\description{
The \code{hMoI} module is an example
of a model that is not extensible. There is now
way to add either dynamic human population density
or mass treatment without violating the model
assumptions.

The model does output observed population density
and here it is
}
\note{
This method dispatches on \code{class(xds_obj$XH_obj)}
}
