% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{skill_set_XH}
\alias{skill_set_XH}
\title{\strong{XH} Skill Set}
\usage{
skill_set_XH(Xname)
}
\arguments{
\item{Xname}{the \strong{XH} module name}
}
\value{
\emph{XH} module skill set, as a list
}
\description{
The \strong{XH} skill set is a list of
an module's capabilities:
\itemize{
\item \strong{\code{H_dynamics}} - if FALSE, \eqn{H} is static. The module doesn't have a port for demographic change
\item \strong{\code{mda}}:  if FALSE, the module doesn't have an \code{mda} port for mass drug administration
\item \strong{\code{msat}}:  if FALSE, the module doesn't have an \code{msat} port for mass
\item \strong{\code{malaria}}: if FALSE, the model probably shouldn't be used for malaria
\item \strong{\code{pr_obs}}: if TRUE, the model outputs a value for observed PR
\item \strong{\code{pf_rdt}}: if TRUE, the model outputs a value for prevalence by RDT
\item \strong{\code{pf_lm}}: if TRUE, the model outputs a value for prevalence by light microscopy
\item \strong{\code{pf_pcr}}: if FALSE, the model outputs a value for prevalence by PCR
}
}
