% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-basicL.R
\name{steady_state_L.basicL_ode}
\alias{steady_state_L.basicL_ode}
\title{Compute the Steady State of \code{dLdt.basicL} (\strong{L} Component)}
\usage{
\method{steady_state_L}{basicL_ode}(eta, xds_obj, s = 1)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
the values of \eqn{L} at the steady state
}
\description{
Given an egg deposition rate \verb{eta,}
return a steady state value for the equations in \link{dLdt.basicL}
}
\note{
This function does not use deSolve
}
