% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-interface.R
\name{steady_state_M}
\alias{steady_state_M}
\title{Compute steady states for \strong{M}}
\usage{
steady_state_M(Lambda, xds_obj, s = 1)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
the steady states
}
\description{
Given the emergence rate of adult mosquitoes (\eqn{\Lambda} or \code{Lambda}) compute
the population density of adult mosquitoes at the steady state
}
