% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeRM.R
\name{steady_state_MY.GeRM}
\alias{steady_state_MY.GeRM}
\title{Compute the steady states as a function of the daily EIR}
\usage{
\method{steady_state_MY}{GeRM}(Lambda, kappa, MY_obj)
}
\arguments{
\item{Lambda}{the daily emergence rate of adult mosquitoes}

\item{kappa}{net infectiousness}

\item{MY_obj}{a list that defines an adult model}
}
\value{
none
}
\description{
This method dispatches on the type of \code{MY_obj}.
}
