% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{steady_state_X.SIS}
\alias{steady_state_X.SIS}
\title{Compute the steady states for the SIS model as a function of the daily EIR}
\usage{
\method{steady_state_X}{SIS}(foi, H, XH_obj)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{XH_obj}{an \strong{XH} model object}
}
\value{
the steady states as a named vector
}
\description{
Compute the steady state of the SIS model as a function of the daily eir.
}
