% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-interface.R
\name{steady_state_XH}
\alias{steady_state_XH}
\title{Steady States for \strong{XH}}
\usage{
steady_state_XH(foi, xds_obj, i = 1)
}
\arguments{
\item{foi}{the daily FoI}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{i}{the vector species index}
}
\value{
steady states
}
\description{
Compute the steady states as a function of the daily FoI for a
static value of human population density
}
