% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_baits.R
\name{SugarBaitEffectSizes.none}
\alias{SugarBaitEffectSizes.none}
\title{No sugar baits}
\usage{
\method{SugarBaitEffectSizes}{none}(t, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{vector species index}
}
\value{
\link{list}
}
\description{
The null model for sugar baits
}
