% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugar_bait-effectsizes-interface.R
\name{SugarBaitEffectSizes}
\alias{SugarBaitEffectSizes}
\title{Compute sugar bait effect sizes}
\usage{
SugarBaitEffectSizes(t, pars)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Set the effect sizes of sugar baits
}
