% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{TimeSpent.static}
\alias{TimeSpent.static}
\title{Compute time spent objects: static models}
\usage{
\method{TimeSpent}{static}(t, y, xds_obj)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
the unmodified \code{xds} object
}
\description{
Return the time spent objects unmodified
}
