% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission.dynamic}
\alias{Transmission.dynamic}
\title{Compute transmission, the dynamic case}
\usage{
\method{Transmission}{dynamic}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
Compute transmission terms with a dynamic mixing matrix
}
