% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission}
\alias{Transmission}
\title{Compute the mixing matrix and transmission terms}
\usage{
Transmission(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{pars$beta}
}
