% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission}
\alias{Transmission}
\title{Compute the mixing matrix and transmission terms}
\usage{
Transmission(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \code{xds} object
}
\description{
This method dispatches on the type of \code{xds_obj$beta}
}
\keyword{internal}
