% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{Transmission.setup}
\alias{Transmission.setup}
\title{Compute transmission, the static case}
\usage{
\method{Transmission}{setup}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \code{xds} object}
}
\value{
the modified \code{xds} object
}
\description{
Set up and compute transmission terms with a static mixing matrix
}
\details{
The \code{setup} case is called whenever any parameter affecting the mixing matrix
in a static model is changed
}
