% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission.R
\name{transmission_dynamics}
\alias{transmission_dynamics}
\title{Compute transmission, the dynamic case}
\usage{
transmission_dynamics(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \code{xds} object}
}
\value{
an \code{xds} object
}
\description{
Compute transmission terms with a dynamic mixing matrix
}
