% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_eir.R
\name{travel_eir}
\alias{travel_eir}
\title{Travel EIR}
\usage{
travel_eir(t, pars, i)
}
\arguments{
\item{t}{the time}

\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}
}
\value{
the daily travel EIR, a \link{numeric} vector
}
\description{
Compute the EIR while traveling
}
