% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{traveling.dynamic}
\alias{traveling.dynamic}
\title{Time Spent Traveling}
\usage{
\method{traveling}{dynamic}(t, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}#}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets the value of a parameter
describing time spent traveling. It is computed in \link{make_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
