% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveling.R
\name{traveling.setup}
\alias{traveling.setup}
\title{Time Spent Traveling}
\usage{
\method{traveling}{setup}(t, pars, i)
}
\arguments{
\item{t}{current simulation time}

\item{pars}{a \link{list}#}

\item{i}{host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
This function sets a static value for the parameter
describing time spent traveling, resets the class to \code{static} and
then triggers an update to the blood feeding model.
It is computed in \link{make_TaR}
and used to compute availability. It is also used in \link{Exposure}
}
