% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-L-trivial.R
\name{trivial_L}
\alias{trivial_L}
\title{\code{trivial} --- \strong{L} module}
\description{
The trivial module outputs the emergence
rate of adult mosquitoes by calling a
\emph{trace function,}
\deqn{F_\alpha(t) = \Lambda \times F_S(t, V_s) \times F_T(t, V_t) \times F_K(t, V_k)}
where
\itemize{
\item \eqn{\Lambda} is the mean daily emergence rate of adult mosquitoes
\item \eqn{F_S(t,V_s)} or \code{F_season} is a seasonal pattern function
\item \eqn{F_T(t,V_t)} or \code{F_trend} is a trend pattern function
\item \eqn{F_K(t,V_k)} or \code{F_shock} is a perturbation function
}

The variables \eqn{V_s}, \eqn{V_t}, and \eqn{V_t} are
called by \link{get_variables}, which dispatches on the class of
\code{season_par} or \code{trend_par} or \code{shock_par}.
}
\section{Parameters}{

\describe{
\item{\code{Lambda}}{the mean daily emergence rate}
\item{\code{F_season}}{a seasonal pattern function, \eqn{{F_S(t,V_s)}}}
\item{\code{F_trend}}{a trend function, \eqn{F_T(t,V_t)}}
\item{\code{F_shock}}{a shock function, \eqn{F_K(t,V_k)}}
\item{\code{season_par}}{dispatches \link{get_variables} to get \eqn{V_s}}
\item{\code{trend_par}}{dispatches \link{get_variables} to get \eqn{V_t}}
\item{\code{shock_par}}{dispatches \link{get_variables} to get \eqn{V_k}}
}

The default values are \code{F_season=F_trend=F_shock=F_one} and the classes of the
objects \code{season_par} and \code{trend_par} and \code{shock_par} are all \code{list}
}

\section{Get}{

\itemize{
\item \code{get_L_pars} --- the \code{trivial} method returns all the parameters
\item \code{get_mean_forcing} --- get \code{Lambda}
\item \code{get_season} --- returns \code{season_par}
\item \code{get_trend}  --- returns \code{trend_par}
\item \code{get_shock}  --- returns \code{shock_par}
}
}

\section{Change}{

\itemize{
\item \code{change_L_pars} --- change parameters by name
\item \code{change_mean_forcing} --- changes \code{Lambda}
\item \code{change_season} --- changes elements of \code{season_par}
\item \code{change_trend}  --- changes elements of \code{trend_par}
\item \code{change_shock}  --- changes elements of \code{shock_par}
}
}

\section{Notes}{

\enumerate{
\item The module has no state variables.
\item The size of an object saved by \code{saveRDS} balloons if it saves a function,
so \code{saveXDS} function strips the functions and \code{readRDS} remakes the function
from the stored parameters.
\code{F_season}, \code{F_trend}, and \code{F_shock} can be set up manually by passing any
user defined function. If so, the user should use \code{saveRDS} and \code{readRDS}
rather than \code{saveXDS} and \code{readXDS}
\item Setup expects that \verb{membership= c(1:nPatches),} but any membership vector
works.
}
}

