% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-trivial.R
\name{trivial_XH}
\alias{trivial_XH}
\title{\code{trivial} --- \strong{XH} module}
\description{
The trivial module outputs the net infectiousness, \eqn{\kappa},
by calling a
\emph{trace function,}
\deqn{F_\kappa(t) = \kappa S(t) T(t) K(t)}
where
\itemize{
\item \eqn{\kappa} or \code{kappa} is the mean net infectiousness
\item \eqn{S(t)} or \code{F_season} is a seasonal pattern
\item \eqn{T(t)} or \code{F_trend} is a trend pattern
\item \eqn{K(t)} or \code{F_shock} is a perturbation
}

Note: \eqn{0 \leq F_\kappa(t) \leq 1}
}
\section{Parameters}{

\describe{
\item{\code{kappa}}{the net infectiousness}
\item{\code{F_season}}{a seasonal pattern function, \eqn{{S(t)}}}
\item{\code{F_trend}}{a trend function, \eqn{T(t)}}
\item{\code{F_shock}}{a shock function, \eqn{K(t)}}#' }

Setup also adds the objects \code{season_par} and \code{trend_par} and \code{shock_par} for
use by \code{ramp.trace}
}

\section{Get}{

\itemize{
\item \code{get_XH_pars} --- the \code{trivial} method returns all the parameters
}
}

\section{Change}{

\itemize{
\item \code{change_XH_pars} --- change parameters by name
}
}

\section{Notes}{

\enumerate{
\item The module has no state variables.
\item The size of an object saved by \code{saveRDS} balloons if it saves a function,
so \code{saveXDS} function strips the functions and \code{readRDS} remakes the function
from the stored parameters.
\item \code{F_season}, \code{F_trend}, and \code{F_shock} can be set up manually by passing any
user defined function. If so, the user should use \code{saveRDS} and \code{readRDS}
rather than \code{saveXDS} and \code{readXDS}
\item Setup expects that \verb{membership= c(1:nPatches),} but any membership vector
works.
}
}

