% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-search_weights.R
\name{update_blood_search_weights}
\alias{update_blood_search_weights}
\title{Update Blood Search Weights}
\usage{
update_blood_search_weights(xds_obj, s, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for blood feeding search weights, \eqn{\omega}.
Dispatches on \code{class(xds_obj$XH_obj[[i]]$search_obj)}.
}
\keyword{internal}
