% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-search_weights.R
\name{update_blood_search_weights.static}
\alias{update_blood_search_weights.static}
\title{Update Blood Search Weights (static)}
\usage{
\method{update_blood_search_weights}{static}(xds_obj, s, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns \code{xds_obj} unmodified; search weights are static.
}
\keyword{internal}
