% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-traps.R
\name{update_blood_traps}
\alias{update_blood_traps}
\title{Update the blood_traps}
\usage{
update_blood_traps(t, y, xds_obj, s)
}
\arguments{
\item{t}{the time}

\item{y}{the state variables}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the blood_traps, \eqn{\Theta}.
Dispatches on \code{class(xds_obj$XY_interface$traps_obj[[s]])}.
}
\keyword{internal}
