% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-search.R
\name{update_habitat_search_weights}
\alias{update_habitat_search_weights}
\title{Update habitat search weights}
\usage{
update_habitat_search_weights(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for habitat feeding search weights, \eqn{\omega}.
Dispatches on \code{class(xds_obj$L_obj[[s]]$search_obj)}.
}
\keyword{internal}
