% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-search.R
\name{update_habitat_search_weights.static}
\alias{update_habitat_search_weights.static}
\title{Update habitat search weights (static)}
\usage{
\method{update_habitat_search_weights}{static}(xds_obj, s)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the vector species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns \code{xds_obj} unmodified; search weights are static.
}
\keyword{internal}
