% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{update_Linits.basicL}
\alias{update_Linits.basicL}
\title{Update inits for the basicL aquatic mosquito competition model}
\usage{
\method{update_Linits}{basicL}(pars, y0, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
none
}
\description{
Update inits for the basicL aquatic mosquito competition model
}
