% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{update_Linits}
\alias{update_Linits}
\title{Set Initial Values for the \strong{L} Component Module from a State Vector}
\usage{
update_Linits(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
Set Initial Values for the \strong{L} Component Module from a State Vector
}
