% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{update_Linits}
\alias{update_Linits}
\title{Update \strong{L} Component Initial Values}
\usage{
update_Linits(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
an \code{xds} object
}
\description{
This sets initial values for an \strong{L} Component
module from a state variables vector \eqn{y}
}
