% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{update_Linits.trivial}
\alias{update_Linits.trivial}
\title{Update Initial Values for \code{basicL} from a state vector \code{y}}
\usage{
\method{update_Linits}{trivial}(pars, y, s)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{y}{the state variables}

\item{s}{the species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns the unmodified \strong{\code{xds}} object
}
