% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-basicL-xde.R
\name{Update_Lt.basicL}
\alias{Update_Lt.basicL}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
Update_Lt.basicL(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Lt} for the basicL competition model.
}
