% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R, R/runtime.R
\name{Update_Lt}
\alias{Update_Lt}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
Update_Lt(t, y, pars, s)

Update_Lt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector of length \code{pars$L_ix}

a \link{vector} containing the vector of all state derivatives
}
\description{
This method dispatches on the type of \code{pars$Lpar}.

Compute and update the state variables for
a model with only humans
}
