% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{Update_Lt}
\alias{Update_Lt}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
Update_Lt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector of length \code{pars$L_ix}
}
\description{
This method dispatches on the type of \code{pars$Lpar}.
}
