% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{Update_Lt}
\alias{Update_Lt}
\title{Update States for \strong{L}}
\usage{
Update_Lt(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
the states, a \link{numeric} vector of length \eqn{n_q=}\code{nHabitats}
}
\description{
This method updates the state variables
for \strong{L} Component modules
}
\note{
Dispatches on the class of \code{pars$Lpar[[s]]}
}
