% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trace.R
\name{Update_Lt.trace}
\alias{Update_Lt.trace}
\title{Derivatives for aquatic stage mosquitoes}
\usage{
Update_Lt.trace(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the vector species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Lt} for the trace (forced emergence) model.
}
