% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{Update_Lt.trivial}
\alias{Update_Lt.trivial}
\title{Update State Variables for \code{trivial} (\strong{L} Component)}
\usage{
\method{Update_Lt}{trivial}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_Lt} for the trivial (forced emergence) model.
}
