% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-GeM.R
\name{Update_MYt.GeM}
\alias{Update_MYt.GeM}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYt}{GeM}(t, y, xds_obj, s)

\method{Update_MYt}{GeM}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a \link{numeric} vector

a \link{numeric} vector
}
\description{
Implements \link{Update_MYt} for the GeM_dts model.

Implements \link{Update_MYt} for the GeM_dts model.
}
\keyword{internal}
