% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-MY-trivial.R
\name{Update_MYt.trivial}
\alias{Update_MYt.trivial}
\title{Handle state updates for the \code{trivial} \strong{MY} module}
\usage{
\method{Update_MYt}{trivial}(t, y, xds_obj, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} model object}

\item{s}{the species index}
}
\value{
a null vector
}
\description{
Implements \link{Update_MYt} for the trivial (forced emergence) model.
}
