% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SI.R
\name{update_MYZinits.SI}
\alias{update_MYZinits.SI}
\title{Make inits for \code{SI} adult mosquito model}
\usage{
\method{update_MYZinits}{SI}(pars, y0, s)
}
\arguments{
\item{pars}{a \link{list}}

\item{y0}{a vector of variable values from a simulation}

\item{s}{the species index}
}
\value{
a \link{list}
}
\description{
Make inits for \code{SI} adult mosquito model
}
