% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-SEI.R
\name{Update_MYZt.SEI}
\alias{Update_MYZt.SEI}
\title{Derivatives for adult mosquitoes}
\usage{
\method{Update_MYZt}{SEI}(t, y, pars, s)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}

\item{s}{the species index}
}
\value{
a \link{numeric} vector
}
\description{
Implements \link{Update_MYZt} for the SEI model.
}
