% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{update_time_away}
\alias{update_time_away}
\title{Update Time Away}
\usage{
update_time_away(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the fraction of time spent outside the
spatial domain, \code{time_away}.
Dispatches on \code{class(xds_obj$XY_interface$time_away_obj[[i]])}.
}
\keyword{internal}
