% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel.R
\name{update_time_away.static}
\alias{update_time_away.static}
\title{Update Time Away (static)}
\usage{
\method{update_time_away}{static}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns \code{xds_obj} unmodified; \code{time_away} is static.
}
\keyword{internal}
