% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{update_timespent}
\alias{update_timespent}
\title{Update the time spent Matrix}
\usage{
update_timespent(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Port function for the time spent matrix, \eqn{\Theta}.
Dispatches on \code{class(xds_obj$XY_interface$timespent_obj[[i]])}.
}
\keyword{internal}
