% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{update_timespent.static}
\alias{update_timespent.static}
\title{Update the time spent Matrix (static)}
\usage{
\method{update_timespent}{static}(xds_obj, i)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}
}
\value{
an \strong{\code{xds}} object
}
\description{
Returns \code{xds_obj} unmodified; the time spent matrix is static.
}
\keyword{internal}
