% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\title{Variables}
\description{
The term \emph{variables} is overloaded.
}
\section{Terminology}{


In \strong{\verb{ramp.xds,}}  there are several kinds of variables and terms:
\itemize{
\item the \strong{independent variable} is usually time, but it is sometimes age
\item the \strong{dependent variables} or \emph{state variables} describe the \emph{state} of the system
- most dependent variables are computed as part of a \strong{Core Dynamical Components}

\if{html}{\out{<div class="sourceCode">}}\preformatted{- **other state variables** are not can be set up and computed along with state variables (see [Other_State_Variables])
}\if{html}{\out{</div>}}
\item \strong{accessory variables} are computed for other reasons. For example, accessory variables are used in delay differential equations to internalize computation of lagged terms using \link[deSolve:timelags]{deSolve::lagderiv} (see \href{https://dd-harp.github.io/ramp.xds/articles/Delays.html}{Delay Equations})
}
}

