% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl.dynamic}
\alias{VectorControl.dynamic}
\title{Distribute vector control, the null model}
\usage{
\method{VectorControl}{dynamic}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
a named \link{list}
}
\description{
Implements \link{VectorControl} for the control model of vector control (do nothing)
}
