% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl.none}
\alias{VectorControl.none}
\title{Implement No Vector Control}
\usage{
\method{VectorControl}{none}(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{a \strong{\code{ramp.xds}} model object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
With no vector control,
the \strong{\code{xds}} object is returned unmodified
}
\note{
No control or \code{none} is the default setting
}
