% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControl}
\alias{VectorControl}
\title{Implement Vector Control}
\usage{
VectorControl(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
The generic function to
implement vector control.
}
\note{
This a junction to implement various modes of
vector control.
Non-trivial vector control modules are in
\href{https://github.com/dd-harp/ramp.control}{\strong{\code{ramp.control}}}.
}
