% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControlEffectSizes.none}
\alias{VectorControlEffectSizes.none}
\title{Set the values of exogenous variables}
\usage{
\method{VectorControlEffectSizes}{none}(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
After none setup, no exogenous
variables are configured so vector_control returns
the \strong{\code{xds}} object without modification
}
