% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_control.R
\name{VectorControlEffectSizes}
\alias{VectorControlEffectSizes}
\title{Vector control effect sizes}
\usage{
VectorControlEffectSizes(t, y, xds_obj)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
This method dispatches on the type of \code{xds_obj$VECTOR_CONTROL}.
}
