% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egg_laying.R
\name{view_habitat_matrix}
\alias{view_habitat_matrix}
\title{View habitat membership}
\usage{
view_habitat_matrix(pars)
}
\arguments{
\item{pars}{an \code{xds} object}
}
\value{
a \code{nHabitats} \eqn{\times 2} \link{matrix}: col 1 is the habitat index, and col 2 is the patch index
}
\description{
Shows the habitat membership information from \eqn{\cal N}
}
\seealso{
\link{create_habitat_matrix}
}
