% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitats.R
\name{view_habitat_matrix}
\alias{view_habitat_matrix}
\title{View habitat membership, \eqn{N}}
\usage{
view_habitat_matrix(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
a \link{list}
}
\description{
Output the habitat membership information as a list
}
\seealso{
\code{\link[=make_habitat_matrix]{make_habitat_matrix()}}
}
