% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding.R
\name{view_residence_matrix}
\alias{view_residence_matrix}
\title{View residence membership}
\usage{
view_residence_matrix(pars, i = 1)
}
\arguments{
\item{pars}{an \code{xds} object}

\item{i}{the host species index}
}
\value{
a named \link{list}
}
\description{
Shows the residence membership information (from the residence matrix)
}
\seealso{
\link{create_residence_matrix}
}
