% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_equations.R
\name{xde_derivatives}
\alias{xde_derivatives}
\title{Compute Derivatives}
\usage{
xde_derivatives(t, y, pars)
}
\arguments{
\item{t}{current simulation time}

\item{y}{state vector}

\item{pars}{a \link{list}}
}
\value{
a \link{list} containing the vector of all state derivatives
}
\description{
The function to compute the derivatives dispatches on \code{class(frame)}
}
\seealso{
\link{xds_solve}
}
