% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xde_solving.R
\name{xde_steady.ode}
\alias{xde_steady.ode}
\title{Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}}
\usage{
\method{xde_steady}{ode}(pars)
}
\arguments{
\item{pars}{a \link{list} that defines a model}
}
\value{
a \link{list}
}
\description{
This method dispatches on the type of \code{pars$xde}
}
