% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-interface.R
\name{xde_steady_state_L}
\alias{xde_steady_state_L}
\title{Compute steady states for \eqn{\cal L} models}
\usage{
xde_steady_state_L(eta, Lpar)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{Lpar}{a list that defines an xde model}
}
\value{
a named \link{list}: values of the state variables at the steady state
}
\description{
For differential equation models, compute
steady states as a function of daily eggs laid, \eqn{\eta}
}
\note{
This method dispatches on the class of \code{Lpar}.
}
