% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquatic-trivial.R
\name{xde_steady_state_L.trivial}
\alias{xde_steady_state_L.trivial}
\title{Compute the Steady State of \code{dLdt.trivial} (\strong{L} Component)}
\usage{
\method{xde_steady_state_L}{trivial}(eta, Lpar)
}
\arguments{
\item{eta}{the egg-laying rate}

\item{Lpar}{a list that defines an xde model}
}
\value{
none
}
\description{
Return a numeric vector of length 0
}
