% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{xde_steady_state_X}
\alias{xde_steady_state_X}
\title{Compute Steady States for an \strong{X}-Component Module}
\usage{
xde_steady_state_X(foi, H, Xpar)
}
\arguments{
\item{foi}{the daily FoI}

\item{H}{human / host population density}

\item{Xpar}{a list that defines an \strong{X} model object}
}
\value{
none
}
\description{
Compute the steady states as a function of the daily FoI for a
static value of human population density
}
