% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds-template.R
\name{xds_dde.dts}
\alias{xds_dde.dts}
\title{Set \code{xds} for \code{dts}}
\usage{
\method{xds_dde}{dts}(xds_obj)
}
\arguments{
\item{xds_obj}{a \strong{\code{ramp.xds}} modelobject}
}
\value{
a \strong{\code{ramp.xds}} model object
}
\description{
Do not change \code{xds}
}
