% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_help.R
\name{xds_help}
\alias{xds_help}
\title{\strong{\code{ramp.xds}}: Getting Help}
\description{
\strong{\code{ramp.xds}} concepts and functions are well documented.
Use these guides to find a concept or function.
\itemize{
\item \link{xds_help_basic_setup} -- an overview of basic setup
\item \link{xds_help_setup_options} -- setup for other features
\item \strong{\code{get_}} functions inspect a model component or outputs
\item \strong{\code{change_}} functions modify part of a model component
\item \strong{\code{setup_}} functions replace a model component
}
}
\section{XH Component}{

Help guides for human / host demography, infection, & immunity.

\code{?XH_}
\itemize{
\item \code{\link{XH_module_list}} -- list of \strong{XH} Modules
\item \code{\link{XH_functions}} -- list of functions for \strong{XH} Modules
\item \code{\link{XH_get}} -- list of \code{get_} functions for \strong{XH} Modules
\item \code{\link{XH_change}} -- list of \code{change_} functions for \strong{XH} Modules
\item \code{\link{XH_setup}} -- list of \code{setup_} functions for \strong{XH} Modules
}
}

\section{MY Component}{

Help guides
for adult mosquito ecology & infection dynamics.

\code{?MY_}
\itemize{
\item \code{\link{MY_module_list}} -- list of \strong{MY} Modules
\item \code{\link{MY_functions}} -- list of functions for \strong{MY} Modules
\item \code{\link{MY_get}} -- list of \code{get_} functions for \strong{MY} Modules
\item \code{\link{MY_change}} -- list of \code{change_} functions for \strong{MY} Modules
\item \code{\link{MY_setup}} -- list of \code{setup_} functions for \strong{MY} Modules
}
}

\section{L Component}{

Help guides
for immature mosquito ecology.

\code{?L_}
\itemize{
\item \code{\link{L_module_list}} -- list of \strong{L} Modules
\item \code{\link{L_functions}} -- list of functions for \strong{L} Modules
\item \code{\link{L_get}} -- list of \code{get_} functions for \strong{L} Modules
\item \code{\link{L_change}} -- list of \code{change_} functions for \strong{L} Modules
\item \code{\link{L_setup}} -- list of \code{setup_} functions for \strong{L} Modules
}
}

