% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ports.R
\name{xds_info_port}
\alias{xds_info_port}
\title{Ports}
\description{
\itemize{
\item \strong{port} -- a parameter or variable whose value is set by a function call with a standardized, flexible implementation
\item junction -- a function that handles a bundle of ports (see \link{xds_info_junction})
}

The implementation of a port follows a design template.
That template establishes a protocol for writing modules with
\emph{features} that could be implemented in several ways to
add realism:
heterogeneous transmission, exogenous forcing (\emph{eg} weather),
malaria control, malaria importation, habitat dynamics, or something else.

By implementing a parameter or variable as a port, a module or
other component becomes extensible with negligible computational costs:
default
values assigned during \link{xds_info_basic_setup} have no effect;
but features can be
configured after basic setup (see \link{xds_info_setup_options}).

In developing ports, the mathematical consequences
\emph{ought to be} fully considered. If some implementations
of a port could be mathematically inconsistent, the port
documentation should include warnings.
}
\section{Port Value}{

Each port is a flexible method for assigning a value parameter or variable (\code{par}).
The current value of the parameter
or variable has a location on a base object (\code{base_obj}), and it is
always retrieved from that location:

\code{xds_obj$base_obj$par}
}

\section{Port Object}{

A port is set up and implemented through an object (\code{par_obj}) that
is located on the same base object:

\code{xds_obj$base_obj$par_obj}

The main purpose of \code{class(par_obj)} is
to dispatch the port function, so \code{par_obj}
could be an empty list.
If an empty list would do, it's more useful
for it to be informative. For example:

\code{par_obj <- this_case}

where \code{this_case = class(par_obj)}

In some cases, \code{par_obj} is a named list
that holds other parameter values.
}

\section{Port Function}{


The value of the parameter is set by a function call with
a standard form:

\code{xds_obj <- this_port(t, xds_obj, ...)}

where
\itemize{
\item \code{t} is time
\item The \code{xds_obj} is passed,
\item \code{this_port} dispatches on \verb{class(xds_obj$base_obj$par_obj).}
\item \code{...} other arguments often include the index of the host or vector species.
}

Inside \code{this_port.this_case}, the parameter value
can be set by a function call:

\code{xds_obj$base_obj$par <- F_this_port(t, ...)}

where \code{F_this_port} is not constrained.
Since \code{xds_obj} is passed to \code{this_port},
functions can use
any variables that have been defined anywhere
on the \code{xds_obj}.
}

\section{Default Case}{

The default case is:

\code{class(xds_obj$base_obj$par_obj) =  "static"}

and \code{this_port.static} returns \code{xds_obj}
unmodified.
}

\section{Trace Function Case}{

For most ports, an alternative case is to configure
a trace function using package standard
\emph{composable time
series} trace function (see \code{ramp.func} amd \code{ramp.forcing}).
}

\section{Setup}{

All methods should have well-defined setup functions:

\code{xds_obj <- setup_this_port(case_name, xds_obj, options, ...)}

where \code{case_name} is used to dispatch \code{setup_this_port}

In cases where changing \code{par} should trigger function
calls to
change other parameters that depend on it,
\code{setup_this_port.static} should set

\code{class(xds_obj$base_obj$par_obj) =  "setup"}

The "setup" method
\itemize{
\item calls those other functions
\item sets \code{class(par_obj) = "static"}
}
}

\section{Extensibility}{


Any port can be modified by adding a new \code{S3} method. Since most
of these examples involve some form of mechanistic forcing, new
cases are generally added to \strong{\code{ramp.forcing}.} All new methods
should be fully documented.

For every port, the mathematical consequences of \emph{exogenous forcing}
should be carefully examined.
For any
parameter or term that is implemented as a port, the supporting
mathematics should be fully worked out, so the user can
\emph{set it and forget it.}
}

\section{Example}{


In \link{macdonald}, the bionomic parameters are assigned constant values:
those parameters are not ports. That module does not have the \emph{skill set}
to handle exogenous forcing.

In the generalized Macdonald \link{GeM}, several bionomic parameters
are set by calling an \code{S3} function. In this case, the equations have
been set up to handle time-varying survival through a time-varying EIP
(see \link{xds_info_mosquito_bionomics}).
}

\seealso{
\link{xds_info_setup_options} | \link{xds_info_junction}
}
\concept{port}
