% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{xds_info_time_at_risk}
\alias{xds_info_time_at_risk}
\title{Time at Risk}
\description{
Mosquitoes have a daily activity pattern.
If people spend time in different locations at different
times of the day, biting risk in a location is weighted by the
mosquitoes relative activity patterns while there.
}
