% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_spent.R
\name{xds_info_time_spent}
\alias{xds_info_time_spent}
\title{time spent}
\description{
The risk of exposure to mosquito-borne pathogens
is related to time spent in places where vectors are
blood feeding. Each human / host population resides
in a patch. Let \eqn{N_p} denote the number of patches
and \eqn{N_h} the number of strata.
The time spent matrix, \eqn{\Theta}, is an \eqn{N_p \times N_h} matrix:
each columns describes the fraction of time spent by a single
population stratum in each patch.
It is expected that most
time is spent in the patch where
the stratum resides.

In \strong{\code{ramp.xds},} the time spent matrix is static. In \strong{\code{xds.forcing}},
time spent can have a daily pattern, and time at risk weights
time spent by a function describing mosquito daily activity rates
(see \link{xds_info_time_at_risk}).
}
