% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{xds_info_visitors}
\alias{xds_info_visitors}
\title{Visitors}
\description{
Malaria can be imported into a spatial domain
visitors. To set up visitors, the
\itemize{
\item the availability of visitors is a port: \link{xds_port_visitors}
\item the infectiousness of visitors is a port: \link{xds_port_visitors_kappa}
}
}
\seealso{
\link{xds_info_malaria_importation}
}
