% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-plotting.R
\name{xds_lines_EIR}
\alias{xds_lines_EIR}
\title{Add lines for the EIR \emph{vs.} time}
\usage{
xds_lines_EIR(xds_obj, i = 1, clrs = "black", lty = 1, annual = TRUE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{lty}{an integer (or integers) to set the \code{lty} for plotting}

\item{annual}{if true, plot as an annualized rate}
}
\value{
eir, invisibly
}
\description{
Add lines for the EIR \emph{vs.} time
}
