% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_M}
\alias{xds_lines_M}
\title{Add lines for adult mosquito population density}
\usage{
xds_lines_M(MYZ, pars, clrs = "darkblue", llty = 1)
}
\arguments{
\item{MYZ}{a list the output of \code{ramp.xde::parse_deout()}}

\item{pars}{a list that defines an \code{ramp.xde} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for adult mosquito population density
}
