% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-adult.R
\name{xds_lines_Y}
\alias{xds_lines_Y}
\title{Add lines for the density of infected and infective mosquitoes}
\usage{
xds_lines_Y(times, MY, xds_obj, clrs = "purple", llty = 1)
}
\arguments{
\item{times}{a sequence of times when variables were output}

\item{MY}{xds_objed ouptuts}

\item{xds_obj}{an \strong{\code{xds}} object}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the density of infected and infective mosquitoes
}
