% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_lines_Y_fracs}
\alias{xds_lines_Y_fracs}
\title{Add lines for the fraction of infected and infective mosquitoes}
\usage{
xds_lines_Y_fracs(times, MYZ, pars, clrs = "purple", llty = 1)
}
\arguments{
\item{times}{times points for the observations}

\item{MYZ}{parsed ouptuts}

\item{pars}{an \strong{\code{xds}} object}

\item{clrs}{a vector of colors for infected mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}
}
\description{
Add lines for the fraction of infected and infective mosquitoes
}
