% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-plotting.R
\name{xds_plot_EIR}
\alias{xds_plot_EIR}
\title{Plot the EIR \emph{vs.} time}
\usage{
xds_plot_EIR(
  xds_obj,
  i = 1,
  clrs = "black",
  llty = 1,
  add = FALSE,
  annual = TRUE
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{a logical: plot axes only if FALSE}

\item{annual}{if true, plot as an annualized rate}
}
\value{
eir, invisibly
}
\description{
Plot the EIR \emph{vs.} time
}
