% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-terms.R
\name{xds_plot_PR}
\alias{xds_plot_PR}
\title{Plot the prevalence / parasite rate (PR) from a model of human infection and immunity}
\usage{
xds_plot_PR(xds_obj, i = 1, clrs = "black", llty = 1, add = FALSE)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) that specifies \code{lty} for plotting}

\item{add}{a logical: plot axes only if FALSE}
}
\description{
Plot the prevalence / parasite rate (PR) from a model of human infection and immunity
}
