% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-interface.R
\name{xds_plot_X}
\alias{xds_plot_X}
\title{Basic plotting for epidemiological models}
\usage{
xds_plot_X(pars, i = 1, clrs = "black", llty = 1, add_axes = TRUE)
}
\arguments{
\item{pars}{an \strong{\code{xds}} object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Basic plotting for epidemiological models
}
