% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-XH-SIS.R
\name{xds_plot_X.SIS}
\alias{xds_plot_X.SIS}
\title{Plot the density of infected individuals for the SIS model}
\usage{
\method{xds_plot_X}{SIS}(
  xds_obj,
  i = 1,
  clrs = c("darkblue", "darkred"),
  llty = 1,
  add = FALSE
)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{i}{the host species index}

\item{clrs}{a vector of colors}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{add}{plot axes only if FALSE}
}
\description{
Plot the density of infected individuals for the SIS model
}
\keyword{internal}
