% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adult-plotting.R
\name{xds_plot_Z_fracs}
\alias{xds_plot_Z_fracs}
\title{Plot the fraction infective}
\usage{
xds_plot_Z_fracs(
  pars,
  s = 1,
  Zclrs = "darkred",
  llty = 1,
  stable = FALSE,
  add_axes = TRUE
)
}
\arguments{
\item{pars}{a list that defines an \code{ramp.xde} model (\emph{e.g.},  generated by \code{xde_setup()})}

\item{s}{the vector species index}

\item{Zclrs}{a vector of colors for infective mosquitoes}

\item{llty}{an integer (or integers) to set the \code{lty} for plotting}

\item{stable}{a logical: set to FALSE for \code{orbits} and TRUE for \code{stable_orbits}}

\item{add_axes}{a logical: plot axes only if TRUE}
}
\description{
Plot the fraction infective
}
