% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blood_feeding-concepts.R
\name{xds_port_other_blood_hosts}
\alias{xds_port_other_blood_hosts}
\title{Other Blood Hosts}
\description{
The \strong{XY} - interface has a port for the availability of other blood hosts,
defined as any vertebrate host that attracts blood feeding mosquitoes but
that are not defined
}
