% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria.R
\name{xds_port_time_away}
\alias{xds_port_time_away}
\title{Time Away}
\description{
A variable \code{time_away} is used in computing
\link{Transmission} to weight
\code{time_spent} and in \link{Exposure} to weight the local \emph{vs.}
travel EIR
}
