% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/travel_malaria.R
\name{xds_port_travel_eir}
\alias{xds_port_travel_eir}
\title{Travel EIR}
\description{
Travel malaria is defined as exposure to malaria
acquired outside of the spatial domain represented
by the patches. The user can configure two ports:
\itemize{
\item time spent away / here
\item the travel EIR
}
}
