% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visitors.R
\name{xds_port_visitor_kappa}
\alias{xds_port_visitor_kappa}
\title{Visitors Infectiousness}
\description{
There are two parameters built into the transmission interface to handle
biting on non-resident humans:
\itemize{
\item \code{visitors} is the availability of non-resident humans
\item \code{visitor_kappa} is the net infectiousness of the visitors
}
}
