% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup}
\alias{xds_setup}
\title{Build a Model and Configure All Components}
\usage{
xds_setup(
  xds = "ode",
  Xname = "SIS",
  XHoptions = list(),
  MYname = "SI",
  MYoptions = list(),
  Lname = "trivial",
  Loptions = list(),
  nPatches = 1,
  HPop = 1000,
  residence = 1,
  TimeSpent = list(),
  searchB = 1,
  membership = 1,
  Koptions = list(Kname = "no_setup"),
  searchQ = 1,
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{Xname}{a character string defining a \strong{X} Component module}

\item{XHoptions}{a list to configure the \strong{X} Component module}

\item{MYname}{a character string defining an \strong{MY} module}

\item{MYoptions}{options to set up the \strong{MY} component}

\item{Lname}{a character string defining a \strong{L} Component module}

\item{Loptions}{a list to configure the \strong{L} Component module}

\item{nPatches}{is the number of patches}

\item{HPop}{is the number of humans in each patch}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{searchB}{is a vector of search weights for blood feeding}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{Koptions}{a K matrix, or options for \link{setup_K_matrix} (see \link{mosquito_dispersal})}

\item{searchQ}{is a vector of search weights for egg laying}

\item{BFopts}{a list to configure the blood feeding model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax
Make an \strong{\code{xds}} \emph{model object}:
\itemize{
\item Define a module for each dynamical component (see \link{dynamical_components}):
\itemize{
\item \strong{XH} Component -- human / host infection dynamics
\item \strong{MY} Component -- adult mosquito ecology and infection dynamics
\item \strong{L} Component -- aquatic mosquito ecology
}
\item Define basic structural parameters for a single host and vector population:
\itemize{
\item \code{nPatches} - the number of patches (see \link{patch_dynamics})
\item \code{membership} - the habitat membership vector (see \link{aquatic_habitats})
\item \code{nHabitats = length(membership)}
\item \code{residence} - the human residence vector (see \link{human_populations})
\item \code{HPop} - the human population size (see \link{human_populations})
\item \code{nStrata = length(residence) = length(HPop)}
}
\item Configure some of the basic elements:
\itemize{
\item Search Weights (see \link{search_weights})
\item Mosquito Dispersal matrix (if \code{nPatches>1}; see \link{mosquito_dispersal})
\item Time Spent matrix (if \code{nPatches>1}; see \link{time_spent})
}
}
}
\details{
\enumerate{
\item \link{make_xds_object_template} returns an object template with
a properly configured interface for blood feeding
and egg laying (see \link{xds_object}):
\itemize{
\item \code{nPatches} is passed as a parameter
\item \code{nHabitats = length(membership)}
\item \code{nStrata = length(residence) = length(HPop)}
\item \code{nHostSpecies=1} (basic setup handles only the first host species)
\item \code{nVectorSpecies=1} (basic setup handles only the first vector species)
\item \code{class(xds_obj$frame) = 'full'}
}
\item Each one of the dynamical components is configured.
\itemize{
\item \strong{\code{xds_obj$XH_obj[[1]]}} defines a model for human / host infection dynamics of class \code{Xname} (the \strong{X} component). The parameter values passed in a named list, \code{XHoptions.}
\item \strong{\code{xds_obj$MY_obj[[1]]}} defines a model for adult mosquito ecology & infection dynamics of class \code{MYname} (the \strong{MY} Component). The parameter values are passed in a named list, \code{MYoptions.}
\item \strong{\code{xds_obj$L_obj[[1]]}} defines a model for aquatic mosquito ecology of class \code{Lname} (The \strong{L} Component). The parameter values are passed in a named list, \code{Loptions.}
}
\item After configuring the dynamical components, several structural parameters can be configured at the command line:
\itemize{
\item Habitat search weights can be set
\item Host population search weights can be set
\item A mosquito dispersal matrix, \eqn{\cal Koptions}, can be set
\item A time spent matrix, \eqn{\Theta}, can be set
}
}
}
\note{
Other options can be configured after basic setup (see \link{xds_help_setup_options}). If the \strong{MY} Component is the \code{trivial} module, consider using \link{xds_setup_aquatic} or \link{xds_setup_human} or \link{xds_setup_eir}. Models for mosquito
ecology need not include states describing infection status (see \link{xds_setup_mosy}).
}
\seealso{
\link{make_xds_object_template}, \link{xds_help_basic_setup}
}
