% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup}
\alias{xds_setup}
\title{Set up an \code{xds} model object}
\usage{
xds_setup(
  xds = "xde",
  dlay = "ode",
  MYZname = "RM",
  Xname = "SIS",
  Lname = "trace",
  nPatches = 1,
  HPop = 1000,
  residence = 1,
  membership = 1,
  Xday = 1,
  MYZday = 1,
  Lday = 1,
  searchB = 1,
  TimeSpent = list(),
  calK = list(),
  searchQ = 1,
  MYZopts = list(),
  Xopts = list(),
  Lopts = list(),
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{dlay}{is either "ode" or "dde"}

\item{MYZname}{is a character string defining a MYZ model}

\item{Xname}{is a character string defining a X model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{HPop}{is the number of humans in each patch}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{MYZopts}{a list to configure the MYZ model}

\item{Xopts}{a list to configure the X model}

\item{Lopts}{a list to configure the L model}

\item{BFopts}{a list to configure the blood feeding model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
a \link{list}
}
\description{
Streamlines setup for \code{xds} model objects
}
