% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup}
\alias{xds_setup}
\title{Set up and configure an \strong{\code{xds}} model object}
\usage{
xds_setup(
  xds = "xde",
  dlay = "ode",
  MYZname = "RM",
  Xname = "SIS",
  Lname = "trace",
  nPatches = 1,
  membership = 1,
  residence = 1,
  HPop = 1000,
  Xday = 1,
  MYZday = 1,
  Lday = 1,
  searchB = 1,
  TimeSpent = list(),
  calK = list(),
  searchQ = 1,
  MYZopts = list(),
  Xopts = list(),
  Lopts = list(),
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{dlay}{is either "ode" or "dde"}

\item{MYZname}{is a character string defining a MYZ model}

\item{Xname}{is a character string defining a X model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{HPop}{is the number of humans in each patch}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{MYZopts}{a list to configure the MYZ model}

\item{Xopts}{a list to configure the X model}

\item{Lopts}{a list to configure the L model}

\item{BFopts}{a list to configure the blood feeding model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
a \link{list}
}
\description{
This sets up an \strong{\code{xds}} model object with \code{pars$frame = 'full'} each
dynamical component fully configured for a single host and vector species:
\itemize{
\item \strong{\code{Xpar}} defines the human / host infection dynamics, \eqn{\cal XH}, of class \code{Xname}
\item \strong{\code{MYZpar}} defines the adult mosquito ecology & infection dynamics, \eqn{\cal MYZ}, of class \code{MYZname}
\item \strong{\code{Lpar}} defines the aquatic mosquito ecology, \eqn{\cal L}, of class \code{Lname}
\item Parameter values for the models are passed as named lists: \code{Xopts}, \code{MYZopts}, and \code{Lopts}
\item Search weights, the mosquito dispersal matrix, and the time spent matrix can also be
configured using parameters passed at the command line.
\item Runtime parameters for \code{dts} models can be configured by passing values: \code{Xday}, \code{MYZday}, and \code{Lday}
}

Advanced features must be configured later, including:
\itemize{
\item multiple-host species or multiple-vector species
\item exogenous forcing by weather, resources, or other factors
\item vector control, vaccines, or other mass
}
}
