% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup}
\alias{xds_setup}
\title{Basic Setup: Make a Fully Defined \strong{\code{xds}} Object}
\usage{
xds_setup(
  xds = "ode",
  Xname = "SIS",
  Xopts = list(),
  MYZname = "RM",
  MYZopts = list(),
  Lname = "trivial",
  Lopts = list(),
  nPatches = 1,
  HPop = 1000,
  residence = 1,
  membership = 1,
  searchB = 1,
  TimeSpent = list(),
  calK = list(),
  searchQ = 1,
  Xday = 1,
  MYZday = 1,
  Lday = 1,
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{Xname}{a character string defining a \eqn{\cal X} model}

\item{Xopts}{a list to configure the X model}

\item{MYZname}{a character string defining a \eqn{\cal MYZ} model}

\item{MYZopts}{a list to configure the MYZ model}

\item{Lname}{a character string defining a \eqn{\cal L} model}

\item{Lopts}{a list to configure the L model}

\item{nPatches}{is the number of patches}

\item{HPop}{is the number of humans in each patch}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{BFopts}{a list to configure the blood feeding model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
Make an \strong{\code{xds}} \emph{model object}:
\itemize{
\item Define the dynamical components:
\itemize{
\item \eqn{\cal XH} - a model for human / host infection dynamics of class \strong{\code{Xname}} with trivial demographics
\item \eqn{\cal MYZ} - a model for adult mosquito ecology and infection dynamics of class \strong{\code{MYZname}}
\item \eqn{\cal L} - a model for aquatic mosquito ecology of class \strong{\code{Lname}}
}
\item Define basic structural parameters for a single host and vector population:
\itemize{
\item \eqn{n_p} or \code{nPatches} - the number of patches
\item \eqn{n_q} or \code{nHabitats <- length(membership)} - the number and locations of aquatic habitats
\item \eqn{n_h} or \code{nStrata <- length(residence)} - the number of human / host population strata and basic demographic information
}
\item Configure some of the basic elements
\itemize{
\item Search weights for human population strata
\item Search weights for aquatic habitats
\item The mosquito dispersal matrix, \eqn{\cal K}
\item The time spent matrix \eqn{\Theta}
}
\item Configure runtime parameters for discrete-time systems
}

Advanced options can be configured after basic setup.
}
\details{
\enumerate{
\item Using the basic structural parameters, a basic template is created by \link{make_xds_template} with
a properly configured interface for blood feeding and egg laying, and \code{pars$frame = class(pars$frame) = 'full'} .
\itemize{
\item \code{nPatches} is passed as a parameter
\item \code{nHabitats} is configured by passing the habitat \code{membership} vector, and \code{nHabitats <- length(membership)}
\item \code{nStrata} is configured by passing a vector of human population densities and a residence vector, and \code{nStrata <- length(residence) <- length(HPop)}
\item \code{nHosts=1} -- basic setup handles the first host species
\item \code{nVectors=1} -- basic setup handles the first vector species
}
\item Each one of the dynamical components is configured.
\itemize{
\item \strong{\code{pars$Xpar[[1]]}} defines a model for human / host infection dynamics of class \code{Xname} (the \eqn{\cal XH} dynamical component). The parameter values passed in a named list, \code{Xopts.}
\item \strong{\code{pars$MYZpar[[1]]}} defines a model for adult mosquito ecology & infection dynamics of class \code{MYZname} (the \eqn{\cal MYZ} dynamical component). The parameter values are passed in a named list, \code{MYZopts.}
\item \strong{\code{pars$Lpar[[1]]}} defines a model for aquatic mosquito ecology of class \code{Lname} (The \eqn{\cal L} dynamical component). The parameter values are passed in a named list, \code{Lopts.}
}
\item After configuring the dynamical components, several structural parameters can be configured at the command line:
\itemize{
\item Habitat search weights can be set
\item Host population search weights can be set
\item A mosquito dispersal matrix, \eqn{\cal K}, can be set
\item A time spent matrix, \eqn{\Theta}, can be set
\item Runtime parameters for \code{dts} models can be configured by passing values: \code{Xday}, \code{MYZday}, and \code{Lday}
}
}

Advanced features must be configured later, including:
\itemize{
\item multiple-host species or multiple-vector species
\item exogenous forcing by weather, resources, or other factors
\item vector control, vaccines, or other mass
}
}
\note{
If \eqn{\cal MYZ} is the trivial model, consider using \link{xds_setup_aquatic} or \link{xds_setup_human} or \link{xds_setup_cohort}. Models for mosquito
ecology need not include states describing infection status (see \link{xds_setup_mosy}).
}
\seealso{
\link{make_xds_template}
}
