% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup}
\alias{xds_setup}
\title{Build a Model and Configure All Components}
\usage{
xds_setup(
  xds = "ode",
  Xname = "SIS",
  XHoptions = list(),
  MYname = "macdonald",
  MYoptions = list(),
  Lname = "trivial",
  Loptions = list(),
  nPatches = 1,
  HPop = 1000,
  residence = 1,
  membership = 1,
  searchB = 1,
  TimeSpent = list(),
  K_matrix = list(),
  searchQ = 1,
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{Xname}{a character string defining a \strong{X} Component module}

\item{XHoptions}{a list to configure the \strong{X} Component module}

\item{MYname}{a character string defining an \strong{MY} module}

\item{MYoptions}{options to set up the \strong{MY} component}

\item{Lname}{a character string defining a \strong{L} Component module}

\item{Loptions}{a list to configure the \strong{L} Component module}

\item{nPatches}{is the number of patches}

\item{HPop}{is the number of humans in each patch}

\item{residence}{is a vector that describes the patch where each human stratum lives}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{searchB}{is a vector of search weights for blood feeding}

\item{TimeSpent}{is either a TimeSpent matrix or a string to call a function that sets it up}

\item{K_matrix}{is either a K_matrix matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{BFopts}{a list to configure the blood feeding model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax
Make an \strong{\code{xds}} \emph{model object}:
\itemize{
\item Define the dynamical components:
\itemize{
\item \strong{XH} - Component modules for human / host infection dynamics of class \strong{\code{Xname}} with trivial demographics
\item \strong{MY} - Component modules for adult mosquito ecology and infection dynamics of class \strong{\code{MYname}}
\item \strong{L} Component modules for aquatic mosquito ecology of class \strong{\code{Lname}}
}
\item Define basic structural parameters for a single host and vector population:
\itemize{
\item \eqn{n_p} or \code{nPatches} - the number of patches
\item \eqn{n_q} or \code{nHabitats = length(membership)} - the numer and locations of aquatic habitats
\item \eqn{n_h} or \code{nStrata = length(residence)} - the number of human / host population strata and basic demographic information
}
\item Configure some of the basic elements
\itemize{
\item Search weights for human population strata
\item Search weights for aquatic habitats
\item The mosquito dispersal matrix, \eqn{K}
\item The time spent matrix \eqn{\Theta}
}
}

Advanced options can be configured after basic setup.
}
\details{
\enumerate{
\item Using the basic structural parameters, a basic template is created by \link{make_xds_object_template} with
a properly configured interface for blood feeding and egg laying, and \code{xds_obj$frame = class(xds_obj$frame) = 'full'} .
\itemize{
\item \code{nPatches} is passed as a parameter
\item \code{nHabitats} is configured by passing the habitat \code{membership} vector, and \code{nHabitats = length(membership)}
\item \code{nStrata} is configured by passing a vector of human population densities and a residence vector, and \code{nStrata = length(residence) = length(HPop)}
\item \code{nHostSpecies=1} (basic setup handles only the first host species)
\item \code{nVectorSpecies=1} (basic setup handles only the first vector species)
}
}

.
2. Each one of the dynamical components is configured.
\itemize{
\item \strong{\code{xds_obj$XH_obj[[1]]}} defines a model for human / host infection dynamics of class \code{Xname} (the \strong{X} component). The parameter values passed in a named list, \code{XHoptions.}
\item \strong{\code{xds_obj$MY_obj[[1]]}} defines a model for adult mosquito ecology & infection dynamics of class \code{MYname} (the \strong{MY} Component). The parameter values are passed in a named list, \code{MYoptions.}
\item \strong{\code{xds_obj$L_obj[[1]]}} defines a model for aquatic mosquito ecology of class \code{Lname} (The \strong{L} Component). The parameter values are passed in a named list, \code{Loptions.}
}
\enumerate{
\item After configuring the dynamical components, several structural parameters can be configured at the command line:
\itemize{
\item Habitat search weights can be set
\item Host population search weights can be set
\item A mosquito dispersal matrix, \eqn{\cal K}, can be set
\item A time spent matrix, \eqn{\Theta}, can be set
}
}

Advanced features must be configured later, including:
\itemize{
\item multiple-host species or multiple-vector species
\item exogenous forcing by weather, resources, or other factors
\item vector control, vaccines, or other mass
}
}
\note{
If the \strong{MY} Component is the \code{trivial} module, consider using \link{xds_setup_aquatic} or \link{xds_setup_human} or \link{xds_setup_cohort}. Models for mosquito
ecology need not include states describing infection status (see \link{xds_setup_mosy}).
}
\seealso{
\link{make_xds_object_template}
}
