% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_aquatic}
\alias{xds_setup_aquatic}
\title{Make an \strong{\code{xds}} Object to Study Aquatic Mosquito Ecology}
\usage{
xds_setup_aquatic(
  xds = "xde",
  dlay = "ode",
  nHabitats = 1,
  Lname = "basicL",
  Lday = 1,
  Lopts = list(),
  MYZopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{dlay}{is either "ode" or "dde"}

\item{nHabitats}{is the number of habitats}

\item{Lname}{is a character string defining a L model}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{Lopts}{a list to configure the L model}

\item{MYZopts}{a list to configure F_eggs from the trivial model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
A modified version of \link{xds_setup} that streamlines setup for a model that
with a trivial \eqn{\cal MYZ} component. The model also configures a trivial \eqn{\cal X} component.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'aquatic'}
to dispatch \link{xde_solve.aquatic} or \link{dts_solve.aquatic}
}
\seealso{
\link{xds_setup}
}
