% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_aquatic}
\alias{xds_setup_aquatic}
\title{Build a Model of Immature Mosquito Ecology}
\usage{
xds_setup_aquatic(
  xds = "ode",
  nHabitats = 1,
  Lname = "basicL",
  Loptions = list(),
  MYoptions = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{nHabitats}{is the number of habitats}

\item{Lname}{is a character string defining a \strong{L} Component module}

\item{Loptions}{a list to configure the \strong{L} Component module}

\item{MYoptions}{a list to configure \link{F_eggs.trivial}}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax

A modified version of \link{xds_setup} that streamlines setup for an \strong{L} Component
when the \strong{MY} Component is set to \code{trivial.} The model
also sets \strong{X} Component to the \code{trivial} module.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'aquatic'}
to dispatch functions that compute derivatives (\code{xde}), update variables (\code{dts}),
and parse outputs
}
\seealso{
\link{xds_setup}
}
