% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_aquatic}
\alias{xds_setup_aquatic}
\title{Make an \strong{\code{xds}} Object to Study Aquatic Mosquito Ecology}
\usage{
xds_setup_aquatic(
  xds = "ode",
  nHabitats = 1,
  Lname = "basicL",
  Lday = 1,
  Lopts = list(),
  MYZopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{nHabitats}{is the number of habitats}

\item{Lname}{is a character string defining a \strong{L} Component module}

\item{Lday}{is the run-time time step for \strong{L} Component (in days): integer or 1/integer}

\item{Lopts}{a list to configure the \strong{L} Component module}

\item{MYZopts}{a list to configure \link{F_eggs.trivial}}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
A modified version of \link{xds_setup} that streamlines setup for an \strong{L} Component
when the \strong{MYZ} Component is set to \code{trivial.} The model
also sets \strong{X} Component to the \code{trivial} module.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'aquatic'}
to dispatch \link{xde_derivatives.aquatic} or \link{dts_update.aquatic}
}
\seealso{
\link{xds_setup}
}
