% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_eir}
\alias{xds_setup_eir}
\title{Build a Model for a single Human / Host Epidemiology forced by the EIR}
\usage{
xds_setup_eir(
  model_name = "unnamed",
  eir = 1,
  F_season = F_one,
  F_trend = F_one,
  F_shock = F_one,
  F_age = F_one,
  season_par = list(name = "F_one"),
  trend_par = list(name = "F_one"),
  age_par = list(name = "F_one"),
  shock_par = list(name = "F_one"),
  xds = "ode",
  Xname = "SIS",
  XHoptions = list(),
  HPop = 1000,
  searchB = 1
)
}
\arguments{
\item{model_name}{is a name for the model (arbitrary)}

\item{eir}{is the entomological inoculation rate}

\item{F_season}{the seasonal pattern function}

\item{F_trend}{the trend function}

\item{F_shock}{the shock function}

\item{F_age}{the relative biting rate by age}

\item{season_par}{a list of options for F_season}

\item{trend_par}{a list of options for F_trend}

\item{age_par}{a list of options for F_age}

\item{shock_par}{a list of options for F_shock}

\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{Xname}{is a character string specifying an \strong{X} Component module}

\item{XHoptions}{a list to configure the \strong{X} Component module}

\item{HPop}{is the number of humans in each stratum}

\item{searchB}{is a vector of search weights for blood feeding}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax

A modified version of \link{xds_setup} to setup up
studies of malaria epidemiology, defined in a narrow sense, to
examine patterns in populations forced by the EIR.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'eir'}
to dispatch functions that compute derivatives (\code{xde}), update variables (\code{dts}),
and parse outputs

The interface includes options to configure a function
describing \code{F_eir} as a function of time, with seasonal components
and a trend.

This can be used to model a cohort as it ages;
a function is set up to modify exposure by age.
}
\seealso{
\link{xds_setup} and \link{xds_setup_human}
}
