% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_eir}
\alias{xds_setup_eir}
\title{Build a Model for a single Human / Host Epidemiology forced by the EIR}
\usage{
xds_setup_eir(
  eir = 1,
  season_par = makepar_F_one(),
  trend_par = makepar_F_one(),
  age_par = makepar_F_one(),
  shock_par = makepar_F_one(),
  xds = "ode",
  Xname = "SIS",
  XHoptions = list(),
  HPop = 1000,
  searchB = 1,
  model_name = "unnamed"
)
}
\arguments{
\item{eir}{is the entomological inoculation rate}

\item{season_par}{parameters to configure a seasonality function using \link{make_function}}

\item{trend_par}{parameters to configure a trends function using \link{make_function}}

\item{age_par}{parameters to configure an age weights function using \link{make_function}}

\item{shock_par}{parameters to configure a shock using \link{make_function}}

\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{Xname}{is a character string specifying an \strong{X} Component module}

\item{XHoptions}{a list to configure the \strong{X} Component module}

\item{HPop}{is the number of humans in each stratum}

\item{searchB}{is a vector of search weights for blood feeding}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax

A modified version of \link{xds_setup} to setup up
studies of malaria epidemiology, defined in a narrow sense, to
examine patterns in populations forced by the EIR.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'eir'}
to dispatch \link{xde_derivatives.eir} or \link{dts_update.eir}

The interface includes options to configure a function
describing \code{F_eir} as a function of time, with seasonal components
and a trend.

This can be used to model a cohort as it ages;
a function is set up to modify exposure by age.
}
\seealso{
\link{xds_setup} and \link{xds_setup_human}
}
