% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_human}
\alias{xds_setup_human}
\title{Build a Model of Human / Host Epidemiology}
\usage{
xds_setup_human(
  model_name = "unnamed",
  Xname = "SIS",
  XHoptions = list(),
  xds = "ode",
  nPatches = 1,
  residence = 1,
  HPop = 1000,
  searchB = list(),
  TSoptions = list(name = "no_setup"),
  MYoptions = list(),
  BFopts = list()
)
}
\arguments{
\item{model_name}{a name for the model}

\item{Xname}{a character string defining a \strong{X} Component module}

\item{XHoptions}{a named list to configure the \strong{X} Component module}

\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{nPatches}{the number of patches}

\item{residence}{a vector that describes the patch where each human stratum lives}

\item{HPop}{the number of humans in each patch}

\item{searchB}{a vector of search weights for blood feeding}

\item{TSoptions}{either a TimeSpent matrix or a string to call a function that sets it up}

\item{MYoptions}{list to configure the \strong{MY} Component module}

\item{BFopts}{list to configure the blood feeding model}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax

A modified version of \link{xds_setup} that
streamlines setup for models with a trivial \strong{MY} Component.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'human'}
to dispatch functions that compute derivatives (\code{xde}), update variables (\code{dts}),
and parse outputs

The \strong{MY} Component module is set to \code{trivial}. The function \link{F_fqZ.trivial} is called to compute passes the
density of infectious adult mosquitoes, and \eqn{f} and \eqn{q} can still be
configured. In this case, the daily EIR is computed using the blood feeding
interface, including \link{Exposure}, in the same way as a model with a non-trivial
\strong{MY} Component module.

To study human cohort dynamics by passing a function that computes the daily EIR,
consider using \link{xds_setup_eir}.
}
\seealso{
\link{xds_setup} and \link{xds_setup_eir}
}
