% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_human}
\alias{xds_setup_human}
\title{Set up an \strong{\code{xds}} object to model human / host infection dynamics}
\usage{
xds_setup_human(
  xds = "xde",
  dlay = "ode",
  Xname = "SIS",
  nPatches = 1,
  residence = 1,
  HPop = 1000,
  Xday = 1,
  searchB = 1,
  TimeSpent = list(),
  MYZopts = list(),
  Xopts = list(),
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations'}

\item{dlay}{either "ode" or "dde"}

\item{Xname}{a character string defining a X model}

\item{nPatches}{the number of patches}

\item{residence}{a vector that describes the patch where each human stratum lives}

\item{HPop}{the number of humans in each patch}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{searchB}{a vector of search weights for blood feeding}

\item{TimeSpent}{either a TimeSpent matrix or a string to call a function that sets it up}

\item{MYZopts}{list to configure the MYZ model}

\item{Xopts}{list to configure the X model}

\item{BFopts}{list to configure the blood feeding model}

\item{model_name}{a name for the model}
}
\value{
a \link{list}
}
\description{
Set up a model to explore human dynamics forced by infective mosquitoes
}
\details{
Write me
}
