% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_human}
\alias{xds_setup_human}
\title{Make an \strong{\code{xds}} Object to Study Human / Host Epidemiology}
\usage{
xds_setup_human(
  Xname = "SIS",
  Xopts = list(),
  xds = "xde",
  dlay = "ode",
  nPatches = 1,
  residence = 1,
  HPop = 1000,
  Xday = 1,
  searchB = 1,
  TimeSpent = list(),
  MYZopts = list(),
  BFopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{Xname}{a character string defining a X model}

\item{Xopts}{list to configure the X model}

\item{xds}{is \code{xde}/\code{dts} for differential / difference equations'}

\item{dlay}{either "ode" or "dde"}

\item{nPatches}{the number of patches}

\item{residence}{a vector that describes the patch where each human stratum lives}

\item{HPop}{the number of humans in each patch}

\item{Xday}{is the run-time time step for X component (in days): integer or 1/integer}

\item{searchB}{a vector of search weights for blood feeding}

\item{TimeSpent}{either a TimeSpent matrix or a string to call a function that sets it up}

\item{MYZopts}{list to configure the MYZ model}

\item{BFopts}{list to configure the blood feeding model}

\item{model_name}{a name for the model}
}
\value{
an \strong{\code{xds}} object
}
\description{
A modified version of \link{xds_setup} that
streamlines setup for models with a trival \eqn{\cal MYZ} component.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'human'}
to dispatch \link{xde_derivatives.human} or \link{dts_update.human} and associated functions.

The \eqn{\cal MYZ} model is trivial. The funcion \link{F_fqZ} still passes the
density of infectious adult mosquitoes, so \eqn{f} and \eqn{q} can still be
configured. In this case, the daily EIR is computed using the blood feeding
interface, including \link{Exposure}, in the same way as a model with a non-trivial
\eqn{\cal MYZ} component.

To study human cohort dynamics by passing a function that computes the daily EIR,
consider using \link{xds_setup_cohort}.
}
\seealso{
\link{xds_setup} and \link{xds_setup_cohort}
}
