% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_mosy}
\alias{xds_setup_mosy}
\title{Build a Model of Mosquito Ecology}
\usage{
xds_setup_mosy(
  xds = "ode",
  MYZname = "basicM",
  Lname = "basicL",
  nPatches = 1,
  membership = 1,
  HPop = 1000,
  MYZday = 1,
  Lday = 1,
  calK = list(),
  searchQ = 1,
  kappa = 0,
  MYZopts = list(),
  Lopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{MYZname}{is a character string defining a \strong{MYZ} Component module}

\item{Lname}{is a character string defining a \strong{L} Component module}

\item{nPatches}{is the number of patches}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{HPop}{is the human / host population density}

\item{MYZday}{is the run-time time step for \strong{MYZ} Component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for \strong{L} Component (in days): integer or 1/integer}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{kappa}{is a vector describing net infectiousness}

\item{MYZopts}{a list to configure the \strong{MYZ} Component module}

\item{Lopts}{a list to configure the \strong{L} Component module}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
\loadmathjax
A modified version of \link{xds_setup} that streamlines setup for models without parasite / pathogen infection
dynamics. These models lack the \strong{Y} component and an \strong{X} component, but they will often need an \strong{H}
component (host density).

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'mosy'}
to dispatch \link{xde_derivatives.mosy} or \link{dts_update.mosy} and associated functions.

The \strong{X} Component module is \code{trivial}, but since humans / vertebrate hosts can be a
resource, \code{HPop} must be set.
}
\seealso{
\link{xds_setup} and \link{dMYZdt.basicM}
}
