% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_mosy}
\alias{xds_setup_mosy}
\title{Make an \strong{\code{xds}} Object to Study Mosquito Ecology}
\usage{
xds_setup_mosy(
  xds = "ode",
  MYZname = "basicM",
  Lname = "basicL",
  nPatches = 1,
  membership = 1,
  HPop = 1000,
  MYZday = 1,
  Lday = 1,
  calK = list(),
  searchQ = 1,
  kappa = 0,
  MYZopts = list(),
  Lopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{ode} or \code{dde} or \code{dts} for ordinary OR delay differential OR difference equations}

\item{MYZname}{is a character string defining a MYZ model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{HPop}{is the human population density}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{kappa}{is a vector describing net infectiousness}

\item{MYZopts}{a list to configure the MYZ model}

\item{Lopts}{a list to configure the L model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
an \strong{\code{xds}} object
}
\description{
A modified version of \link{xds_setup} that streamlines setup for models that do not require a
component describing parasite / pathogen infection dynamics in the adult mosquito population.

The \strong{\code{xds}} object defines \code{frame = class(frame) = 'mosy'}
to dispatch \link{xde_derivatives.mosy} or \link{dts_update.mosy} and associated functions.

The \eqn{\cal X} model is trivial, but since humans / vertebrate hosts can be a
resource, \code{HPop} must be set. Notably, it is possible to set values of NI \code{kappa} but
}
\seealso{
\link{xds_setup} and \link{dMYZdt.basicM}
}
