% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_setup.R
\name{xds_setup_mosy}
\alias{xds_setup_mosy}
\title{Set up an \code{xds} model object for mosquito ecology}
\usage{
xds_setup_mosy(
  xds = "xde",
  dlay = "ode",
  MYZname = "basicM",
  Lname = "basicL",
  nPatches = 1,
  HPop = 1000,
  membership = 1,
  MYZday = 1,
  Lday = 1,
  calK = list(),
  searchQ = 1,
  kappa = 0,
  MYZopts = list(),
  Lopts = list(),
  model_name = "unnamed"
)
}
\arguments{
\item{xds}{is \code{xde}/\code{dts} for differential / difference equations}

\item{dlay}{is either "ode" or "dde"}

\item{MYZname}{is a character string defining a MYZ model}

\item{Lname}{is a character string defining a L model}

\item{nPatches}{is the number of patches}

\item{HPop}{is the human population density}

\item{membership}{is a vector that describes the patch where each aquatic habitat is found}

\item{MYZday}{is the run-time time step for MYZ component (in days): integer or 1/integer}

\item{Lday}{is the run-time time step for L component (in days): integer or 1/integer}

\item{calK}{is either a calK matrix or a string that defines how to set it up}

\item{searchQ}{is a vector of search weights for egg laying}

\item{kappa}{is a vector describing net infectiousness}

\item{MYZopts}{a list to configure the MYZ model}

\item{Lopts}{a list to configure the L model}

\item{model_name}{is a name for the model (arbitrary)}
}
\value{
a \link{list}
}
\description{
Set up an \code{xds} model object for mosquito ecology
}
