% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{xds_shrink.cohort}
\alias{xds_shrink.cohort}
\title{Shrink an xds xds_obj object}
\usage{
\method{xds_shrink}{cohort}(xds_obj)
}
\arguments{
\item{xds_obj}{an xds xds_obj object}
}
\value{
a smaller xds xds_obj object
}
\description{
Shrink an xds xds_obj object
}
