% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve.dde}
\alias{xds_solve.dde}
\title{Solve a system of equations using \link[deSolve:dede]{deSolve::dede}}
\usage{
\method{xds_solve}{dde}(pars, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{pars}{a \link{list} that defines a model}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time interval for outputs}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
Implements \link{xds_solve} for delay differential equations
}
