% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve.dts}
\alias{xds_solve.dts}
\title{Solve a Discrete-Time System}
\usage{
\method{xds_solve}{dts}(xds_obj, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time step}

\item{times}{the times}
}
\value{
a \link{list}
}
\description{
Implements \link{xds_solve} for discrete time systems.
}
\note{
For discrete time systems, the values in \code{times} need not be
a sequence. If \code{times} is not null, the system iterates up to \code{max(times)}
and then returns the solutions at the subset specified in \code{times.}
}
