% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve.ode}
\alias{xds_solve.ode}
\title{Solve a System of Ordinary Differential Equations}
\usage{
\method{xds_solve}{ode}(xds_obj, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time step}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method solves a system of equations and computes the values of
the dependent variables at times \eqn{t=0, dt, 2dt, ... Tmax}. The function dispatches
on \code{xds_obj$dlay} to call:
\itemize{
\item \link[deSolve:ode]{deSolve::ode} if \code{class(dlay) == 'ode'}
\item \link[deSolve:dede]{deSolve::dede} if \code{class(dlay) == 'dde'}
}

Note that the call to \link{xde_derivatives} dispatches on \code{xds_obj$frame}

Implements \link{xds_solve} for ordinary differential equations
}
