% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_solve.R
\name{xds_solve}
\alias{xds_solve}
\title{Solve a Dynamical System}
\usage{
xds_solve(xds_obj, Tmax = 365, dt = 1, times = NULL)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{Tmax}{the last time point, run from 0...Tmax}

\item{dt}{the time step}

\item{times}{the times}
}
\value{
an \strong{\code{xds}} object
}
\description{
Besides solving, this function does several housekeeping tasks:
\itemize{
\item runs \link{check_models}
\item sets up the vector of times when outputs are wanted: see \link{make_times_xde} or \link{make_times_dts}
\item sets up \eqn{y_0,} the initial values vector (see \link{get_inits})
\item solves the system
\item parses and attaches outputs (see \link{parse_outputs})
}
}
\note{
The function \code{\link[=xds_solve]{xds_solve()}} dispatches on \code{xds_obj$xde}
}
\seealso{
\code{\link[=xde_derivatives]{xde_derivatives()}} or \code{\link[=dts_update]{dts_update()}}
}
