% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_stable_orbit}
\alias{xds_stable_orbit}
\title{Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}}
\usage{
xds_stable_orbit(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}

\item{options}{a list of configurable options}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$xde}
}
