% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_stable_orbit}
\alias{xds_stable_orbit}
\title{Compute stable orbits}
\usage{
xds_stable_orbit(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of configurable options}
}
\value{
an \strong{\code{xds}} model object
}
\description{
Run the system for 10 years, and save the last year's orbits.
}
