% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_stable_orbit.xde}
\alias{xds_stable_orbit.xde}
\title{Compute stable orbit for a system of differential equations}
\usage{
\method{xds_stable_orbit}{xde}(xds_obj, options = list())
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}

\item{options}{a list of configurable options}
}
\value{
an \strong{\code{xds}} object
}
\description{
This method dispatches on the type of \code{xds_obj$xde}.
}
