% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_steady.ode}
\alias{xds_steady.ode}
\title{Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}}
\usage{
\method{xds_steady}{ode}(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Solve for the steady state of a system of equations using \link[rootSolve:steady]{rootSolve::steady}
}
\note{
This method dispatches on \code{class(dlay)}
}
