% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xds_steady.R
\name{xds_steady}
\alias{xds_steady}
\title{Solve for the steady state}
\usage{
xds_steady(xds_obj)
}
\arguments{
\item{xds_obj}{an \strong{\code{xds}} model object}
}
\value{
an \strong{\code{xds}} object
}
\description{
Compute the steady state for a
system of differential equations.
}
