% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human-help.R
\name{XH_setup}
\alias{XH_setup}
\title{Setup functions for the XH component}
\description{
Set up functions for \strong{XH}

\code{?setup_XH}
\itemize{
\item \code{\link{setup_XH_obj}} -- set up the XH module object
\item \code{\link{setup_XH_ix}} -- set up variable indices
\item \code{\link{setup_XH_inits}} -- set up initial values
}
}
